==> 1_notgoogle.Rmd <==
---
title: "1_notgoogle"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(tidyverse)
library(janitor)
library(readxl)
library(lubridate)
library(stringr)
```

```{r making employment impact by county!}
# cleaning the industry data
employ_dirty <- read_csv('data/county_stats/cbp2019.csv', show_col_types = FALSE) %>%
  mutate(fipstate = as.character(fipstate),
         fipscty = as.character(fipscty)) %>%
  mutate(fipstate = str_pad(fipstate, width = 2, side = "left", pad = "0"),
         fipscty = str_pad(fipscty, width = 3, side = "left", pad = "0")) %>%
  
  # cleaning up the naics codes
  mutate(naics = gsub('------', 'all', naics)) %>%
  mutate(naics = gsub('/', '', naics)) %>%
  mutate(naics = gsub('-', '', naics)) %>%
  select(fipstate, fipscty, naics, emp) %>%
  filter(naics == "all" | str_length(naics) == 3)

propensity <- read_csv('data/county_stats/pone.0239113.s001.csv') %>%
  select(industry_code, affected_share) %>%
  mutate(affected_share = affected_share/100) %>%
  mutate(industry_code = as.character(industry_code))

#matching propensity scores

employ_propensity <- left_join(employ_dirty, propensity, by = c("naics" = "industry_code")) %>%
  mutate(fips = str_c(fipstate, fipscty, sep = "")) %>%
  select(-fipstate, -fipscty) 

all_employ <- employ_propensity %>%
  filter(naics == "all") %>%
  select(fips, emp) %>%
  rename("total_emp" = "emp")

employ_propensity_2 <- employ_propensity %>%
  filter(naics != "all") %>%
  left_join(all_employ, by = "fips") %>%
  mutate(affected_emp = emp * affected_share)

employ_propensity_3 <- employ_propensity_2 %>%
  group_by(fips) %>%
  mutate(affected_emp = replace_na(affected_emp, 0)) %>%
  summarize(affect_emp = sum(affected_emp), total_emp = total_emp) %>%
  distinct() %>%
  mutate(affected_share = affect_emp/total_emp)

# saved this as an RDS

# Now let's match this with DMAs

county_propensity <- readRDS('clean_data/employment_propensity.rds')

dma_codes <- read_csv('dma_codes.csv') %>%
  select(google_code, statefp, cntyfp) %>%
  mutate(statefp = str_pad(statefp, width = 2, side = "left", pad = "0"),
         cntyfp = str_pad(cntyfp, width = 3, side = "left", pad = 0)) %>%
  mutate(fips = str_c(statefp, cntyfp, sep = "")) %>%
  select(google_code, fips)

dma_propensity <- left_join(county_propensity, dma_codes, by = "fips") %>%
  drop_na(google_code) %>%
  select(-fips) %>%
  group_by(google_code) %>%
  select(-affected_share) %>%
  summarize(affect_emp = sum(affect_emp),
            total_emp = sum(total_emp)) %>%
  mutate(affected_share = affect_emp/total_emp)
  


```
```{r}
# let's do liquor stores


```

```{r}
# Try this again with just hospitality

employ_dirty <- read_csv('data/county_stats/cbp2019.csv', show_col_types = FALSE) %>%
  mutate(fipstate = as.character(fipstate),
         fipscty = as.character(fipscty)) %>%
  mutate(fipstate = str_pad(fipstate, width = 2, side = "left", pad = "0"),
         fipscty = str_pad(fipscty, width = 3, side = "left", pad = "0")) %>%
  
  # cleaning up the naics codes
  mutate(naics = gsub('------', 'all', naics)) %>%
  mutate(naics = gsub('/', '', naics)) %>%
  mutate(naics = gsub('-', '', naics)) %>%
  select(fipstate, fipscty, naics, emp) %>%
  filter(naics == "all" | naics == "72")

all_emp <- employ_dirty %>%
  filter(naics == "all") %>%
  select(fipstate, fipscty, emp) %>%
  rename("all_emp" = "emp")

hospitality_emp <- employ_dirty %>%
  filter(naics == "72") %>%
  select(-naics) %>%
  left_join(all_emp, by = c("fipstate", "fipscty")) %>%
  mutate(fips = str_c(fipstate, fipscty, sep = "")) %>%
  select(-fipstate, -fipscty)

# rerun the dma codes bit in the previous cell

dma_hospitality <- left_join(hospitality_emp, dma_codes, by = "fips") %>%
  drop_na(google_code) %>%
  select(-fips) %>%
  group_by(google_code) %>%
  summarize(emp = sum(emp),
            all_emp = sum(all_emp)) %>%
  mutate(affected_share = emp/all_emp)
```


```{r}
# Vaccinations
vaccinations_dirty <- read_csv('data/vaccinations.csv', show_col_types = FALSE) %>%
  clean_names() %>%
  select(date, fips, series_complete_pop_pct, series_complete_yes,
         administered_dose1_recip, administered_dose1_pop_pct)

# Let's match this by week!

week_begin <- as_tibble(readRDS('week_begin.rds')) %>%
  mutate(week_end = as_date(date) + days(6))
  # We want data from the END of each week rather than the beginning


# Read in the daily vaccination data

vaccinations_daily <- readRDS('data/vaccinations_daily.rds') %>%
  mutate(date = mdy(date)) %>%
  filter(date <= as_date("2020-12-31")) %>%
  filter(fips != "UNK") %>%
  filter(date %in% pull(week_begin, week_end))

# Vax data is pretty insignificant at the end of 2020
```


```{r}
# politics

dma_codes <- read_csv('dma_codes.csv') %>%
  select(google_code, statefp, cntyfp) %>%
  mutate(statefp = str_pad(statefp, width = 2, side = "left", pad = "0"),
         cntyfp = str_pad(cntyfp, width = 3, side = "left", pad = 0)) %>%
  mutate(fips = str_c(statefp, cntyfp, sep = "")) %>%
  select(google_code, fips)

elections <- read_csv('data/county_stats/countypres_2000-2020.csv',
                      show_col_types = FALSE) %>%
  filter(year == 2020, party == "REPUBLICAN") %>%
  select(county_fips, candidatevotes, totalvotes) %>%
  mutate(county_fips = str_pad(county_fips, width = 5, side = "left", pad = "0")) %>%
  left_join(., dma_codes, by = c("county_fips" = "fips")) %>%
  drop_na(google_code) %>%
  select(google_code, candidatevotes, totalvotes) %>%
  group_by(google_code) %>%
  summarize(candidatevotes = sum(candidatevotes),
            totalvotes = sum(totalvotes)) %>% 
  mutate(gop_pct = candidatevotes/totalvotes * 100) %>%
  select(google_code, gop_pct)
  



  
```
```{r}
# COVID cases
# Read in week begin first

covid_county <- read_csv('data/covid_county.csv', show_col_types = FALSE) %>%
  select(-county, -state) %>%
  filter(date %in% pull(week_begin, week_end)) %>%
  group_by(fips) %>%
  arrange(date) %>%
  mutate(new_cases = cases - lag(cases),
         new_deaths = deaths - lag(deaths)) %>%
  mutate(new_cases = ifelse(is.na(new_cases), cases, new_cases),
         new_deaths = ifelse(is.na(new_deaths), deaths, new_deaths)) %>%
  select(-cases, -deaths) %>%
  # lag the dates
  mutate(date = date + days(1))

# Read in some population data first
pop <- read_csv('data/county_stats/pop_race.csv', skip = 1) %>% 
  clean_names() %>%
  select(id, estimate_sex_and_age_total_population) %>%
  mutate(fips = str_sub(id, str_length(id) - 4, str_length(id))) %>%
  select(-id) %>%
  left_join(., dma_codes, by = "fips") %>%
  drop_na(google_code) %>%
  group_by(google_code) %>%
  summarize(total_pop = sum(estimate_sex_and_age_total_population))

covid_dma <- left_join(covid_county, dma_codes, by = "fips") %>%
  drop_na(google_code) %>%
  ungroup() %>%
  group_by(google_code, date) %>%
  summarize(new_cases = sum(new_cases),
            new_deaths = sum(new_deaths)) %>%
  left_join(., pop, by = "google_code") %>%
  drop_na(total_pop) %>%
  mutate(cases_per_100k = new_cases / (total_pop/100000),
         deaths_per_100k = new_deaths / (total_pop/100000)) %>%
  select(-new_cases, -new_deaths)
```

```{r}
# Let's cleannnnn

demo <- read_csv('data/county_stats/social_explorer_data.csv') %>%
  rename("pop_den" = "SE_A00002_002",
         "pop" = "SE_A01001_001",
         "black"= "SE_A03001_003",
         "white" = "SE_A03001_002",
         "over_25" = "SE_B12001_001",
         "bachelors" = "SE_B12001_004",
         "median_income" = "SE_A14006_001",
         "gini" = "SE_A14028_001",
         "noninsti" = "SE_A20001_001",
         "uninsured" = "SE_A20001_002") %>% 
  clean_names() %>% 
  select(geo_fips, pop_den, pop, black, white, over_25, bachelors, median_income, 
         gini, noninsti, uninsured) %>% 
  mutate(white_pct = white / pop,
         black_pct = black / pop,
         bachelor_pct = bachelors/over_25,
         uninsured_pct = uninsured/noninsti) %>% 
  select(geo_fips, pop_den, white_pct, black_pct, bachelor_pct, uninsured_pct, median_income,
         pop, gini)

# match with DMA
  
demo_dma <- left_join(demo, dma_codes, by = c("geo_fips" = "fips")) %>% 
  drop_na(google_code) %>% 
  group_by(google_code) %>% 
  summarize(across(c(pop_den:median_income, gini), ~ weighted.mean(., pop))) %>% 
  left_join(elections, by = "google_code")

searches_norm3_v18 <- left_join(searches_norm3_v17, demo_dma, by = c("dma" = "google_code"))

searches_norm3_v18_5 <- searches_norm3_v18 %>%
  filter(dma %in% dma_5)

searches_v22 <- left_join(searches_v21, demo_dma, by = c("dma" = "google_code"))

searches_v22_5 <- searches_v22 %>%
   filter(dma %in% dma_5)
```

```{r}
# Merge in some of the mental health data!

mental_health <- read_excel('data/county_stats/2021 County Health Rankings Data - v1.xlsx',
                            sheet = "Ranked Measure Data", skip = 1) %>% 
  clean_names() %>% 
  select(fips, state, county, mental_health_provider_rate)

state_mental <- mental_health %>% 
  filter(is.na(county)) %>% 
  mutate(state = state2abbr(state)) %>% 
  select(-fips, -county)

pop <- read_csv('data/county_stats/pop_race.csv', skip = 1) %>% 
  clean_names() %>%
  select(id, estimate_sex_and_age_total_population) %>%
  mutate(fips = str_sub(id, str_length(id) - 4, str_length(id))) %>%
  select(-id) %>% 
  rename("pop" = "estimate_sex_and_age_total_population")

dma_mental_health <- left_join(mental_health, pop, by = "fips") %>% 
  drop_na(county) %>% 
  left_join(dma_codes, by = "fips") %>% 
  group_by(google_code) %>% 
  summarize(mental_health_provider_rate = weighted.mean(mental_health_provider_rate,
                                                        w = pop, na.rm = TRUE)) %>% 
  drop_na(google_code)

searches_norm3_v24_5 <- left_join(searches_norm3_v23_5, dma_mental_health, by = c("dma" = "google_code"))

hps_all_v19 <- left_join(hps_all_v18, state_mental, by = "state")
  
counseling_1 <- felm(log(counseling) ~ 
                       factor(c2_workplace_closing) * mental_health_provider_rate + 
               initclaims_combined + 
               contclaims_combined + cases_per_100k_5day + 
               factor(e1_income_support) + factor(e2_debt_contract_relief) + 
                log(covid_norm+1) + log(lockdown_norm+1)
             |factor(period) + factor(state) + factor(year)|0|state,
             data = hps_all_v19 %>% drop_na(pop),
             weights = na.omit(hps_all_v19$pop))

therapy_bus <- felm(I(log(therapy + 1 )) ~  
                 busclose_first * year_indicator + 
                  busclose_first:year_indicator:mental_health_provider_rate + 
                busopen_most_pop * year_indicator + 
                  busopen_most_pop:year_indicator:mental_health_provider_rate + 
                        online_only * year_indicator + 
                        hybrid * year_indicator + 
                        other * year_indicator + 
                        cases_per_100k + unemploy_rate + temp + precip + 
                       log(lockdown + 1) + log(covid + 1) + 
                       factor(income_support) + factor(debt_contract_relief)
                       |factor(week) + factor(dma) + factor(year)|0|dma, 
             weights = na.omit(searches_norm3_v24_5$pop),
                  data = searches_norm3_v24_5 %>% filter(is.na(pop) == FALSE))


```
