==> 2_google.Rmd <==
---
title: "2_google"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(tidyverse)
library(janitor)
library(gtrendsR)
library(readxl)

dma_codes <- read_csv('dma_codes.csv', col_types = cols()) %>%
  select(google_code) %>%
  mutate(google_code = ifelse(google_code == "US-D)-511",
                              "US-DC-511", google_code)) %>%
  distinct()

dma_codes <- pull(dma_codes, google_code)

dma_codes <- sort(dma_codes)

state_codes <- read_rds('state_codes.rds')

state_codes <- c(state_codes$sub_code)
```

Search terms: (don't use anymore -- look at the google doc)

Depression: 640
Anxiety and Stress: 639
Mental Health: 437

Headaches & Migraines: 631
Sleep Disorders: 633

Movies: 34
Comics & Animation: 316
Books & Literature: 22
Music & Audio: 35
Online Media: 613
TV & Video: 36
Games: 8

Substance Abuse: 257
Subcats:
- Drug & Alcohol Treatment: 1350
- Smoking & Smoking Cessation: 1237

Alcoholic Beverages: 277
Tobacco Products: 123

```{r setup}
# Don't need to run this -- this is replication code from the junior paper, I
# have the results already

data(countries)

us_codes <- countries %>%
  filter(country_code == "US") %>%
  slice(20267:20476) %>%
  mutate(name = tolower(name)) %>%
  select(-country_code) %>%
  clean_names()

google_crosswalk <- read_csv('data/GoogleTrends_CountyDMA_Mapping.csv') %>%
  clean_names() %>%
  mutate(google_dma = tolower(google_dma))

dma_codes <- left_join(google_crosswalk, us_codes, by = c("google_dma" = "name"))

dma_codes_csv <- dma_codes %>%
  drop_na(sub_code) %>%
  mutate(state_dma = str_sub(google_dma, start = str_length(google_dma) - 1),
         dma_code = str_sub(sub_code, start = str_length(sub_code) - 2)) %>%
  mutate(google_code = str_c("US", state_dma, dma_code, sep = "-")) %>%
  mutate(google_code = toupper(google_code))
```

```{r}
# START HERE

dma_codes <- read_csv('dma_codes.csv', col_types = cols()) %>%
  select(google_code) %>%
  mutate(google_code = ifelse(google_code == "US-D)-511",
                              "US-DC-511", google_code)) %>%
  distinct()

dma_codes <- pull(dma_codes, google_code)

dma_codes <- sort(dma_codes)
```

```{r}
# And start reading in data here -- this is for individual search terms
# Sadness here
search_term <- 'sad'

sad <- tibble()

```


```{r}
# Run this

for(i in 1:54) {
  
  if(i < 53) {
    # Exception case for when we need to search twice for NY
    if(i != 32) {
       temp <- gtrends(keyword = search_term, geo = c('US-NY-501', dma_codes[(i-1)*4 + 1],
                                                 dma_codes[(i-1)*4 + 2],
                                                 dma_codes[(i-1)*4 + 3],
                                                 dma_codes[(i-1)*4 + 4]),
                            time = '2015-12-25 2020-12-31', 
                         onlyInterest = TRUE)$interest_over_time
       
       temp <- temp %>% 
          mutate(hits = ifelse(hits == "<1", "0.5", hits)) %>%
          mutate(hits = as.numeric(hits)) %>%
         pivot_wider(id_cols = date, names_from = geo,
                                    values_from = hits) %>%
                        mutate_at(vars(3:6), ~ . * max(.)/max(`US-NY-501`))
    }
    else {
      temp <- gtrends(keyword = search_term, geo = c('US-NY-501', dma_codes[(i-1)*4 + 1],
                                                 dma_codes[(i-1)*4 + 2],
                                                 dma_codes[(i-1)*4 + 3],
                                                 dma_codes[(i-1)*4 + 4]),
                            time = '2015-12-25 2020-12-31', 
                         onlyInterest = TRUE)$interest_over_time %>%
        distinct() %>%
        mutate(hits = ifelse(hits == "<1", "0.5", hits)) %>%
        mutate(hits = as.numeric(hits)) %>%
        pivot_wider(id_cols = date, names_from = geo,
                values_from = hits) %>%
        mutate_at(vars(3:5), ~ . * max(.)/max(`US-NY-501`))
    }
   
  }
  
  else if(i == 53) {
    temp <- gtrends(keyword = search_term, geo = c('US-NY-501', dma_codes[(i-1)*4 + 1]),
                            time = '2015-12-25 2020-12-31', 
                         onlyInterest = TRUE)$interest_over_time %>%
      mutate(hits = ifelse(hits == "<1", "0.5", hits)) %>%
      mutate(hits = as.numeric(hits)) %>%
      pivot_wider(id_cols = date, names_from = geo,
                values_from = hits) %>%
      mutate_at(vars(3), ~ . * max(.)/max(`US-NY-501`))
  }
  
  else{
    temp <- gtrends(keyword = search_term, geo = c('US-NY-501'),
                            time = '2015-12-25 2020-12-31', 
                         onlyInterest = TRUE)$interest_over_time %>%
      mutate(hits = ifelse(hits == "<1", "0.5", hits)) %>%
      mutate(hits = as.numeric(hits)) %>%
      pivot_wider(id_cols = date, names_from = geo,
                  values_from = hits) %>%
      rename(`US-NY-501-FINAL` = `US-NY-501`)
  }
  
  if(i == 1) {
    sad <- temp
  }
  else {
    sad <- cbind(sad, temp)
  }
  Sys.sleep(3)
  
  print(i)
}

sad_clean <- sad[, -which(duplicated(names(sad)))] %>%
  select(-`US-NY-501`) %>%
  rename(`US-NY-501` = `US-NY-501-FINAL`) %>% 
  pivot_longer(cols = `US-AK-743`:`US-NY-501`,
               names_to = "dma", values_to = "sad")

saveRDS(sad_clean, file = "sad_scaled.rds")
```


```{r}
# lagging quickly

lockdown_clean_2 <- lockdown_clean %>%
  group_by(dma) %>%
  arrange(date) %>%
  mutate(lockdown = lag(lockdown)) %>%
  drop_na(lockdown)
```

```{r}
# Pivot longer really quickly

exercise_long <- exercise_clean %>%
  pivot_longer(cols = `US-AK-743`:`US-NY-501`,
               names_to = "dma", values_to = "exercise")

saveRDS(exercise_long, file = "exercise_scaled.rds")


```

```{r}
# Getting the week beginning dates -- each date is the BEGINNING of the week

week_begin <- sadness_clean %>%
  select(date)

```

```{r}
# And this is for categories

category <- '639'

depression <- tibble()
```


```{r}
# Run this

for(i in 1:54) {
  
  if(i < 53) {
    temp <- gtrends(cat = category, geo = c('US-NY-501', dma_codes[(i-1)*4 + 1],
                                                 dma_codes[(i-1)*4 + 2],
                                                 dma_codes[(i-1)*4 + 3],
                                                 dma_codes[(i-1)*4 + 4]),
                            time = '2016-01-01 2020-12-31', 
                         onlyInterest = TRUE)$interest_over_time %>%
    mutate(hits = ifelse(hits == "<1", "0.5", hits)) %>%
    mutate(hits = as.numeric(hits)) %>%
    pivot_wider(id_cols = date, names_from = geo,
                values_from = hits) %>%
    mutate_at(vars(3:6), ~ . * max(.)/max(`US-NY-501`))
  }
  
  else if(i == 53) {
    temp <- gtrends(cat = category, geo = c('US-NY-501', dma_codes[(i-1)*4 + 1]),
                            time = '2016-01-01 2020-12-31', 
                         onlyInterest = TRUE)$interest_over_time %>%
    mutate(hits = ifelse(hits == "<1", "0.5", hits)) %>%
    mutate(hits = as.numeric(hits)) %>%
    pivot_wider(id_cols = date, names_from = geo,
                values_from = hits) %>%
    mutate_at(vars(3), ~ . * max(.)/max(`US-NY-501`))
  }
  
  else{
    temp <- gtrends(cat = category, geo = c('US-NY-501'),
                            time = '2016-01-01 2020-12-31', 
                         onlyInterest = TRUE)$interest_over_time %>%
      mutate(hits = ifelse(hits == "<1", "0.5", hits)) %>%
      mutate(hits = as.numeric(hits)) %>%
      pivot_wider(id_cols = date, names_from = geo,
                  values_from = hits) %>%
      rename(`US-NY-501-FINAL` = `US-NY-501`)
  }
  
  if(i == 1) {
    depression <- temp
  }
  else {
    depression <- cbind(depression, temp)
  }
  Sys.sleep(3)
  
  print(i)
}

depression_clean <- depression[, -which(duplicated(names(depression)))] %>%
  select(-`US-NY-501`) %>%
  rename(`US-NY-501` = `US-NY-501-FINAL`)

saveRDS(depression_clean, file = "depression_scaled.rds")
```
```{r}
# For only one DMA at a time ... 

for(i in 1:209) {
  temp <- gtrends(cat = category, geo = dma_codes[i],
                            time = '2016-01-01 2020-12-31', 
                         onlyInterest = TRUE)$interest_over_time %>%
                pivot_wider(id_cols = date, names_from = geo,
                values_from = hits)
  
  if(i == 1) {
    depression <- temp
  }
  else {
    temp <- temp %>%
        select(-date)
    
    depression <- cbind(depression, temp)
  }
  Sys.sleep(3)
  
  print(i)
}

# depression_clean <- depression[, -which(duplicated(names(depression)))] %>%
#   select(-`US-NY-501`) %>%
#   rename(`US-NY-501` = `US-NY-501-FINAL`)
```

```{r}
# Internet user data by DMA

dma_codes <- read_csv('dma_codes.csv') %>%
  select(google_code, statefp, cntyfp) %>%
  mutate(statefp = str_pad(statefp, width = 2, side = "left", pad = "0"),
         cntyfp = str_pad(cntyfp, width = 3, side = "left", pad = 0)) %>%
  mutate(fips = str_c(statefp, cntyfp, sep = "")) %>%
  select(google_code, fips)

# internet <- read_excel("data/county_stats/county_connections_dec_2018.xlsx",
#                        sheet = "County Connections Dec 2018") %>%
#   mutate_at(c("consumer", "non_consumer", "all", "ratio"),
#             ~ ifelse(. == -9999, 0, .))

internet <- read_excel("data/county_stats/broadband_long2000-2018rev.xlsx") %>%
  filter(year == 2018) %>%
  select(id, broadband, cfips) %>%
  mutate(cfips = str_pad(cfips, width = 5, side = "left", pad = "0"))

# match with population

population <- read_csv("data/county_stats/pop_race.csv", skip = 1) %>%
  clean_names() %>%
  select(id, estimate_sex_and_age_total_population, 
         estimate_sex_and_age_total_population_5_to_9_years,
         estimate_sex_and_age_total_population_5_to_9_years,
         estimate_sex_and_age_total_population_10_to_14_years) %>%
  mutate(adult_pop = estimate_sex_and_age_total_population - 
           estimate_sex_and_age_total_population_5_to_9_years - 
           estimate_sex_and_age_total_population_10_to_14_years) %>%
  select(id, adult_pop)

internet_users <- left_join(internet, population, by = "id") %>%
  drop_na(adult_pop) %>%
  mutate(internet_users = broadband * adult_pop / 1000) %>%
  select(cfips, internet_users)

# Match with DMA

dma_internet <- left_join(internet_users, dma_codes, by = c("cfips" = "fips")) %>%
  drop_na(google_code) %>%
  select(-cfips) %>%
  group_by(google_code) %>%
  summarize(internet_users = sum(internet_users))
  

  
```

```{r}
# Reading in our anchor index

anchor <- read_excel('anchor_index.xlsx')

anchor_list <- c(anchor$Code)
```


```{r}
# Googling our anchor index
# anchor_index <- tibble()

for(j in 44:length(anchor_list)) {
  
  keyword <- tibble()
  
  for(i in 1:54) {
  
    if(i < 53) {
      # Exception case for when we need to search twice for NY
      if(i != 32) {
         temp <- gtrends(keyword = anchor_list[j], geo = c('US-NY-501', dma_codes[(i-1)*4 + 1],
                                                   dma_codes[(i-1)*4 + 2],
                                                   dma_codes[(i-1)*4 + 3],
                                                   dma_codes[(i-1)*4 + 4]),
                              time = '2016-01-01 2020-12-31', 
                           onlyInterest = TRUE)$interest_over_time
         
         temp <- temp %>% 
           mutate(hits = ifelse(hits == "<1", "0.5", hits)) %>%
           mutate(hits = as.numeric(hits)) %>%
           pivot_wider(id_cols = date, names_from = geo,
                                      values_from = hits) %>%
            mutate_at(vars(3:6), ~ . * max(.)/max(`US-NY-501`))
      }
      else {
        temp <- gtrends(keyword = anchor_list[j], geo = c('US-NY-501', dma_codes[(i-1)*4 + 1],
                                                   dma_codes[(i-1)*4 + 2],
                                                   dma_codes[(i-1)*4 + 3],
                                                   dma_codes[(i-1)*4 + 4]),
                              time = '2016-01-01 2020-12-31', 
                           onlyInterest = TRUE)$interest_over_time %>%
          distinct() %>%
          mutate(hits = ifelse(hits == "<1", "0.5", hits)) %>%
          mutate(hits = as.numeric(hits)) %>%
          pivot_wider(id_cols = date, names_from = geo,
                  values_from = hits) %>%
          mutate_at(vars(3:5), ~ . * max(.)/max(`US-NY-501`))
      }
     
    }
    
    else if(i == 53) {
      temp <- gtrends(keyword = anchor_list[j], geo = c('US-NY-501', dma_codes[(i-1)*4 + 1]),
                              time = '2016-01-01 2020-12-31', 
                           onlyInterest = TRUE)$interest_over_time %>%
      mutate(hits = ifelse(hits == "<1", "0.5", hits)) %>%
          mutate(hits = as.numeric(hits)) %>%
      pivot_wider(id_cols = date, names_from = geo,
                  values_from = hits) %>%
      mutate_at(vars(3), ~ . * max(.)/max(`US-NY-501`))
    }
    
    else {
      temp <- gtrends(keyword = anchor_list[j], geo = c('US-NY-501'),
                              time = '2016-01-01 2020-12-31', 
                           onlyInterest = TRUE)$interest_over_time %>%
        mutate(hits = ifelse(hits == "<1", "0.5", hits)) %>%
        mutate(hits = as.numeric(hits)) %>%
        pivot_wider(id_cols = date, names_from = geo,
                    values_from = hits) %>%
        rename(`US-NY-501-FINAL` = `US-NY-501`)
    }
    
    if(i == 1) {
      keyword <- temp
    }
    else {
      keyword <- cbind(keyword, temp)
    }
    Sys.sleep(3)
    
    # print(i)
  }

  keyword_clean <- keyword[, -which(duplicated(names(keyword)))] %>%
    select(-`US-NY-501`) %>%
    rename(`US-NY-501` = `US-NY-501-FINAL`) %>%
    pivot_longer(cols = `US-AK-743`:`US-NY-501`,
                 names_to = "dma", values_to = anchor[[j,2]])
  
  if(j == 1) {
    anchor_index <- keyword_clean
  }
  
  else {
    anchor_index <- left_join(anchor_index, keyword_clean, by = c("dma", "date"))
  }
  
  print(j)
  
  Sys.sleep(10)

}

anchor_index_2 <- anchor_index %>%
  clean_names()


```
```{r}
# Rolling average

anchor_index <- anchor_index_2 %>%
  mutate(across(c(`honey`:`japan`), ~ . * 100 /max(.))) %>%
  rowwise() %>%
  mutate(average = mean(c_across(`honey`:`japan`))) %>%
  ungroup() %>%
  group_by(dma) %>%
  arrange(date) %>%
  mutate(average_3 = rollmean(average, k = 3, fill = "extend"))

searches_v9 <- left_join(searches_v8, 
                         anchor_index %>% mutate(date = as_date(date)) %>% 
                                                   select(date, dma, average, average_3), 
                         by = c("dma", "date")) 



searches_norm3 <- searches_v9 %>%
  mutate(across(c(`alcohol`:`worry`, depression), ~ ./average_3)) %>%
  # renormalize everything to 100 %>%
  mutate(across(c(`alcohol`:`worry`, depression), ~ . * 100 / max(.)))

```
```{r}
# Redoing average excluding the terms w/ more than 15% of zeroes

# zero_percent_anchor <- anchor_index %>%
#   group_by(dma) %>%
#   select(dma, `honey`:`japan`) %>% 
#   summarize(across(.cols = c(`honey`:`japan`), ~ sum(.==0)),
#             weeks = n()) %>%
#   mutate(across(.cols = c(`honey`:`japan`), ~ . * 100 / weeks))
# 
# zero_summary_anchor <- zero_percent_anchor %>%
#   summarize(across(.cols = `honey`:`japan`, mean)) %>%
#   pivot_longer(cols = `honey`:`japan`, names_to = "term", values_to= "zero_percent")

anchor_index_2 <- anchor_index %>%
  select(-elevator, -cereal, -waste, -recycling, -debit_card, -soft_drinks, -french_fries,
         -grain, -currency) %>%
  mutate(across(c(`honey`:`japan`), ~ . * 100 /max(.))) %>%
  rowwise() %>%
  mutate(average = mean(c_across(`honey`:`japan`))) %>%
  ungroup() %>%
  group_by(dma) %>%
  arrange(date) %>%
  mutate(average_3 = rollmean(average, k = 3, fill = "extend"))

searches_norm3_v3 <- left_join(searches_v10 %>% select(-average, -average_3), 
                         anchor_index_2 %>% mutate(date = as_date(date)) %>% 
                                                   select(date, dma, average, average_3), 
                         by = c("dma", "date")) %>% 
  mutate(across(c(`alcohol`:`worry`, depression), ~ ./average_3)) %>%
  # renormalize everything to 100
  mutate(across(c(`alcohol`:`worry`, depression), ~ . * 100 / max(.)))

```

```{r}
# Merge in new search terms that we're interested in

new_search <- cannabis_scaled %>%
  left_join(., cigarette_scaled, by = c("dma", "date")) %>%
  left_join(., drug_scaled, by = c("dma", "date")) %>% 
  left_join(., e_cig_scaled, by = c("dma", "date")) %>% 
  left_join(., exercise_scaled, by = c("dma", "date")) %>% 
  left_join(., lonely_scaled, by = c("dma", "date")) %>% 
  left_join(., sleep_scaled, by = c("dma", "date")) %>% 
  left_join(., stress_scaled, by = c("dma", "date")) %>% 
  left_join(., yoga_scaled, by = c("dma", "date"))

searches_norm3_v4 <- searches_norm3_v3 %>%
  select(-loneliness, -stress) %>%
  left_join(., new_search, by = c("dma", "date")) %>%
  mutate(across(`cigarette`:`yoga`, ~ ./average_3)) %>%
  mutate(across(`cigarette`:`yoga`, ~ . * 100 / max(.)))

searches_v11 <- searches_v10 %>%
  select(-loneliness, -stress) %>%
  left_join(., new_search, by = c("dma", "date"))

# Joining in mental health searches

searches_norm3_v5 <- searches_norm3_v4 %>%
  left_join(., mental_health_clean, by = c("dma", "date")) %>%
  mutate(mental_health =  mental_health /average_3) %>%
  mutate(mental_health = mental_health * 100 / max(mental_health))

searches_v12 <- searches_v11 %>%
  left_join(., mental_health_clean, by = c("dma", "date"))

searches_norm3_v6 <- searches_norm3_v5 %>%
  left_join(., covid_clean, by = c("dma", "date")) %>%
  mutate(covid =  covid /average_3) %>%
  mutate(covid = covid * 100 / max(covid)) %>%
  left_join(., lockdown_clean, by = c("dma", "date")) %>%
  mutate(lockdown =  lockdown /average_3) %>%
  mutate(lockdown = lockdown * 100 / max(lockdown))

searches_v13 <- searches_v12 %>%
  left_join(., covid_clean, by = c("dma", "date")) %>%
  left_join(., lockdown_clean, by = c("dma", "date"))

searches_norm3_v7 <- searches_norm3_v6 %>%
  left_join(., streaming_media_clean, by = c("dma", "date")) %>%
  mutate(streaming_media =  streaming_media /average_3) %>%
  mutate(streaming_media = streaming_media * 100 / max(streaming_media))

searches_v14 <- searches_v13 %>%
  left_join(., streaming_media_clean, by = c("dma", "date"))

searches_norm3_v8 <- searches_norm3_v7 %>%
  select(-covid, -lockdown) %>% 
  left_join(., covid_clean_2, by = c("dma", "date")) %>%
  mutate(covid =  covid /average_3) %>%
  mutate(covid = covid * 100 / max(covid)) %>% 
  left_join(., lockdown_clean_2, by = c("dma", "date")) %>%
  mutate(lockdown =  lockdown /average_3) %>%
  mutate(lockdown = lockdown * 100 / max(lockdown))

searches_v15 <- searches_v14 %>%
  select(-covid, -lockdown) %>% 
  left_join(., covid_clean_2, by = c("dma", "date")) %>%
  left_join(., lockdown_clean_2, by = c("dma", "date"))

```
```{r}
# second batch of huge search joins

searches_norm3_v11 <- searches_norm3_v10 %>%
  left_join(., alone_clean, by = c("dma", "date")) %>%
  mutate(alone = alone/average_3) %>%
  mutate(alone = alone * 100 / max(alone)) %>%
  select(-depression) %>% 
  left_join(., depression_clean, by = c("dma", "date")) %>%
  mutate(depression = depression/average_3) %>%
  mutate(depression = depression * 100 / max(depression)) %>% 
  left_join(., insomnia_clean, by = c("dma", "date")) %>%
  mutate(insomnia = insomnia/average_3) %>%
  mutate(insomnia = insomnia * 100 / max(insomnia)) %>% 
  left_join(., nervous_clean, by = c("dma", "date")) %>%
  mutate(nervous = nervous/average_3) %>%
  mutate(nervous = nervous * 100 / max(nervous)) %>% 
  select(-panic) %>%
  left_join(., panic_clean, by = c("dma", "date")) %>%
  mutate(panic = panic/average_3) %>%
  mutate(panic = panic * 100 / max(panic)) %>% 
  select(-stress) %>%
  left_join(., stress_clean, by = c("dma", "date")) %>%
  mutate(stress = stress/average_3) %>%
  mutate(stress = stress * 100 / max(stress)) %>% 
  left_join(., tired_clean, by = c("dma", "date")) %>%
  mutate(tired = tired/average_3) %>%
  mutate(tired = tired * 100 / max(tired)) %>% 
  select(-worry) %>%
  left_join(., worry_clean, by = c("dma", "date")) %>%
  mutate(worry = worry/average_3) %>%
  mutate(worry = worry * 100 / max(worry)) %>% 
  select(-yoga) %>%
  left_join(., yoga_clean, by = c("dma", "date")) %>%
  mutate(yoga = yoga/average_3) %>%
  mutate(yoga = yoga * 100 / max(yoga)) %>%
  select(-sadness) %>%
  left_join(., sad_clean, by = c("dma", "date")) %>%
  mutate(sad = sad/average_3) %>%
  mutate(sad = sad * 100 / max(sad))

searches_v18 <- searches_v17 %>%
  select(-depression, -panic, -worry, -yoga, -sadness, -stress) %>%
  left_join(., alone_clean, by = c("dma", "date")) %>%
  left_join(., depression_clean, by = c("dma", "date")) %>%
  left_join(., insomnia_clean, by = c("dma", "date")) %>%
  left_join(., nervous_clean, by = c("dma", "date")) %>%
  left_join(., panic_clean, by = c("dma", "date")) %>%
  left_join(., stress_clean, by = c("dma", "date")) %>%
  left_join(., tired_clean, by = c("dma", "date")) %>%
  left_join(., worry_clean, by = c("dma", "date")) %>%
  left_join(., yoga_clean, by = c("dma", "date")) %>%
  left_join(., sad_clean, by = c("dma", "date"))
```
```{r}
# for state codes 
state_codes <- read_rds('state_codes.rds')

state_codes <- c(state_codes$sub_code)
```

```{r}
search_term <- "/m/0k_9"

anxiety <- tibble()

search_term <- "depression"

depression <- tibble()

search_term <- "lockdown"

lockdown <- tibble()

search_term <- "/g/11j2cc_qll"

search_term <- "/m/03c1dkx"

therapy <- tibble()

search_term <- "tired"

tired <- tibble()



  
  
```

```{r}
# State changes

for(i in 1:14) {
  
  if(i < 13) {
    # Exception case for when we need to search twice for NY
    if(i != 8) {
       temp <- gtrends(keyword = search_term, geo = c('US-NY', state_codes[(i-1)*4 + 1],
                                                 state_codes[(i-1)*4 + 2],
                                                 state_codes[(i-1)*4 + 3],
                                                 state_codes[(i-1)*4 + 4]),
                            time = '2020-04-19 2021-09-27', 
                         onlyInterest = TRUE)$interest_over_time
       
       temp <- temp %>% 
          mutate(hits = ifelse(hits == "<1", "0.5", hits)) %>%
          mutate(hits = as.numeric(hits)) %>%
         pivot_wider(id_cols = date, names_from = geo,
                                    values_from = hits) %>%
                        mutate_at(vars(3:6), ~ . * max(.)/max(`US-NY`))
    }
    else {
      temp <- gtrends(keyword = search_term, geo = c('US-NY', state_codes[(i-1)*4 + 1],
                                                 state_codes[(i-1)*4 + 2],
                                                 state_codes[(i-1)*4 + 3],
                                                 state_codes[(i-1)*4 + 4]),
                            time = '2020-04-19 2021-09-27', 
                         onlyInterest = TRUE)$interest_over_time %>%
        distinct() %>%
        mutate(hits = ifelse(hits == "<1", "0.5", hits)) %>%
        mutate(hits = as.numeric(hits)) %>%
        pivot_wider(id_cols = date, names_from = geo,
                values_from = hits) %>%
        mutate_at(vars(3:5), ~ . * max(.)/max(`US-NY`))
    }
   
  }
  
  else if(i == 13) {
    temp <- gtrends(keyword = search_term, geo = c('US-NY', state_codes[(i-1)*4 + 1], 
                                                   state_codes[(i-1)*4 + 2], 
                                                   state_codes[(i-1)*4 + 3]),
                            time = '2020-04-19 2021-09-27', 
                         onlyInterest = TRUE)$interest_over_time %>%
      mutate(hits = ifelse(hits == "<1", "0.5", hits)) %>%
      mutate(hits = as.numeric(hits)) %>%
      pivot_wider(id_cols = date, names_from = geo,
                values_from = hits) %>%
      mutate_at(vars(3:5), ~ . * max(.)/max(`US-NY`))
  }
  
  else{
    temp <- gtrends(keyword = search_term, geo = c('US-NY'),
                            time = '2020-04-19 2021-09-27', 
                         onlyInterest = TRUE)$interest_over_time %>%
      mutate(hits = ifelse(hits == "<1", "0.5", hits)) %>%
      mutate(hits = as.numeric(hits)) %>%
      pivot_wider(id_cols = date, names_from = geo,
                  values_from = hits) %>%
      rename(`US-NY-FINAL` = `US-NY`)
  }
  
  if(i == 1) {
    tired <- temp
  }
  else {
    tired <- cbind(tired, temp)
  }
  Sys.sleep(3)
  
  print(i)
}

tired_clean <- tired[, -which(duplicated(names(tired)))] %>%
  select(-`US-NY`) %>%
  rename(`US-NY` = `US-NY-FINAL`) %>% 
  pivot_longer(cols = `US-AL`:`US-NY`,
               names_to = "state", values_to = "tired")

saveRDS(tired_clean, file = "tired_state.rds")
```


```{r}
# Looped version for multiple terms

search_term <- c("nervous", "stress", "worry", "panic")

search_meaning <- search_term

# search_term <- c('/m/0463cq4', 'lonely', 'sad', '/m/06z5s')
# 
# search_meaning <- c('crying', 'lonely', 'sad', 'suicide')

for(j in 1:length(search_term)) {
  for(i in 1:14) {
  
    if(i < 13) {
      # Exception case for when we need to search twice for NY
      if(i != 8) {
         temp <- gtrends(keyword = search_term[j], geo = c('US-NY', state_codes[(i-1)*4 + 1],
                                                   state_codes[(i-1)*4 + 2],
                                                   state_codes[(i-1)*4 + 3],
                                                   state_codes[(i-1)*4 + 4]),
                              time = '2020-04-19 2021-09-27', 
                           onlyInterest = TRUE)$interest_over_time
         
         temp <- temp %>% 
            mutate(hits = ifelse(hits == "<1", "0.5", hits)) %>%
            mutate(hits = as.numeric(hits)) %>%
           pivot_wider(id_cols = date, names_from = geo,
                                      values_from = hits) %>%
                          mutate_at(vars(3:6), ~ . * max(.)/max(`US-NY`))
      }
      else {
        temp <- gtrends(keyword = search_term[j], geo = c('US-NY', state_codes[(i-1)*4 + 1],
                                                   state_codes[(i-1)*4 + 2],
                                                   state_codes[(i-1)*4 + 3],
                                                   state_codes[(i-1)*4 + 4]),
                              time = '2020-04-19 2021-09-27', 
                           onlyInterest = TRUE)$interest_over_time %>%
          distinct() %>%
          mutate(hits = ifelse(hits == "<1", "0.5", hits)) %>%
          mutate(hits = as.numeric(hits)) %>%
          pivot_wider(id_cols = date, names_from = geo,
                  values_from = hits) %>%
          mutate_at(vars(3:5), ~ . * max(.)/max(`US-NY`))
      }
     
    }
    
    else if(i == 13) {
      temp <- gtrends(keyword = search_term[j], geo = c('US-NY', state_codes[(i-1)*4 + 1], 
                                                     state_codes[(i-1)*4 + 2], 
                                                     state_codes[(i-1)*4 + 3]),
                              time = '2020-04-19 2021-09-27', 
                           onlyInterest = TRUE)$interest_over_time %>%
        mutate(hits = ifelse(hits == "<1", "0.5", hits)) %>%
        mutate(hits = as.numeric(hits)) %>%
        pivot_wider(id_cols = date, names_from = geo,
                  values_from = hits) %>%
        mutate_at(vars(3:5), ~ . * max(.)/max(`US-NY`))
    }
    
    else{
      temp <- gtrends(keyword = search_term[j], geo = c('US-NY'),
                              time = '2020-04-19 2021-09-27', 
                           onlyInterest = TRUE)$interest_over_time %>%
        mutate(hits = ifelse(hits == "<1", "0.5", hits)) %>%
        mutate(hits = as.numeric(hits)) %>%
        pivot_wider(id_cols = date, names_from = geo,
                    values_from = hits) %>%
        rename(`US-NY-FINAL` = `US-NY`)
    }
    
    if(i == 1) {
      total <- temp
    }
    else {
      total <- cbind(total, temp)
    }
    Sys.sleep(3)
    
  }
  
  total_clean <- total[, -which(duplicated(names(total)))] %>%
    select(-`US-NY`) %>%
    rename(`US-NY` = `US-NY-FINAL`) %>% 
    pivot_longer(cols = `US-AL`:`US-NY`,
                 names_to = "state", values_to = paste(search_meaning[j]))
  
  if(j == 1) {
      total_total <- total_clean
  }
  else {
      total_total <- left_join(total_total, total_clean, by = c("date", "state"))
  }

  if(j == length(search_term)) {
    saveRDS(total_total, file = "anxiety_terms_state.rds")
    
  }
  print(j)
  Sys.sleep(1)
  
}




```

```{r}
# for the covid/lockdown searches - combine and then lag by a week

for(i in 1:14) {
  
  if(i < 13) {
    # Exception case for when we need to search twice for NY
    if(i != 8) {
       temp <- gtrends(keyword = search_term, geo = c('US-NY', state_codes[(i-1)*4 + 1],
                                                 state_codes[(i-1)*4 + 2],
                                                 state_codes[(i-1)*4 + 3],
                                                 state_codes[(i-1)*4 + 4]),
                            time = '2020-04-12 2021-09-27', 
                         onlyInterest = TRUE)$interest_over_time
       
       temp <- temp %>% 
          mutate(hits = ifelse(hits == "<1", "0.5", hits)) %>%
          mutate(hits = as.numeric(hits)) %>%
         pivot_wider(id_cols = date, names_from = geo,
                                    values_from = hits) %>%
                        mutate_at(vars(3:6), ~ . * max(.)/max(`US-NY`))
    }
    else {
      temp <- gtrends(keyword = search_term, geo = c('US-NY', state_codes[(i-1)*4 + 1],
                                                 state_codes[(i-1)*4 + 2],
                                                 state_codes[(i-1)*4 + 3],
                                                 state_codes[(i-1)*4 + 4]),
                            time = '2020-04-12 2021-09-27', 
                         onlyInterest = TRUE)$interest_over_time %>%
        distinct() %>%
        mutate(hits = ifelse(hits == "<1", "0.5", hits)) %>%
        mutate(hits = as.numeric(hits)) %>%
        pivot_wider(id_cols = date, names_from = geo,
                values_from = hits) %>%
        mutate_at(vars(3:5), ~ . * max(.)/max(`US-NY`))
    }
   
  }
  
  else if(i == 13) {
    temp <- gtrends(keyword = search_term, geo = c('US-NY', state_codes[(i-1)*4 + 1], 
                                                   state_codes[(i-1)*4 + 2], 
                                                   state_codes[(i-1)*4 + 3]),
                            time = '2020-04-12 2021-09-27', 
                         onlyInterest = TRUE)$interest_over_time %>%
      mutate(hits = ifelse(hits == "<1", "0.5", hits)) %>%
      mutate(hits = as.numeric(hits)) %>%
      pivot_wider(id_cols = date, names_from = geo,
                values_from = hits) %>%
      mutate_at(vars(3:5), ~ . * max(.)/max(`US-NY`))
  }
  
  else{
    temp <- gtrends(keyword = search_term, geo = c('US-NY'),
                            time = '2020-04-12 2021-09-27', 
                         onlyInterest = TRUE)$interest_over_time %>%
      mutate(hits = ifelse(hits == "<1", "0.5", hits)) %>%
      mutate(hits = as.numeric(hits)) %>%
      pivot_wider(id_cols = date, names_from = geo,
                  values_from = hits) %>%
      rename(`US-NY-FINAL` = `US-NY`)
  }
  
  if(i == 1) {
    covid <- temp
  }
  else {
    covid <- cbind(covid, temp)
  }
  Sys.sleep(3)
  
  print(i)
}

covid_clean <- covid[, -which(duplicated(names(covid)))] %>%
  select(-`US-NY`) %>%
  rename(`US-NY` = `US-NY-FINAL`) %>% 
  pivot_longer(cols = `US-AL`:`US-NY`,
               names_to = "state", values_to = "covid")

saveRDS(covid_clean, file = "covid_state.rds")
```

```{r}
#l lag the other searches

lockdown_clean_2 <- left_join(lockdown_clean, covid_clean, by = c("date", "state")) %>%
  group_by(state) %>%
  arrange(date) %>%
  mutate(lockdown = lag(lockdown), covid = lag(covid)) %>%
  drop_na(lockdown)
```

```{r}
anchor <- read_excel('anchor_index.xlsx')

anchor_list <- c(anchor$Code)
```


```{r}
# State anchor bank

for(j in 16:length(anchor_list)) {
  
  keyword <- tibble()
  
  for(i in 1:14) {
  
    if(i < 13) {
      # Exception case for when we need to search twice for NY
      if(i != 8) {
         temp <- gtrends(keyword = anchor_list[j], geo = c('US-NY', state_codes[(i-1)*4 + 1],
                                                   state_codes[(i-1)*4 + 2],
                                                   state_codes[(i-1)*4 + 3],
                                                   state_codes[(i-1)*4 + 4]),
                              time = '2020-04-19 2021-09-27', 
                           onlyInterest = TRUE)$interest_over_time
         
         temp <- temp %>% 
           mutate(hits = ifelse(hits == "<1", "0.5", hits)) %>%
           mutate(hits = as.numeric(hits)) %>%
           pivot_wider(id_cols = date, names_from = geo,
                                      values_from = hits) %>%
            mutate_at(vars(3:6), ~ . * max(.)/max(`US-NY`))
      }
      else {
        temp <- gtrends(keyword = anchor_list[j], geo = c('US-NY', state_codes[(i-1)*4 + 1],
                                                   state_codes[(i-1)*4 + 2],
                                                   state_codes[(i-1)*4 + 3],
                                                   state_codes[(i-1)*4 + 4]),
                              time = '2020-04-19 2021-09-27', 
                           onlyInterest = TRUE)$interest_over_time %>%
          distinct() %>%
          mutate(hits = ifelse(hits == "<1", "0.5", hits)) %>%
          mutate(hits = as.numeric(hits)) %>%
          pivot_wider(id_cols = date, names_from = geo,
                  values_from = hits) %>%
          mutate_at(vars(3:5), ~ . * max(.)/max(`US-NY`))
      }
     
    }
    
    else if(i == 13) {
      temp <- gtrends(keyword = anchor_list[j], geo = c('US-NY', state_codes[(i-1)*4 + 1],
                                                        state_codes[(i-1)*4 + 2], 
                                                   state_codes[(i-1)*4 + 3]),
                              time = '2020-04-19 2021-09-27', 
                           onlyInterest = TRUE)$interest_over_time %>%
      mutate(hits = ifelse(hits == "<1", "0.5", hits)) %>%
          mutate(hits = as.numeric(hits)) %>%
      pivot_wider(id_cols = date, names_from = geo,
                  values_from = hits) %>%
      mutate_at(vars(3:5), ~ . * max(.)/max(`US-NY`))
    }
    
    else {
      temp <- gtrends(keyword = anchor_list[j], geo = c('US-NY'),
                              time = '2020-04-19 2021-09-27', 
                           onlyInterest = TRUE)$interest_over_time %>%
        mutate(hits = ifelse(hits == "<1", "0.5", hits)) %>%
        mutate(hits = as.numeric(hits)) %>%
        pivot_wider(id_cols = date, names_from = geo,
                    values_from = hits) %>%
        rename(`US-NY-FINAL` = `US-NY`)
    }
    
    if(i == 1) {
      keyword <- temp
    }
    else {
      keyword <- cbind(keyword, temp)
    }
    Sys.sleep(3)
    
    # print(i)
  }

  keyword_clean <- keyword[, -which(duplicated(names(keyword)))] %>%
    select(-`US-NY`) %>%
    rename(`US-NY` = `US-NY-FINAL`) %>%
    pivot_longer(cols = `US-AL`:`US-NY`,
                 names_to = "state", values_to = anchor[[j,2]])
  
  if(j == 1) {
    anchor_index <- keyword_clean
  }
  
  else {
    anchor_index <- left_join(anchor_index, keyword_clean, by = c("state", "date"))
  }
  
  print(j)
  
  Sys.sleep(10)

}


anchor_index_2 <- anchor_index %>%
  clean_names() %>% 
  select(-elevator, -cereal, -waste, -recycling, -debit_card, -soft_drinks, -french_fries,
         -grain, -currency) %>%
  mutate(across(c(`honey`:`japan`), ~ . * 100 /max(.))) %>%
  rowwise() %>%
  mutate(average = mean(c_across(`honey`:`japan`))) %>%
  ungroup() %>%
  group_by(state) %>%
  arrange(date) %>%
  mutate(average_3 = rollmean(average, k = 3, fill = "extend"))

saveRDS(anchor_index_2, file = "anchor_index_state.rds")
```


```{r}

state_searches <- anxiety_clean %>%
  left_join(.,depression_clean, by = c("state", "date")) %>%
  left_join(., anchor_index_2 %>% select(state, date, average, average_3), 
            by = c("state", "date"))

state_searches_norm <- state_searches %>%
  mutate(depression = depression / average_3, 
         anxiety = anxiety / average_3) %>% 
  mutate(depression = depression * 100 / max(depression),
         anxiety = anxiety * 100 / max(anxiety))
```

```{r}
# State anchor bank for alcohol

for(j in 1:length(anchor_list)) {
  
  keyword <- tibble()
  
  for(i in 1:14) {
  
    if(i < 13) {
      # Exception case for when we need to search twice for NY
      if(i != 8) {
         temp <- gtrends(keyword = anchor_list[j], geo = c('US-NY', state_codes[(i-1)*4 + 1],
                                                   state_codes[(i-1)*4 + 2],
                                                   state_codes[(i-1)*4 + 3],
                                                   state_codes[(i-1)*4 + 4]),
                              time = '2017-01-01 2020-12-31', 
                           onlyInterest = TRUE)$interest_over_time
         
         temp <- temp %>% 
           mutate(hits = ifelse(hits == "<1", "0.5", hits)) %>%
           mutate(hits = as.numeric(hits)) %>%
           pivot_wider(id_cols = date, names_from = geo,
                                      values_from = hits) %>%
            mutate_at(vars(3:6), ~ . * max(.)/max(`US-NY`))
      }
      else {
        temp <- gtrends(keyword = anchor_list[j], geo = c('US-NY', state_codes[(i-1)*4 + 1],
                                                   state_codes[(i-1)*4 + 2],
                                                   state_codes[(i-1)*4 + 3],
                                                   state_codes[(i-1)*4 + 4]),
                              time = '2017-01-01 2020-12-31', 
                           onlyInterest = TRUE)$interest_over_time %>%
          distinct() %>%
          mutate(hits = ifelse(hits == "<1", "0.5", hits)) %>%
          mutate(hits = as.numeric(hits)) %>%
          pivot_wider(id_cols = date, names_from = geo,
                  values_from = hits) %>%
          mutate_at(vars(3:5), ~ . * max(.)/max(`US-NY`))
      }
     
    }
    
    else if(i == 13) {
      temp <- gtrends(keyword = anchor_list[j], geo = c('US-NY', state_codes[(i-1)*4 + 1],
                                                        state_codes[(i-1)*4 + 2], 
                                                   state_codes[(i-1)*4 + 3]),
                              time = '2017-01-01 2020-12-31', 
                           onlyInterest = TRUE)$interest_over_time %>%
      mutate(hits = ifelse(hits == "<1", "0.5", hits)) %>%
          mutate(hits = as.numeric(hits)) %>%
      pivot_wider(id_cols = date, names_from = geo,
                  values_from = hits) %>%
      mutate_at(vars(3:5), ~ . * max(.)/max(`US-NY`))
    }
    
    else {
      temp <- gtrends(keyword = anchor_list[j], geo = c('US-NY'),
                              time = '2017-01-01 2020-12-31', 
                           onlyInterest = TRUE)$interest_over_time %>%
        mutate(hits = ifelse(hits == "<1", "0.5", hits)) %>%
        mutate(hits = as.numeric(hits)) %>%
        pivot_wider(id_cols = date, names_from = geo,
                    values_from = hits) %>%
        rename(`US-NY-FINAL` = `US-NY`)
    }
    
    if(i == 1) {
      keyword <- temp
    }
    else {
      keyword <- cbind(keyword, temp)
    }
    Sys.sleep(3)
    
    # print(i)
  }

  keyword_clean <- keyword[, -which(duplicated(names(keyword)))] %>%
    select(-`US-NY`) %>%
    rename(`US-NY` = `US-NY-FINAL`) %>%
    pivot_longer(cols = `US-AL`:`US-NY`,
                 names_to = "state", values_to = anchor[[j,2]])
  
  if(j == 1) {
    anchor_index <- keyword_clean
  }
  
  else {
    anchor_index <- left_join(anchor_index, keyword_clean, by = c("state", "date"))
  }
  
  print(j)
  
  Sys.sleep(10)

}


anchor_index_2 <- anchor_index %>%
  clean_names() %>% 
  select(-elevator, -cereal, -waste, -recycling, -debit_card, -soft_drinks, -french_fries,
         -grain, -currency) %>%
  mutate(across(c(`honey`:`japan`), ~ . * 100 /max(.))) %>%
  rowwise() %>%
  mutate(average = mean(c_across(`honey`:`japan`))) %>%
  ungroup() %>%
  mutate(month = month(date),
         year = year(date)) %>% 
  group_by(state, month, year) %>%
  summarize(average = mean(average)) %>%
  mutate(state = str_sub(state, start = 4, end = 5)) %>%
  left_join(., state_fips, by = c("state" = "postal_code"))  %>% 
  drop_na(fips) %>%
  mutate(fips = as.character(fips)) %>% 
  mutate(fips = str_pad(fips, width = 2, side = "left", pad = "0"))

saveRDS(anchor_index_2, file = "anchor_index_state_alc.rds")
```

```{r}
# defining search term

search_term <- "/m/02vxs3q"

search_term <- "/m/012jc" # alcoholism

alcohol <- tibble()

# read in state codes

state_codes <- c(state_codes$sub_code)

# read in state fips

state_fips <- read_excel('statefips.xlsx') %>%
  select(-ame) %>%
  clean_names()
```

```{r}
# New searches for alc

for(i in 1:14) {
  
  if(i < 13) {
    # Exception case for when we need to search twice for NY
    if(i != 8) {
       temp <- gtrends(keyword = search_term, geo = c('US-NY', state_codes[(i-1)*4 + 1],
                                                 state_codes[(i-1)*4 + 2],
                                                 state_codes[(i-1)*4 + 3],
                                                 state_codes[(i-1)*4 + 4]),
                            time = '2017-01-01 2020-12-31', 
                         onlyInterest = TRUE)$interest_over_time
       
       temp <- temp %>% 
          mutate(hits = ifelse(hits == "<1", "0.5", hits)) %>%
          mutate(hits = as.numeric(hits)) %>%
         pivot_wider(id_cols = date, names_from = geo,
                                    values_from = hits) %>%
                        mutate_at(vars(3:6), ~ . * max(.)/max(`US-NY`))
    }
    else {
      temp <- gtrends(keyword = search_term, geo = c('US-NY', state_codes[(i-1)*4 + 1],
                                                 state_codes[(i-1)*4 + 2],
                                                 state_codes[(i-1)*4 + 3],
                                                 state_codes[(i-1)*4 + 4]),
                            time = '2017-01-01 2020-12-31', 
                         onlyInterest = TRUE)$interest_over_time %>%
        distinct() %>%
        mutate(hits = ifelse(hits == "<1", "0.5", hits)) %>%
        mutate(hits = as.numeric(hits)) %>%
        pivot_wider(id_cols = date, names_from = geo,
                values_from = hits) %>%
        mutate_at(vars(3:5), ~ . * max(.)/max(`US-NY`))
    }
   
  }
  
  else if(i == 13) {
    temp <- gtrends(keyword = search_term, geo = c('US-NY', state_codes[(i-1)*4 + 1], 
                                                   state_codes[(i-1)*4 + 2], 
                                                   state_codes[(i-1)*4 + 3]),
                            time = '2017-01-01 2020-12-31', 
                         onlyInterest = TRUE)$interest_over_time %>%
      mutate(hits = ifelse(hits == "<1", "0.5", hits)) %>%
      mutate(hits = as.numeric(hits)) %>%
      pivot_wider(id_cols = date, names_from = geo,
                values_from = hits) %>%
      mutate_at(vars(3:5), ~ . * max(.)/max(`US-NY`))
  }
  
  else{
    temp <- gtrends(keyword = search_term, geo = c('US-NY'),
                            time = '2017-01-01 2020-12-31', 
                         onlyInterest = TRUE)$interest_over_time %>%
      mutate(hits = ifelse(hits == "<1", "0.5", hits)) %>%
      mutate(hits = as.numeric(hits)) %>%
      pivot_wider(id_cols = date, names_from = geo,
                  values_from = hits) %>%
      rename(`US-NY-FINAL` = `US-NY`)
  }
  
  if(i == 1) {
    alcohol <- temp
  }
  else {
    alcohol <- cbind(alcohol, temp)
  }
  Sys.sleep(3)
  
  print(i)
}

alcohol_clean <- alcohol[, -which(duplicated(names(alcohol)))] %>%
  select(-`US-NY`) %>%
  rename(`US-NY` = `US-NY-FINAL`) %>% 
  pivot_longer(cols = `US-AL`:`US-NY`,
               names_to = "state", values_to = "alcohol") %>% 
  mutate(year = year(date),
         month = month(date)) %>%
  select(-date) %>%
  group_by(month, year, state) %>%
  summarize(alcohol = mean(alcohol)) %>% 
  mutate(state = str_sub(state, start = 4, end = 5)) %>%
  left_join(., state_fips, by = c("state" = "postal_code"))  %>% 
  drop_na(fips) %>%
  mutate(fips = as.character(fips)) %>% 
  mutate(fips = str_pad(fips, width = 2, side = "left", pad = "0"))
# this doesn't include DC but alc data doesn't have it anw

saveRDS(alcohol_clean, file = "alcohol_state.rds")
```
```{r}
# Normalizing alcohol searches against this

alcohol_norm <- left_join(alcohol_clean, anchor_index_2, by = c("month", "year", "state", "fips")) %>%
  ungroup() %>% 
  mutate(alcohol_norm = alcohol/average) %>%
  mutate(alcohol_norm = alcohol_norm * 100 / max(alcohol_norm))

# alcoholism 
alcohol_v2 <- alcohol_final %>%
  left_join(alcohol_clean %>% rename("alcoholism" = "alcohol")) %>%
  mutate(alcoholism_norm = alcoholism/average) %>% 
  mutate(alcoholism_norm = alcoholism_norm * 100 / max(alcoholism_norm))
```

```{r}
search_term <- "/m/02vxs3q" # smoking

search_term <- "/m/09k76" # cigarettes

search_term <- "/m/02vlckk" # e-cigarettes



smoking <- tibble()
```

```{r}

# New searches for smoking


for(i in 1:14) {
  
  if(i < 13) {
    # Exception case for when we need to search twice for NY
    if(i != 8) {
       temp <- gtrends(keyword = search_term, geo = c('US-NY', state_codes[(i-1)*4 + 1],
                                                 state_codes[(i-1)*4 + 2],
                                                 state_codes[(i-1)*4 + 3],
                                                 state_codes[(i-1)*4 + 4]),
                            time = '2019-01-01 2021-07-31', 
                         onlyInterest = TRUE)$interest_over_time
       
       temp <- temp %>% 
          mutate(hits = ifelse(hits == "<1", "0.5", hits)) %>%
          mutate(hits = as.numeric(hits)) %>%
         pivot_wider(id_cols = date, names_from = geo,
                                    values_from = hits) %>%
                        mutate_at(vars(3:6), ~ . * max(.)/max(`US-NY`))
    }
    else {
      temp <- gtrends(keyword = search_term, geo = c('US-NY', state_codes[(i-1)*4 + 1],
                                                 state_codes[(i-1)*4 + 2],
                                                 state_codes[(i-1)*4 + 3],
                                                 state_codes[(i-1)*4 + 4]),
                            time = '2019-01-01 2021-07-31', 
                         onlyInterest = TRUE)$interest_over_time %>%
        distinct() %>%
        mutate(hits = ifelse(hits == "<1", "0.5", hits)) %>%
        mutate(hits = as.numeric(hits)) %>%
        pivot_wider(id_cols = date, names_from = geo,
                values_from = hits) %>%
        mutate_at(vars(3:5), ~ . * max(.)/max(`US-NY`))
    }
   
  }
  
  else if(i == 13) {
    temp <- gtrends(keyword = search_term, geo = c('US-NY', state_codes[(i-1)*4 + 1], 
                                                   state_codes[(i-1)*4 + 2], 
                                                   state_codes[(i-1)*4 + 3]),
                            time = '2019-01-01 2021-07-31', 
                         onlyInterest = TRUE)$interest_over_time %>%
      mutate(hits = ifelse(hits == "<1", "0.5", hits)) %>%
      mutate(hits = as.numeric(hits)) %>%
      pivot_wider(id_cols = date, names_from = geo,
                values_from = hits) %>%
      mutate_at(vars(3:5), ~ . * max(.)/max(`US-NY`))
  }
  
  else{
    temp <- gtrends(keyword = search_term, geo = c('US-NY'),
                            time = '2019-01-01 2021-07-31', 
                         onlyInterest = TRUE)$interest_over_time %>%
      mutate(hits = ifelse(hits == "<1", "0.5", hits)) %>%
      mutate(hits = as.numeric(hits)) %>%
      pivot_wider(id_cols = date, names_from = geo,
                  values_from = hits) %>%
      rename(`US-NY-FINAL` = `US-NY`)
  }
  
  if(i == 1) {
    smoking <- temp
  }
  else {
    smoking <- cbind(smoking, temp)
  }
  Sys.sleep(3)
  
  print(i)
}

smoking_clean <- smoking[, -which(duplicated(names(smoking)))] %>%
  select(-`US-NY`) %>%
  rename(`US-NY` = `US-NY-FINAL`) %>% 
  pivot_longer(cols = `US-AL`:`US-NY`,
               names_to = "state", values_to = "smoking") %>% 
  mutate(year = year(date),
         month = month(date)) %>%
  select(-date) %>%
  group_by(month, year, state) %>%
  summarize(smoking = mean(smoking)) %>% 
  mutate(state = str_sub(state, start = 4, end = 5)) %>%
  left_join(., state_fips, by = c("state" = "postal_code"))  %>% 
  drop_na(fips) %>%
  mutate(fips = as.character(fips)) %>% 
  mutate(fips = str_pad(fips, width = 2, side = "left", pad = "0"))
# this doesn't include DC but alc data doesn't have it anw

saveRDS(smoking_clean, file = "smoking_state.rds")
```

```{r}
# State anchor bank for smoking

for(j in 4:length(anchor_list)) {
  
  keyword <- tibble()
  
  for(i in 1:14) {
  
    if(i < 13) {
      # Exception case for when we need to search twice for NY
      if(i != 8) {
         temp <- gtrends(keyword = anchor_list[j], geo = c('US-NY', state_codes[(i-1)*4 + 1],
                                                   state_codes[(i-1)*4 + 2],
                                                   state_codes[(i-1)*4 + 3],
                                                   state_codes[(i-1)*4 + 4]),
                              time = '2019-01-01 2021-07-31', 
                           onlyInterest = TRUE)$interest_over_time
         
         temp <- temp %>% 
           mutate(hits = ifelse(hits == "<1", "0.5", hits)) %>%
           mutate(hits = as.numeric(hits)) %>%
           pivot_wider(id_cols = date, names_from = geo,
                                      values_from = hits) %>%
            mutate_at(vars(3:6), ~ . * max(.)/max(`US-NY`))
      }
      else {
        temp <- gtrends(keyword = anchor_list[j], geo = c('US-NY', state_codes[(i-1)*4 + 1],
                                                   state_codes[(i-1)*4 + 2],
                                                   state_codes[(i-1)*4 + 3],
                                                   state_codes[(i-1)*4 + 4]),
                              time = '2019-01-01 2021-07-31', 
                           onlyInterest = TRUE)$interest_over_time %>%
          distinct() %>%
          mutate(hits = ifelse(hits == "<1", "0.5", hits)) %>%
          mutate(hits = as.numeric(hits)) %>%
          pivot_wider(id_cols = date, names_from = geo,
                  values_from = hits) %>%
          mutate_at(vars(3:5), ~ . * max(.)/max(`US-NY`))
      }
     
    }
    
    else if(i == 13) {
      temp <- gtrends(keyword = anchor_list[j], geo = c('US-NY', state_codes[(i-1)*4 + 1],
                                                        state_codes[(i-1)*4 + 2], 
                                                   state_codes[(i-1)*4 + 3]),
                              time = '2019-01-01 2021-07-31', 
                           onlyInterest = TRUE)$interest_over_time %>%
      mutate(hits = ifelse(hits == "<1", "0.5", hits)) %>%
          mutate(hits = as.numeric(hits)) %>%
      pivot_wider(id_cols = date, names_from = geo,
                  values_from = hits) %>%
      mutate_at(vars(3:5), ~ . * max(.)/max(`US-NY`))
    }
    
    else {
      temp <- gtrends(keyword = anchor_list[j], geo = c('US-NY'),
                              time = '2019-01-01 2021-07-31', 
                           onlyInterest = TRUE)$interest_over_time %>%
        mutate(hits = ifelse(hits == "<1", "0.5", hits)) %>%
        mutate(hits = as.numeric(hits)) %>%
        pivot_wider(id_cols = date, names_from = geo,
                    values_from = hits) %>%
        rename(`US-NY-FINAL` = `US-NY`)
    }
    
    if(i == 1) {
      keyword <- temp
    }
    else {
      keyword <- cbind(keyword, temp)
    }
    Sys.sleep(3)
    
    # print(i)
  }

  keyword_clean <- keyword[, -which(duplicated(names(keyword)))] %>%
    select(-`US-NY`) %>%
    rename(`US-NY` = `US-NY-FINAL`) %>%
    pivot_longer(cols = `US-AL`:`US-NY`,
                 names_to = "state", values_to = anchor[[j,2]])
  
  if(j == 1) {
    anchor_index <- keyword_clean
  }
  
  else {
    anchor_index <- left_join(anchor_index, keyword_clean, by = c("state", "date"))
  }
  
  print(j)
  
  Sys.sleep(5)

}


anchor_index_2 <- anchor_index %>%
  clean_names() %>% 
  select(-elevator, -cereal, -waste, -recycling, -debit_card, -soft_drinks, -french_fries,
         -grain, -currency) %>%
  mutate(across(c(`honey`:`japan`), ~ . * 100 /max(.))) %>%
  rowwise() %>%
  mutate(average = mean(c_across(`honey`:`japan`))) %>%
  ungroup() %>%
  mutate(month = month(date),
         year = year(date)) %>% 
  group_by(state, month, year) %>%
  summarize(average = mean(average)) %>%
  mutate(state = str_sub(state, start = 4, end = 5)) %>%
  left_join(., state_fips, by = c("state" = "postal_code"))  %>% 
  drop_na(fips) %>%
  mutate(fips = as.character(fips)) %>% 
  mutate(fips = str_pad(fips, width = 2, side = "left", pad = "0"))

saveRDS(anchor_index_2, file = "anchor_index_state_smoking.rds")
```

```{r}
# Merge all these together

# read in cig sales and smoking state

cig_sales <- cig_sales %>% 
  mutate(month = as.numeric(month))

# Read in population data for per capita work

pop <- read_csv('data/state_data/R12941084_SL040.csv') %>% 
  clean_names() %>% 
  select(geo_fips, se_a00001_001) %>% 
  rename('total_pop' = 'se_a00001_001')

smoking <- left_join(smoking_state, cig_sales, by = c("fips" = "fips_state",
                                                      "state",
                                             "month", "year")) %>% 
  drop_na(packs_mil) %>% 
  filter(packs_mil >= 0) %>% 
  select(-date) %>% 
  left_join(anchor_index_2, by = c("state", "month", "year", "fips")) %>% 
  mutate(smoking_norm = smoking /average) %>% 
  mutate(smoking_norm = smoking_norm * 100 / max(smoking_norm)) %>% 
  left_join(pop, by = c("fips" = "geo_fips")) %>% 
  mutate(packs_per_capita = packs_mil * 1000000 / total_pop)


```

```{r}
# Adding in the new smoking searches

smoking_v2 <- smoking_final %>%
  left_join(smoking_clean %>% rename("cigarettes" = "smoking")) %>% 
  mutate(cigarettes_norm = cigarettes / average) %>% 
  mutate(cigarettes_norm = cigarettes_norm * 100 / max(cigarettes_norm))

smoking_v3 <- smoking_v2 %>%
  left_join(smoking_clean %>% rename("e_cig" = "smoking")) %>% 
  mutate(e_cig_norm = e_cig / average) %>% 
  mutate(e_cig_norm = e_cig_norm * 100 / max(e_cig_norm))
```

```{r}
# Adding in addiction

search_term <- "/m/0l08" # addiction

search_term <- "/m/0278r2" # drug rehab

addiction <- tibble()


```

```{r}
# Run this

for(i in 1:54) {
  
  if(i < 53) {
    # Exception case for when we need to search twice for NY
    if(i != 32) {
       temp <- gtrends(keyword = search_term, geo = c('US-NY-501', dma_codes[(i-1)*4 + 1],
                                                 dma_codes[(i-1)*4 + 2],
                                                 dma_codes[(i-1)*4 + 3],
                                                 dma_codes[(i-1)*4 + 4]),
                            time = '2015-12-25 2020-12-31', 
                         onlyInterest = TRUE)$interest_over_time
       
       temp <- temp %>% 
          mutate(hits = ifelse(hits == "<1", "0.5", hits)) %>%
          mutate(hits = as.numeric(hits)) %>%
         pivot_wider(id_cols = date, names_from = geo,
                                    values_from = hits) %>%
                        mutate_at(vars(3:6), ~ . * max(.)/max(`US-NY-501`))
    }
    else {
      temp <- gtrends(keyword = search_term, geo = c('US-NY-501', dma_codes[(i-1)*4 + 1],
                                                 dma_codes[(i-1)*4 + 2],
                                                 dma_codes[(i-1)*4 + 3],
                                                 dma_codes[(i-1)*4 + 4]),
                            time = '2015-12-25 2020-12-31', 
                         onlyInterest = TRUE)$interest_over_time %>%
        distinct() %>%
        mutate(hits = ifelse(hits == "<1", "0.5", hits)) %>%
        mutate(hits = as.numeric(hits)) %>%
        pivot_wider(id_cols = date, names_from = geo,
                values_from = hits) %>%
        mutate_at(vars(3:5), ~ . * max(.)/max(`US-NY-501`))
    }
   
  }
  
  else if(i == 53) {
    temp <- gtrends(keyword = search_term, geo = c('US-NY-501', dma_codes[(i-1)*4 + 1]),
                            time = '2015-12-25 2020-12-31', 
                         onlyInterest = TRUE)$interest_over_time %>%
      mutate(hits = ifelse(hits == "<1", "0.5", hits)) %>%
      mutate(hits = as.numeric(hits)) %>%
      pivot_wider(id_cols = date, names_from = geo,
                values_from = hits) %>%
      mutate_at(vars(3), ~ . * max(.)/max(`US-NY-501`))
  }
  
  else{
    temp <- gtrends(keyword = search_term, geo = c('US-NY-501'),
                            time = '2015-12-25 2020-12-31', 
                         onlyInterest = TRUE)$interest_over_time %>%
      mutate(hits = ifelse(hits == "<1", "0.5", hits)) %>%
      mutate(hits = as.numeric(hits)) %>%
      pivot_wider(id_cols = date, names_from = geo,
                  values_from = hits) %>%
      rename(`US-NY-501-FINAL` = `US-NY-501`)
  }
  
  if(i == 1) {
    addiction <- temp
  }
  else {
    addiction <- cbind(addiction, temp)
  }
  Sys.sleep(3)
  
  print(i)
}

addiction_clean <- addiction[, -which(duplicated(names(addiction)))] %>%
  select(-`US-NY-501`) %>%
  rename(`US-NY-501` = `US-NY-501-FINAL`) %>% 
  pivot_longer(cols = `US-AK-743`:`US-NY-501`,
               names_to = "dma", values_to = "addiction")

saveRDS(addiction_clean, file = "addiction_scaled.rds")
```

```{r}
# Add it in 

# this is for rehab

addiction_clean <- addiction_clean %>%
  rename('rehab' = 'addiction')

searches_norm3_v20 <- searches_norm3_v19 %>%
  left_join(addiction_clean, by = c("dma", "date")) %>% 
  mutate(rehab = rehab / average_3) %>% 
  mutate(rehab = rehab * 100 / max(rehab))

searches_norm3_v20_5 <- searches_norm3_v20 %>%
  filter(dma %in% dma_5)

searches_v24 <- searches_v23 %>%
  left_join(addiction_clean, by = c("dma", "date"))

searches_v24_5 <- searches_v24 %>%
  filter(dma %in% dma_5)

```
