==> 3_lockdowndates.Rmd <==
---
title: "3_lockdowndates"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(tidyverse)
library(janitor)
library(readxl)
library(lubridate)
```

```{r}
interventions_dirty <- 
read_excel('data/policy/Copy-of-Copy-of-Local-Policy-Responses-to-COVID-19.fin_ (1).xlsx',
           sheet = 'Local Policies') %>%
  # Get rid of city-specific policies
  filter(is.na(cityname)) %>%
  select(stfips, 
         countyfips, stsipstart, stsipend, localsipstart, localsipend,
         dummysipstart, dummysipend, stbusclose, localbusclose, stbusopen, 
         localbusopen, dummybusclose, dummybusopen) %>%
  # Pick the earliest starts and the latest ends
  mutate(sipstart = ifelse(dummysipstart == 1, as_date(localsipstart), as_date(stsipstart)),
         sipend = ifelse(dummysipend == 1, as_date(localsipend), as_date(stsipend)),
         busclose = ifelse(dummybusclose == 1, as_date(localbusclose), as_date(stbusclose)),
         busopen = ifelse(dummybusopen == 1, as_date(localbusopen), as_date(stbusopen))) %>%
  select(stfips, countyfips, sipstart, sipend, busclose, busopen) %>%
  mutate(stfips = str_pad(stfips, width = 2, side = "left", pad = "0"),
         countyfips = str_pad(countyfips, width = 3, side = "left", pad = "0")) %>%
  mutate(fips = str_c(stfips, countyfips, sep = "")) %>%
  select(-stfips, -countyfips) %>%
  mutate(sipstart = as_date(sipstart),
         sipend = as_date(sipend),
         busclose = as_date(busclose),
         busopen = as_date(busopen)) %>%
  # Replace NAs with 2021 dates that never trigger to stop NA errors
  mutate(sipstart = replace_na(sipstart, as_date("2021-01-01")),
         sipend = replace_na(sipend, as_date("2021-01-01")),
         busclose = replace_na(busclose, as_date("2021-01-01")),
         busopen = replace_na(busopen, as_date("2021-01-01")))
```


```{r}
# Let's try to aggregate this by DMA and (maybe) week -- read in 
# goolsbee interventions

dma_codes <- read_csv('dma_codes.csv') %>%
  select(google_code, statefp, cntyfp) %>%
  mutate(statefp = str_pad(statefp, width = 2, side = "left", pad = "0"),
         cntyfp = str_pad(cntyfp, width = 3, side = "left", pad = 0)) %>%
  mutate(fips = str_c(statefp, cntyfp, sep = "")) %>%
  select(google_code, fips)

goolsbee_interventions <- readRDS('clean_data/county/goolsbee_interventions.rds') %>%
  mutate(across(c(`sipstart`:`busopen`), ~ ifelse(. == as_date("2021-01-01"),
                                                   as_date("2030-01-01"), .))) %>%
  mutate(across(c(`sipstart`:`busopen`), as_date))

# Unfinished

# Let's make a tibble of every FIPS-date combination

week_begin <- readRDS('week_begin.rds')

fips <- tibble(goolsbee_interventions$fips)

merged <- merge(week_begin, fips) %>%
  clean_names() %>%
  left_join(., goolsbee_interventions, by = c("goolsbee_interventions_fips" = "fips"))

# population data here

pop <- read_csv('data/county_stats/pop_race.csv', skip = 1) %>% 
  clean_names() %>%
  select(id, estimate_sex_and_age_total_population) %>%
  mutate(fips = str_sub(id, str_length(id) - 4, str_length(id))) %>%
  select(-id)

# This gives us a dataframe of all county implementation dates

merged2 <- merged %>%
  as_tibble() %>%
  mutate(date = as_date(date)) %>%
  mutate(sip = ifelse(sipstart <= (date + days(6)) & sipend >= date + days(6),
                      1, 0),
         busclose = ifelse(busclose <= (date + days(6)) & busopen >= date + days(6),
                           1, 0)) %>%
  select(date, goolsbee_interventions_fips, sip, busclose) %>%
  # merge in population data
  left_join(., pop, by = c("goolsbee_interventions_fips" = "fips")) %>%
  mutate(sip_pop = sip * estimate_sex_and_age_total_population,
         busclose_pop = busclose * estimate_sex_and_age_total_population)

# Now let's group this by DMA too -- read in the DMA codes here

merged3 <- left_join(merged2, dma_codes, by = c("goolsbee_interventions_fips" = "fips")) %>%
  drop_na(google_code) %>%
  group_by(date, google_code) %>%
  summarize(dmas = n(),
            sip_sum = sum(sip),
            busclose_sum = sum(busclose),
            pop = sum(estimate_sex_and_age_total_population),
            sip_pop = sum(sip_pop),
            busclose_pop = sum(busclose_pop)) %>%
  mutate(sip_prop = sip_sum / dmas,
         busclose_prop = busclose_sum / dmas,
         sip_first = ifelse(sip_sum > 0, 1, 0),
         busclose_first = ifelse(busclose_sum > 0, 1, 0),
         sip_pop_pct = sip_pop/pop,
         busclose_pop_pct = busclose_pop/pop) %>%
  mutate(sip_most_dma = ifelse(sip_prop >= 0.5, 1, 0),
         busclose_most_dma = ifelse(busclose_prop >= 0.5, 1, 0),
         sip_most_pop = ifelse(sip_pop_pct >= 0.5, 1, 0),
         busclose_most_pop = ifelse(busclose_pop_pct >= 0.5, 1, 0)) %>%
  select(date, google_code, sip_prop, busclose_prop,
         sip_first, busclose_first, sip_most_dma, busclose_most_dma,
         sip_pop_pct, busclose_pop_pct, sip_most_pop,
         busclose_most_pop)
```

```{r}
# Recoding this for both lockdown ending and beginning

# start with the merge from previous chunk

merged2 <- merged %>%
  as_tibble() %>%
  mutate(date = as_date(date)) %>%
  mutate(sip_start = ifelse(sipstart <= (date + days(6)),
                      1, 0),
         sip_end = ifelse(sipend <= date + days(6), 1, 0),
         busclose = ifelse(busclose <= (date + days(6)),
                           1, 0),
         busopen = ifelse(busopen <= date + days(6), 1, 0)) %>%
  select(date, goolsbee_interventions_fips, sip_start, sip_end, busclose, busopen) %>%
  # merge in population data
  left_join(., pop, by = c("goolsbee_interventions_fips" = "fips")) %>%
  mutate(sip_pop = sip_start * estimate_sex_and_age_total_population,
         no_sip_pop = sip_end * estimate_sex_and_age_total_population,
         busclose_pop = busclose * estimate_sex_and_age_total_population,
         busopen_pop = busopen * estimate_sex_and_age_total_population)

merged3 <- left_join(merged2, dma_codes, by = c("goolsbee_interventions_fips" = "fips")) %>%
  drop_na(google_code) %>%
  group_by(date, google_code) %>% 
  summarize(dmas = n(),
            sip_sum = sum(sip_start),
            no_sip_sum = sum(sip_end),
            busclose_sum = sum(busclose),
            busopen_sum = sum(busopen),
            pop = sum(estimate_sex_and_age_total_population),
            sip_pop = sum(sip_pop),
            no_sip_pop = sum(no_sip_pop),
            busclose_pop = sum(busclose_pop),
            busopen_pop = sum(busopen_pop)) %>% 
  mutate(sip_prop = sip_sum / dmas,
         no_sip_prop = no_sip_sum/dmas,
         busclose_prop = busclose_sum / dmas,
         busopen_prop = busopen_sum/dmas,
         sip_first = ifelse(sip_sum > 0, 1, 0),
         no_sip_first = ifelse(no_sip_sum > 0, 1, 0),
         busclose_first = ifelse(busclose_sum > 0, 1, 0),
         busopen_first = ifelse(busopen_sum > 0, 1, 0),
         sip_pop_pct = sip_pop/pop,
         no_sip_pop_pct = no_sip_pop / pop,
         busclose_pop_pct = busclose_pop/pop,
         busopen_pop_pct = busopen_pop/pop) %>% 
   mutate(sip_most_dma = ifelse(sip_prop >= 0.5, 1, 0),
          no_sip_most_dma = ifelse(no_sip_prop >= 0.5, 1, 0),
         busclose_most_dma = ifelse(busclose_prop >= 0.5, 1, 0),
         busopen_most_dma = ifelse(busopen_prop >= 0.5, 1, 0),
         sip_most_pop = ifelse(sip_pop_pct >= 0.5, 1, 0),
         no_sip_most_pop = ifelse(no_sip_pop_pct >= 0.5, 1, 0),
         busclose_most_pop = ifelse(busclose_pop_pct >= 0.5, 1, 0),
         busopen_most_pop = ifelse(busopen_pop_pct >= 0.5, 1, 0)) %>% 
  select(-dmas, -sip_sum, -no_sip_sum, -busclose_sum, -busopen_sum,
         -sip_pop, -no_sip_pop, -busclose_pop, -busopen_pop)
  
```

```{r}
# now let's make a slight modification to our lockdowns indicator by making it
# for every year

dma_interventions <- readRDS('clean_data/dma_interventions_v3.rds') %>%
  mutate(year = year(date)) %>%
  group_by(year, google_code) %>%
  mutate(week = 1:n()) %>%
  ungroup()

dma_interventions_20 <- dma_interventions %>%
  filter(year == 2020) %>%
  select(-date, -year)

# Note: version 2 just complies with our regression model estimation thing

dma_interventions_2 <- dma_interventions %>%
  select(date, google_code, year, week) %>%
  left_join(., dma_interventions_20, by = c("google_code", "week")) %>%
  group_by(google_code) %>%
  arrange(date) %>%
  fill(sip_prop:busopen_most_pop, .direction = "down")

searches_norm3_v9 <- searches_norm3_v8 %>%
  select(-week, -busclose_prop, -sip_prop, -sip_first, -busclose_first,
         -sip_most_dma, -busclose_most_dma, -sip_pop_pct, -busclose_pop_pct,
         -sip_pop_pct, -busclose_pop_pct, -sip_most_pop, -busclose_most_pop, -total_pop) %>% 
  left_join(dma_interventions_2, by = c("dma" = "google_code", "date"))

searches_v16 <- searches_v15 %>%
  select(-week, -busclose_prop, -sip_prop, -sip_first, -busclose_first,
         -sip_most_dma, -busclose_most_dma, -sip_pop_pct, -busclose_pop_pct,
         -sip_pop_pct, -busclose_pop_pct, -sip_most_pop, -busclose_most_pop, -total_pop) %>% 
  left_join(dma_interventions_2, by = c("dma" = "google_code", "date"))

# Add in DC data -- nvm guess we've excluded it b/c no crosswalk idk

dc <- readRDS('week_begin.rds') %>%
  as_tibble() %>%
  mutate(date = as_date(date)) %>% 
  mutate(year = year(date)) %>%
  group_by(year) %>%
  mutate(week = 1:n()) %>%
  ungroup() %>%
  mutate(sip_first = ifelse(date >= as_date("2020-04-01") - days(6), 1, 0),
         no_sip_first = ifelse(date >= as_date("2020-05-29") - days(6), 1, 0),
         busclose_first = ifelse(date >= as_date("2020-03-29") - days(6), 1, 0),
         busopen_first = ifelse(date >= as_date("2020-05-29") - days(6), 1, 0)) %>% 
  mutate(sip_prop = sip_first,
         no_sip_prop = no_sip_first,
         busclose_prop = busclose_first,
         busopen_prop = busopen_first,
         sip_pop_pct = sip_first,
         no_sip_pop_pct = no_sip_first,
         sip_most_dma = sip_first,
         no_sip_most_dma = no_sip_first,
         busclose_most_dma = busclose_first,
         busopen_most_dma = busopen_first,
         sip_most_pop = sip_first,
         no_sip_most_pop = no_sip_first,
         busclose_most_pop = busclose_first,
         busopen_most_pop = busopen_first)
  
```

```{r}
# Let's include school closures

# school <- read_csv('data/policy/school.csv', col_types = cols()) %>% 
#   clean_names() %>% 
#   mutate(fips = as.character(fips)) %>% 
#   mutate(fips = str_pad(fips, width = 5, side = "left", pad = "0")) %>%
#   select(-date_partial_school_closed) %>% 
#   mutate(date_all_school_closed = as_date(date_all_school_closed),
#          earliest_date_school_closed = as_date(earliest_date_school_closed)) %>%
#   rename("school_close_all" = "date_all_school_closed",
#          "school_close_first" = "earliest_date_school_closed") %>%
#   mutate(school_close_all = ifelse(is.na(school_close_all), as_date("2030-01-01"),
#                                    school_close_all),
#          school_close_first = ifelse(is.na(school_close_first), as_date("2030-01-01"),
#                                      school_close_first)) %>%
#   mutate(school_close_all = as_date(school_close_all),
#          school_close_first = as_date(school_close_first))

# Use newer data

school <- read_csv('data/policy/interventions.csv') %>% 
  clean_names() %>% 
  select(fips, public_schools) %>% 
  drop_na() %>% 
  mutate(public_schools = as.Date(public_schools, origin = "01-01-01")) %>% 
  mutate(public_schools = as_date(public_schools)) %>% 
  rename('school_close' = 'public_schools')

# Don't need to run again if you've already done it
pop <- read_csv('data/county_stats/pop_race.csv', skip = 1) %>% 
  clean_names() %>%
  select(id, estimate_sex_and_age_total_population) %>%
  mutate(fips = str_sub(id, str_length(id) - 4, str_length(id))) %>%
  select(-id)

week_begin <- readRDS('week_begin.rds')

fips <- tibble(school$fips)

# merged <- merge(week_begin, fips) %>%
#   clean_names() %>%
#   left_join(., school, by = c("school_fips" = "fips")) %>%
#   # ignore the time zone error i don't think it matters
#   mutate(school_first = ifelse(school_close_first <= date + days(6), 1, 0),
#          school_all = ifelse(school_close_all <= date + days(6), 1, 0)) %>%
#   left_join(., pop, by = c("school_fips" = "fips")) %>%
#   mutate(school_pop_first = school_first * estimate_sex_and_age_total_population,
#          school_pop_all = school_all * estimate_sex_and_age_total_population) %>% 
#   select(date, school_fips, school_first, school_all, estimate_sex_and_age_total_population,
#          school_pop_first, school_pop_all)

merged <- merge(week_begin, fips) %>% 
  clean_names() %>% 
  left_join(., school, by = c("school_fips" = "fips")) %>% 
  mutate(school_close = ifelse(school_close <= date + days(6), 1, 0)) %>% 
  left_join(., pop, by = c("school_fips" = "fips")) %>% 
  mutate(school_pop = school_close * estimate_sex_and_age_total_population) %>% 
  drop_na(estimate_sex_and_age_total_population) %>% 
  select(date, school_fips, school_close, school_pop, 
         estimate_sex_and_age_total_population)
  

merged2 <- left_join(merged, dma_codes, by = c("school_fips" = "fips")) %>% 
  drop_na(google_code) %>% 
  group_by(date, google_code) %>% 
  summarize(dmas = n(),
            school = sum(school_close),
            school_pop = sum(school_pop), 
            pop = sum(estimate_sex_and_age_total_population)) %>% 
  mutate(school_close = ifelse(school > 0, 1, 0),
         school_pop_pct = school_pop/pop,
         school_prop = school/dmas) %>% 
  mutate(school_most_pop = ifelse(school_pop_pct >= 0.5, 1, 0),
         school_most_dma = ifelse(school_prop >= 0.5, 1, 0)) %>% 
  select(-dmas, -school, -school_pop, -pop)

# merged2 <- left_join(merged, dma_codes, by = c("school_fips" = "fips")) %>%
#   drop_na(google_code) %>%
#   group_by(date, google_code) %>%
#   summarize(dmas = n(),
#          school_first = sum(school_first),
#          school_all = sum(school_all),
#          school_pop_first = sum(school_pop_first),
#          school_pop_all = sum(school_pop_all),
#          pop = sum(estimate_sex_and_age_total_population)) %>% 
#   mutate(school_first_first = ifelse(school_first > 0, 1, 0),
#          school_all_first = ifelse(school_all > 0, 1, 0),
#          school_first_pop_pct = school_pop_first / pop,
#          school_all_pop_pct = school_pop_all / pop,
#          school_first_prop = school_first / dmas,
#          school_all_prop = school_all / dmas) %>% 
#   mutate(school_first_most_pop = ifelse(school_first_pop_pct >= 0.5, 1, 0),
#          school_all_most_pop = ifelse(school_all_pop_pct >= 0.5, 1, 0),
#          school_first_most_dma = ifelse(school_first_prop >= 0.5, 1, 0),
#          school_all_most_dma = ifelse(school_all_prop >= 0.5, 1, 0)) %>% 
#   select(-dmas, -pop)

merged3 <- merged2 %>%
  mutate(year = year(date)) %>%
  group_by(year, google_code) %>%
  mutate(week = 1:n()) %>%
  ungroup()

school_20 <- merged3  %>%
  filter(year == 2020) %>%
  select(-date, -year)

searches_norm3_v13 <- left_join(searches_norm3_v12, school_20, by = c("dma" = "google_code",
                                                                     "week")) %>% 
  select(-c(school_first:school_all_most_dma))

searches_v20 <- left_join(searches_v19, school_20, by = c("dma" = "google_code",
                                                                     "week")) %>% 
  select(-c(school_first:school_all_most_dma))

searches_norm3_v13


``` 

```{r}
# School closures v2 from MCH data

school <- read_excel('data/policy/covid-school-data.xlsx') %>% 
  clean_names() %>% 
  select(district_nces,
         physical_state, open_date, teaching_method, enrollment)

crosswalk <- read_excel('data/policy/sdlist-20.xls', skip = 2) %>% 
  clean_names() %>% 
  mutate(district_nces = str_c(state_fips, district_id_number, sep = ""),
         county_fips = str_c(state_fips, county_fips, sep = "")) %>% 
  select(-district_id_number, -school_district_name, -county_names)

# Read in dma 

dma_codes <- read_csv('dma_codes.csv') %>%
  select(google_code, statefp, cntyfp) %>%
  mutate(statefp = str_pad(statefp, width = 2, side = "left", pad = "0"),
         cntyfp = str_pad(cntyfp, width = 3, side = "left", pad = 0)) %>%
  mutate(fips = str_c(statefp, cntyfp, sep = "")) %>%
  select(google_code, fips)

school_2 <- left_join(school, crosswalk, by = "district_nces") %>% 
  drop_na(county_fips, open_date) %>% 
  filter(teaching_method != "Pending",
         teaching_method != "Unknown") %>% 
  select(-physical_state) %>% 
  left_join(dma_codes, by = c("county_fips" = "fips")) %>% 
  drop_na(google_code)

# read in week_begin

week_begin <- week_begin %>%
  mutate(online_only = 0, on_premises = 0, hybrid = 0, other = 0)

school_weekly <- expand_grid(week_begin, school_2) %>% 
  select(-district_nces) %>% 
  mutate(enrollment = as.numeric(enrollment),
         open_date = mdy(open_date)) %>% 
  drop_na(enrollment) %>% 
  mutate(online_only = ifelse(teaching_method == "Online Only"  & open_date <= date,
                              online_only + enrollment, online_only),
         on_premises = ifelse(teaching_method == "On Premises" & open_date <= date,
                               on_premises + enrollment, on_premises),
         hybrid = ifelse(teaching_method == "Hybrid" & open_date <= date,
                          hybrid + enrollment, hybrid),
         other = ifelse(teaching_method == "Other" & open_date <= date,
                        other + enrollment, other)) %>% 
  group_by(date, google_code) %>% 
  summarize(across(online_only:other, sum)) %>% 
  mutate(enrollment = online_only + on_premises + hybrid + other) %>% 
  mutate(across(online_only:other, ~ ifelse(enrollment != 0, ./enrollment, .))) %>% 
  select(-enrollment)

school_weekly_20 <- school_weekly %>% 
  filter(year(date) == 2020) %>% 
  group_by(google_code) %>% 
  mutate(week = 1:n())
  

searches_norm3_v14 <- searches_norm3_v13 %>%
  left_join(school_weekly_20, by = c("dma" = "google_code", "week")) %>% 
  rename('school_first' = 'school_close') %>% 
  select(-date.y) %>% 
  rename('date' = 'date.x') %>% 
  group_by(dma) %>% 
  arrange(week) %>% 
  fill(c(school_first:other), .direction = "down") %>% 
  ungroup()

searches_v21 <- searches_v20 %>%
  left_join(school_weekly_20, by = c("dma" = "google_code", "week")) %>% 
  rename('school_first' = 'school_close') %>% 
  select(-date.y) %>% 
  rename('date' = 'date.x') %>% 
  group_by(dma) %>% 
  arrange(week) %>% 
  fill(c(school_first:other), .direction = "down") %>% 
  ungroup()
  

```


```{r}
# State policy spreadsheets -- you need to add a state column to other things first

cgrt <- read_csv('data/policy/OxCGRT_latest.csv', col_types = cols()) %>%
  clean_names() %>%
  filter(country_code == "USA",
         jurisdiction == "STATE_TOTAL") %>%
  select(region_code, date, e1_income_support,
         e2_debt_contract_relief, economic_support_index) %>%
  # Based on previous AK data we replace the only missing values which are from
  # AK
  mutate(e1_income_support = replace_na(e1_income_support, 0),
         economic_support_index = replace_na(economic_support_index, 25)) %>%
  filter(date <= 20210105) %>%
  mutate(region_code = str_sub(region_code, start = 4),
         date = as.character(date)) %>%
  mutate(date = ymd(date))

# Let's do this weekly now 

week_begin <- readRDS('week_begin.rds') %>%
  mutate(date = as_date(date)) %>%
  # we want policy by the week's end
  mutate(week_end = date + days(6))

cgrt_weekly <- cgrt %>%
  filter(date %in% c(week_begin$week_end)) %>%
  mutate(date = date - days(6))
```


```{r}
# Join this onto searches v2 to make v3

searches_v3 <- readRDS('clean_data/search/searches_v2.rds') %>%
  left_join(., cgrt_weekly, by = c("state" = "region_code", "date")) %>%
  mutate(e1_income_support = replace_na(e1_income_support, 0),
         e2_debt_contract_relief = replace_na(e2_debt_contract_relief, 0),
         economic_support_index = replace_na(economic_support_index, 0)) %>%
  rename('income_support' = 'e1_income_support',
         'debt_contract_relief'= 'e2_debt_contract_relief') %>%
  mutate(year = as.factor(year),
         year_indicator = ifelse(year == "2020", 1, 0))

# Fixing searches v3 to have lagged covid cases instead of covid cases

searches_v3 <- readRDS('clean_data/search/searches_v3.rds') %>%
  group_by(dma) %>%
  arrange(date) %>%
  mutate(cases_per_100k = lag(cases_per_100k),
         deaths_per_100k = lag(deaths_per_100k)) %>%
  fill(c(`cases_per_100k`, `deaths_per_100k`), .direction = "up") %>%
  ungroup()

# to do: merge in lockdown dates and make some graphs
```


```{r}
# Note we are missing DC data!!!! -- let's do this manually

dc <- tibble(dma = "US-DC-511", sipstart = as_date("2020-04-01"),
             sipend = as_date("2020-05-29"), busclose = as_date("2020-03-25"),
             busopen = as_date("2020-05-29"))

# Look at the dc unemployment data from here

dc_unemploy <- dc_unemploy %>%
  mutate(year = as.numeric(year))

dma_interventions_v2 <- readRDS('clean_data/dma_interventions_v2.rds') %>%
  select(-year)

searches_v4 <- left_join(searches_v3, dma_interventions_v2, 
                         by = c("date", "dma" = "google_code"))

searches_dc <- searches_v4 %>%
  filter(dma == "US-DC-511") %>%
  select(-unemploy_rate) %>%
  group_by(year) %>%
  arrange(date) %>%
  mutate(week = 1:n()) %>%
  # join in our unemployment rate
  left_join(., dc_unemploy, by = c("year", "month" = "period", "dma")) %>%
  rename("unemploy_rate" = "value") %>%
  mutate(sip_first = ifelse(dc$sipstart[1] >= (date + days(6)) & 
                              dc$sipend[1] <= (date + days(6)),
                            1, 0),
         busclose_first = ifelse(dc$busclose[1] >= (date + days(6)) & 
                              dc$busclose[1] <= (date + days(6)),
                            1, 0)) %>%
  mutate(sip_prop = sip_first,
         busclose_prop = busclose_first,
         sip_most_dma = sip_first,
         busclose_most_dma = busclose_first,
         sip_pop_pct = sip_first,
         busclose_pop_pct = busclose_first,
         sip_most_pop = sip_first,
         busclose_most_pop = busclose_first)

# This is actually searches_v4
searches_v5 <- searches_v4 %>%
  filter(dma != "US-DC-511") %>%
  rbind(searches_dc) %>%
  mutate(year_indicator = ifelse(year == 2020, 1, 0),
         year = as.factor(year))
```


```{r}
# Fixing the depression data

depression <- readRDS('depression_scaled.rds') %>%
  mutate(date = as_date(date))

searches_v4 <- readRDS('searches_v4.rds') %>%
  select(-depression) %>%
  left_join(depression, by = c("date", "dma"))
```


```{r}

# Recoding these to match start and end dates


```

```{r}
# Let's do state telehealth laws?

telehealth <- read_excel('data/policy/COVID-19 US state policy database 9_27_2021.xlsx',
                         sheet = "state_policy") %>%
  clean_names() %>% 
  select(postcode, fips, med1135w, acaenrol, prevtlhl, tlhlaud, tlhlmed, alcrest, alcdeliv)

telehealth_2 <- telehealth %>% 
  mutate(across(c(acaenrol, tlhlaud:alcdeliv), as.numeric)) %>% 
  mutate(across(c(acaenrol, tlhlaud:alcdeliv), ~ ifelse(. == 0, 47484, .))) %>% 
  mutate(across(c(acaenrol, tlhlaud:alcdeliv), ~ as.Date(., origin = "1899-12-30"))) %>% 
  mutate(fips = str_pad(fips, width = 2, side = "left", pad = "0")) %>% 
  mutate(takeout_indicator = ifelse(alcrest == as_date("2030-01-01"), 0, 1),
         delivery_indicator = ifelse(alcdeliv == as_date("2030-01-01"), 0, 1))
```

```{r}
# Convert to weekly

# read in week_begin again and do the expand grid

telehealth_weekly <- expand_grid(telehealth_2, week_begin) %>% 
  mutate(across(c(med1135w:acaenrol, tlhlaud:alcdeliv), ~ ifelse(date >= ., 1, 0)))
```
```{r}
# Read in my custom payment parity handcoded data

pp <- read_excel('data/policy/payment_parity.xlsx') %>% 
  clean_names %>% 
  mutate(pre_covid_pp = pre_covid + pre_covid_medicaid_only,
         covid_pp = during_covid + during_covid_medicaid_only) %>% 
  mutate(pre_covid_pp = as.factor(pre_covid_pp),
         covid_pp = as.factor(covid_pp)) %>% 
  mutate(state = state2abbr(state))
```


```{r}
# Let's merge!

searches_norm3_v12 <- left_join(searches_norm3_v11, telehealth_weekly %>% select(-fips), 
                                by = c("date", "state" = "postcode")) %>%
  left_join(., pp, by = "state")

searches_norm3_v12_5 <- left_join(searches_norm3_v11_5, telehealth_weekly %>% select(-fips), 
                                by = c("date", "state" = "postcode")) %>%
  left_join(., pp, by = "state")

searches_v19 <-  left_join(searches_v18, telehealth_weekly %>% select(-fips), 
                                by = c("date", "state" = "postcode")) %>%
  left_join(., pp, by = "state")
```
