==> 5_merging.Rmd <==
---
title: "5_merging"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(tidyverse)
library(janitor)
library(lubridate)
```

```{r}
# start with searches v6

# read in covid_dma
covid_dma_v2 <- readRDS('clean_data/covid_dma_v2.rds')
# don't need to lag

searches_v7 <- searches_v6 %>%
  select(-cases_per_100k, -deaths_per_100k) %>%
  left_join(., covid_dma_v2, by = c("dma" = "google_code", "date")) %>%
  mutate(cases_per_100k = replace_na(cases_per_100k, 0),
         deaths_per_100k = replace_na(deaths_per_100k, 0))
```


```{r}
# Make some numerical counters for when lockdowns happen
sip_first <- searches_v7 %>%
  filter(sip_first == 1) %>%
  group_by(dma) %>%
  arrange(date) %>%
  slice(1) %>%
  select(dma, week) %>%
  rename('start_week' = 'week')

busclose_first <- searches_v7 %>%
  filter(busclose_first == 1) %>%
  group_by(dma) %>%
  arrange(date) %>%
  slice(1) %>%
  select(dma, week) %>%
  rename('start_week' = 'week')

sip_most_pop <- searches_v7 %>%
  filter(sip_most_pop == 1) %>%
  group_by(dma) %>%
  arrange(date) %>%
  slice(1) %>%
  select(dma, week) %>%
  rename('start_week' = 'week')

busclose_most_pop <- searches_v7 %>%
  filter(busclose_most_pop == 1) %>%
  group_by(dma) %>%
  arrange(date) %>%
  slice(1) %>%
  select(dma, week) %>%
  rename('start_week' = 'week')

busclose_most_dma <- searches_v7 %>%
  filter(busclose_most_dma == 1) %>%
  group_by(dma) %>%
  arrange(date) %>%
  slice(1) %>%
  select(dma, week) %>%
  rename('start_week' = 'week')

sip_most_dma <- searches_v7 %>%
  filter(sip_most_dma == 1) %>%
  group_by(dma) %>%
  arrange(date) %>%
  slice(1) %>%
  select(dma, week) %>%
  rename('start_week' = 'week')

searches_v8 <- searches_v7 %>%
  left_join(., sip_first, by = "dma") %>%
  mutate(sip_first_counter = week - start_week) %>%
  select(-start_week) %>%
  left_join(., busclose_first, by = "dma") %>%
  mutate(busclose_first_counter = week - start_week) %>%
  select(-start_week) %>%
  left_join(., sip_most_pop, by = "dma") %>%
  mutate(sip_most_pop_counter = week - start_week) %>%
  select(-start_week) %>% 
  left_join(., busclose_most_pop, by = "dma") %>%
  mutate(busclose_most_pop_counter = week - start_week) %>%
  select(-start_week) %>%
  left_join(., sip_most_dma, by = "dma") %>%
  mutate(sip_most_dma_counter = week - start_week) %>%
  select(-start_week) %>% 
  left_join(., busclose_most_dma, by = "dma") %>%
  mutate(busclose_most_dma_counter = week - start_week) %>%
  select(-start_week)
```


```{r}
# Let's change up some of the coding practices

searches_norm3_v2 <- searches_norm3 %>%
  mutate(sip_first_2 = ifelse(sip_first == 0, NA, sip_first)) %>%
  group_by(dma, year) %>%
  arrange(date) %>%
  fill(sip_first_2, .direction = "down") %>%
  mutate(sip_first_2 = replace_na(sip_first_2, 0))
```

```{r}
# Doing the counters for sip ends too

sip <- searches_norm3_v20 %>%
  filter(no_sip_most_pop == 1) %>%
  group_by(dma) %>%
  arrange(date) %>%
  slice(1) %>%
  select(dma, week) %>%
  rename('start_week' = 'week')

bus <- searches_norm3_v20 %>%
  filter(busopen_most_pop == 1) %>%
  group_by(dma) %>%
  arrange(date) %>%
  slice(1) %>%
  select(dma, week) %>%
  rename('start_week' = 'week')

searches_norm3_v21 <- searches_norm3_v20 %>%
  left_join(., sip) %>% 
  mutate(no_sip_most_pop_counter = week - start_week) %>% 
  select(-start_week) %>% 
  left_join(., bus) %>% 
  mutate(busopen_most_pop_counter = week - start_week)

searches_norm3_v21_5 <- searches_norm3_v21 %>%
  filter(dma %in% dma_5)

# Let's try making them ordered b/c otherwise hard to interpret


```

```{r}
# Merging weather in!

searches_norm3_v2 <- searches_norm3 %>%
  mutate(year = as.character(year)) %>%
  left_join(., weather_4, by = c("dma" = "google_code", "year", "month")) %>% 
  drop_na(precip, temp)

searches_v10 <- searches_v9 %>%
  mutate(year = as.character(year)) %>%
  left_join(., weather_4, by = c("dma" = "google_code", "year", "month")) %>% 
  drop_na(precip, temp)

```
