==> 6_regressions.Rmd <==
---
title: "5_regressions"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(tidyverse)
library(lfe)
library(texreg)
library(fixest)
```

```{r}
searches_v7 <- readRDS('clean_data/search/searches_v9.rds')

# Let's fuck around a bit and se what happens

searches_postlockdown <- searches_norm3 %>%
  filter(week >= 15)
```

```{r}

# Summary stats

summary_stats <- searches_norm3_v14 %>%
  select(`alcohol`:`TV`, cannabis:streaming_media, alone:sad) %>% 
  summarize(across(alcohol:sad, list(mean = mean, sd = sd))) %>% 
  pivot_longer(cols = alcohol_mean: sad_sd, names_to = "term", values_to = "value")

# Percentage of zeroes

zero_percent <- searches_norm3_v14 %>%
  group_by(dma) %>%
  select(dma, `alcohol`:`TV`, cannabis:streaming_media, alone:sad) %>% 
  summarize(across(.cols = c(`alcohol`:sad), ~ sum(.==0)),
            weeks = n()) %>%
  mutate(across(.cols = c(alcohol:sad), ~ . * 100 / weeks)) %>%
  group_by(dma) %>%
  mutate(mean = mean(alcohol:sad))

zero_percent <- searches_norm3_v20 %>%
  group_by(dma) %>%
  select(dma, addiction, rehab) %>% 
  summarize(across(.cols = c(addiction:rehab), ~ sum(.==0)),
            weeks = n()) %>%
  mutate(across(.cols = c(addiction:rehab), ~ . * 100 / weeks)) %>%
  group_by(dma) %>%
  mutate(mean = mean(addiction:rehab))

zero_percent %>% 
  ungroup() %>% 
  summarize(across(alcohol:sad, list(mean = ~ mean(.), sd = sd))) %>% 
  pivot_longer(cols = alcohol_mean:sad_sd, names_to = "term", values_to = "value")

summary_stats <- zero_percent %>% 
  ungroup() %>% 
  summarize(across(addiction:rehab, list(mean = ~ mean(.), sd = sd)))

# Read in dma covid and merge

dma_pop_zero_percent <- left_join(dma_pop, zero_percent, by = c("google_code" = "dma")) %>%
  drop_na()

dma_5 <- zero_percent %>% 
  filter(mean <= 5) %>% 
  select(dma) %>%
  distinct()

dma_5 <- c(dma_5$dma)

missing_graph <- dma_pop_zero_percent %>% 
  ggplot(mapping = aes(x = total_pop, y = mean)) + 
    geom_point() + 
    scale_x_continuous(trans='log10') + 
    labs(x = "DMA Population (log scale)", y = "Mean Percent Zeroes") + 
    theme_classic()

zero_percent_summary <- zero_percent %>%
  ungroup() %>%
  summarize(across(.cols = alcohol:sad, ~ mean(., na.rm = TRUE))) %>%
  pivot_longer(cols = alcohol:sad, names_to = "term", values_to = "percent_zero")

# testing anchor index too 

zero_percent_anchor <- anchor_index %>%
  group_by(dma) %>%
  select(dma, `honey`:`japan`) %>% 
  summarize(across(.cols = c(`honey`:`japan`), ~ sum(.==0)),
            weeks = n()) %>%
  mutate(across(.cols = c(`honey`:`japan`), ~ . * 100 / weeks))

zero_summary_anchor <- zero_percent_anchor %>%
  summarize(across(.cols = `honey`:`japan`, mean)) %>%
  pivot_longer(cols = `honey`:`japan`, names_to = "term", values_to= "zero_percent")

# Maybe I should revise the anchor index adjusted terms

```

```{r}
# Run only regressions with less than 15 percent zeroes on average?

searches_norm3_v11_5 <- searches_norm3_v11 %>%
  filter(dma %in% dma_5)
```

```{r}
# read in searches v4 or later

test <- felm(I(log(depression + 1)) ~  sip_first * year_indicator + 
               deaths_per_100k + cases_per_100k +  
               unemploy_rate + economic_support_index|factor(dma):factor(year) + week|0|dma, 
             data = searches_v9)

test3 <-  felm(I(log(depression + 1)) ~  
                 sip_first * year_indicator + no_sip_first * year_indicator +
               cases_per_100k +
                 unemploy_rate + factor(debt_contract_relief) + factor(income_support) + 
                 precip + temp + covid + lockdown|
                 factor(dma):factor(year) + factor(week)|0|dma, 
             data = searches_norm3_v9)

test3 <-  felm(I(suicide) ~  sip_first * year_indicator + 
               deaths_per_100k + cases_per_100k +  
               unemploy_rate + economic_support_index|factor(dma):factor(year) + week|0|dma, 
             data = searches_norm3)
```
```{r}
# With re-scaling, depression

dep_sip_scaled <-  felm(I(log(depression + 1)) ~  
                 sip_first * year_indicator + no_sip_most_pop * year_indicator +
               cases_per_100k +
                 unemploy_rate + factor(debt_contract_relief) + factor(income_support) + 
                 precip + temp + covid + lockdown|
                 factor(dma):factor(year) + factor(week)|0|dma, 
             data = searches_norm3_v10)

dep_sip_noscale <- felm(I(log(depression + 1)) ~  
                 sip_first * year_indicator + no_sip_most_pop * year_indicator +
               cases_per_100k +
                 unemploy_rate + factor(debt_contract_relief) + factor(income_support) + 
                 precip + temp + covid + lockdown|
                 factor(dma):factor(year) + factor(week)|0|dma, 
             data = searches_v17)

dep_bus_scaled <- felm(I(log(depression + 1)) ~  
                 busclose_first * year_indicator + busopen_most_pop * year_indicator +
               cases_per_100k +
                 unemploy_rate + factor(debt_contract_relief) + factor(income_support) + 
                 precip + temp + covid + lockdown|
                 factor(dma):factor(year) + factor(week)|0|dma, 
             data = searches_norm3_v10)

dep_bus_noscale <- felm(I(log(depression + 1)) ~  
                 busclose_first * year_indicator + busopen_most_pop * year_indicator +
               cases_per_100k +
                 unemploy_rate + factor(debt_contract_relief) + factor(income_support) + 
                 precip + temp + covid + lockdown|
                 factor(dma):factor(year) + factor(week)|0|dma, 
             data = searches_v17)
```


```{r}

# Let's check school closures just for fun

dep_school_scaled <- felm(I(log(depression + 1)) ~  
                 school_all_first * year_indicator +
               cases_per_100k +
                 unemploy_rate + factor(debt_contract_relief) + factor(income_support) + 
                 precip + temp + covid + lockdown|
                 factor(dma):factor(year) + factor(week)|0|dma, 
             data = searches_norm3_v10)

dep_sip_scaled <-  felm(I(log(depression + 1)) ~  
                 sip_first * year_indicator + no_sip_most_pop * year_indicator +
               cases_per_100k +
                 unemploy_rate + factor(debt_contract_relief) + factor(income_support) + 
                 precip + temp + covid + lockdown|
                 factor(dma):factor(year) + factor(week)|0|dma, 
             data = searches_norm3_v10_5)
```


```{r}
anx_sip_scaled <-  felm(I(log(anxiety + 1)) ~  
                 sip_first * year_indicator + no_sip_most_pop * year_indicator +
               cases_per_100k +
                 unemploy_rate + factor(debt_contract_relief) + factor(income_support) + 
                 precip + temp + covid + lockdown|
                 factor(dma):factor(year) + factor(week)|0|dma, 
             data = searches_norm3_v10)

anx_sip_noscale <- felm(I(log(anxiety + 1)) ~  
                 sip_first * year_indicator + no_sip_most_pop * year_indicator +
               cases_per_100k +
                 unemploy_rate + factor(debt_contract_relief) + factor(income_support) + 
                 precip + temp + covid + lockdown|
                 factor(dma):factor(year) + factor(week)|0|dma, 
             data = searches_v17)

anx_bus_scaled <- felm(I(log(anxiety + 1)) ~  
                 busclose_first * year_indicator + busopen_most_pop * year_indicator +
               cases_per_100k +
                 unemploy_rate + factor(debt_contract_relief) + factor(income_support) + 
                 precip + temp + covid + lockdown|
                 factor(dma):factor(year) + factor(week)|0|dma, 
             data = searches_norm3_v10)

anx_bus_noscale <- felm(I(log(anxiety + 1)) ~  
                 busclose_first * year_indicator + busopen_most_pop * year_indicator +
               cases_per_100k +
                 unemploy_rate + factor(debt_contract_relief) + factor(income_support) + 
                 precip + temp + covid + lockdown|
                 factor(dma):factor(year) + factor(week)|0|dma, 
             data = searches_v17)

htmlreg(l = list(anx_sip_scaled, anx_bus_scaled, anx_sip_noscale, anx_bus_noscale), 
        digits = 4, include.ci = FALSE, file = "anxiety.doc")
```


```{r}
lonely_sip_scaled <-  felm(I(log(lonely + 1)) ~  
                 sip_first * year_indicator + no_sip_most_pop * year_indicator +
               cases_per_100k +
                 unemploy_rate + factor(debt_contract_relief) + factor(income_support) + 
                 precip + temp + covid + lockdown|
                 factor(dma):factor(year) + factor(week)|0|dma, 
             data = searches_norm3_v10)

lonely_sip_noscale <- felm(I(log(lonely + 1)) ~  
                 sip_first * year_indicator + no_sip_most_pop * year_indicator +
               cases_per_100k +
                 unemploy_rate + factor(debt_contract_relief) + factor(income_support) + 
                 precip + temp + covid + lockdown|
                 factor(dma):factor(year) + factor(week)|0|dma, 
             data = searches_v17)

lonely_bus_scaled <- felm(I(log(lonely + 1)) ~  
                 busclose_first * year_indicator + busopen_most_pop * year_indicator +
               cases_per_100k +
                 unemploy_rate + factor(debt_contract_relief) + factor(income_support) + 
                 precip + temp + covid + lockdown|
                 factor(dma):factor(year) + factor(week)|0|dma, 
             data = searches_norm3_v10)

lonely_bus_noscale <- felm(I(log(lonely + 1)) ~  
                 busclose_first * year_indicator + busopen_most_pop * year_indicator +
               cases_per_100k +
                 unemploy_rate + factor(debt_contract_relief) + factor(income_support) + 
                 precip + temp + covid + lockdown|
                 factor(dma):factor(year) + factor(week)|0|dma, 
             data = searches_v17)

htmlreg(l = list(lonely_sip_scaled, lonely_bus_scaled, lonely_sip_noscale, lonely_bus_noscale), 
        digits = 4, include.ci = FALSE, file = "lonely.doc")
```
```{r}
lonely_sip_scaled <-  felm(I(log(lonely + 1)) ~  
                 sip_first * year_indicator + no_sip_most_pop * year_indicator +
               cases_per_100k +
                 unemploy_rate + factor(debt_contract_relief) + factor(income_support) + 
                 precip + temp + covid + lockdown|
                 factor(dma):factor(year) + factor(week)|0|dma, 
             data = searches_norm3_v10)

lonely_sip_noscale <- felm(I(log(lonely + 1)) ~  
                 sip_first * year_indicator + no_sip_most_pop * year_indicator +
               cases_per_100k +
                 unemploy_rate + factor(debt_contract_relief) + factor(income_support) + 
                 precip + temp + covid + lockdown|
                 factor(dma):factor(year) + factor(week)|0|dma, 
             data = searches_v17)

lonely_bus_scaled <- felm(I(log(lonely + 1)) ~  
                 busclose_first * year_indicator + busopen_most_pop * year_indicator +
               cases_per_100k +
                 unemploy_rate + factor(debt_contract_relief) + factor(income_support) + 
                 precip + temp + covid + lockdown|
                 factor(dma):factor(year) + factor(week)|0|dma, 
             data = searches_norm3_v10)

lonely_bus_noscale <- felm(I(log(lonely + 1)) ~  
                 busclose_first * year_indicator + busopen_most_pop * year_indicator +
               cases_per_100k +
                 unemploy_rate + factor(debt_contract_relief) + factor(income_support) + 
                 precip + temp + covid + lockdown|
                 factor(dma):factor(year) + factor(week)|0|dma, 
             data = searches_v17)

htmlreg(l = list(lonely_sip_scaled, lonely_bus_scaled, lonely_sip_noscale, lonely_bus_noscale), 
        digits = 4, include.ci = FALSE, file = "lonely.doc")
```
