==> 7_graphs.Rmd <==
---
title: "graphing"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(tidyverse)
library(ggplot2)
library(zoo)
```

```{r}
depression_busclose <- searches_norm3_v6 %>%
  group_by(year_indicator, busclose_first_counter) %>%
  filter(dma == "US-NY-501") %>%
  summarize(depression = mean(depression)) %>%
  mutate(depression = rollmean(depression, k = 5, fill = NA)) %>%
  ggplot(mapping = aes(x = busclose_first_counter, y = depression, color = factor(year_indicator))) +
    geom_line() + 
    theme_classic(base_size = 14) +
    scale_color_discrete(name = "Year", labels = c("2016-2019", "2020")) +
    labs(title = "Depression vs. Business Closure", y = "SVI", x = "Weeks vs. Lockdown") +
    scale_x_continuous(breaks = seq(-10, 
                                    max(searches_norm3_v6$busclose_first_counter, na.rm = TRUE), 
                                    by = 10))

depression_sip <- searches_norm3 %>%
  group_by(year_indicator, sip_first_counter) %>%
  summarize(depression = mean(depression)) %>%
  mutate(depression = rollmean(depression, k = 5, fill = NA)) %>%
  ggplot(mapping = aes(x = sip_first_counter, y = depression, color = factor(year_indicator))) +   
    geom_line() + 
    theme_classic(base_size = 14) +
    scale_color_discrete(name = "Year", labels = c("2016-2019", "2020")) +
    labs(title = "Depression vs. Shelter-in-place", y = "SVI", x = "Weeks vs. Lockdown") +
    scale_x_continuous(breaks = seq(-10, 
                                    max(searches_norm3$sip_first_counter, na.rm = TRUE), 
                                    by = 10))


anxiety_sip <- searches_norm3 %>%
  group_by(year_indicator, sip_first_counter) %>%
  summarize(anxiety = mean(anxiety)) %>%
  mutate(anxiety = rollmean(anxiety, k = 5, fill = NA)) %>%
  ggplot(mapping = aes(x = sip_first_counter, y = anxiety, color = factor(year_indicator))) +   
    geom_line() + 
    theme_classic(base_size = 14) +
    scale_color_discrete(name = "Year", labels = c("2016-2019", "2020")) +
    labs(title = "anxiety vs. Shelter-in-place", y = "SVI", x = "Weeks vs. Lockdown") +
    scale_x_continuous(breaks = seq(-10, 
                                    max(searches_norm3$sip_first_counter, na.rm = TRUE), 
                                    by = 10))

anxiety_busclose <- searches_norm3 %>%
  group_by(year_indicator, busclose_first_counter) %>%
  summarize(anxiety = mean(anxiety)) %>%
  mutate(anxiety = rollmean(anxiety, k = 5, fill = NA)) %>%
  ggplot(mapping = aes(x = busclose_first_counter, y = anxiety, color = factor(year_indicator))) +   
    geom_line() + 
    theme_classic(base_size = 14) +
    scale_color_discrete(name = "Year", labels = c("2016-2019", "2020")) +
    labs(title = "anxiety vs. Business Closure", y = "SVI", x = "Weeks vs. Lockdown") +
    scale_x_continuous(breaks = seq(-10, 
                                    max(searches_norm3$busclose_first_counter, na.rm = TRUE), 
                                    by = 10))

worry_sip <- searches_norm3 %>%
  group_by(year_indicator, sip_first_counter) %>%
  summarize(worry = mean(worry)) %>%
  mutate(worry = rollmean(worry, k = 5, fill = NA)) %>%

  ggplot(mapping = aes(x = sip_first_counter, y = worry, color = factor(year_indicator))) +   
    geom_line() + 
    theme_classic(base_size = 14) +
    scale_color_discrete(name = "Year", labels = c("2016-2019", "2020")) +
    labs(title = "worry vs. Shelter-in-place", y = "SVI", x = "Weeks vs. Lockdown") +
    scale_x_continuous(breaks = seq(-10, 
                                    max(searches_norm3$sip_first_counter, na.rm = TRUE), 
                                    by = 10))

worry_busclose <- searches_norm3 %>%
  group_by(year_indicator, busclose_first_counter) %>%
  summarize(worry = mean(worry)) %>%
  mutate(worry = rollmean(worry, k = 5, fill = NA)) %>%
  ggplot(mapping = aes(x = busclose_first_counter, y = worry, color = factor(year_indicator))) +   
    geom_line() + 
    theme_classic(base_size = 14) +
    scale_color_discrete(name = "Year", labels = c("2016-2019", "2020")) +
    labs(title = "worry vs. Business Closure", y = "SVI", x = "Weeks vs. Lockdown") +
    scale_x_continuous(breaks = seq(-10, 
                                    max(searches_norm3$busclose_first_counter, na.rm = TRUE), 
                                    by = 10))

loneliness_sip <- searches_norm3 %>%
  group_by(year_indicator, sip_first_counter) %>%
  summarize(loneliness = mean(loneliness)) %>%
  mutate(loneliness = rollmean(loneliness, k = 5, fill = NA)) %>%
  ggplot(mapping = aes(x = sip_first_counter, y = loneliness, color = factor(year_indicator))) +   
    geom_line() + 
    theme_classic(base_size = 14) +
    scale_color_discrete(name = "Year", labels = c("2016-2019", "2020")) +
    labs(title = "loneliness vs. Shelter-in-place", y = "SVI", x = "Weeks vs. Lockdown") +
    scale_x_continuous(breaks = seq(-10, 
                                    max(searches_norm3$sip_first_counter, na.rm = TRUE), 
                                    by = 10))

loneliness_busclose <- searches_norm3 %>%
  group_by(year_indicator, busclose_first_counter) %>%
  summarize(loneliness = mean(loneliness)) %>%
  mutate(loneliness = rollmean(loneliness, k = 5, fill = NA)) %>%
  ggplot(mapping = aes(x = busclose_first_counter, y = loneliness, color = factor(year_indicator))) +   
    geom_line() + 
    theme_classic(base_size = 14) +
    scale_color_discrete(name = "Year", labels = c("2016-2019", "2020")) +
    labs(title = "loneliness vs. Business Closure", y = "SVI", x = "Weeks vs. Lockdown") +
    scale_x_continuous(breaks = seq(-10, 
                                    max(searches_norm3$busclose_first_counter, na.rm = TRUE), 
                                    by = 10))

crying_sip <- searches_norm3 %>%
  group_by(year_indicator, sip_first_counter) %>%
  summarize(crying = mean(crying)) %>%
  mutate(crying = rollmean(crying, k = 5, fill = NA)) %>%
  ggplot(mapping = aes(x = sip_first_counter, y = crying, color = factor(year_indicator))) +   
    geom_line() + 
    theme_classic(base_size = 14) +
    scale_color_discrete(name = "Year", labels = c("2016-2019", "2020")) +
    labs(title = "crying vs. Shelter-in-place", y = "SVI", x = "Weeks vs. Lockdown") +
    scale_x_continuous(breaks = seq(-10, 
                                    max(searches_norm3$sip_first_counter, na.rm = TRUE), 
                                    by = 10))

crying_busclose <- searches_norm3 %>%
  group_by(year_indicator, busclose_first_counter) %>%
  summarize(crying = mean(crying)) %>%
  mutate(crying = rollmean(crying, k = 5, fill = NA)) %>%
  ggplot(mapping = aes(x = busclose_first_counter, y = crying, color = factor(year_indicator))) +   
    geom_line() + 
    theme_classic(base_size = 14) +
    scale_color_discrete(name = "Year", labels = c("2016-2019", "2020")) +
    labs(title = "crying vs. Business Closure", y = "SVI", x = "Weeks vs. Lockdown") +
    scale_x_continuous(breaks = seq(-10, 
                                    max(searches_norm3$busclose_first_counter, na.rm = TRUE), 
                                    by = 10))

stress_sip <- searches_norm3 %>%
  group_by(year_indicator, sip_first_counter) %>%
  summarize(stress = mean(stress)) %>%
  mutate(stress = rollmean(stress, k = 5, fill = NA)) %>%
  ggplot(mapping = aes(x = sip_first_counter, y = stress, color = factor(year_indicator))) +   
    geom_line() + 
    theme_classic(base_size = 14) +
    scale_color_discrete(name = "Year", labels = c("2016-2019", "2020")) +
    labs(title = "stress vs. Shelter-in-place", y = "SVI", x = "Weeks vs. Lockdown") +
    scale_x_continuous(breaks = seq(-10, 
                                    max(searches_norm3$sip_first_counter, na.rm = TRUE), 
                                    by = 10))

stress_busclose <- searches_norm3 %>%
  group_by(year_indicator, busclose_first_counter) %>%
  summarize(stress = mean(stress)) %>%
  mutate(stress = rollmean(stress, k = 5, fill = NA)) %>%
  ggplot(mapping = aes(x = busclose_first_counter, y = stress, color = factor(year_indicator))) +   
    geom_line() + 
    theme_classic(base_size = 14) +
    scale_color_discrete(name = "Year", labels = c("2016-2019", "2020")) +
    labs(title = "stress vs. Business Closure", y = "SVI", x = "Weeks vs. Lockdown") +
    scale_x_continuous(breaks = seq(-10, 
                                    max(searches_norm3$busclose_first_counter, na.rm = TRUE), 
                                    by = 10))


    
```
```{r}
# Substance abuse

anxiety_sip <- searches_norm3 %>%
  group_by(year_indicator, sip_first_counter) %>%
  summarize(anxiety = mean(anxiety)) %>%
  mutate(anxiety = rollmean(anxiety, k = 5, fill = NA)) %>%
  ggplot(mapping = aes(x = sip_first_counter, y = anxiety, color = factor(year_indicator))) +   
    geom_line() + 
    theme_classic(base_size = 14) +
    scale_color_discrete(name = "Year", labels = c("2016-2019", "2020")) +
    labs(title = "anxiety vs. Shelter-in-place", y = "SVI", x = "Weeks vs. Lockdown") +
    scale_x_continuous(breaks = seq(-10, 
                                    max(searches_norm3$sip_first_counter, na.rm = TRUE), 
                                    by = 10))

anxiety_busclose <- searches_norm3 %>%
  group_by(year_indicator, busclose_first_counter) %>%
  summarize(anxiety = mean(anxiety)) %>%
  mutate(anxiety = rollmean(anxiety, k = 5, fill = NA)) %>%
  ggplot(mapping = aes(x = busclose_first_counter, y = anxiety, color = factor(year_indicator))) +   
    geom_line() + 
    theme_classic(base_size = 14) +
    scale_color_discrete(name = "Year", labels = c("2016-2019", "2020")) +
    labs(title = "anxiety vs. Business Closure", y = "SVI", x = "Weeks vs. Lockdown") +
    scale_x_continuous(breaks = seq(-10, 
                                    max(searches_norm3$busclose_first_counter, na.rm = TRUE), 
                                    by = 10))

smoking_sip <- searches_norm3 %>%
  group_by(year_indicator, sip_first_counter) %>%
  summarize(smoking = mean(smoking)) %>%
  mutate(smoking = rollmean(smoking, k = 5, fill = NA)) %>%
  ggplot(mapping = aes(x = sip_first_counter, y = smoking, color = factor(year_indicator))) +   
    geom_line() + 
    theme_classic(base_size = 14) +
    scale_color_discrete(name = "Year", labels = c("2016-2019", "2020")) +
    labs(title = "smoking vs. Shelter-in-place", y = "SVI", x = "Weeks vs. Lockdown") +
    scale_x_continuous(breaks = seq(-10, 
                                    max(searches_norm3$sip_first_counter, na.rm = TRUE), 
                                    by = 10))

smoking_busclose <- searches_norm3 %>%
  group_by(year_indicator, busclose_first_counter) %>%
  summarize(smoking = mean(smoking)) %>%
  mutate(smoking = rollmean(smoking, k = 5, fill = NA)) %>%
  ggplot(mapping = aes(x = busclose_first_counter, y = smoking, color = factor(year_indicator))) +   
    geom_line() + 
    theme_classic(base_size = 14) +
    scale_color_discrete(name = "Year", labels = c("2016-2019", "2020")) +
    labs(title = "smoking vs. Business Closure", y = "SVI", x = "Weeks vs. Lockdown") +
    scale_x_continuous(breaks = seq(-10, 
                                    max(searches_norm3$busclose_first_counter, na.rm = TRUE), 
                                    by = 10))

substance_abuse_sip <- searches_norm3 %>%
  group_by(year_indicator, sip_first_counter) %>%
  summarize(substance_abuse = mean(substance_abuse)) %>%
  mutate(substance_abuse = rollmean(substance_abuse, k = 5, fill = NA)) %>%
  ggplot(mapping = aes(x = sip_first_counter, y = substance_abuse, color = factor(year_indicator))) +   
    geom_line() + 
    theme_classic(base_size = 14) +
    scale_color_discrete(name = "Year", labels = c("2016-2019", "2020")) +
    labs(title = "substance_abuse vs. Shelter-in-place", y = "SVI", x = "Weeks vs. Lockdown") +
    scale_x_continuous(breaks = seq(-10, 
                                    max(searches_norm3$sip_first_counter, na.rm = TRUE), 
                                    by = 10))

substance_abuse_busclose <- searches_norm3 %>%
  group_by(year_indicator, busclose_first_counter) %>%
  summarize(substance_abuse = mean(substance_abuse)) %>%
  mutate(substance_abuse = rollmean(substance_abuse, k = 5, fill = NA)) %>%
  ggplot(mapping = aes(x = busclose_first_counter, y = substance_abuse, color = factor(year_indicator))) +   
    geom_line() + 
    theme_classic(base_size = 14) +
    scale_color_discrete(name = "Year", labels = c("2016-2019", "2020")) +
    labs(title = "substance_abuse vs. Business Closure", y = "SVI", x = "Weeks vs. Lockdown") +
    scale_x_continuous(breaks = seq(-10, 
                                    max(searches_norm3$busclose_first_counter, na.rm = TRUE), 
                                    by = 10))
```


```{r}
# Leisure

books_sip <- searches_norm3 %>%
  group_by(year_indicator, sip_first_counter) %>%
  summarize(books = mean(books)) %>%
  mutate(books = rollmean(books, k = 5, fill = NA)) %>%
  ggplot(mapping = aes(x = sip_first_counter, y = books, color = factor(year_indicator))) +   
    geom_line() + 
    theme_classic(base_size = 14) +
    scale_color_discrete(name = "Year", labels = c("2016-2019", "2020")) +
    labs(title = "books vs. Shelter-in-place", y = "SVI", x = "Weeks vs. Lockdown") +
    scale_x_continuous(breaks = seq(-10, 
                                    max(searches_norm3$sip_first_counter, na.rm = TRUE), 
                                    by = 10))

books_busclose <- searches_norm3 %>%
  group_by(year_indicator, busclose_first_counter) %>%
  summarize(books = mean(books)) %>%
  mutate(books = rollmean(books, k = 5, fill = NA)) %>%
  ggplot(mapping = aes(x = busclose_first_counter, y = books, color = factor(year_indicator))) +   
    geom_line() + 
    theme_classic(base_size = 14) +
    scale_color_discrete(name = "Year", labels = c("2016-2019", "2020")) +
    labs(title = "books vs. Business Closure", y = "SVI", x = "Weeks vs. Lockdown") +
    scale_x_continuous(breaks = seq(-10, 
                                    max(searches_norm3$busclose_first_counter, na.rm = TRUE), 
                                    by = 10))

film_sip <- searches_norm3 %>%
  group_by(year_indicator, sip_first_counter) %>%
  summarize(film = mean(film)) %>%
  mutate(film = rollmean(film, k = 5, fill = NA)) %>%
  ggplot(mapping = aes(x = sip_first_counter, y = film, color = factor(year_indicator))) +   
    geom_line() + 
    theme_classic(base_size = 14) +
    scale_color_discrete(name = "Year", labels = c("2016-2019", "2020")) +
    labs(title = "film vs. Shelter-in-place", y = "SVI", x = "Weeks vs. Lockdown") +
    scale_x_continuous(breaks = seq(-10, 
                                    max(searches_norm3$sip_first_counter, na.rm = TRUE), 
                                    by = 10))

film_busclose <- searches_norm3 %>%
  group_by(year_indicator, busclose_first_counter) %>%
  summarize(film = mean(film)) %>%
  mutate(film = rollmean(film, k = 5, fill = NA)) %>%
  ggplot(mapping = aes(x = busclose_first_counter, y = film, color = factor(year_indicator))) +   
    geom_line() + 
    theme_classic(base_size = 14) +
    scale_color_discrete(name = "Year", labels = c("2016-2019", "2020")) +
    labs(title = "film vs. Business Closure", y = "SVI", x = "Weeks vs. Lockdown") +
    scale_x_continuous(breaks = seq(-10, 
                                    max(searches_norm3$busclose_first_counter, na.rm = TRUE), 
                                    by = 10))

games_sip <- searches_norm3 %>%
  group_by(year_indicator, sip_first_counter) %>%
  summarize(games = mean(games)) %>%
  mutate(games = rollmean(games, k = 5, fill = NA)) %>%
  ggplot(mapping = aes(x = sip_first_counter, y = games, color = factor(year_indicator))) +   
    geom_line() + 
    theme_classic(base_size = 14) +
    scale_color_discrete(name = "Year", labels = c("2016-2019", "2020")) +
    labs(title = "games vs. Shelter-in-place", y = "SVI", x = "Weeks vs. Lockdown") +
    scale_x_continuous(breaks = seq(-10, 
                                    max(searches_norm3$sip_first_counter, na.rm = TRUE), 
                                    by = 10))

games_busclose <- searches_norm3 %>%
  group_by(year_indicator, busclose_first_counter) %>%
  summarize(games = mean(games)) %>%
  mutate(games = rollmean(games, k = 5, fill = NA)) %>%
  ggplot(mapping = aes(x = busclose_first_counter, y = games, color = factor(year_indicator))) +   
    geom_line() + 
    theme_classic(base_size = 14) +
    scale_color_discrete(name = "Year", labels = c("2016-2019", "2020")) +
    labs(title = "games vs. Business Closure", y = "SVI", x = "Weeks vs. Lockdown") +
    scale_x_continuous(breaks = seq(-10, 
                                    max(searches_norm3$busclose_first_counter, na.rm = TRUE), 
                                    by = 10))

music_sip <- searches_norm3 %>%
  group_by(year_indicator, sip_first_counter) %>%
  summarize(music = mean(music)) %>%
  mutate(music = rollmean(music, k = 5, fill = NA)) %>%
  ggplot(mapping = aes(x = sip_first_counter, y = music, color = factor(year_indicator))) +   
    geom_line() + 
    theme_classic(base_size = 14) +
    scale_color_discrete(name = "Year", labels = c("2016-2019", "2020")) +
    labs(title = "music vs. Shelter-in-place", y = "SVI", x = "Weeks vs. Lockdown") +
    scale_x_continuous(breaks = seq(-10, 
                                    max(searches_norm3$sip_first_counter, na.rm = TRUE), 
                                    by = 10))

music_busclose <- searches_norm3 %>%
  group_by(year_indicator, busclose_first_counter) %>%
  summarize(music = mean(music)) %>%
  mutate(music = rollmean(music, k = 5, fill = NA)) %>%
  ggplot(mapping = aes(x = busclose_first_counter, y = music, color = factor(year_indicator))) +   
    geom_line() + 
    theme_classic(base_size = 14) +
    scale_color_discrete(name = "Year", labels = c("2016-2019", "2020")) +
    labs(title = "music vs. Business Closure", y = "SVI", x = "Weeks vs. Lockdown") +
    scale_x_continuous(breaks = seq(-10, 
                                    max(searches_norm3$busclose_first_counter, na.rm = TRUE), 
                                    by = 10))



```


```{r}
ggsave(plot = alcohol_sip, file = "alcohol_sip.png")

ggsave(plot = alcohol_busclose, file = "alcohol_busclose.png")

ggsave(plot = anxiety_sip, file = "anxiety_sip.png")

ggsave(plot = anxiety_busclose, file = "anxiety_busclose.png")

ggsave(plot = books_sip, file = "books_sip.png")

ggsave(plot = books_busclose, file = "books_busclose.png")

ggsave(plot = crying_sip, file = "crying_sip.png")

ggsave(plot = crying_busclose, file = "crying_busclose.png")

ggsave(plot = depression_sip, file = "depression_sip.png")

ggsave(plot = depression_busclose, file = "depression_busclose.png")

ggsave(plot = film_sip, file = "film_sip.png")

ggsave(plot = film_busclose, file = "film_busclose.png")

ggsave(plot = games_sip, file = "games_sip.png")

ggsave(plot = games_busclose, file = "games_busclose.png")

ggsave(plot = loneliness_sip, file = "loneliness_sip.png")

ggsave(plot = loneliness_busclose, file = "loneliness_busclose.png")

ggsave(plot = music_sip, file = "music_sip.png")

ggsave(plot = music_busclose, file = "music_busclose.png")

ggsave(plot = smoking_sip, file = "smoking_sip.png")

ggsave(plot = smoking_busclose, file = "smoking_busclose.png")

ggsave(plot = stress_sip, file = "stress_sip.png")

ggsave(plot = stress_busclose, file = "stress_busclose.png")

ggsave(plot = substance_abuse_sip, file = "substance_abuse_sip.png")

ggsave(plot = substance_abuse_busclose, file = "substance_abuse_busclose.png")

ggsave(plot = worry_sip, file = "worry_sip.png")

ggsave(plot = worry_busclose, file = "worry_busclose.png")
```

```{r}
# Try ratio

depression_busclose <- searches_norm3 %>%
  group_by(year_indicator, busclose_first_counter) %>%
  summarize(depression = mean(depression)) %>%
  pivot_wider(names_from = year_indicator, 
              values_from = depression) %>%
  mutate(ratio = `1`/`0`) %>%
  ggplot(mapping = aes(x = busclose_first_counter, y = ratio)) +   
    geom_line() + 
    theme_bw() +
    labs(title = "Depression vs. Business Closure", y = "SVI ratio (2020 to pre-2020)", 
         x = "Weeks vs. Lockdown") +
    scale_x_continuous(breaks = seq(-10, 
                                    max(searches_norm3$busclose_first_counter, na.rm = TRUE), 
                                    by = 10))

depression_sip <- searches_norm3 %>%
  group_by(year_indicator, sip_first_counter) %>%
  summarize(depression = mean(depression)) %>%
  pivot_wider(names_from = year_indicator, 
              values_from = depression) %>%
  mutate(ratio = `1`/`0`) %>%
  ggplot(mapping = aes(x = sip_first_counter, y = ratio)) +   
    geom_line() + 
    theme_bw() +
    labs(title = "Depression vs. Business Closure", y = "SVI ratio (2020 to pre-2020)", 
         x = "Weeks vs. Lockdown") +
    scale_x_continuous(breaks = seq(-10, 
                                    max(searches_norm3$sip_first_counter, na.rm = TRUE), 
                                    by = 10))

alcohol_busclose <- searches_norm3 %>%
  group_by(year_indicator, busclose_first_counter) %>%
  summarize(alcohol = mean(alcohol)) %>%
  pivot_wider(names_from = year_indicator, 
              values_from = alcohol) %>%
  mutate(ratio = `1`/`0`) %>%
  ggplot(mapping = aes(x = busclose_first_counter, y = ratio)) +   
    geom_line() + 
    theme_bw() +
    labs(title = "alcohol vs. Business Closure", y = "SVI ratio (2020 to pre-2020)", 
         x = "Weeks vs. Lockdown") +
    scale_x_continuous(breaks = seq(-10, 
                                    max(searches_norm3$busclose_first_counter, na.rm = TRUE), 
                                    by = 10))

alcohol_sip <- searches_norm3 %>%
  group_by(year_indicator, sip_first_counter) %>%
  summarize(alcohol = mean(alcohol)) %>%
  pivot_wider(names_from = year_indicator, 
              values_from = alcohol) %>%
  mutate(ratio = `1`/`0`) %>%
  ggplot(mapping = aes(x = sip_first_counter, y = ratio)) +   
    geom_line() + 
    theme_bw() +
    labs(title = "alcohol vs. Business Closure", y = "SVI ratio (2020 to pre-2020)", 
         x = "Weeks vs. Lockdown") +
    scale_x_continuous(breaks = seq(-10, 
                                    max(searches_norm3$sip_first_counter, na.rm = TRUE), 
                                    by = 10))

anxiety_busclose <- searches_norm3 %>%
  group_by(year_indicator, busclose_first_counter) %>%
  summarize(anxiety = mean(anxiety)) %>%
  pivot_wider(names_from = year_indicator, 
              values_from = anxiety) %>%
  mutate(ratio = `1`/`0`) %>%
  ggplot(mapping = aes(x = busclose_first_counter, y = ratio)) +   
    geom_line() + 
    theme_bw() +
    labs(title = "anxiety vs. Business Closure", y = "SVI ratio (2020 to pre-2020)", 
         x = "Weeks vs. Lockdown") +
    scale_x_continuous(breaks = seq(-10, 
                                    max(searches_norm3$busclose_first_counter, na.rm = TRUE), 
                                    by = 10))

anxiety_sip <- searches_norm3 %>%
  group_by(year_indicator, sip_first_counter) %>%
  summarize(anxiety = mean(anxiety)) %>%
  pivot_wider(names_from = year_indicator, 
              values_from = anxiety) %>%
  mutate(ratio = `1`/`0`) %>%
  ggplot(mapping = aes(x = sip_first_counter, y = ratio)) +   
    geom_line() + 
    theme_bw() +
    labs(title = "anxiety vs. Business Closure", y = "SVI ratio (2020 to pre-2020)", 
         x = "Weeks vs. Lockdown") +
    scale_x_continuous(breaks = seq(-10, 
                                    max(searches_norm3$sip_first_counter, na.rm = TRUE), 
                                    by = 10))

books_busclose <- searches_norm3 %>%
  group_by(year_indicator, busclose_first_counter) %>%
  summarize(books = mean(books)) %>%
  pivot_wider(names_from = year_indicator, 
              values_from = books) %>%
  mutate(ratio = `1`/`0`) %>%
  ggplot(mapping = aes(x = busclose_first_counter, y = ratio)) +   
    geom_line() + 
    theme_bw() +
    labs(title = "books vs. Business Closure", y = "SVI ratio (2020 to pre-2020)", 
         x = "Weeks vs. Lockdown") +
    scale_x_continuous(breaks = seq(-10, 
                                    max(searches_norm3$busclose_first_counter, na.rm = TRUE), 
                                    by = 10))

books_sip <- searches_norm3 %>%
  group_by(year_indicator, sip_first_counter) %>%
  summarize(books = mean(books)) %>%
  pivot_wider(names_from = year_indicator, 
              values_from = books) %>%
  mutate(ratio = `1`/`0`) %>%
  ggplot(mapping = aes(x = sip_first_counter, y = ratio)) +   
    geom_line() + 
    theme_bw() +
    labs(title = "books vs. Business Closure", y = "SVI ratio (2020 to pre-2020)", 
         x = "Weeks vs. Lockdown") +
    scale_x_continuous(breaks = seq(-10, 
                                    max(searches_norm3$sip_first_counter, na.rm = TRUE), 
                                    by = 10))

crying_busclose <- searches_norm3 %>%
  group_by(year_indicator, busclose_first_counter) %>%
  summarize(crying = mean(crying)) %>%
  pivot_wider(names_from = year_indicator, 
              values_from = crying) %>%
  mutate(ratio = `1`/`0`) %>%
  ggplot(mapping = aes(x = busclose_first_counter, y = ratio)) +   
    geom_line() + 
    theme_bw() +
    labs(title = "crying vs. Business Closure", y = "SVI ratio (2020 to pre-2020)", 
         x = "Weeks vs. Lockdown") +
    scale_x_continuous(breaks = seq(-10, 
                                    max(searches_norm3$busclose_first_counter, na.rm = TRUE), 
                                    by = 10))

crying_sip <- searches_norm3 %>%
  group_by(year_indicator, sip_first_counter) %>%
  summarize(crying = mean(crying)) %>%
  pivot_wider(names_from = year_indicator, 
              values_from = crying) %>%
  mutate(ratio = `1`/`0`) %>%
  ggplot(mapping = aes(x = sip_first_counter, y = ratio)) +   
    geom_line() + 
    theme_bw() +
    labs(title = "crying vs. Business Closure", y = "SVI ratio (2020 to pre-2020)", 
         x = "Weeks vs. Lockdown") +
    scale_x_continuous(breaks = seq(-10, 
                                    max(searches_norm3$sip_first_counter, na.rm = TRUE), 
                                    by = 10))

film_busclose <- searches_norm3 %>%
  group_by(year_indicator, busclose_first_counter) %>%
  summarize(film = mean(film)) %>%
  pivot_wider(names_from = year_indicator, 
              values_from = film) %>%
  mutate(ratio = `1`/`0`) %>%
  ggplot(mapping = aes(x = busclose_first_counter, y = ratio)) +   
    geom_line() + 
    theme_bw() +
    labs(title = "film vs. Business Closure", y = "SVI ratio (2020 to pre-2020)", 
         x = "Weeks vs. Lockdown") +
    scale_x_continuous(breaks = seq(-10, 
                                    max(searches_norm3$busclose_first_counter, na.rm = TRUE), 
                                    by = 10))

film_sip <- searches_norm3 %>%
  group_by(year_indicator, sip_first_counter) %>%
  summarize(film = mean(film)) %>%
  pivot_wider(names_from = year_indicator, 
              values_from = film) %>%
  mutate(ratio = `1`/`0`) %>%
  ggplot(mapping = aes(x = sip_first_counter, y = ratio)) +   
    geom_line() + 
    theme_bw() +
    labs(title = "film vs. Business Closure", y = "SVI ratio (2020 to pre-2020)", 
         x = "Weeks vs. Lockdown") +
    scale_x_continuous(breaks = seq(-10, 
                                    max(searches_norm3$sip_first_counter, na.rm = TRUE), 
                                    by = 10))

games_busclose <- searches_norm3 %>%
  group_by(year_indicator, busclose_first_counter) %>%
  summarize(games = mean(games)) %>%
  pivot_wider(names_from = year_indicator, 
              values_from = games) %>%
  mutate(ratio = `1`/`0`) %>%
  ggplot(mapping = aes(x = busclose_first_counter, y = ratio)) +   
    geom_line() + 
    theme_bw() +
    labs(title = "games vs. Business Closure", y = "SVI ratio (2020 to pre-2020)", 
         x = "Weeks vs. Lockdown") +
    scale_x_continuous(breaks = seq(-10, 
                                    max(searches_norm3$busclose_first_counter, na.rm = TRUE), 
                                    by = 10))

games_sip <- searches_norm3 %>%
  group_by(year_indicator, sip_first_counter) %>%
  summarize(games = mean(games)) %>%
  pivot_wider(names_from = year_indicator, 
              values_from = games) %>%
  mutate(ratio = `1`/`0`) %>%
  ggplot(mapping = aes(x = sip_first_counter, y = ratio)) +   
    geom_line() + 
    theme_bw() +
    labs(title = "games vs. Business Closure", y = "SVI ratio (2020 to pre-2020)", 
         x = "Weeks vs. Lockdown") +
    scale_x_continuous(breaks = seq(-10, 
                                    max(searches_norm3$sip_first_counter, na.rm = TRUE), 
                                    by = 10))

loneliness_busclose <- searches_norm3 %>%
  group_by(year_indicator, busclose_first_counter) %>%
  summarize(loneliness = mean(loneliness)) %>%
  pivot_wider(names_from = year_indicator, 
              values_from = loneliness) %>%
  mutate(ratio = `1`/`0`) %>%
  ggplot(mapping = aes(x = busclose_first_counter, y = ratio)) +   
    geom_line() + 
    theme_bw() +
    labs(title = "loneliness vs. Business Closure", y = "SVI ratio (2020 to pre-2020)", 
         x = "Weeks vs. Lockdown") +
    scale_x_continuous(breaks = seq(-10, 
                                    max(searches_norm3$busclose_first_counter, na.rm = TRUE), 
                                    by = 10))

loneliness_sip <- searches_norm3 %>%
  group_by(year_indicator, sip_first_counter) %>%
  summarize(loneliness = mean(loneliness)) %>%
  pivot_wider(names_from = year_indicator, 
              values_from = loneliness) %>%
  mutate(ratio = `1`/`0`) %>%
  ggplot(mapping = aes(x = sip_first_counter, y = ratio)) +   
    geom_line() + 
    theme_bw() +
    labs(title = "loneliness vs. Business Closure", y = "SVI ratio (2020 to pre-2020)", 
         x = "Weeks vs. Lockdown") +
    scale_x_continuous(breaks = seq(-10, 
                                    max(searches_norm3$sip_first_counter, na.rm = TRUE), 
                                    by = 10))

smoking_busclose <- searches_norm3 %>%
  group_by(year_indicator, busclose_first_counter) %>%
  summarize(smoking = mean(smoking)) %>%
  pivot_wider(names_from = year_indicator, 
              values_from = smoking) %>%
  mutate(ratio = `1`/`0`) %>%
  ggplot(mapping = aes(x = busclose_first_counter, y = ratio)) +   
    geom_line() + 
    theme_bw() +
    labs(title = "smoking vs. Business Closure", y = "SVI ratio (2020 to pre-2020)", 
         x = "Weeks vs. Lockdown") +
    scale_x_continuous(breaks = seq(-10, 
                                    max(searches_norm3$busclose_first_counter, na.rm = TRUE), 
                                    by = 10))

smoking_sip <- searches_norm3 %>%
  group_by(year_indicator, sip_first_counter) %>%
  summarize(smoking = mean(smoking)) %>%
  pivot_wider(names_from = year_indicator, 
              values_from = smoking) %>%
  mutate(ratio = `1`/`0`) %>%
  ggplot(mapping = aes(x = sip_first_counter, y = ratio)) +   
    geom_line() + 
    theme_bw() +
    labs(title = "smoking vs. Business Closure", y = "SVI ratio (2020 to pre-2020)", 
         x = "Weeks vs. Lockdown") +
    scale_x_continuous(breaks = seq(-10, 
                                    max(searches_norm3$sip_first_counter, na.rm = TRUE), 
                                    by = 10))

stress_busclose <- searches_norm3 %>%
  group_by(year_indicator, busclose_first_counter) %>%
  summarize(stress = mean(stress)) %>%
  pivot_wider(names_from = year_indicator, 
              values_from = stress) %>%
  mutate(ratio = `1`/`0`) %>%
  ggplot(mapping = aes(x = busclose_first_counter, y = ratio)) +   
    geom_line() + 
    theme_bw() +
    labs(title = "stress vs. Business Closure", y = "SVI ratio (2020 to pre-2020)", 
         x = "Weeks vs. Lockdown") +
    scale_x_continuous(breaks = seq(-10, 
                                    max(searches_norm3$busclose_first_counter, na.rm = TRUE), 
                                    by = 10))

stress_sip <- searches_norm3 %>%
  group_by(year_indicator, sip_first_counter) %>%
  summarize(stress = mean(stress)) %>%
  pivot_wider(names_from = year_indicator, 
              values_from = stress) %>%
  mutate(ratio = `1`/`0`) %>%
  ggplot(mapping = aes(x = sip_first_counter, y = ratio)) +   
    geom_line() + 
    theme_bw() +
    labs(title = "stress vs. Business Closure", y = "SVI ratio (2020 to pre-2020)", 
         x = "Weeks vs. Lockdown") +
    scale_x_continuous(breaks = seq(-10, 
                                    max(searches_norm3$sip_first_counter, na.rm = TRUE), 
                                    by = 10))

music_busclose <- searches_norm3 %>%
  group_by(year_indicator, busclose_first_counter) %>%
  summarize(music = mean(music)) %>%
  pivot_wider(names_from = year_indicator, 
              values_from = music) %>%
  mutate(ratio = `1`/`0`) %>%
  ggplot(mapping = aes(x = busclose_first_counter, y = ratio)) +   
    geom_line() + 
    theme_bw() +
    labs(title = "music vs. Business Closure", y = "SVI ratio (2020 to pre-2020)", 
         x = "Weeks vs. Lockdown") +
    scale_x_continuous(breaks = seq(-10, 
                                    max(searches_norm3$busclose_first_counter, na.rm = TRUE), 
                                    by = 10))

music_sip <- searches_norm3 %>%
  group_by(year_indicator, sip_first_counter) %>%
  summarize(music = mean(music)) %>%
  pivot_wider(names_from = year_indicator, 
              values_from = music) %>%
  mutate(ratio = `1`/`0`) %>%
  ggplot(mapping = aes(x = sip_first_counter, y = ratio)) +   
    geom_line() + 
    theme_bw() +
    labs(title = "music vs. Business Closure", y = "SVI ratio (2020 to pre-2020)", 
         x = "Weeks vs. Lockdown") +
    scale_x_continuous(breaks = seq(-10, 
                                    max(searches_norm3$sip_first_counter, na.rm = TRUE), 
                                    by = 10))

substance_abuse_busclose <- searches_norm3 %>%
  group_by(year_indicator, busclose_first_counter) %>%
  summarize(substance_abuse = mean(substance_abuse)) %>%
  pivot_wider(names_from = year_indicator, 
              values_from = substance_abuse) %>%
  mutate(ratio = `1`/`0`) %>%
  ggplot(mapping = aes(x = busclose_first_counter, y = ratio)) +   
    geom_line() + 
    theme_bw() +
    labs(title = "substance_abuse vs. Business Closure", y = "SVI ratio (2020 to pre-2020)", 
         x = "Weeks vs. Lockdown") +
    scale_x_continuous(breaks = seq(-10, 
                                    max(searches_norm3$busclose_first_counter, na.rm = TRUE), 
                                    by = 10))

substance_abuse_sip <- searches_norm3 %>%
  group_by(year_indicator, sip_first_counter) %>%
  summarize(substance_abuse = mean(substance_abuse)) %>%
  pivot_wider(names_from = year_indicator, 
              values_from = substance_abuse) %>%
  mutate(ratio = `1`/`0`) %>%
  ggplot(mapping = aes(x = sip_first_counter, y = ratio)) +   
    geom_line() + 
    theme_bw() +
    labs(title = "substance_abuse vs. Business Closure", y = "SVI ratio (2020 to pre-2020)", 
         x = "Weeks vs. Lockdown") +
    scale_x_continuous(breaks = seq(-10, 
                                    max(searches_norm3$sip_first_counter, na.rm = TRUE), 
                                    by = 10))

worry_busclose <- searches_norm3 %>%
  group_by(year_indicator, busclose_first_counter) %>%
  summarize(worry = mean(worry)) %>%
  pivot_wider(names_from = year_indicator, 
              values_from = worry) %>%
  mutate(ratio = `1`/`0`) %>%
  ggplot(mapping = aes(x = busclose_first_counter, y = ratio)) +   
    geom_line() + 
    theme_bw() +
    labs(title = "worry vs. Business Closure", y = "SVI ratio (2020 to pre-2020)", 
         x = "Weeks vs. Lockdown") +
    scale_x_continuous(breaks = seq(-10, 
                                    max(searches_norm3$busclose_first_counter, na.rm = TRUE), 
                                    by = 10))

worry_sip <- searches_norm3 %>%
  group_by(year_indicator, sip_first_counter) %>%
  summarize(worry = mean(worry)) %>%
  pivot_wider(names_from = year_indicator, 
              values_from = worry) %>%
  mutate(ratio = `1`/`0`) %>%
  ggplot(mapping = aes(x = sip_first_counter, y = ratio)) +   
    geom_line() + 
    theme_bw() +
    labs(title = "worry vs. Business Closure", y = "SVI ratio (2020 to pre-2020)", 
         x = "Weeks vs. Lockdown") +
    scale_x_continuous(breaks = seq(-10, 
                                    max(searches_norm3$sip_first_counter, na.rm = TRUE), 
                                    by = 10))
```

```{r}
searches_norm3_v6 %>%
  mutate(year_indicator = as.character(year_indicator)) %>% 
  group_by(busclose_first_counter, year_indicator) %>%
  arrange(date) %>%
  summarize(alcohol = mean(alcohol, na.rm = TRUE)) %>%
  #mutate(alcohol = rollmean(alcohol, k = 3, fill = "extend")) %>%
  ggplot(mapping = aes(x = busclose_first_counter, y = alcohol)) +
    geom_line(aes(color = year_indicator)) + 
    theme_classic(base_size = 14) +
    scale_color_discrete(name = "Year", labels = c("2016-2019", "2020")) +
    labs(title = "Alcohol vs. Business Closure", y = "SVI", x = "Weeks vs. Lockdown") +
    scale_x_continuous(breaks = seq(-10, 
                                    max(searches_norm3_v6$busclose_first_counter, na.rm = TRUE), 
                                    by = 10)) + 
    annotate("text", x = 33, y = 2.5, label = "Holiday season") + 
    geom_vline(xintercept = 0, linetype = "dashed")

searches_norm3_v6 %>%
  mutate(year_indicator = as.character(year_indicator)) %>% 
  group_by(busclose_first_counter, year_indicator) %>%
  arrange(date) %>%
  summarize(smoking = mean(smoking, na.rm = TRUE)) %>%
  #mutate(smoking = rollmean(smoking, k = 3, fill = "extend")) %>%
  ggplot(mapping = aes(x = busclose_first_counter, y = smoking)) +
    geom_line(aes(color = year_indicator)) + 
    theme_classic(base_size = 14) +
    scale_color_discrete(name = "Year", labels = c("2016-2019", "2020")) +
    labs(title = "smoking vs. Business Closure", y = "SVI", x = "Weeks vs. Lockdown") +
    scale_x_continuous(breaks = seq(-10, 
                                    max(searches_norm3_v6$busclose_first_counter, na.rm = TRUE), 
                                    by = 10)) + 
    geom_vline(xintercept = 0, linetype = "dashed")

searches_norm3_v6 %>%
  mutate(year_indicator = as.character(year_indicator)) %>% 
  group_by(busclose_first_counter, year_indicator) %>%
  arrange(date) %>%
  summarize(cannabis = mean(cannabis, na.rm = TRUE)) %>%
  #mutate(cannabis = rollmean(cannabis, k = 3, fill = "extend")) %>%
  ggplot(mapping = aes(x = busclose_first_counter, y = cannabis)) +
    geom_line(aes(color = year_indicator)) + 
    theme_classic(base_size = 14) +
    scale_color_discrete(name = "Year", labels = c("2016-2019", "2020")) +
    labs(title = "cannabis vs. Business Closure", y = "SVI", x = "Weeks vs. Lockdown") +
    scale_x_continuous(breaks = seq(-10, 
                                    35, 
                                    by = 10)) + 
    xlim(-10, 35) +
    ylim(40, 60) +
    geom_vline(xintercept = 0, linetype = "dashed")

searches_norm3_v6 %>%
  mutate(year_indicator = as.character(year_indicator)) %>% 
  group_by(busclose_first_counter, year_indicator) %>%
  arrange(date) %>%
  summarize(e_cig = mean(e_cig, na.rm = TRUE)) %>%
  #mutate(e_cig = rollmean(e_cig, k = 3, fill = "extend")) %>%
  ggplot(mapping = aes(x = busclose_first_counter, y = e_cig)) +
    geom_line(aes(color = year_indicator)) + 
    theme_classic(base_size = 14) +
    scale_color_discrete(name = "Year", labels = c("2016-2019", "2020")) +
    labs(title = "e_cig vs. Business Closure", y = "SVI", x = "Weeks vs. Lockdown") +
    scale_x_continuous(breaks = seq(-10, 
                                    max(searches_norm3_v6$busclose_first_counter, na.rm = TRUE), 
                                    by = 10)) + 
    geom_vline(xintercept = 0, linetype = "dashed")

searches_norm3_v6 %>%
  mutate(year_indicator = as.character(year_indicator)) %>% 
  group_by(busclose_first_counter, year_indicator) %>%
  arrange(date) %>%
  summarize(alcoholism = mean(alcoholism, na.rm = TRUE)) %>%
  #mutate(alcoholism = rollmean(alcoholism, k = 3, fill = "extend")) %>%
  ggplot(mapping = aes(x = busclose_first_counter, y = alcoholism)) +
    geom_line(aes(color = year_indicator)) + 
    theme_classic(base_size = 14) +
    scale_color_discrete(name = "Year", labels = c("2016-2019", "2020")) +
    labs(title = "alcoholism vs. Business Closure", y = "SVI", x = "Weeks vs. Lockdown") +
    scale_x_continuous(breaks = seq(-10, 
                                    max(searches_norm3_v6$busclose_first_counter, na.rm = TRUE), 
                                    by = 10)) + 
    geom_vline(xintercept = 0, linetype = "dashed")

searches_norm3_v6 %>%
  mutate(year_indicator = as.character(year_indicator)) %>% 
  group_by(busclose_first_counter, year_indicator) %>%
  arrange(date) %>%
  summarize(cigarette = mean(cigarette, na.rm = TRUE)) %>%
  #mutate(cigarette = rollmean(cigarette, k = 3, fill = "extend")) %>%
  ggplot(mapping = aes(x = busclose_first_counter, y = cigarette)) +
    geom_line(aes(color = year_indicator)) + 
    theme_classic(base_size = 14) +
    scale_color_discrete(name = "Year", labels = c("2016-2019", "2020")) +
    labs(title = "cigarette vs. Business Closure", y = "SVI", x = "Weeks vs. Lockdown") +
    scale_x_continuous(breaks = seq(-10, 
                                    max(searches_norm3_v6$busclose_first_counter, na.rm = TRUE), 
                                    by = 10)) + 
    geom_vline(xintercept = 0, linetype = "dashed")

searches_norm3_v6 %>%
  mutate(year_indicator = as.character(year_indicator)) %>% 
  group_by(busclose_first_counter, year_indicator) %>%
  arrange(date) %>%
  summarize(therapy = mean(therapy, na.rm = TRUE)) %>%
  #mutate(therapy = rollmean(therapy, k = 3, fill = "extend")) %>%
  ggplot(mapping = aes(x = busclose_first_counter, y = therapy)) +
    geom_line(aes(color = year_indicator)) + 
    theme_classic(base_size = 14) +
    scale_color_discrete(name = "Year", labels = c("2016-2019", "2020")) +
    labs(title = "therapy vs. Business Closure", y = "SVI", x = "Weeks vs. Lockdown") +
    scale_x_continuous(breaks = seq(-10, 
                                    max(searches_norm3_v6$busclose_first_counter, na.rm = TRUE), 
                                    by = 10)) + 
    geom_vline(xintercept = 0, linetype = "dashed")

searches_norm3_v6 %>%
  mutate(year_indicator = as.character(year_indicator)) %>% 
  group_by(busclose_first_counter, year_indicator) %>%
  arrange(date) %>%
  summarize(exercise = mean(exercise, na.rm = TRUE)) %>%
  #mutate(exercise = rollmean(exercise, k = 3, fill = "extend")) %>%
  ggplot(mapping = aes(x = busclose_first_counter, y = exercise)) +
    geom_line(aes(color = year_indicator)) + 
    theme_classic(base_size = 14) +
    scale_color_discrete(name = "Year", labels = c("2016-2019", "2020")) +
    labs(title = "exercise vs. Business Closure", y = "SVI", x = "Weeks vs. Lockdown") +
    scale_x_continuous(breaks = seq(-10, 
                                    max(searches_norm3_v6$busclose_first_counter, na.rm = TRUE), 
                                    by = 10)) + 
    geom_vline(xintercept = 0, linetype = "dashed")

searches_norm3_v6 %>%
  mutate(year_indicator = as.character(year_indicator)) %>% 
  group_by(busclose_first_counter, year_indicator) %>%
  arrange(date) %>%
  summarize(yoga = mean(yoga, na.rm = TRUE)) %>%
  #mutate(yoga = rollmean(yoga, k = 3, fill = "extend")) %>%
  ggplot(mapping = aes(x = busclose_first_counter, y = yoga)) +
    geom_line(aes(color = year_indicator)) + 
    theme_classic(base_size = 14) +
    scale_color_discrete(name = "Year", labels = c("2016-2019", "2020")) +
    labs(title = "yoga vs. Business Closure", y = "SVI", x = "Weeks vs. Lockdown") +
    scale_x_continuous(breaks = seq(-10, 
                                    max(searches_norm3_v6$busclose_first_counter, na.rm = TRUE), 
                                    by = 10)) + 
    geom_vline(xintercept = 0, linetype = "dashed")

searches_norm3_v6 %>%
  mutate(year_indicator = as.character(year_indicator)) %>% 
  group_by(busclose_first_counter, year_indicator) %>%
  arrange(date) %>%
  summarize(mental_health = mean(mental_health, na.rm = TRUE)) %>%
  #mutate(mental_health = rollmean(mental_health, k = 3, fill = "extend")) %>%
  ggplot(mapping = aes(x = busclose_first_counter, y = mental_health)) +
    geom_line(aes(color = year_indicator)) + 
    theme_classic(base_size = 14) +
    scale_color_discrete(name = "Year", labels = c("2016-2019", "2020")) +
    labs(title = "mental_health vs. Business Closure", y = "SVI", x = "Weeks vs. Lockdown") +
    scale_x_continuous(breaks = seq(-10, 
                                    max(searches_norm3_v6$busclose_first_counter, na.rm = TRUE), 
                                    by = 10)) + 
    geom_vline(xintercept = 0, linetype = "dashed")
```


```{r}
# Leisure

searches_norm3_v6 %>%
  mutate(year_indicator = as.character(year_indicator)) %>% 
  group_by(busclose_first_counter, year_indicator) %>%
  arrange(date) %>%
  summarize(books = mean(books, na.rm = TRUE)) %>%
  #mutate(books = rollmean(books, k = 3, fill = "extend")) %>%
  ggplot(mapping = aes(x = busclose_first_counter, y = books)) +
    geom_line(aes(color = year_indicator)) + 
    theme_classic(base_size = 14) +
    scale_color_discrete(name = "Year", labels = c("2016-2019", "2020")) +
    labs(title = "books vs. Business Closure", y = "SVI", x = "Weeks vs. Lockdown") +
    scale_x_continuous(breaks = seq(-10, 
                                    max(searches_norm3_v6$busclose_first_counter, na.rm = TRUE), 
                                    by = 10)) + 
    geom_vline(xintercept = 0, linetype = "dashed")

searches_norm3_v6 %>%
  mutate(year_indicator = as.character(year_indicator)) %>% 
  group_by(busclose_first_counter, year_indicator) %>%
  arrange(date) %>%
  summarize(film = mean(film, na.rm = TRUE)) %>%
  #mutate(film = rollmean(film, k = 3, fill = "extend")) %>%
  ggplot(mapping = aes(x = busclose_first_counter, y = film)) +
    geom_line(aes(color = year_indicator)) + 
    theme_classic(base_size = 14) +
    scale_color_discrete(name = "Year", labels = c("2016-2019", "2020")) +
    labs(title = "film vs. Business Closure", y = "SVI", x = "Weeks vs. Lockdown") +
    scale_x_continuous(breaks = seq(-10, 
                                    max(searches_norm3_v6$busclose_first_counter, na.rm = TRUE), 
                                    by = 10)) + 
    geom_vline(xintercept = 0, linetype = "dashed")

games <- searches_norm3_v8 %>%
  mutate(year_indicator = as.character(year_indicator)) %>% 
  group_by(busclose_first_counter, year_indicator) %>%
  arrange(date) %>%
  summarize(games = mean(games, na.rm = TRUE)) %>%
  #mutate(games = rollmean(games, k = 3, fill = "extend")) %>%
  ggplot(mapping = aes(x = busclose_first_counter, y = games)) +
    geom_line(aes(color = year_indicator)) + 
    theme_classic(base_size = 14) +
    scale_color_discrete(name = "Year", labels = c("2016-2019", "2020")) +
    labs(title = "games vs. Business Closure", y = "SVI", x = "Weeks vs. Lockdown") +
    scale_x_continuous(breaks = seq(-10, 
                                    max(searches_norm3_v8$busclose_first_counter, na.rm = TRUE), 
                                    by = 10)) + 
    geom_vline(xintercept = 0, linetype = "dashed")

searches_norm3_v6 %>%
  mutate(year_indicator = as.character(year_indicator)) %>% 
  group_by(busclose_first_counter, year_indicator) %>%
  arrange(date) %>%
  summarize(TV = mean(TV, na.rm = TRUE)) %>%
  #mutate(TV = rollmean(TV, k = 3, fill = "extend")) %>%
  ggplot(mapping = aes(x = busclose_first_counter, y = TV)) +
    geom_line(aes(color = year_indicator)) + 
    theme_classic(base_size = 14) +
    scale_color_discrete(name = "Year", labels = c("2016-2019", "2020")) +
    labs(title = "TV vs. Business Closure", y = "SVI", x = "Weeks vs. Lockdown") +
    scale_x_continuous(breaks = seq(-10, 
                                    max(searches_norm3_v6$busclose_first_counter, na.rm = TRUE), 
                                    by = 10)) + 
    geom_vline(xintercept = 0, linetype = "dashed")

searches_norm3_v6 %>%
  mutate(year_indicator = as.character(year_indicator)) %>% 
  group_by(busclose_first_counter, year_indicator) %>%
  arrange(date) %>%
  summarize(music = mean(music, na.rm = TRUE)) %>%
  #mutate(music = rollmean(music, k = 3, fill = "extend")) %>%
  ggplot(mapping = aes(x = busclose_first_counter, y = music)) +
    geom_line(aes(color = year_indicator)) + 
    theme_classic(base_size = 14) +
    scale_color_discrete(name = "Year", labels = c("2016-2019", "2020")) +
    labs(title = "music vs. Business Closure", y = "SVI", x = "Weeks vs. Lockdown") +
    scale_x_continuous(breaks = seq(-10, 
                                    max(searches_norm3_v6$busclose_first_counter, na.rm = TRUE), 
                                    by = 10)) + 
    geom_vline(xintercept = 0, linetype = "dashed")

searches_norm3_v7 %>%
  mutate(year_indicator = as.character(year_indicator)) %>% 
  group_by(busclose_first_counter, year_indicator) %>%
  arrange(date) %>%
  summarize(streaming_media = mean(streaming_media, na.rm = TRUE)) %>%
  #mutate(streaming_media = rollmean(streaming_media, k = 3, fill = "extend")) %>%
  ggplot(mapping = aes(x = busclose_first_counter, y = streaming_media)) +
    geom_line(aes(color = year_indicator)) + 
    theme_classic(base_size = 14) +
    scale_color_discrete(name = "Year", labels = c("2016-2019", "2020")) +
    labs(title = "streaming_media vs. Business Closure", y = "SVI", x = "Weeks vs. Lockdown") +
    scale_x_continuous(breaks = seq(-10, 
                                    max(searches_norm3_v6$busclose_first_counter, na.rm = TRUE), 
                                    by = 10)) + 
    geom_vline(xintercept = 0, linetype = "dashed")

# Anchor terms

searches_norm3_v8 %>%
  mutate(year_indicator = as.character(year_indicator),
         date = as_date(date)) %>% 
  group_by(date) %>% 
  summarize(average_3 = mean(average_3, na.rm = TRUE)) %>%
  #mutate(average_3 = rollmean(average_3, k = 3, fill = "extend")) %>%
  ggplot(mapping = aes(x = date, y = average_3)) +
    geom_line() + 
    theme_classic(base_size = 14) +
    scale_x_date(breaks = "6 months", date_labels = "%m/%y") +
    labs(title = "Anchor Term Average (3-week rolling)", y = "SVI", x = "Date")

searches_norm3_v8 %>%
  mutate(year_indicator = as.character(year_indicator),
         date = as_date(date),
         average_5 = rollmean(average, k = 5, fill = "extend")) %>% 
  group_by(date) %>% 
  summarize(average_5 = mean(average_5, na.rm = TRUE)) %>%
  #mutate(average_3 = rollmean(average_3, k = 3, fill = "extend")) %>%
  ggplot(mapping = aes(x = date, y = average_5)) +
    geom_line() + 
    theme_classic(base_size = 14) +
    scale_x_date(breaks = "6 months", date_labels = "%m/%y") +
    labs(title = "Date", y = "SVI", x = "Weeks vs. Lockdown")

anchor_index %>% 
  mutate(date = as_date(date)) %>% 
  group_by(date) %>% 
  summarize(rice = mean(rice, na.rm = TRUE)) %>% 
  ggplot(mapping = aes(x = date, y = rice)) + 
    geom_line() + 
    theme_classic(base_size = 14) + 
    scale_x_date(breaks = "6 months", date_labels = "%m/%y") + 
    labs(title = "rice", x = "Date", y = "SVI")

```

```{r}
searches_norm3_v10 %>%
  mutate(year_indicator = as.character(year_indicator)) %>% 
  group_by(week, year_indicator) %>%
  summarize(depression = mean(depression)) %>%
  ungroup() %>%
  group_by(year_indicator) %>%
  mutate(depression = rollmean(depression, k = 3, fill = "extend")) %>%
  ggplot(mapping = aes(x = week, y = depression)) + geom_line(aes(color = year_indicator)) + 
    labs(x = "Week of Year", y = "SVI", title = "Depression", 
         subtitle = "Dashed line indicates approx. lockdown start date") + 
    scale_color_discrete(name = "Year", labels = c("2016-2019", "2020")) + 
    scale_x_continuous(breaks = seq(from = 0, to = 53, by = 10)) + 
    geom_vline(xintercept = 13, linetype = "dashed") + 
    theme_classic(base_size = 14)

searches_norm3_v10_5 %>%
  mutate(year_indicator = as.character(year_indicator)) %>% 
  group_by(week, year_indicator) %>%
  summarize(depression = mean(depression)) %>%
  ungroup() %>%
  group_by(year_indicator) %>%
  mutate(depression = rollmean(depression, k = 3, fill = "extend")) %>%
  ggplot(mapping = aes(x = week, y = depression)) + geom_line(aes(color = year_indicator)) + 
    labs(x = "Week of Year", y = "SVI", title = "Depression (excl. low-volume regions)", 
         subtitle = "Dashed line indicates approx. lockdown start date") + 
    scale_color_discrete(name = "Year", labels = c("2016-2019", "2020")) + 
    scale_x_continuous(breaks = seq(from = 0, to = 53, by = 10)) + 
    geom_vline(xintercept = 13, linetype = "dashed") + 
    theme_classic(base_size = 14)

anxiety <- searches_norm3_v10 %>%
  mutate(year_indicator = as.character(year_indicator)) %>% 
  group_by(week, year_indicator) %>%
  summarize(anxiety = mean(anxiety)) %>%
  ungroup() %>%
  group_by(year_indicator) %>%
  mutate(anxiety = rollmean(anxiety, k = 3, fill = "extend")) %>%
  ggplot(mapping = aes(x = week, y = anxiety)) + geom_line(aes(color = year_indicator)) + 
    labs(x = "Week of Year", y = "SVI", title = "anxiety", 
         subtitle = "Dashed line indicates approx. lockdown start date") + 
    scale_color_discrete(name = "Year", labels = c("2016-2019", "2020")) + 
    scale_x_continuous(breaks = seq(from = 0, to = 53, by = 10)) + 
    geom_vline(xintercept = 13, linetype = "dashed") + 
    theme_classic(base_size = 14)

anxiety_filter <- searches_norm3_v10_5 %>%
  mutate(year_indicator = as.character(year_indicator)) %>% 
  group_by(week, year_indicator) %>%
  summarize(anxiety = mean(anxiety)) %>%
  ungroup() %>%
  group_by(year_indicator) %>%
  mutate(anxiety = rollmean(anxiety, k = 3, fill = "extend")) %>%
  ggplot(mapping = aes(x = week, y = anxiety)) + geom_line(aes(color = year_indicator)) + 
    labs(x = "Week of Year", y = "SVI", title = "anxiety (excl. low-volume regions)", 
         subtitle = "Dashed line indicates approx. lockdown start date") + 
    scale_color_discrete(name = "Year", labels = c("2016-2019", "2020")) + 
    scale_x_continuous(breaks = seq(from = 0, to = 53, by = 10)) + 
    geom_vline(xintercept = 13, linetype = "dashed") + 
    theme_classic(base_size = 14)
```
```{r}
crying <- searches_norm3_v10 %>%
  mutate(year_indicator = as.character(year_indicator)) %>% 
  group_by(week, year_indicator) %>%
  summarize(crying = mean(crying)) %>%
  ungroup() %>%
  group_by(year_indicator) %>%
  mutate(crying = rollmean(crying, k = 3, fill = "extend")) %>%
  ggplot(mapping = aes(x = week, y = crying)) + geom_line(aes(color = year_indicator)) + 
    labs(x = "Week of Year", y = "SVI", title = "crying", 
         subtitle = "Dashed line indicates approx. lockdown start date") + 
    scale_color_discrete(name = "Year", labels = c("2016-2019", "2020")) + 
    scale_x_continuous(breaks = seq(from = 0, to = 53, by = 10)) + 
    geom_vline(xintercept = 13, linetype = "dashed") + 
    theme_classic(base_size = 14)

crying_filter <- searches_norm3_v10_5 %>%
  mutate(year_indicator = as.character(year_indicator)) %>% 
  group_by(week, year_indicator) %>%
  summarize(crying = mean(crying)) %>%
  ungroup() %>%
  group_by(year_indicator) %>%
  mutate(crying = rollmean(crying, k = 3, fill = "extend")) %>%
  ggplot(mapping = aes(x = week, y = crying)) + geom_line(aes(color = year_indicator)) + 
    labs(x = "Week of Year", y = "SVI", title = "crying (excl. low-volume regions)", 
         subtitle = "Dashed line indicates approx. lockdown start date") + 
    scale_color_discrete(name = "Year", labels = c("2016-2019", "2020")) + 
    scale_x_continuous(breaks = seq(from = 0, to = 53, by = 10)) + 
    geom_vline(xintercept = 13, linetype = "dashed") + 
    theme_classic(base_size = 14)


```
```{r}
# read in dma_5 and filter for it

tired <- searches_norm3_v11 %>%
  mutate(year_indicator = as.character(year_indicator)) %>% 
  group_by(week, year_indicator) %>%
  summarize(tired = mean(tired)) %>%
  ungroup() %>%
  group_by(year_indicator) %>%
  mutate(tired = rollmean(tired, k = 3, fill = "extend")) %>%
  ggplot(mapping = aes(x = week, y = tired)) + geom_line(aes(color = year_indicator)) + 
    labs(x = "Week of Year", y = "SVI", title = "tired", 
         subtitle = "Dashed line indicates approx. lockdown start date") + 
    scale_color_discrete(name = "Year", labels = c("2016-2019", "2020")) + 
    scale_x_continuous(breaks = seq(from = 0, to = 53, by = 10)) + 
    geom_vline(xintercept = 13, linetype = "dashed") + 
    theme_classic(base_size = 14)

tired_filter <- searches_norm3_v11_5 %>%
  mutate(year_indicator = as.character(year_indicator)) %>% 
  group_by(week, year_indicator) %>%
  summarize(tired = mean(tired)) %>%
  ungroup() %>%
  group_by(year_indicator) %>%
  mutate(tired = rollmean(tired, k = 3, fill = "extend")) %>%
  ggplot(mapping = aes(x = week, y = tired)) + geom_line(aes(color = year_indicator)) + 
    labs(x = "Week of Year", y = "SVI", title = "tired (excl. low-volume regions)", 
         subtitle = "Dashed line indicates approx. lockdown start date") + 
    scale_color_discrete(name = "Year", labels = c("2016-2019", "2020")) + 
    scale_x_continuous(breaks = seq(from = 0, to = 53, by = 10)) + 
    geom_vline(xintercept = 13, linetype = "dashed") + 
    theme_classic(base_size = 14)
  
```

```{r}
# Let's do some pre-trends graphing

part1 <- searches_norm3_v14 %>% 
  mutate(year_indicator = as.character(year_indicator)) %>% 
  group_by(week, year_indicator) %>%
  summarize(all_mean = mean(c(depression, crying, lonely, sad,
                               stress, suicide, tired))) %>% 
  ungroup() %>% 
  group_by(year_indicator) %>% 
  arrange(week) %>% 
  mutate(roll_mean = rollmean(all_mean, k = 3, fill = "extend")) %>% 
  ggplot(mapping = aes(x = week, y = roll_mean)) + 
    geom_line(aes(color = year_indicator)) + 
    labs(x = "Week of Year", y = "SVI", title = "Average Depression-related SVI") +
    scale_color_discrete(name = "Year", labels = c("2016-2019", "2020")) +
    scale_x_continuous(breaks = seq(from = 0, to = 53, by = 10)) + 
    geom_vline(xintercept = 13, linetype = "dashed") + 
    theme_classic(base_size = 14)

part2 <- searches_norm3_v14 %>% 
  mutate(year_indicator = as.character(year_indicator)) %>% 
  group_by(week, year_indicator) %>%
  summarize(all_mean = mean(c(anxiety, nervous, panic,
                               stress, worry))) %>% 
  ungroup() %>% 
  group_by(year_indicator) %>% 
  arrange(week) %>% 
  mutate(roll_mean = rollmean(all_mean, k = 3, fill = "extend")) %>% 
  ggplot(mapping = aes(x = week, y = roll_mean)) + 
    geom_line(aes(color = year_indicator)) + 
    labs(x = "Week of Year", y = "SVI", title = "Average Anxiety SVI") +
    scale_color_discrete(name = "Year", labels = c("2016-2019", "2020")) +
    scale_x_continuous(breaks = seq(from = 0, to = 53, by = 10)) + 
    geom_vline(xintercept = 13, linetype = "dashed") + 
    theme_classic(base_size = 14)

part3 <- searches_norm3_v14 %>% 
  mutate(year_indicator = as.character(year_indicator)) %>% 
  group_by(week, year_indicator) %>%
  summarize(all_mean = mean(c(exercise, therapy, yoga))) %>% 
  ungroup() %>% 
  group_by(year_indicator) %>% 
  arrange(week) %>% 
  mutate(roll_mean = rollmean(all_mean, k = 3, fill = "extend")) %>% 
  ggplot(mapping = aes(x = week, y = roll_mean)) + 
    geom_line(aes(color = year_indicator)) + 
    labs(x = "Week of Year", y = "SVI", title = "Average Coping SVI") +
    scale_color_discrete(name = "Year", labels = c("2016-2019", "2020")) +
    scale_x_continuous(breaks = seq(from = 0, to = 53, by = 10)) + 
    geom_vline(xintercept = 13, linetype = "dashed") + 
    theme_classic(base_size = 14)

part4 <- searches_norm3_v14 %>% 
  mutate(year_indicator = as.character(year_indicator)) %>% 
  group_by(week, year_indicator) %>%
  summarize(all_mean = mean(c(alcohol, alcoholism, cannabis, cigarette,
                              e_cig, substance_abuse, smoking))) %>% 
  ungroup() %>% 
  group_by(year_indicator) %>% 
  arrange(week) %>% 
  mutate(roll_mean = rollmean(all_mean, k = 3, fill = "extend")) %>% 
  ggplot(mapping = aes(x = week, y = roll_mean)) + 
    geom_line(aes(color = year_indicator)) + 
    labs(x = "Week of Year", y = "SVI", title = "Average Substance Use SVI") +
    scale_color_discrete(name = "Year", labels = c("2016-2019", "2020")) +
    scale_x_continuous(breaks = seq(from = 0, to = 53, by = 10)) + 
    geom_vline(xintercept = 13, linetype = "dashed") + 
    theme_classic(base_size = 14)

combined <- ggarrange(plotlist = list(part1, part2, part3, part4))
```

```{r}
# Coefficient plots!!!

depression_sip<- felm(I(log(0.485 * crying + 0.782 * depression + 
                                0.127 * lonely + 0.320 * sad + 0.184 * tired + 1)) ~ 
                                     sip_first * year_indicator + 
                                     no_sip_most_pop * year_indicator + 
                        online_only * year_indicator + 
                        hybrid * year_indicator + 
                        other * year_indicator + 
                        cases_per_100k + unemploy_rate + temp + precip + 
                       log(lockdown + 1) + log(covid + 1) + 
                       factor(income_support) + factor(debt_contract_relief)
                       |factor(week) + factor(dma) + factor(year)|0|dma, 
             data = searches_norm3_v22_5 %>% drop_na(pop),
             weights = na.omit(searches_norm3_v22_5$pop))

anxiety_sip <- felm(I(log(0.924 * anxiety + 0.213 * nervous + 
                                0.133 * stress + 0.198 * worry + 0.211 * panic + 1)) ~ 
                                  sip_first * year_indicator + 
                                     no_sip_most_pop * year_indicator + 
                        online_only * year_indicator + 
                        hybrid * year_indicator + 
                        other * year_indicator + 
                        cases_per_100k + unemploy_rate + temp + precip + 
                       log(lockdown + 1) + log(covid + 1) + 
                       factor(income_support) + factor(debt_contract_relief)
                       |factor(week) + factor(dma) + factor(year)|0|dma, 
             weights = na.omit(searches_norm3_v22_5$pop),
                  data = searches_norm3_v22_5 %>% filter(is.na(pop) == FALSE))

therapy_sip <- felm(I(log(therapy + 1 )) ~  
                 sip_first * year_indicator + 
                no_sip_most_pop * year_indicator + 
                        online_only * year_indicator + 
                        hybrid * year_indicator + 
                        other * year_indicator + 
                        cases_per_100k + unemploy_rate + temp + precip + 
                       log(lockdown + 1) + log(covid + 1) + 
                       factor(income_support) + factor(debt_contract_relief)
                       |factor(week) + factor(dma) + factor(year)|0|dma, 
             weights = na.omit(searches_norm3_v22_5$pop),
                  data = searches_norm3_v22_5 %>% filter(is.na(pop) == FALSE))

therapy_sip <- felm(I(log(therapy + 1 )) ~  
                 sip_first * year_indicator + 
                no_sip_most_pop * year_indicator + 
                        online_only * year_indicator + 
                        hybrid * year_indicator + 
                        other * year_indicator + 
                        cases_per_100k + unemploy_rate + temp + precip + 
                       log(lockdown + 1) + log(covid + 1) + 
                       factor(income_support) + factor(debt_contract_relief)
                       |factor(week) + factor(dma) + factor(year)|0|dma, 
             weights = na.omit(searches_norm3_v22_5$pop),
                  data = searches_norm3_v22_5 %>% filter(is.na(pop) == FALSE))

yoga_sip <- felm(I(log(yoga + 1 )) ~  
                 sip_first * year_indicator + 
                no_sip_most_pop * year_indicator + 
                        online_only * year_indicator + 
                        hybrid * year_indicator + 
                        other * year_indicator + 
                        cases_per_100k + unemploy_rate + temp + precip + 
                       log(lockdown + 1) + log(covid + 1) + 
                       factor(income_support) + factor(debt_contract_relief)
                       |factor(week) + factor(dma) + factor(year)|0|dma, 
             weights = na.omit(searches_norm3_v22_5$pop),
                  data = searches_norm3_v22_5 %>% filter(is.na(pop) == FALSE))


substance_sip <- felm(log(0.144*alcohol + 0.833*cannabis + 0.451*cigarette + 
                            0.172*e_cig + 0.228*smoking + 1) ~ 
                                  sip_first*year_indicator + 
                                     no_sip_most_pop*year_indicator + 
                        online_only*year_indicator + 
                        hybrid*year_indicator + 
                        other*year_indicator + 
                        cases_per_100k + unemploy_rate + temp + precip + 
                       log(lockdown + 1) + log(covid + 1) + 
                       factor(income_support) + factor(debt_contract_relief)
                       |factor(week) + factor(dma) + factor(year)|0|dma, 
             weights = na.omit(searches_norm3_v22_5$pop),
                  data = searches_norm3_v22_5 %>% filter(is.na(pop) == FALSE))

substance_misuse_sip <- felm(log(0.308 * addiction + 0.390 * alcoholism + 0.104 * rehab + 
                             0.861 * substance_abuse + 1) ~ 
                                  sip_first * year_indicator + 
                                    no_sip_most_pop * year_indicator + 
                        online_only * year_indicator + 
                        hybrid * year_indicator + 
                        other * year_indicator + 
                        cases_per_100k + unemploy_rate + temp + precip + 
                       log(lockdown + 1) + log(covid + 1) + 
                       factor(income_support) + factor(debt_contract_relief)
                       |factor(week) + factor(dma) + factor(year)|0|dma, 
              weights = na.omit(searches_norm3_v22_5$pop),
                  data = searches_norm3_v22_5 %>% filter(is.na(pop) == FALSE))
```

```{r}
# 

dwplot(list(depression_sip, anxiety_sip),
       vline = geom_vline(
           xintercept = 0,
           colour = "grey60"
       ),
       vars_order = c("sip_first:year_indicator", "year_indicator:no_sip_most_pop"),
       model_order = c("Model 1", "Model 2"),
       dodge_size = 0.9)%>% 
       relabel_predictors(c("sip_first:year_indicator" = "SIP start",
                          "year_indicator:no_sip_most_pop" = "SIP end")) + 
      theme_classic(base_size = 14) + 
      scale_color_discrete(name = "Outcome", labels = c(
                                                        "Anxiety", "Depression")) +
      scale_x_continuous(labels = scales::percent) + 
      labs(x = "Effect size (approx. % change)") + 
      theme(legend.position = "top")

dwplot(list(therapy_sip, yoga_sip),
       vline = geom_vline(
           xintercept = 0,
           colour = "grey60"
       ),
       vars_order = c("sip_first:year_indicator", "year_indicator:no_sip_most_pop",
                      "year_indicator:online_only", "year_indicator:hybrid"),
       model_order = c("Model 1", "Model 2"),
       dodge_size = 0.9)%>% 
       relabel_predictors(c("sip_first:year_indicator" = "SIP start",
                          "year_indicator:no_sip_most_pop" = "SIP end",
                          "year_indicator:online_only" = "Online school reopening",
                          "year_indicator:hybrid" = "Hybrid school reopening")) + 
      theme_classic(base_size = 14) + 
      scale_color_discrete(name = "Outcome", labels = c(
                                                        "Yoga", "Therapy")) +
      scale_x_continuous(labels = scales::percent) + 
      labs(x = "Effect size (approx. % change)") + 
      theme(legend.position = "top")

dwplot(list(substance_sip, substance_misuse_sip),
       vline = geom_vline(
           xintercept = 0,
           colour = "grey60"
       ),
       vars_order = c("sip_first:year_indicator", "year_indicator:no_sip_most_pop",
                      "year_indicator:online_only", "year_indicator:hybrid"),
       model_order = c("Model 1", "Model 2"),
       dodge_size = 0.9)%>% 
       relabel_predictors(c("sip_first:year_indicator" = "SIP start",
                          "year_indicator:no_sip_most_pop" = "SIP end",
                          "year_indicator:online_only" = "Online school reopening",
                          "year_indicator:hybrid" = "Hybrid school reopening")) + 
      theme_classic(base_size = 14) + 
      scale_color_discrete(name = "Outcome", labels = c(
                                                        "Yoga", "Therapy")) +
      scale_x_continuous(labels = scales::percent) + 
      labs(x = "Effect size (approx. % change)") + 
      theme(legend.position = "top")
```

```{r}
# Try for business closure

depression_bus <- felm(I(log(0.485 * crying + 0.782 * depression + 
                                0.127 * lonely + 0.320 * sad + 0.184 * tired + 1)) ~ 
                                     busclose_first * year_indicator + 
                                     busopen_most_pop * year_indicator + 
                        online_only * year_indicator + 
                        hybrid * year_indicator + 
                        other * year_indicator + 
                        cases_per_100k + unemploy_rate + temp + precip + 
                       log(lockdown + 1) + log(covid + 1) + 
                       factor(income_support) + factor(debt_contract_relief)
                       |factor(week) + factor(dma) + factor(year)|0|dma, 
             data = searches_norm3_v22_5 %>% drop_na(pop),
             weights = na.omit(searches_norm3_v22_5$pop))

anxiety_bus <- felm(I(log(0.924 * anxiety + 0.213 * nervous + 
                                0.133 * stress + 0.198 * worry + 0.211 * panic + 1)) ~ 
                                  busclose_first * year_indicator + 
                                     busopen_most_pop * year_indicator + 
                        online_only * year_indicator + 
                        hybrid * year_indicator + 
                        other * year_indicator + 
                        cases_per_100k + unemploy_rate + temp + precip + 
                       log(lockdown + 1) + log(covid + 1) + 
                       factor(income_support) + factor(debt_contract_relief)
                       |factor(week) + factor(dma) + factor(year)|0|dma, 
             weights = na.omit(searches_norm3_v22_5$pop),
                  data = searches_norm3_v22_5 %>% filter(is.na(pop) == FALSE))
```

```{r}
dwplot(list(depression_bus, anxiety_bus),
       vline = geom_vline(
           xintercept = 0,
           colour = "grey60"
       ),
       vars_order = c("busclose_first:year_indicator", "year_indicator:busopen_most_pop"),
       model_order = c("Model 1", "Model 2")) %>% 
       relabel_predictors(c("busclose_first:year_indicator" = "Closure start",
                          "year_indicator:busopen_most_pop" = "Closure end")) + 
      theme_classic(base_size = 14) + 
      scale_color_discrete(name = "Outcome", labels = c("Anxiety", "Depression")) +
      scale_x_continuous(labels = scales::percent) + 
      labs(x = "Effect size (approx. % change)") + 
      theme(legend.position = "top")
```

```{r}
# therapy and yoga

therapy_bus <- felm(I(log(therapy + 1 )) ~  
                 busclose_first * year_indicator + 
                busopen_most_pop * year_indicator + 
                        online_only * year_indicator + 
                        hybrid * year_indicator + 
                        other * year_indicator + 
                        cases_per_100k + unemploy_rate + temp + precip + 
                       log(lockdown + 1) + log(covid + 1) + 
                       factor(income_support) + factor(debt_contract_relief)
                       |factor(week) + factor(dma) + factor(year)|0|dma, 
             weights = na.omit(searches_norm3_v22_5$pop),
                  data = searches_norm3_v22_5 %>% filter(is.na(pop) == FALSE))

yoga_bus <- felm(I(log(yoga + 1 )) ~  
                 busclose_first * year_indicator + 
                busopen_most_pop * year_indicator + 
                        online_only * year_indicator + 
                        hybrid * year_indicator + 
                        other * year_indicator + 
                        cases_per_100k + unemploy_rate + temp + precip + 
                       log(lockdown + 1) + log(covid + 1) + 
                       factor(income_support) + factor(debt_contract_relief)
                       |factor(week) + factor(dma) + factor(year)|0|dma, 
             weights = na.omit(searches_norm3_v22_5$pop),
                  data = searches_norm3_v22_5 %>% filter(is.na(pop) == FALSE))

dwplot(list(therapy_bus, yoga_bus),
       vline = geom_vline(
           xintercept = 0,
           colour = "grey60"
       ),
       vars_order = c("busclose_first:year_indicator", "year_indicator:busopen_most_pop"),
       model_order = c("Model 1", "Model 2")) %>% 
       relabel_predictors(c("busclose_first:year_indicator" = "Closure start",
                          "year_indicator:busopen_most_pop" = "Closure end",
                          "year_indicator:online_only" = "Online school reopening",
                          "year_indicator:hybrid" = "Hybrid school reopening")) + 
      theme_classic(base_size = 14) + 
      scale_color_discrete(name = "Outcome", labels = c("Yoga", "Therapy")) +
      scale_x_continuous(labels = scales::percent) + 
      labs(x = "Effect size (approx. % change)") + 
      theme(legend.position = "top")
```



```{r}
# substance misuse

substance_busclose <- felm(log(0.144*alcohol + 0.833*cannabis + 0.451*cigarette + 
                            0.172*e_cig + 0.228*smoking + 1) ~ 
                                  busclose_first*year_indicator + 
                                     busopen_most_pop*year_indicator + 
                        online_only*year_indicator + 
                        hybrid*year_indicator + 
                        other*year_indicator + 
                        cases_per_100k + unemploy_rate + temp + precip + 
                       log(lockdown + 1) + log(covid + 1) + 
                       factor(income_support) + factor(debt_contract_relief)
                       |factor(week) + factor(dma) + factor(year)|0|dma, 
             weights = na.omit(searches_norm3_v22_5$pop),
                  data = searches_norm3_v22_5 %>% filter(is.na(pop) == FALSE))

substance_misuse_busclose <- felm(log(0.308 * addiction + 0.390 * alcoholism + 0.104 * rehab + 
                             0.861 * substance_abuse + 1) ~ 
                                  busclose_first * year_indicator + 
                                     busopen_most_pop * year_indicator + 
                        online_only * year_indicator + 
                        hybrid * year_indicator + 
                        other * year_indicator + 
                        cases_per_100k + unemploy_rate + temp + precip + 
                       log(lockdown + 1) + log(covid + 1) + 
                       factor(income_support) + factor(debt_contract_relief)
                       |factor(week) + factor(dma) + factor(year)|0|dma, 
              weights = na.omit(searches_norm3_v22_5$pop),
                  data = searches_norm3_v22_5 %>% filter(is.na(pop) == FALSE))

dwplot(list(substance_busclose, substance_misuse_busclose),
       vline = geom_vline(
           xintercept = 0,
           colour = "grey60"
       ),
       vars_order = c("busclose_first:year_indicator", "year_indicator:busopen_most_pop",
                      "year_indicator:online_only", "year_indicator:hybrid"),
       model_order = c("Model 1", "Model 2")) %>% 
       relabel_predictors(c("busclose_first:year_indicator" = "Closure start",
                          "year_indicator:busopen_most_pop" = "Closure end",
                          "year_indicator:online_only" = "Online school reopening",
                          "year_indicator:hybrid" = "Hybrid school reopening")) + 
      theme_classic(base_size = 14) + 
      scale_color_discrete(name = "Outcome", labels = c("Substance Misuse", "Substance Use")) +
      scale_x_continuous(labels = scales::percent) + 
      labs(x = "Effect size (approx. % change)") + 
      theme(legend.position = "top")


```

```{r}
substance_sip <- felm(log(0.144*alcohol + 0.833*cannabis + 0.451*cigarette + 
                            0.172*e_cig + 0.228*smoking + 1) ~ 
                                  sip_first*year_indicator + 
                                     no_sip_most_pop*year_indicator + 
                        online_only*year_indicator + 
                        hybrid*year_indicator + 
                        other*year_indicator + 
                        cases_per_100k + unemploy_rate + temp + precip + 
                       log(lockdown + 1) + log(covid + 1) + 
                       factor(income_support) + factor(debt_contract_relief)
                       |factor(week) + factor(dma) + factor(year)|0|dma, 
             weights = na.omit(searches_norm3_v22_5$pop),
                  data = searches_norm3_v22_5 %>% filter(is.na(pop) == FALSE))

substance_misuse_sip <- felm(log(0.308 * addiction + 0.390 * alcoholism + 0.104 * rehab + 
                             0.861 * substance_abuse + 1) ~ 
                                  sip_first * year_indicator + 
                                    no_sip_most_pop * year_indicator + 
                        online_only * year_indicator + 
                        hybrid * year_indicator + 
                        other * year_indicator + 
                        cases_per_100k + unemploy_rate + temp + precip + 
                       log(lockdown + 1) + log(covid + 1) + 
                       factor(income_support) + factor(debt_contract_relief)
                       |factor(week) + factor(dma) + factor(year)|0|dma, 
              weights = na.omit(searches_norm3_v22_5$pop),
                  data = searches_norm3_v22_5 %>% filter(is.na(pop) == FALSE))

dwplot(list(substance_sip, substance_misuse_sip),
       vline = geom_vline(
           xintercept = 0,
           colour = "grey60"
       ),
       vars_order = c("sip_first:year_indicator", "year_indicator:no_sip_most_pop",
                      "year_indicator:online_only", "year_indicator:hybrid"),
       model_order = c("Model 1", "Model 2")) %>% 
       relabel_predictors(c("sip_first:year_indicator" = "SIP start",
                          "year_indicator:no_sip_most_pop" = "SIP end",
                          "year_indicator:online_only" = "Online school reopening",
                          "year_indicator:hybrid" = "Hybrid school reopening")) + 
      theme_classic(base_size = 14) + 
      scale_color_discrete(name = "Outcome", labels = c("Substance Misuse", "Substance Use")) +
      scale_x_continuous(labels = scales::percent) + 
      labs(x = "Effect size (approx. % change)") + 
      theme(legend.position = "top")

```

```{r}
# HPS - depression and anxiety

depression_1 <- felm(log(hps_depression) ~ c1_post + c1_future +
               initclaims_combined + 
               contclaims_combined + cases_per_100k_5day + 
               factor(e1_income_support) + factor(e2_debt_contract_relief) + 
                log(covid_norm+1) + log(lockdown_norm+1)
             |factor(period) + factor(state) + factor(year)|0|state,
             data = hps_all_v19 %>% drop_na(pop),
             weights = na.omit(hps_all_v19$pop))

depression_2 <- felm(log(hps_depression) ~ factor(c1_school_closing) +
               initclaims_combined + 
               contclaims_combined + cases_per_100k_5day + 
               factor(e1_income_support) + factor(e2_debt_contract_relief) + 
                log(covid_norm+1) + log(lockdown_norm+1)
             |factor(period) + factor(state) + factor(year)|0|state,
             data = hps_all_v19 %>% drop_na(pop),
             weights = na.omit(hps_all_v19$pop))

anxiety_1 <- felm(log(hps_anxiety) ~ c1_post + c1_future +
               initclaims_combined + 
               contclaims_combined + cases_per_100k_5day + 
               factor(e1_income_support) + factor(e2_debt_contract_relief) + 
                log(covid_norm+1) + log(lockdown_norm+1)
             |factor(period) + factor(state) + factor(year)|0|state,
             data = hps_all_v19 %>% drop_na(pop),
             weights = na.omit(hps_all_v19$pop))

anxiety_2 <- felm(log(hps_anxiety) ~ factor(c1_school_closing) +
               initclaims_combined + 
               contclaims_combined + cases_per_100k_5day + 
               factor(e1_income_support) + factor(e2_debt_contract_relief) + 
                log(covid_norm+1) + log(lockdown_norm+1)
             |factor(period) + factor(state) + factor(year)|0|state,
             data = hps_all_v19 %>% drop_na(pop),
             weights = na.omit(hps_all_v19$pop))

dwplot(list(depression_1, depression_2, anxiety_1, anxiety_2),
       vline = geom_vline(
           xintercept = 0,
           colour = "grey60",
       ),
       vars_order = c("c1_post", "c1_future",
                      "factor(c1_school_closing)1", "factor(c1_school_closing)2",
                      "factor(c1_school_closing)3"),
       model_order = c("Model 1", "Model 3", "Model 2", "Model 4"),
       dodge_size = 1) %>% 
    relabel_predictors(c("c1_post" = "Closure end",
                         "c1_future" = "Subsequent closure start",
                         "factor(c1_school_closing)1" = "Closure level 1",
                         "factor(c1_school_closing)2" = "Closure level 2",
                         "factor(c1_school_closing)3" = "Closure level 3")) +
    theme_classic(base_size = 14) +
    scale_color_discrete(name = "Model", labels = c("Anxiety (strictness)",
                                                      "Depression (strictness)",
                                                      "Anxiety (start/end)", 
                                                      "Depression (start/end)")) +
      scale_x_continuous(labels = scales::percent) + 
      labs(x = "Effect size (approx. % change)") + 
      theme(legend.position = "top") + 
      guides(color=guide_legend(nrow=2, byrow=TRUE))
                        
```

```{r}


depression_1 <- felm(log(hps_depression) ~ c2_post + c2_future +
               initclaims_combined + 
               contclaims_combined + cases_per_100k_5day + 
               factor(e1_income_support) + factor(e2_debt_contract_relief) + 
                log(covid_norm+1) + log(lockdown_norm+1)
             |factor(period) + factor(state) + factor(year)|0|state,
             data = hps_all_v19 %>% drop_na(pop),
             weights = na.omit(hps_all_v19$pop))

depression_2 <- felm(log(hps_depression) ~ factor(c2_workplace_closing) +
               initclaims_combined + 
               contclaims_combined + cases_per_100k_5day + 
               factor(e1_income_support) + factor(e2_debt_contract_relief) + 
                log(covid_norm+1) + log(lockdown_norm+1)
             |factor(period) + factor(state) + factor(year)|0|state,
             data = hps_all_v19 %>% drop_na(pop),
             weights = na.omit(hps_all_v19$pop))

anxiety_1 <- felm(log(hps_anxiety) ~ c2_post + c2_future +
               initclaims_combined + 
               contclaims_combined + cases_per_100k_5day + 
               factor(e1_income_support) + factor(e2_debt_contract_relief) + 
                log(covid_norm+1) + log(lockdown_norm+1)
             |factor(period) + factor(state) + factor(year)|0|state,
             data = hps_all_v19 %>% drop_na(pop),
             weights = na.omit(hps_all_v19$pop))

anxiety_2 <- felm(log(hps_anxiety) ~ factor(c2_workplace_closing) +
               initclaims_combined + 
               contclaims_combined + cases_per_100k_5day + 
               factor(e1_income_support) + factor(e2_debt_contract_relief) + 
                log(covid_norm+1) + log(lockdown_norm+1)
             |factor(period) + factor(state) + factor(year)|0|state,
             data = hps_all_v19 %>% drop_na(pop),
             weights = na.omit(hps_all_v19$pop))

dwplot(list(depression_1, depression_2, anxiety_1, anxiety_2),
       vline = geom_vline(
           xintercept = 0,
           colour = "grey60",
       ),
       vars_order = c("c2_post", "c2_future",
                      "factor(c2_workplace_closing)1", "factor(c2_workplace_closing)2",
                      "factor(c2_workplace_closing)3"),
       model_order = c("Model 1", "Model 3", "Model 2", "Model 4"),
       dodge_size = 1) %>% 
    relabel_predictors(c("c2_post" = "Closure end",
                         "c2_future" = "Subsequent closure start",
                         "factor(c2_workplace_closing)1" = "Closure level 1",
                         "factor(c2_workplace_closing)2" = "Closure level 2",
                         "factor(c2_workplace_closing)3" = "Closure level 3")) +
    theme_classic(base_size = 14) +
    scale_color_discrete(name = "Model", labels = c("Anxiety (strictness)",
                                                      "Depression (strictness)",
                                                      "Anxiety (start/end)", 
                                                      "Depression (start/end)")) +
      scale_x_continuous(labels = scales::percent) + 
      labs(x = "Effect size (approx. % change)") + 
      theme(legend.position = "top") + 
      guides(color=guide_legend(nrow=2, byrow=TRUE))
```

```{r}


depression_1 <- felm(log(hps_depression) ~ c6_post + c6_future +
               initclaims_combined + 
               contclaims_combined + cases_per_100k_5day + 
               factor(e1_income_support) + factor(e2_debt_contract_relief) + 
                log(covid_norm+1) + log(lockdown_norm+1)
             |factor(period) + factor(state) + factor(year)|0|state,
             data = hps_all_v19 %>% drop_na(pop),
             weights = na.omit(hps_all_v19$pop))

depression_2 <- felm(log(hps_depression) ~ factor(c6_stay_at_home_requirements) +
               initclaims_combined + 
               contclaims_combined + cases_per_100k_5day + 
               factor(e1_income_support) + factor(e2_debt_contract_relief) + 
                log(covid_norm+1) + log(lockdown_norm+1)
             |factor(period) + factor(state) + factor(year)|0|state,
             data = hps_all_v19 %>% drop_na(pop),
             weights = na.omit(hps_all_v19$pop))

anxiety_1 <- felm(log(hps_anxiety) ~ c6_post + c6_future +
               initclaims_combined + 
               contclaims_combined + cases_per_100k_5day + 
               factor(e1_income_support) + factor(e2_debt_contract_relief) + 
                log(covid_norm+1) + log(lockdown_norm+1)
             |factor(period) + factor(state) + factor(year)|0|state,
             data = hps_all_v19 %>% drop_na(pop),
             weights = na.omit(hps_all_v19$pop))

anxiety_2 <- felm(log(hps_anxiety) ~ factor(c6_stay_at_home_requirements) +
               initclaims_combined + 
               contclaims_combined + cases_per_100k_5day + 
               factor(e1_income_support) + factor(e2_debt_contract_relief) + 
                log(covid_norm+1) + log(lockdown_norm+1)
             |factor(period) + factor(state) + factor(year)|0|state,
             data = hps_all_v19 %>% drop_na(pop),
             weights = na.omit(hps_all_v19$pop))

dwplot(list(depression_1, depression_2, anxiety_1, anxiety_2),
       vline = geom_vline(
           xintercept = 0,
           colour = "grey60",
       ),
       vars_order = c("c6_post", "c6_future",
                      "factor(c6_stay_at_home_requirements)1", "factor(c6_stay_at_home_requirements)2"),
       model_order = c("Model 1", "Model 3", "Model 2", "Model 4"),
       dodge_size = 1) %>% 
    relabel_predictors(c("c6_post" = "Order end",
                         "c6_future" = "Subsequent order start",
                         "factor(c6_stay_at_home_requirements)1" = "Order level 1",
                         "factor(c6_stay_at_home_requirements)2" = "Order level 2")) +
    theme_classic(base_size = 14) +
    scale_color_discrete(name = "Model", labels = c("Anxiety (strictness",
                                                      "Depression (strictness)",
                                                      "Anxiety (start/end)", 
                                                      "Depression (start/end)")) +
      scale_x_continuous(labels = scales::percent) + 
      labs(x = "Effect size (approx. % change)") + 
      theme(legend.position = "top") + 
      guides(color=guide_legend(nrow=2, byrow=TRUE))
```

```{r}
therapy_1 <- felm(log(counseling) ~ c1_post + c1_future +
               initclaims_combined + 
               contclaims_combined + cases_per_100k_5day + 
               factor(e1_income_support) + factor(e2_debt_contract_relief) + 
                log(covid_norm+1) + log(lockdown_norm+1)
             |factor(period) + factor(state) + factor(year)|0|state,
             data = hps_all_v19 %>% drop_na(pop),
             weights = na.omit(hps_all_v19$pop))

therapy_2 <- felm(log(medication) ~ c1_post + c1_future +
               initclaims_combined + 
               contclaims_combined + cases_per_100k_5day + 
               factor(e1_income_support) + factor(e2_debt_contract_relief) + 
                log(covid_norm+1) + log(lockdown_norm+1)
             |factor(period) + factor(state) + factor(year)|0|state,
             data = hps_all_v19 %>% drop_na(pop),
             weights = na.omit(hps_all_v19$pop))

therapy_3 <- felm(log(no_counseling) ~ c1_post + c1_future +
               initclaims_combined + 
               contclaims_combined + cases_per_100k_5day + 
               factor(e1_income_support) + factor(e2_debt_contract_relief) + 
                log(covid_norm+1) + log(lockdown_norm+1)
             |factor(period) + factor(state) + factor(year)|0|state,
             data = hps_all_v19 %>% drop_na(pop),
             weights = na.omit(hps_all_v19$pop))

therapy_4 <- felm(log(counseling) ~ factor(c1_school_closing) +
               initclaims_combined + 
               contclaims_combined + cases_per_100k_5day + 
               factor(e1_income_support) + factor(e2_debt_contract_relief) + 
                log(covid_norm+1) + log(lockdown_norm+1)
             |factor(period) + factor(state) + factor(year)|0|state,
             data = hps_all_v19 %>% drop_na(pop),
             weights = na.omit(hps_all_v19$pop))

therapy_5 <- felm(log(medication) ~ factor(c1_school_closing) +
               initclaims_combined + 
               contclaims_combined + cases_per_100k_5day + 
               factor(e1_income_support) + factor(e2_debt_contract_relief) + 
                log(covid_norm+1) + log(lockdown_norm+1)
             |factor(period) + factor(state) + factor(year)|0|state,
             data = hps_all_v19 %>% drop_na(pop),
             weights = na.omit(hps_all_v19$pop))

therapy_6 <- felm(log(no_counseling) ~ factor(c1_school_closing) +
               initclaims_combined + 
               contclaims_combined + cases_per_100k_5day + 
               factor(e1_income_support) + factor(e2_debt_contract_relief) + 
                log(covid_norm+1) + log(lockdown_norm+1)
             |factor(period) + factor(state) + factor(year)|0|state,
             data = hps_all_v19 %>% drop_na(pop),
             weights = na.omit(hps_all_v19$pop))

dwplot(list(therapy_1, therapy_2, therapy_3),
       vline = geom_vline(
           xintercept = 0,
           colour = "grey60"
       ),
       vars_order = c("c1_post", "c1_future",
                      "factor(c1_school_closing)1", "factor(c1_school_closing)2",
                      "factor(c1_school_closing)3"),
       model_order = c("Model 1", "Model 2", "Model 3", "Model 4", "Model 5", "Model 6"),
       dodge_size = 1) %>% 
    relabel_predictors(c("c1_post" = "Closure end",
                         "c1_future" = "Subsequent closure start",
                         "factor(c1_school_closing)1" = "Closure level 1",
                         "factor(c1_school_closing)2" = "Closure level 2",
                         "factor(c1_school_closing)3" = "Closure level 3")) +
    theme_classic(base_size = 14) +
    scale_color_discrete(name = "Model", labels = c("No counseling (strictness)",
                                                      "Medication (strictness)",
                                                      "Counseling (strictness)", 
                                                      "No counseling (start/end)",
                                                    "Medication (start/end)",
                                                    "Counseling (start/end)")) +
      scale_x_continuous(labels = scales::percent) + 
      labs(x = "Effect size (approx. % change)") + 
      theme(legend.position = "top") + 
      guides(color=guide_legend(nrow=2, byrow=TRUE))

dwplot(list(therapy_1, therapy_2, therapy_3, therapy_4, therapy_5, therapy_6),
       vline = geom_vline(
           xintercept = 0,
           colour = "grey60"
       ),
       vars_order = c("c1_post", "c1_future",
                      "factor(c1_school_closing)1", "factor(c1_school_closing)2",
                      "factor(c1_school_closing)3"),
       model_order = c("Model 1", "Model 2", "Model 3", "Model 4", "Model 5", "Model 6"),
       dodge_size = 1) %>% 
    relabel_predictors(c("c1_post" = "Closure end",
                         "c1_future" = "Subsequent closure start",
                         "factor(c1_school_closing)1" = "Closure level 1",
                         "factor(c1_school_closing)2" = "Closure level 2",
                         "factor(c1_school_closing)3" = "Closure level 3")) +
    theme_classic(base_size = 14) +
    scale_color_discrete(name = "Model", labels = c("No counseling (strictness)",
                                                      "Medication (strictness)",
                                                      "Counseling (strictness)", 
                                                      "No counseling (start/end)",
                                                    "Medication (start/end)",
                                                    "Counseling (start/end)")) +
      scale_x_continuous(labels = scales::percent) + 
      labs(x = "Effect size (approx. % change)") + 
      theme(legend.position = "top") + 
      guides(color=guide_legend(nrow=2, byrow=TRUE))
```


```{r}

therapy_1 <- felm(log(counseling) ~ c2_post + c2_future +
               initclaims_combined + 
               contclaims_combined + cases_per_100k_5day + 
               factor(e1_income_support) + factor(e2_debt_contract_relief) + 
                log(covid_norm+1) + log(lockdown_norm+1)
             |factor(period) + factor(state) + factor(year)|0|state,
             data = hps_all_v19 %>% drop_na(pop),
             weights = na.omit(hps_all_v19$pop))

therapy_2 <- felm(log(medication) ~ c2_post + c2_future +
               initclaims_combined + 
               contclaims_combined + cases_per_100k_5day + 
               factor(e1_income_support) + factor(e2_debt_contract_relief) + 
                log(covid_norm+1) + log(lockdown_norm+1)
             |factor(period) + factor(state) + factor(year)|0|state,
             data = hps_all_v19 %>% drop_na(pop),
             weights = na.omit(hps_all_v19$pop))

therapy_3 <- felm(log(no_counseling) ~ c2_post + c2_future +
               initclaims_combined + 
               contclaims_combined + cases_per_100k_5day + 
               factor(e1_income_support) + factor(e2_debt_contract_relief) + 
                log(covid_norm+1) + log(lockdown_norm+1)
             |factor(period) + factor(state) + factor(year)|0|state,
             data = hps_all_v19 %>% drop_na(pop),
             weights = na.omit(hps_all_v19$pop))

therapy_4 <- felm(log(counseling) ~ factor(c2_workplace_closing) +
               initclaims_combined + 
               contclaims_combined + cases_per_100k_5day + 
               factor(e1_income_support) + factor(e2_debt_contract_relief) + 
                log(covid_norm+1) + log(lockdown_norm+1)
             |factor(period) + factor(state) + factor(year)|0|state,
             data = hps_all_v19 %>% drop_na(pop),
             weights = na.omit(hps_all_v19$pop))

therapy_5 <- felm(log(medication) ~ factor(c2_workplace_closing) +
               initclaims_combined + 
               contclaims_combined + cases_per_100k_5day + 
               factor(e1_income_support) + factor(e2_debt_contract_relief) + 
                log(covid_norm+1) + log(lockdown_norm+1)
             |factor(period) + factor(state) + factor(year)|0|state,
             data = hps_all_v19 %>% drop_na(pop),
             weights = na.omit(hps_all_v19$pop))

therapy_6 <- felm(log(no_counseling) ~ factor(c2_workplace_closing) +
               initclaims_combined + 
               contclaims_combined + cases_per_100k_5day + 
               factor(e1_income_support) + factor(e2_debt_contract_relief) + 
                log(covid_norm+1) + log(lockdown_norm+1)
             |factor(period) + factor(state) + factor(year)|0|state,
             data = hps_all_v19 %>% drop_na(pop),
             weights = na.omit(hps_all_v19$pop))

dwplot(list(therapy_1, therapy_2, therapy_3, therapy_4, therapy_5, therapy_6),
       vline = geom_vline(
           xintercept = 0,
           colour = "grey60"
       ),
       vars_order = c("c2_post", "c2_future",
                      "factor(c2_workplace_closing)1", "factor(c2_workplace_closing)2",
                      "factor(c2_workplace_closing)3"),
       model_order = c("Model 1", "Model 2", "Model 3", "Model 4", "Model 5", "Model 6"),
       dodge_size = 1) %>% 
    relabel_predictors(c("c2_post" = "Closure end",
                         "c2_future" = "Subsequent closure start",
                         "factor(c2_workplace_closing)1" = "Closure level 1",
                         "factor(c2_workplace_closing)2" = "Closure level 2",
                         "factor(c2_workplace_closing)3" = "Closure level 3")) +
    theme_classic(base_size = 14) +
    scale_color_discrete(name = "Model", labels = c("No counseling (strictness)",
                                                      "Medication (strictness)",
                                                      "Counseling (strictness)", 
                                                      "No counseling (start/end)",
                                                    "Medication (start/end)",
                                                    "Counseling (start/end)")) +
      scale_x_continuous(labels = scales::percent) + 
      labs(x = "Effect size (approx. % change)") + 
      theme(legend.position = "top") + 
      guides(color=guide_legend(nrow=2, byrow=TRUE))

dwplot(list(therapy_1, therapy_2, therapy_3),
       vline = geom_vline(
           xintercept = 0,
           colour = "grey60"
       ),
       vars_order = c("c2_post", "c2_future"),
       model_order = c("Model 1", "Model 2", "Model 3"),
       dodge_size = 1) %>% 
    relabel_predictors(c("c2_post" = "Closure end",
                         "c2_future" = "Subsequent closure start")) +
    theme_classic(base_size = 14) +
    scale_color_discrete(name = "Model", labels = c( 
                                                      "No counseling (start/end)",
                                                    "Medication (start/end)",
                                                    "Counseling (start/end)")) +
      scale_x_continuous(labels = scales::percent) + 
      labs(x = "Effect size (approx. % change)") + 
      theme(legend.position = "top") + 
      guides(color=guide_legend(nrow=2, byrow=TRUE))
```

```{r}

therapy_1 <- felm(log(counseling) ~ c6_post + c6_future +
               initclaims_combined + 
               contclaims_combined + cases_per_100k_5day + 
               factor(e1_income_support) + factor(e2_debt_contract_relief) + 
                log(covid_norm+1) + log(lockdown_norm+1)
             |factor(period) + factor(state) + factor(year)|0|state,
             data = hps_all_v19 %>% drop_na(pop),
             weights = na.omit(hps_all_v19$pop))

therapy_2 <- felm(log(medication) ~ c6_post + c6_future +
               initclaims_combined + 
               contclaims_combined + cases_per_100k_5day + 
               factor(e1_income_support) + factor(e2_debt_contract_relief) + 
                log(covid_norm+1) + log(lockdown_norm+1)
             |factor(period) + factor(state) + factor(year)|0|state,
             data = hps_all_v19 %>% drop_na(pop),
             weights = na.omit(hps_all_v19$pop))

therapy_3 <- felm(log(no_counseling) ~ c6_post + c6_future +
               initclaims_combined + 
               contclaims_combined + cases_per_100k_5day + 
               factor(e1_income_support) + factor(e2_debt_contract_relief) + 
                log(covid_norm+1) + log(lockdown_norm+1)
             |factor(period) + factor(state) + factor(year)|0|state,
             data = hps_all_v19 %>% drop_na(pop),
             weights = na.omit(hps_all_v19$pop))

therapy_4 <- felm(log(counseling) ~ factor(c6_stay_at_home_requirements) +
               initclaims_combined + 
               contclaims_combined + cases_per_100k_5day + 
               factor(e1_income_support) + factor(e2_debt_contract_relief) + 
                log(covid_norm+1) + log(lockdown_norm+1)
             |factor(period) + factor(state) + factor(year)|0|state,
             data = hps_all_v19 %>% drop_na(pop),
             weights = na.omit(hps_all_v19$pop))

therapy_5 <- felm(log(medication) ~ factor(c6_stay_at_home_requirements) +
               initclaims_combined + 
               contclaims_combined + cases_per_100k_5day + 
               factor(e1_income_support) + factor(e2_debt_contract_relief) + 
                log(covid_norm+1) + log(lockdown_norm+1)
             |factor(period) + factor(state) + factor(year)|0|state,
             data = hps_all_v19 %>% drop_na(pop),
             weights = na.omit(hps_all_v19$pop))

therapy_6 <- felm(log(no_counseling) ~ factor(c6_stay_at_home_requirements) +
               initclaims_combined + 
               contclaims_combined + cases_per_100k_5day + 
               factor(e1_income_support) + factor(e2_debt_contract_relief) + 
                log(covid_norm+1) + log(lockdown_norm+1)
             |factor(period) + factor(state) + factor(year)|0|state,
             data = hps_all_v19 %>% drop_na(pop),
             weights = na.omit(hps_all_v19$pop))

dwplot(list(therapy_1, therapy_2, therapy_3, therapy_4, therapy_5, therapy_6),
       vline = geom_vline(
           xintercept = 0,
           colour = "grey60"
       ),
       vars_order = c("c6_post", "c6_future",
                      "factor(c6_stay_at_home_requirements)1", "factor(c6_stay_at_home_requirements)2"),
       model_order = c("Model 1", "Model 2", "Model 3", "Model 4", "Model 5", "Model 6"),
       dodge_size = 1) %>% 
    relabel_predictors(c("c6_post" = "Closure end",
                         "c6_future" = "Subsequent closure start",
                         "factor(c6_stay_at_home_requirements)1" = "Order level 1",
                         "factor(c6_stay_at_home_requirements)2" = "Order level 2")) +
    theme_classic(base_size = 14) +
    scale_color_discrete(name = "Model", labels = c("No counseling (strictness)",
                                                      "Medication (strictness)",
                                                      "Counseling (strictness)", 
                                                      "No counseling (start/end)",
                                                    "Medication (start/end)",
                                                    "Counseling (start/end)")) +
      scale_x_continuous(labels = scales::percent) + 
      labs(x = "Effect size (approx. % change)") + 
      theme(legend.position = "top") + 
      guides(color=guide_legend(nrow=2, byrow=TRUE))

```


```{r}
# Workplace closures, HPS, by Race


depression_3 <- felm(log(depression_black + 1) ~ c2_post + c2_future +
               initclaims_combined + 
               contclaims_combined + cases_per_100k_5day + 
               factor(e1_income_support) + factor(e2_debt_contract_relief) + 
                log(covid_norm+1) + log(lockdown_norm+1)
             |factor(period) + factor(state) + factor(year)|0|state,
             data = hps_all_v19 %>% drop_na(pop),
             weights = na.omit(hps_all_v19$pop))

depression_4 <- felm(log(depression_black + 1) ~ factor(c2_workplace_closing) +
               initclaims_combined + 
               contclaims_combined + cases_per_100k_5day + 
               factor(e1_income_support) + factor(e2_debt_contract_relief) + 
                log(covid_norm+1) + log(lockdown_norm+1)
             |factor(period) + factor(state) + factor(year)|0|state,
             data = hps_all_v19 %>% drop_na(pop),
             weights = na.omit(hps_all_v19$pop))

depression_5 <- felm(log(depression_white + 1) ~ c2_post + c2_future +
               initclaims_combined + 
               contclaims_combined + cases_per_100k_5day + 
               factor(e1_income_support) + factor(e2_debt_contract_relief) + 
                log(covid_norm+1) + log(lockdown_norm+1)
             |factor(period) + factor(state) + factor(year)|0|state,
             data = hps_all_v19 %>% drop_na(pop),
             weights = na.omit(hps_all_v19$pop))

depression_6 <- felm(log(depression_white + 1) ~ factor(c2_workplace_closing) +
               initclaims_combined + 
               contclaims_combined + cases_per_100k_5day + 
               factor(e1_income_support) + factor(e2_debt_contract_relief) + 
                log(covid_norm+1) + log(lockdown_norm+1)
             |factor(period) + factor(state) + factor(year)|0|state,
             data = hps_all_v19 %>% drop_na(pop),
             weights = na.omit(hps_all_v19$pop))

dwplot(list(depression_3, depression_4, depression_5, depression_6),
       vline = geom_vline(
           xintercept = 0,
           colour = "grey60",
       ),
       vars_order = c("c2_post", "c2_future",
                      "factor(c2_workplace_closing)1", 
                      "factor(c2_workplace_closing)2",
                      "factor(c2_workplace_closing)3"),
       model_order = c("Model 1", "Model 3", "Model 2", "Model 4"),
       dodge_size = 1) %>% 
    relabel_predictors(c("c2_post" = "Closure end",
                         "c2_future" = "Subsequent closure start",
                         "factor(c2_workplace_closing)1" = "Closure level 1",
                         "factor(c2_workplace_closing)2" = "Closure level 2",
                         "factor(c2_workplace_closing)3" = "Closure level 3")) +
    theme_classic(base_size = 14) +
    
    scale_color_discrete(name = "Model", labels = c(
                                                    "White (strictness)",
                                                      "Black (strictness)",
                                                      "White (start/end)",
                                                      "Black (start/end)"
                                                      )) +
      scale_x_continuous(labels = scales::percent) + 
      labs(x = "Effect size (approx. % change)") + 
      theme(legend.position = "top") + 
      guides(color=guide_legend(nrow=3, byrow=TRUE))
```


```{r}
anxiety_3 <- felm(log(anxiety_agg_black + 1) ~ c2_post + c2_future +
               initclaims_combined + 
               contclaims_combined + cases_per_100k_5day + 
               factor(e1_income_support) + factor(e2_debt_contract_relief) + 
                log(covid_norm+1) + log(lockdown_norm+1)
             |factor(period) + factor(state) + factor(year)|0|state,
             data = hps_all_v19 %>% drop_na(pop),
             weights = na.omit(hps_all_v19$pop))

anxiety_4 <- felm(log(anxiety_agg_black + 1) ~ factor(c2_workplace_closing) +
               initclaims_combined + 
               contclaims_combined + cases_per_100k_5day + 
               factor(e1_income_support) + factor(e2_debt_contract_relief) + 
                log(covid_norm+1) + log(lockdown_norm+1)
             |factor(period) + factor(state) + factor(year)|0|state,
             data = hps_all_v19 %>% drop_na(pop),
             weights = na.omit(hps_all_v19$pop))

anxiety_5 <- felm(log(anxiety_agg_white + 1) ~ c2_post + c2_future +
               initclaims_combined + 
               contclaims_combined + cases_per_100k_5day + 
               factor(e1_income_support) + factor(e2_debt_contract_relief) + 
                log(covid_norm+1) + log(lockdown_norm+1)
             |factor(period) + factor(state) + factor(year)|0|state,
             data = hps_all_v19 %>% drop_na(pop),
             weights = na.omit(hps_all_v19$pop))

anxiety_6 <- felm(log(anxiety_agg_white + 1) ~ factor(c2_workplace_closing) +
               initclaims_combined + 
               contclaims_combined + cases_per_100k_5day + 
               factor(e1_income_support) + factor(e2_debt_contract_relief) + 
                log(covid_norm+1) + log(lockdown_norm+1)
             |factor(period) + factor(state) + factor(year)|0|state,
             data = hps_all_v19 %>% drop_na(pop),
             weights = na.omit(hps_all_v19$pop))

dwplot(list(anxiety_3, anxiety_4, anxiety_5, anxiety_6),
       vline = geom_vline(
           xintercept = 0,
           colour = "grey60",
       ),
       vars_order = c("c2_post", "c2_future",
                      "factor(c2_workplace_closing)1", 
                      "factor(c2_workplace_closing)2",
                      "factor(c2_workplace_closing)3"),
       model_order = c("Model 1", "Model 3", "Model 2", "Model 4"),
       dodge_size = 1) %>% 
    relabel_predictors(c("c2_post" = "Closure end",
                         "c2_future" = "Subsequent closure start",
                         "factor(c2_workplace_closing)1" = "Closure level 1",
                         "factor(c2_workplace_closing)2" = "Closure level 2",
                         "factor(c2_workplace_closing)3" = "Closure level 3")) +
    theme_classic(base_size = 14) +
    
    scale_color_discrete(name = "Model", labels = c(
                                                    "White (strictness)",
                                                      "Black (strictness)",
                                                      "White (start/end)",
                                                      "Black (start/end)"
                                                      )) +
      scale_x_continuous(labels = scales::percent) + 
      labs(x = "Effect size (approx. % change)") + 
      theme(legend.position = "top") + 
      guides(color=guide_legend(nrow=2, byrow=TRUE))
```

```{r}
# School closures
depression_3 <- felm(log(depression_black + 1) ~ c1_post + c1_future +
               initclaims_combined + 
               contclaims_combined + cases_per_100k_5day + 
               factor(e1_income_support) + factor(e2_debt_contract_relief) + 
                log(covid_norm+1) + log(lockdown_norm+1)
             |factor(period) + factor(state) + factor(year)|0|state,
             data = hps_all_v19 %>% drop_na(pop),
             weights = na.omit(hps_all_v19$pop))

depression_4 <- felm(log(depression_black + 1) ~ factor(c1_school_closing) +
               initclaims_combined + 
               contclaims_combined + cases_per_100k_5day + 
               factor(e1_income_support) + factor(e2_debt_contract_relief) + 
                log(covid_norm+1) + log(lockdown_norm+1)
             |factor(period) + factor(state) + factor(year)|0|state,
             data = hps_all_v19 %>% drop_na(pop),
             weights = na.omit(hps_all_v19$pop))

depression_5 <- felm(log(depression_white + 1) ~ c1_post + c1_future +
               initclaims_combined + 
               contclaims_combined + cases_per_100k_5day + 
               factor(e1_income_support) + factor(e2_debt_contract_relief) + 
                log(covid_norm+1) + log(lockdown_norm+1)
             |factor(period) + factor(state) + factor(year)|0|state,
             data = hps_all_v19 %>% drop_na(pop),
             weights = na.omit(hps_all_v19$pop))

depression_6 <- felm(log(depression_white + 1) ~ factor(c1_school_closing) +
               initclaims_combined + 
               contclaims_combined + cases_per_100k_5day + 
               factor(e1_income_support) + factor(e2_debt_contract_relief) + 
                log(covid_norm+1) + log(lockdown_norm+1)
             |factor(period) + factor(state) + factor(year)|0|state,
             data = hps_all_v19 %>% drop_na(pop),
             weights = na.omit(hps_all_v19$pop))

dwplot(list(depression_3, depression_4, depression_5, depression_6),
       vline = geom_vline(
           xintercept = 0,
           colour = "grey60",
       ),
       vars_order = c("c1_post", "c1_future",
                      "factor(c1_school_closing)1", 
                      "factor(c1_school_closing)2",
                      "factor(c1_school_closing)3"),
       model_order = c("Model 1", "Model 3", "Model 2", "Model 4"),
       dodge_size = 1) %>% 
    relabel_predictors(c("c1_post" = "Closure end",
                         "c1_future" = "Subsequent closure start",
                         "factor(c1_school_closing)1" = "Closure level 1",
                         "factor(c1_school_closing)2" = "Closure level 2",
                         "factor(c1_school_closing)3" = "Closure level 3")) +
    theme_classic(base_size = 14) +
    
    scale_color_discrete(name = "Model", labels = c(
                                                    "White (strictness)",
                                                      "Black (strictness)",
                                                      "White (start/end)",
                                                      "Black (start/end)"
                                                      )) +
      scale_x_continuous(labels = scales::percent) + 
      labs(x = "Effect size (approx. % change)") + 
      theme(legend.position = "top") + 
      guides(color=guide_legend(nrow=3, byrow=TRUE))
```

```{r}
anxiety_3 <- felm(log(anxiety_agg_black + 1) ~ c1_post + c1_future +
               initclaims_combined + 
               contclaims_combined + cases_per_100k_5day + 
               factor(e1_income_support) + factor(e2_debt_contract_relief) + 
                log(covid_norm+1) + log(lockdown_norm+1)
             |factor(period) + factor(state) + factor(year)|0|state,
             data = hps_all_v19 %>% drop_na(pop),
             weights = na.omit(hps_all_v19$pop))

anxiety_4 <- felm(log(anxiety_agg_black + 1) ~ factor(c1_school_closing) +
               initclaims_combined + 
               contclaims_combined + cases_per_100k_5day + 
               factor(e1_income_support) + factor(e2_debt_contract_relief) + 
                log(covid_norm+1) + log(lockdown_norm+1)
             |factor(period) + factor(state) + factor(year)|0|state,
             data = hps_all_v19 %>% drop_na(pop),
             weights = na.omit(hps_all_v19$pop))

anxiety_5 <- felm(log(anxiety_agg_white + 1) ~ c1_post + c1_future +
               initclaims_combined + 
               contclaims_combined + cases_per_100k_5day + 
               factor(e1_income_support) + factor(e2_debt_contract_relief) + 
                log(covid_norm+1) + log(lockdown_norm+1)
             |factor(period) + factor(state) + factor(year)|0|state,
             data = hps_all_v19 %>% drop_na(pop),
             weights = na.omit(hps_all_v19$pop))

anxiety_6 <- felm(log(anxiety_agg_white + 1) ~ factor(c1_school_closing) +
               initclaims_combined + 
               contclaims_combined + cases_per_100k_5day + 
               factor(e1_income_support) + factor(e2_debt_contract_relief) + 
                log(covid_norm+1) + log(lockdown_norm+1)
             |factor(period) + factor(state) + factor(year)|0|state,
             data = hps_all_v19 %>% drop_na(pop),
             weights = na.omit(hps_all_v19$pop))

dwplot(list(anxiety_3, anxiety_4, anxiety_5, anxiety_6),
       vline = geom_vline(
           xintercept = 0,
           colour = "grey60",
       ),
       vars_order = c("c1_post", "c1_future",
                      "factor(c1_school_closing)1", 
                      "factor(c1_school_closing)2",
                      "factor(c1_school_closing)3"),
       model_order = c("Model 1", "Model 3", "Model 2", "Model 4"),
       dodge_size = 1) %>% 
    relabel_predictors(c("c1_post" = "Closure end",
                         "c1_future" = "Subsequent closure start",
                         "factor(c1_school_closing)1" = "Closure level 1",
                         "factor(c1_school_closing)2" = "Closure level 2",
                         "factor(c1_school_closing)3" = "Closure level 3")) +
    theme_classic(base_size = 14) +
    
    scale_color_discrete(name = "Model", labels = c(
                                                    "White (strictness)",
                                                      "Black (strictness)",
                                                      "White (start/end)",
                                                      "Black (start/end)"
                                                      )) +
      scale_x_continuous(labels = scales::percent) + 
      labs(x = "Effect size (approx. % change)") + 
      theme(legend.position = "top") + 
      guides(color=guide_legend(nrow=2, byrow=TRUE))
```

```{r}
# School closures
depression_3 <- felm(log(depression_black + 1) ~ c6_post + c6_future +
               initclaims_combined + 
               contclaims_combined + cases_per_100k_5day + 
               factor(e1_income_support) + factor(e2_debt_contract_relief) + 
                log(covid_norm+1) + log(lockdown_norm+1)
             |factor(period) + factor(state) + factor(year)|0|state,
             data = hps_all_v19 %>% drop_na(pop),
             weights = na.omit(hps_all_v19$pop))

depression_4 <- felm(log(depression_black + 1) ~ factor(c6_stay_at_home_requirements) +
               initclaims_combined + 
               contclaims_combined + cases_per_100k_5day + 
               factor(e1_income_support) + factor(e2_debt_contract_relief) + 
                log(covid_norm+1) + log(lockdown_norm+1)
             |factor(period) + factor(state) + factor(year)|0|state,
             data = hps_all_v19 %>% drop_na(pop),
             weights = na.omit(hps_all_v19$pop))

depression_5 <- felm(log(depression_white + 1) ~ c6_post + c6_future +
               initclaims_combined + 
               contclaims_combined + cases_per_100k_5day + 
               factor(e1_income_support) + factor(e2_debt_contract_relief) + 
                log(covid_norm+1) + log(lockdown_norm+1)
             |factor(period) + factor(state) + factor(year)|0|state,
             data = hps_all_v19 %>% drop_na(pop),
             weights = na.omit(hps_all_v19$pop))

depression_6 <- felm(log(depression_white + 1) ~ factor(c6_stay_at_home_requirements) +
               initclaims_combined + 
               contclaims_combined + cases_per_100k_5day + 
               factor(e1_income_support) + factor(e2_debt_contract_relief) + 
                log(covid_norm+1) + log(lockdown_norm+1)
             |factor(period) + factor(state) + factor(year)|0|state,
             data = hps_all_v19 %>% drop_na(pop),
             weights = na.omit(hps_all_v19$pop))

dwplot(list(depression_3, depression_4, depression_5, depression_6),
       vline = geom_vline(
           xintercept = 0,
           colour = "grey60",
       ),
       vars_order = c("c6_post", "c6_future",
                      "factor(c6_stay_at_home_requirements)1", 
                      "factor(c6_stay_at_home_requirements)2"),
       model_order = c("Model 1", "Model 3", "Model 2", "Model 4"),
       dodge_size = 1) %>% 
    relabel_predictors(c("c6_post" = "Closure end",
                         "c6_future" = "Subsequent closure start",
                         "factor(c6_stay_at_home_requirements)1" = "Order level 1",
                         "factor(c6_stay_at_home_requirements)2" = "Order level 2")) +
    theme_classic(base_size = 14) +
    
    scale_color_discrete(name = "Model", labels = c(
                                                    "White (strictness)",
                                                      "Black (strictness)",
                                                      "White (start/end)",
                                                      "Black (start/end)"
                                                      )) +
      scale_x_continuous(labels = scales::percent) + 
      labs(x = "Effect size (approx. % change)") + 
      theme(legend.position = "top") + 
      guides(color=guide_legend(nrow=3, byrow=TRUE))
```

```{r}
anxiety_3 <- felm(log(anxiety_agg_black + 1) ~ c6_post + c6_future +
               initclaims_combined + 
               contclaims_combined + cases_per_100k_5day + 
               factor(e1_income_support) + factor(e2_debt_contract_relief) + 
                log(covid_norm+1) + log(lockdown_norm+1)
             |factor(period) + factor(state) + factor(year)|0|state,
             data = hps_all_v19 %>% drop_na(pop),
             weights = na.omit(hps_all_v19$pop))

anxiety_4 <- felm(log(anxiety_agg_black + 1) ~ factor(c6_stay_at_home_requirements) +
               initclaims_combined + 
               contclaims_combined + cases_per_100k_5day + 
               factor(e1_income_support) + factor(e2_debt_contract_relief) + 
                log(covid_norm+1) + log(lockdown_norm+1)
             |factor(period) + factor(state) + factor(year)|0|state,
             data = hps_all_v19 %>% drop_na(pop),
             weights = na.omit(hps_all_v19$pop))

anxiety_5 <- felm(log(anxiety_agg_white + 1) ~ c6_post + c6_future +
               initclaims_combined + 
               contclaims_combined + cases_per_100k_5day + 
               factor(e1_income_support) + factor(e2_debt_contract_relief) + 
                log(covid_norm+1) + log(lockdown_norm+1)
             |factor(period) + factor(state) + factor(year)|0|state,
             data = hps_all_v19 %>% drop_na(pop),
             weights = na.omit(hps_all_v19$pop))

anxiety_6 <- felm(log(anxiety_agg_white + 1) ~ factor(c6_stay_at_home_requirements) +
               initclaims_combined + 
               contclaims_combined + cases_per_100k_5day + 
               factor(e1_income_support) + factor(e2_debt_contract_relief) + 
                log(covid_norm+1) + log(lockdown_norm+1)
             |factor(period) + factor(state) + factor(year)|0|state,
             data = hps_all_v19 %>% drop_na(pop),
             weights = na.omit(hps_all_v19$pop))

dwplot(list(anxiety_3, anxiety_4, anxiety_5, anxiety_6),
       vline = geom_vline(
           xintercept = 0,
           colour = "grey60",
       ),
       vars_order = c("c6_post", "c6_future",
                      "factor(c6_stay_at_home_requirements)1", 
                      "factor(c6_stay_at_home_requirements)2"),
       model_order = c("Model 1", "Model 3", "Model 2", "Model 4"),
       dodge_size = 1) %>% 
    relabel_predictors(c("c6_post" = "Closure end",
                         "c6_future" = "Subsequent closure start",
                         "factor(c6_stay_at_home_requirements)1" = "Closure level 1",
                         "factor(c6_stay_at_home_requirements)2" = "Closure level 2")) +
    theme_classic(base_size = 14) +
    
    scale_color_discrete(name = "Model", labels = c(
                                                    "White (strictness)",
                                                      "Black (strictness)",
                                                      "White (start/end)",
                                                      "Black (start/end)"
                                                      )) +
      scale_x_continuous(labels = scales::percent) + 
      labs(x = "Effect size (approx. % change)") + 
      theme(legend.position = "top") + 
      guides(color=guide_legend(nrow=2, byrow=TRUE))
```


```{r}
# Verification stuff -- depression

depression_3 <- felm(log(depression_norm) ~ log(hps_depression)|state|0|state, data = hps_all_v19)

depression_4 <- felm(I(log(0.353 * crying_norm + 
                                                   0.163 * depression_norm + 
                                                   0.529 * lonely_norm + 
                                                   0.586 * sad_norm + 
                                                   0.475 * tired_norm))
                     ~ log(hps_depression)|state|0|state, 
                     data = hps_all_v19)

anxiety_3 <- felm(log(anxiety_norm) ~ log(hps_anxiety)|state|0|state, data = hps_all_v19)

anxiety_4 <- felm(I(log(0.509 * anxiety + 
                                                   0.528 * nervous + 
                                                   0.504 * stress + 
                                                   0.399 * worry + 
                                                   0.223 * panic))
                     ~ log(hps_anxiety)|state|0|state, 
                     data = hps_all_v19)


dwplot(list(depression_3, depression_4, anxiety_3, anxiety_4),
       dot_args = list(size = 3),
       # vline = geom_vline(
       #     xintercept = 0,
       #     colour = "grey60",
       # ),
       vars_order = c("log(hps_depression)", "log(hps_anxiety)"),
       model_order = c("Model 1", "Model 2", "Model 3", "Model 4"),
       dodge_size = 0.5) %>% 
    relabel_predictors(c("log(hps_depression)" = "Depression (HPS)",
                         "log(hps_anxiety)" = "Anxiety (HPS)")) +
    theme_classic(base_size = 14) +
    
    scale_color_discrete(name = "Outcome", labels = c(
                                                    "Anxiety index",
                                                      "Anxiety",
                                                      "Depression index",
                                                      "Depression"
                                                      )) +
      # scale_x_continuous(labels = scales::percent) + 
      # labs(x = "Effect size (approx. % change)") + 
      theme(legend.position = "top") + 
      guides(color=guide_legend(nrow=2, byrow=TRUE)) + 
      labs(x = "Coefficient")
```

```{r}
therapy_2 <- felm(log(counseling) ~ log(therapy_norm+1)|state|0|state, data = hps_all_v19)

therapy_4 <- felm(log(no_counseling) ~ log(therapy_norm+1)|state|0|state, data = hps_all_v19)

therapy_6 <- felm(log(medication) ~ log(therapy_norm+1)|state|0|state, data = hps_all_v19)

dwplot(list(therapy_2, therapy_4, therapy_6),
       vline = geom_vline(
           xintercept = 0,
           colour = "grey60",
       ),
       vars_order = c("log(therapy_norm + 1)"),
       model_order = c("Model 1", "Model 2", "Model 3"),
       dodge_size = 0.5) %>% 
    relabel_predictors(c("log(therapy_norm + 1)" = "Therapy SVI")) +
    theme_classic(base_size = 14) +
    scale_color_discrete(name = "Outcome", labels = c(
                                                    "Medication",
                                                      "No counseling",
                                                      "Counseling"
                                                      )) +
      # scale_x_continuous(labels = scales::percent) + 
      # labs(x = "Effect size (approx. % change)") + 
      theme(legend.position = "top") + 
      
  # guides(color=guide_legend(nrow=2, byrow=TRUE)) + 
      labs(x = "Coefficient")

```

```{r}
alc_3<- felm(log(alcohol_norm + 1) ~ log(all_alc)|state + year|0|state, data = alcohol_v2)

alc_7 <- felm(log(alcoholism_norm + 1) ~ log(all_alc)|state + year|0|state, data = alcohol_v2)

cig_3 <- felm(log(cigarettes_norm+1) ~ log(packs_per_capita)|state + year|0|state, data = smoking_v3)

cig_7 <- felm(log(e_cig_norm + cigarettes_norm + smoking_norm+1) 
              ~log(packs_per_capita)|state+year|0|state, data = smoking_v3)

dwplot(list(alc_3, alc_7),
       vline = geom_vline(
           xintercept = 0,
           colour = "grey60",
       ),
       vars_order = c("log(all_alc)"),
       model_order = c("Model 1", "Model 2"),
       dodge_size = 0.5) %>% 
    relabel_predictors(c("log(all_alc)" = "Gallons per capita")) +
    theme_classic(base_size = 14) +
    scale_color_discrete(name = "Outcome", labels = c(
                                                    "Alcoholism SVI",
                                                      "Alcohol SVI"
                                                      )) +
      # scale_x_continuous(labels = scales::percent) + 
      # labs(x = "Effect size (approx. % change)") + 
      theme(legend.position = "top") + 
      
  # guides(color=guide_legend(nrow=2, byrow=TRUE)) + 
      labs(x = "Coefficient")

dwplot(list(cig_3, cig_7),
       vline = geom_vline(
           xintercept = 0,
           colour = "grey60",
       ),
       vars_order = c("log(packs_per_capita)"),
       model_order = c("Model 1", "Model 2"),
       dodge_size = 0.5) %>% 
    relabel_predictors(c(
                         "log(packs_per_capita)" = "Imputed packs per capita")) +
    theme_classic(base_size = 14) +
    scale_color_discrete(name = "Outcome", labels = c("Smoking index SVI",
                                                      "Cigarettes SVI"
                                                      )) +
      # scale_x_continuous(labels = scales::percent) + 
      # labs(x = "Effect size (approx. % change)") + 
      theme(legend.position = "top") + 
      
  # guides(color=guide_legend(nrow=2, byrow=TRUE)) + 
      labs(x = "Coefficient")
```
