==> 8_weather.Rmd <==
---
title: "8_weather"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(tidyverse)
library(geosphere) #Needed for Haversine distance
library(readxl)
library(rgeos)
library(zoo)
library(sp)
library(janitor)
library(lubridate)
```

```{r}
min_weather_end_date <- as.Date("2020-04-01")

gsod_directory_file <- "data/weather/GSOD_directory.txt"
noaa_station_file <- "data/weather/NOAA_GSOD_stations_clean.txt"
```

```{r}
gsod_url <- c("https://www.ncei.noaa.gov/data/global-summary-of-the-day/access/2016",
               "https://www.ncei.noaa.gov/data/global-summary-of-the-day/access/2017",
               "https://www.ncei.noaa.gov/data/global-summary-of-the-day/access/2018",
               "https://www.ncei.noaa.gov/data/global-summary-of-the-day/access/2019",
               "https://www.ncei.noaa.gov/data/global-summary-of-the-day/access/2020")

gsod_filenames <- read.table(gsod_directory_file, header = FALSE,
                             stringsAsFactors = FALSE,
                             col.names = c("file","last_modified","time","size"))

#These stations have lots of missing data or other issues, so ignore them.
bad_stations <- c("72211800482", #Sarasota FL
                  "72334703809", #Dyer, TN
                  "99818099999", #Sanilac, MI
                  "99726099999", #Sanilac MI
                  "72522904858",
                  "72340703953",
                  "72028803711",
                  "74003024103", #Tooele, UT
                  "72575399999", #Tooele, UT
                  "91197721508", #Also in the mountains on Hawaii
                  "99999921514") #On top of a volcano at 11,000' in Hawaii

#Set up filenames for all stations
gsod_filenames <- gsod_filenames %>%
  mutate(station_id = sub(".csv", "", file)) %>%
  select(file, last_modified, station_id)

#Reading this fixed-width file is a mess
noaa_col_names <- c("USAF",
                    "WBAN",
                    "STATION_NAME",
                    "CTRY",
                    "ST",
                    "CALL",
                    "LAT",
                    "LON",
                    "ELEV_M",
                    "BEGIN",
                    "END")
#Get station locations
noaa_stations <- read_fwf(noaa_station_file, 
                              fwf_positions(c(1, 8, 14, 44, 49, 52, 58, 66, 75, 83, 92), 
                                            c(7,13, 43, 46, 51, 56, 65, 74, 82, 91, 99), 
                                            noaa_col_names),
                              skip = 1, col_types = "ccccccccccc")

#Must filter by END > a few days ago
#Also filter by BEGIN < Jan 1
#Finally remove the bad stations

#Join location to file names 
noaa_stations <- noaa_stations %>%
  unite(usaf_wban, USAF, WBAN, sep="") %>%
  mutate(LAT = as.numeric(sub("\\+","", LAT)),
         LON = as.numeric(sub("\\+","", LON)),
         ELEV_M = as.numeric(sub("\\+","", ELEV_M)),
         BEGIN = as.Date(BEGIN, format = "%Y%m%d"),
         END = as.Date(END, format = "%Y%m%d")) %>%
  inner_join(gsod_filenames,
             by = c("usaf_wban" = "station_id")) %>%
  filter(END >= min_weather_end_date) %>%
  filter(BEGIN <= as.Date("2020-01-01")) %>%
  filter(!usaf_wban %in% bad_stations)
```

```{r}
us_county_geometry <- read_csv('data/weather/us_county_geometry.csv')

county_centers <- data.frame(fips = character(nrow(us_county_geometry)), 
                             lat=numeric(nrow(us_county_geometry)), 
                             lon=numeric(nrow(us_county_geometry)),
                             stringsAsFactors = FALSE)

#Use polygons to get (approximate) county center locations
for (i in 1:nrow(us_county_geometry)){
  this_polygon <- readWKT(us_county_geometry$geometry[i])
  county_centers$fips[i] <- us_county_geometry$fips[i]
  
  #Note coordinates() gives (long, lat) for some reason
  county_centers$lat[i] <- coordinates(this_polygon)[2]
  county_centers$lon[i] <- coordinates(this_polygon)[1]
}

#Add NYC and KC
#Data source is Wolfram Alpha
county_centers <- county_centers %>%
  bind_rows(data.frame(fips = "NYC", lat = 40.66, lon = -73.94,
                       stringsAsFactors = FALSE),
            data.frame(fips = "KSC", lat = 39.13, lon = -94.55,
                       stringsAsFactors = FALSE)
            )

```

```{r}
#Weather station distance matrix
noaa_longlat <- cbind(noaa_stations$LON, noaa_stations$LAT)

#For each county...make distance matrix, find nearest

for (i in 1:nrow(us_county_geometry)) {
  #print(i) #for monitoring progress
  
  #For each county...
  this_county_loc <- cbind(county_centers[i,"lon"], 
                           county_centers[i,"lat"]) 
  
  #Get distances to all stations - in km
  distance_to_stations <- distHaversine(this_county_loc, noaa_longlat)
  
  #Note closest station
  closest_ind <- which.min(distance_to_stations)
  closest_station_id <- noaa_stations[closest_ind,"usaf_wban"]
  
  county_centers$closest_station_usaf_wban[i] <- as.character(closest_station_id)
  county_centers$km_to_closest_station[i] <- distance_to_stations[closest_ind]/1000
  
} 

county_centers_2 <- county_centers %>% 
  left_join(noaa_stations, 
            by = c("closest_station_usaf_wban" = "usaf_wban")) %>%
  select(-CTRY, -ST, -LAT, -LON, -BEGIN, -END)

```

```{r}
# This code is sort of fucked, use the next one
year <- 2016

for(j in 1:length(gsod_url)) {
  all_county_weather <- list()
  
  # Ad hoc written for connection closure
  
  if (year == 2016) {
    for (i in 2506:nrow(county_centers_2)){
      #print(i) #Tracks progress
      #For each county, get the daily weather data for 2020
      this_county_fips <- county_centers_2$fips[i]
      
      this_county_weather_file <- county_centers_2$file[i]
      this_county_weather_url <- paste(gsod_url[j], this_county_weather_file, sep="/")
      
      this_county_weather <- read_csv(this_county_weather_url,
                      col_types = cols(
                        STATION = col_character(),
                        DATE = col_date(format = ""),
                        LATITUDE = col_double(),
                        LONGITUDE = col_double(),
                        ELEVATION = col_double(),
                        NAME = col_character(),
                        TEMP = col_double(),
                        TEMP_ATTRIBUTES = col_double(),
                        DEWP = col_double(),
                        DEWP_ATTRIBUTES = col_double(),
                        SLP = col_double(),
                        SLP_ATTRIBUTES = col_double(),
                        STP = col_double(),
                        STP_ATTRIBUTES = col_double(),
                        VISIB = col_double(),
                        VISIB_ATTRIBUTES = col_double(),
                        WDSP = col_double(),
                        WDSP_ATTRIBUTES = col_double(),
                        MXSPD = col_double(),
                        GUST = col_double(),
                        MAX = col_double(),
                        MAX_ATTRIBUTES = col_character(),
                        MIN = col_double(),
                        MIN_ATTRIBUTES = col_character(),
                        PRCP = col_double(),
                        PRCP_ATTRIBUTES = col_character(),
                        SNDP = col_double(),
                        FRSHTT = col_character()
                        ))
                      
      #Only keeping relevant/useful data
      #Note: These are still in Freedom Units, not metric
      clean_weather_data <- this_county_weather %>%
        transmute(station_id = STATION,
                  station_name = NAME,
                  station_lat = LATITUDE,
                  station_lon = LONGITUDE,
                  date = DATE, 
                  mean_temp = TEMP,
                  min_temp = MIN,
                  max_temp = MAX,
                  dewpoint = DEWP,
                  sea_level_pressure = SLP,
                  station_pressure = STP,
                  visibility = VISIB,
                  wind_speed = WDSP,
                  max_wind_speed = MXSPD,
                  wind_gust = GUST,
                  precipitation = PRCP,
                  precip_flag = PRCP_ATTRIBUTES,
                  FRSHTT = FRSHTT) %>%
        separate(FRSHTT, into = c("fog", "rain", "snow", 
                          "hail", "thunder", "tornado"),
                 sep=c(1,2,3,4,5)) %>%
        mutate(county_fips = this_county_fips)
      
      #999.9 is the missing data indicator, also 9999.9 and 99.9
      clean_weather_data[clean_weather_data == 99.9] <- NA
      clean_weather_data[clean_weather_data == 99.99] <- NA
      clean_weather_data[clean_weather_data == 999.9] <- NA
      clean_weather_data[clean_weather_data == 9999.9] <- NA
      
      #Moving averages
      clean_weather_data_with_avgs <- clean_weather_data %>%
        mutate(mean_temp_3d_avg = rollmean(mean_temp, 3, na.pad = TRUE, align = "center")) %>%
        mutate(mean_temp_5d_avg = rollmean(mean_temp, 5, na.pad = TRUE, align = "center")) %>%
        mutate(mean_temp_10d_avg = rollmean(mean_temp, 10, na.pad = TRUE, align = "center")) %>%
        mutate(mean_temp_15d_avg = rollmean(mean_temp, 15, na.pad = TRUE, align = "center")) %>%
        mutate(max_temp_3d_avg = rollmean(max_temp, 3, na.pad = TRUE, align = "center")) %>%
        mutate(max_temp_5d_avg = rollmean(max_temp, 5, na.pad = TRUE, align = "center")) %>%
        mutate(max_temp_10d_avg = rollmean(max_temp, 10, na.pad = TRUE, align = "center")) %>%
        mutate(max_temp_15d_avg = rollmean(max_temp, 15, na.pad = TRUE, align = "center")) %>% 
        mutate(min_temp_3d_avg = rollmean(min_temp, 3, na.pad = TRUE, align = "center")) %>%
        mutate(min_temp_5d_avg = rollmean(min_temp, 5, na.pad = TRUE, align = "center")) %>%
        mutate(min_temp_10d_avg = rollmean(min_temp, 10, na.pad = TRUE, align = "center")) %>%
        mutate(min_temp_15d_avg = rollmean(min_temp, 15, na.pad = TRUE, align = "center")) %>%
        mutate(dewpoint_3d_avg = rollmean(dewpoint, 3, na.pad = TRUE, align = "center")) %>%
        mutate(dewpoint_5d_avg = rollmean(dewpoint, 5, na.pad = TRUE, align = "center")) %>%
        mutate(dewpoint_10d_avg = rollmean(dewpoint, 10, na.pad = TRUE, align = "center")) %>%
        mutate(dewpoint_15d_avg = rollmean(dewpoint, 15, na.pad = TRUE, align = "center"))
      
      #Store in list
      all_county_weather[[i]] <- clean_weather_data_with_avgs
      
      Sys.sleep(1)
    }
  
    
  }
  
  else {
    for (i in 1:nrow(county_centers_2)){
      #print(i) #Tracks progress
      #For each county, get the daily weather data for 2020
      this_county_fips <- county_centers_2$fips[i]
      
      this_county_weather_file <- county_centers_2$file[i]
      this_county_weather_url <- paste(gsod_url[j], this_county_weather_file, sep="/")
      
      this_county_weather <- read_csv(this_county_weather_url,
                      col_types = cols(
                        STATION = col_character(),
                        DATE = col_date(format = ""),
                        LATITUDE = col_double(),
                        LONGITUDE = col_double(),
                        ELEVATION = col_double(),
                        NAME = col_character(),
                        TEMP = col_double(),
                        TEMP_ATTRIBUTES = col_double(),
                        DEWP = col_double(),
                        DEWP_ATTRIBUTES = col_double(),
                        SLP = col_double(),
                        SLP_ATTRIBUTES = col_double(),
                        STP = col_double(),
                        STP_ATTRIBUTES = col_double(),
                        VISIB = col_double(),
                        VISIB_ATTRIBUTES = col_double(),
                        WDSP = col_double(),
                        WDSP_ATTRIBUTES = col_double(),
                        MXSPD = col_double(),
                        GUST = col_double(),
                        MAX = col_double(),
                        MAX_ATTRIBUTES = col_character(),
                        MIN = col_double(),
                        MIN_ATTRIBUTES = col_character(),
                        PRCP = col_double(),
                        PRCP_ATTRIBUTES = col_character(),
                        SNDP = col_double(),
                        FRSHTT = col_character()
                        ))
                      
      #Only keeping relevant/useful data
      #Note: These are still in Freedom Units, not metric
      clean_weather_data <- this_county_weather %>%
        transmute(station_id = STATION,
                  station_name = NAME,
                  station_lat = LATITUDE,
                  station_lon = LONGITUDE,
                  date = DATE, 
                  mean_temp = TEMP,
                  min_temp = MIN,
                  max_temp = MAX,
                  dewpoint = DEWP,
                  sea_level_pressure = SLP,
                  station_pressure = STP,
                  visibility = VISIB,
                  wind_speed = WDSP,
                  max_wind_speed = MXSPD,
                  wind_gust = GUST,
                  precipitation = PRCP,
                  precip_flag = PRCP_ATTRIBUTES,
                  FRSHTT = FRSHTT) %>%
        separate(FRSHTT, into = c("fog", "rain", "snow", 
                          "hail", "thunder", "tornado"),
                 sep=c(1,2,3,4,5)) %>%
        mutate(county_fips = this_county_fips)
      
      #999.9 is the missing data indicator, also 9999.9 and 99.9
      clean_weather_data[clean_weather_data == 99.9] <- NA
      clean_weather_data[clean_weather_data == 99.99] <- NA
      clean_weather_data[clean_weather_data == 999.9] <- NA
      clean_weather_data[clean_weather_data == 9999.9] <- NA
      
      #Moving averages
      clean_weather_data_with_avgs <- clean_weather_data %>%
        mutate(mean_temp_3d_avg = rollmean(mean_temp, 3, na.pad = TRUE, align = "center")) %>%
        mutate(mean_temp_5d_avg = rollmean(mean_temp, 5, na.pad = TRUE, align = "center")) %>%
        mutate(mean_temp_10d_avg = rollmean(mean_temp, 10, na.pad = TRUE, align = "center")) %>%
        mutate(mean_temp_15d_avg = rollmean(mean_temp, 15, na.pad = TRUE, align = "center")) %>%
        mutate(max_temp_3d_avg = rollmean(max_temp, 3, na.pad = TRUE, align = "center")) %>%
        mutate(max_temp_5d_avg = rollmean(max_temp, 5, na.pad = TRUE, align = "center")) %>%
        mutate(max_temp_10d_avg = rollmean(max_temp, 10, na.pad = TRUE, align = "center")) %>%
        mutate(max_temp_15d_avg = rollmean(max_temp, 15, na.pad = TRUE, align = "center")) %>% 
        mutate(min_temp_3d_avg = rollmean(min_temp, 3, na.pad = TRUE, align = "center")) %>%
        mutate(min_temp_5d_avg = rollmean(min_temp, 5, na.pad = TRUE, align = "center")) %>%
        mutate(min_temp_10d_avg = rollmean(min_temp, 10, na.pad = TRUE, align = "center")) %>%
        mutate(min_temp_15d_avg = rollmean(min_temp, 15, na.pad = TRUE, align = "center")) %>%
        mutate(dewpoint_3d_avg = rollmean(dewpoint, 3, na.pad = TRUE, align = "center")) %>%
        mutate(dewpoint_5d_avg = rollmean(dewpoint, 5, na.pad = TRUE, align = "center")) %>%
        mutate(dewpoint_10d_avg = rollmean(dewpoint, 10, na.pad = TRUE, align = "center")) %>%
        mutate(dewpoint_15d_avg = rollmean(dewpoint, 15, na.pad = TRUE, align = "center"))
      
      #Store in list
      all_county_weather[[i]] <- clean_weather_data_with_avgs
      
      Sys.sleep(1)
    }
    
  }

  #Probably a cleverer coder could vectorize or lapply this
  
  
  #Put it all together
  all_county_weather_df <- bind_rows(all_county_weather)
  saveRDS(all_county_weather_df, file = paste(county_weather, 
                                              as.character(year), 
                                              ".rds", sep = ""))
  
  year <- year + 1
  
}

all_county_weather <- bind_rows(county_weather2016,
                                county_weather2017,
                                county_weather2018,
                                county_weather2019,
                                county_weather2020)

```
```{r}

year <- 2016

# it's only length-1 b/c we alr did 2020

for(j in 1:(length(gsod_url)-1)) {
  all_county_weather <- list()

    for (i in 1:nrow(county_centers_2)){
      #print(i) #Tracks progress
      #For each county, get the daily weather data for 2020
      this_county_fips <- county_centers_2$fips[i]
      
      this_county_weather_file <- county_centers_2$file[i]
      this_county_weather_url <- paste(gsod_url[j], this_county_weather_file, sep="/")
      
      this_county_weather <- read_csv(this_county_weather_url,
                      col_types = cols(
                        STATION = col_character(),
                        DATE = col_date(format = ""),
                        LATITUDE = col_double(),
                        LONGITUDE = col_double(),
                        ELEVATION = col_double(),
                        NAME = col_character(),
                        TEMP = col_double(),
                        TEMP_ATTRIBUTES = col_double(),
                        DEWP = col_double(),
                        DEWP_ATTRIBUTES = col_double(),
                        SLP = col_double(),
                        SLP_ATTRIBUTES = col_double(),
                        STP = col_double(),
                        STP_ATTRIBUTES = col_double(),
                        VISIB = col_double(),
                        VISIB_ATTRIBUTES = col_double(),
                        WDSP = col_double(),
                        WDSP_ATTRIBUTES = col_double(),
                        MXSPD = col_double(),
                        GUST = col_double(),
                        MAX = col_double(),
                        MAX_ATTRIBUTES = col_character(),
                        MIN = col_double(),
                        MIN_ATTRIBUTES = col_character(),
                        PRCP = col_double(),
                        PRCP_ATTRIBUTES = col_character(),
                        SNDP = col_double(),
                        FRSHTT = col_character()
                        ))
                      
      #Only keeping relevant/useful data
      #Note: These are still in Freedom Units, not metric
      clean_weather_data <- this_county_weather %>%
        transmute(station_id = STATION,
                  station_name = NAME,
                  station_lat = LATITUDE,
                  station_lon = LONGITUDE,
                  date = DATE, 
                  mean_temp = TEMP,
                  min_temp = MIN,
                  max_temp = MAX,
                  dewpoint = DEWP,
                  sea_level_pressure = SLP,
                  station_pressure = STP,
                  visibility = VISIB,
                  wind_speed = WDSP,
                  max_wind_speed = MXSPD,
                  wind_gust = GUST,
                  precipitation = PRCP,
                  precip_flag = PRCP_ATTRIBUTES,
                  FRSHTT = FRSHTT) %>%
        separate(FRSHTT, into = c("fog", "rain", "snow", 
                          "hail", "thunder", "tornado"),
                 sep=c(1,2,3,4,5)) %>%
        mutate(county_fips = this_county_fips)
      
      #999.9 is the missing data indicator, also 9999.9 and 99.9
      clean_weather_data[clean_weather_data == 99.9] <- NA
      clean_weather_data[clean_weather_data == 99.99] <- NA
      clean_weather_data[clean_weather_data == 999.9] <- NA
      clean_weather_data[clean_weather_data == 9999.9] <- NA
      
      #Moving averages
      clean_weather_data_with_avgs <- clean_weather_data %>%
        mutate(mean_temp_3d_avg = rollmean(mean_temp, 3, na.pad = TRUE, align = "center")) %>%
        mutate(mean_temp_5d_avg = rollmean(mean_temp, 5, na.pad = TRUE, align = "center")) %>%
        mutate(mean_temp_10d_avg = rollmean(mean_temp, 10, na.pad = TRUE, align = "center")) %>%
        mutate(mean_temp_15d_avg = rollmean(mean_temp, 15, na.pad = TRUE, align = "center")) %>%
        mutate(max_temp_3d_avg = rollmean(max_temp, 3, na.pad = TRUE, align = "center")) %>%
        mutate(max_temp_5d_avg = rollmean(max_temp, 5, na.pad = TRUE, align = "center")) %>%
        mutate(max_temp_10d_avg = rollmean(max_temp, 10, na.pad = TRUE, align = "center")) %>%
        mutate(max_temp_15d_avg = rollmean(max_temp, 15, na.pad = TRUE, align = "center")) %>% 
        mutate(min_temp_3d_avg = rollmean(min_temp, 3, na.pad = TRUE, align = "center")) %>%
        mutate(min_temp_5d_avg = rollmean(min_temp, 5, na.pad = TRUE, align = "center")) %>%
        mutate(min_temp_10d_avg = rollmean(min_temp, 10, na.pad = TRUE, align = "center")) %>%
        mutate(min_temp_15d_avg = rollmean(min_temp, 15, na.pad = TRUE, align = "center")) %>%
        mutate(dewpoint_3d_avg = rollmean(dewpoint, 3, na.pad = TRUE, align = "center")) %>%
        mutate(dewpoint_5d_avg = rollmean(dewpoint, 5, na.pad = TRUE, align = "center")) %>%
        mutate(dewpoint_10d_avg = rollmean(dewpoint, 10, na.pad = TRUE, align = "center")) %>%
        mutate(dewpoint_15d_avg = rollmean(dewpoint, 15, na.pad = TRUE, align = "center"))
      
      #Store in list
      all_county_weather[[i]] <- clean_weather_data_with_avgs
      
      Sys.sleep(1)
    }
  
  

  #Probably a cleverer coder could vectorize or lapply this
  
  
  #Put it all together
  all_county_weather_df <- bind_rows(all_county_weather)
  saveRDS(all_county_weather_df, file = paste(county_weather, 
                                              as.character(year), 
                                              ".rds", sep = ""))
  
  year <- year + 1
  
}

all_county_weather <- bind_rows(county_weather2016,
                                county_weather2017,
                                county_weather2018,
                                county_weather2019,
                                county_weather2020)

```

```{r}
# Abandoning that - let's do precipitation and temperature instead

# precipitation
for (i in 1:5) {
  
  for (j in 1:12) {
    
    
  }
}

```

```{r}
# Monthly weather

# https://www.ncdc.noaa.gov/cag/county/mapping/110-pcp-202108-1.csv

precip_begin <- "https://www.ncdc.noaa.gov/cag/county/mapping/110-pcp-"

precip <- tibble()

for(i in 1:5) {
  year <- as.character(2015 + i)
  for(j in 1:12) {
    month_char <- ifelse(str_length(as.character(j)) == 1, str_pad(as.character(j), width = 2,
                                                              side = "left",
                                                              pad = "0"), as.character(j))
    precip_url <- paste(precip_begin, year, month_char, "-1.csv", sep = "")
    
    temp <- read_csv(precip_url, show_col_types = FALSE, skip = 3) %>%
      clean_names() %>%
      select(location_id, location, value) %>%
      mutate(value = ifelse(value == -99, NA, value)) %>%
      rename('precip' = 'value') %>%
      mutate(month = as.numeric(month_char),
             year = as.numeric(`year`))
    
    if (year == "2016" & j == 1) {
      precip <- temp
    }
    
    else {
      precip <- rbind(precip, temp)
    }
    
  }
}
```

```{r}
# temperature

# https://www.ncdc.noaa.gov/cag/county/mapping/110-tavg-202108-1.csv

temp_begin <- "https://www.ncdc.noaa.gov/cag/county/mapping/110-tavg-"

temperature <- tibble()

for(i in 1:5) {
  year <- as.character(2015 + i)
  for(j in 1:12) {
    month_char <- ifelse(str_length(as.character(j)) == 1, str_pad(as.character(j), width = 2,
                                                              side = "left",
                                                              pad = "0"), as.character(j))
    temp_url <- paste(temp_begin, year, month_char, "-1.csv", sep = "")
    
    temp <- read_csv(temp_url, show_col_types = FALSE, skip = 3) %>%
      clean_names() %>%
      select(location_id, location, value) %>%
      mutate(value = ifelse(value == -99, NA, value)) %>%
      rename('temp' = 'value') %>%
      mutate(month = as.numeric(month_char),
             year = as.numeric(`year`))
    
    if (year == "2016" & j == 1) {
      temperature <- temp
    }
    
    else {
      temperature <- rbind(temperature, temp)
    }
    
  }
}

```

```{r}


weather <- full_join(precip, temperature, by = c("location_id", "location",
                                                 "month", "year")) 


weather_2 <- weather %>% 
  mutate(county = str_sub(location_id, start = 4, end = 6),
         state = str_sub(location_id, start = 1, end = 2))

statefips <- read_excel('statefips.xlsx') %>% 
  clean_names() %>%
  select(postal_code, fips) %>%
  mutate(fips = as.character(fips)) %>%
  mutate(fips = str_pad(fips, width = 2, side = "left", pad = "0")) %>%
  rename('stfips' = 'fips')

weather_3 <- left_join(weather_2, statefips, by = c("state" = "postal_code")) %>%
  mutate(fips = str_c(stfips, county, sep = "")) %>%
  select(-location_id, -location, -county, -stfips, -state)

# converting to DMA

dma_codes <- read_csv('dma_codes.csv') %>%
  select(google_code, statefp, cntyfp) %>%
  mutate(statefp = str_pad(statefp, width = 2, side = "left", pad = "0"),
         cntyfp = str_pad(cntyfp, width = 3, side = "left", pad = 0)) %>%
  mutate(fips = str_c(statefp, cntyfp, sep = "")) %>%
  select(google_code, fips)

weather_4 <- left_join(weather_3, dma_codes, by = "fips") %>%
  drop_na(google_code) %>%
  group_by(google_code, year, month) %>%
  summarize(precip = mean(precip, na.rm = TRUE),
            temp = mean(temp, na.rm = TRUE)) %>%
  mutate(year = as.character(year))

```
