==> 9_verification.Rmd <==
---
title: "9_verification"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(tidyverse)
library(texreg)
library(lfe)
library(janitor)
library(lubridate)
library(openintro)
library(readxl)
library(ggpubr)
library(usmap)
library(RColorBrewer)
```

```{r}
# Read in the household pulse file

hps <- read_csv(
'data/verification/Indicators_of_Anxiety_or_Depression_Based_on_Reported_Frequency_of_Symptoms_During_Last_7_Days.csv', col_types = cols()) %>% 
  clean_names() %>% 
  select(-confidence_interval, -quartile_range, -low_ci, -high_ci, -phase, -time_period,
         -time_period_label) %>% 
  mutate(time_period_start_date = mdy(time_period_start_date),
         time_period_end_date = mdy(time_period_end_date)) %>%  
  rename('start_date' = 'time_period_start_date',
         'end_date' = 'time_period_end_date') %>% 
  pivot_wider(names_from = indicator, values_from = value) %>%
  clean_names() %>% 
  rename('hps_depression' = 'symptoms_of_depressive_disorder',
         'hps_anxiety' = 'symptoms_of_anxiety_disorder',
         'hps_either' = 'symptoms_of_anxiety_disorder_or_depressive_disorder')

hps_state <- hps %>%
  filter(state != "United States") %>%
  select(-subgroup, -group) %>%
  mutate(state_2 = state2abbr(state)) %>%
  select(-state)

hps_merged_norm <- expand_grid(hps_state, state_searches_norm) %>%
  mutate(state_3 = str_sub(state, start = 4, end = 5)) %>%
  filter(state_3 == state_2, 
         date + days(6) >= start_date,
         date <= end_date) %>%
  select(-state_3) %>%
  rename('state_code' = 'state') %>%
  rename('state' = 'state_2')

hps_merged <- expand_grid(hps_state, state_searches_v1) %>%
  mutate(state_3 = str_sub(state, start = 4, end = 5)) %>%
  filter(state_3 == state_2, 
         date + days(6) >= start_date,
         date <= end_date) %>%
  select(-state_3) %>%
  rename('state_code' = 'state') %>%
  rename('state' = 'state_2')

hps_merged_2 <- hps_merged %>% 
  group_by(start_date, end_date, state, hps_depression, hps_anxiety) %>% 
  summarize(across(anxiety:average_3, mean)) %>%
  select(-average, -average_3)

hps_merged_norm_v2 <- hps_merged_norm_v1 %>% 
  group_by(start_date, end_date, state, hps_depression, hps_anxiety) %>% 
  summarize(across(anxiety:average_3, mean)) %>%
  rename('anxiety_norm' = 'anxiety',
         'depression_norm' = 'depression')

hps_all_v1 <- left_join(hps_merged_norm_v2, hps_merged_2)


  
```

```{r}
depression_norm <- hps_all_v1 %>% 
  ggplot(mapping = aes(x = hps_depression, y = depression_norm)) + 
    geom_point() + 
    theme_classic(base_size = 14) + 
    labs(x = "% People Reporting Depression (HPS)",
         y = "Depression SVI (normalized)") + 
    geom_smooth(method = 'lm')

depression <- hps_all_v1 %>% 
  ggplot(mapping = aes(x = hps_depression, y = depression)) + 
    geom_point() + 
    theme_classic(base_size = 14) + 
    labs(x = "% People Reporting Depression (HPS)",
         y = "Depression SVI (un-normalized)") + 
    geom_smooth(method = 'lm')

depression_norm_log <- hps_all_v1 %>%
  ggplot(mapping = aes(x = log(hps_depression), y = log(depression_norm+1))) + 
    geom_point() + 
    theme_classic(base_size = 14) + 
    labs(x = "Log % People Reporting Depression (HPS)",
         y = "Log Depression SVI (normalized)") + 
    geom_smooth(method = 'lm')

depression_log <- hps_all_v1 %>% 
  ggplot(mapping = aes(x = log(hps_depression), y = log(depression+1))) + 
    geom_point() + 
    theme_classic(base_size = 14) + 
    labs(x = "Log % People Reporting Depression (HPS)",
         y = "Log Depression SVI (un-normalized)") + 
    geom_smooth(method = 'lm')

depression_total <- ggarrange(depression_norm, depression, depression_norm_log, depression_log, 
          labels = c("Normalized", "Un-Normalized", "Log Normalized", 
                     "Log Un-Normalized"), nrow = 1)

ggsave(depression_total, file = "depression_total.png", width = 10, height = 3,
       units = "in")

```

```{r}

test <- felm(depression_norm ~ hps_depression|state|0|state, data = hps_all_v1)

test2 <- felm(depression ~ hps_depression|state|0|state, data = hps_all_v1)

test3 <- felm(log(depression_norm+1) ~ log(hps_depression+1)|state|0|state, data = hps_all_v1)

test4 <- felm(log(depression + 1) ~ log(hps_depression + 1)|state|0|state, data = hps_all_v1)

htmlreg(l = list(test, test2, test3, test4), include.ci = FALSE, digits = 4, file = "depression_verif.doc")
```
```{r}
anxiety_norm <- hps_all_v1 %>% 
  filter(state == "NY") %>% 
  ggplot(mapping = aes(x = hps_anxiety, y = anxiety_norm)) + 
    geom_point() + 
    theme_classic(base_size = 14) + 
    labs(x = "% People Reporting Anxiety (HPS)",
         y = "Anxiety SVI (normalized)") + 
    geom_smooth(method = 'lm')

anxiety <- hps_all_v1 %>% 
  ggplot(mapping = aes(x = hps_anxiety, y = anxiety)) + 
    geom_point() + 
    theme_classic(base_size = 14) + 
    labs(x = "% People Reporting Anxiety (HPS)",
         y = "Anxiety SVI (Un-normalized)") + 
    geom_smooth(method = 'lm')

anxiety_norm_log <- hps_all_v1 %>% 
  ggplot(mapping = aes(x = log(hps_anxiety), y = log(anxiety_norm+1))) + 
    geom_point() + 
    theme_classic(base_size = 14) + 
    labs(x = "% People Reporting Anxiety (HPS)",
         y = "Anxiety SVI (normalized)") + 
    geom_smooth(method = 'lm')

anxiety_log <- hps_all_v1 %>% 
  ggplot(mapping = aes(x = log(hps_anxiety), y = log(anxiety+1))) + 
    geom_point() + 
    theme_classic(base_size = 14) + 
    labs(x = "% People Reporting Anxiety (HPS)",
         y = "Anxiety SVI (un-normalized)") + 
    geom_smooth(method = 'lm')

anxiety_total <- ggarrange(anxiety_norm, anxiety, anxiety_norm_log, anxiety_log, 
          labels = c("Normalized", "Un-Normalized", "Log Normalized", 
                     "Log Un-Normalized"), nrow = 1)

ggsave(anxiety_total, file = "anxiety_total_2.png", width = 10, height = 3,
       units = "in")

```

```{r}

test <- felm(depression_norm ~ hps_depression|0|0|state, data = hps_all_v3)

test2 <- felm(depression_norm ~ hps_depression|state|0|state, data = hps_all_v3)

test3 <- felm(log(depression_norm+1) ~ log(hps_depression)|0|0|state, data = hps_all_v3)

test4 <- felm(log(depression_norm+1) ~ log(hps_depression)|state|0|state, data = hps_all_v3)

htmlreg(l = list(test, test2, test3, test4), include.ci = FALSE, digits = 4, file = "depression_verif_part1.doc")

# part 2

test <- felm(depression ~ hps_depression|0|0|state, data = hps_all_v3)

test2 <- felm(depression ~ hps_depression|state|0|state, data = hps_all_v3)

test3 <- felm(log(depression+1) ~ log(hps_depression)|0|0|state, data = hps_all_v3)

test4 <- felm(log(depression+1) ~ log(hps_depression)|state|0|state, data = hps_all_v3)

htmlreg(l = list(test, test2, test3, test4), include.ci = FALSE, digits = 4, file = "depression_verif_part2.doc")

```

```{r}

```


```{r}
# alc 

alc <- read_excel('data/verification/state_alcsales_December2020.xlsx', sheet = "Data") %>%
  clean_names() %>% 
  select(-gallons, -ethanol, -population, -per_capita3yr, -pct_change) %>% 
  pivot_wider(names_from = 'beverage', values_from = 'per_capita') %>% 
  rename('spirits' = `1`,
         'wine' = `2`,
         'beer' = `3`,
         'all_alc' = `4`)

# read in alcohol norm

alc_2 <- left_join(alc, alcohol_norm, by = c("year", "month", "fips"))

```
```{r}
# LOOK AT THIS ... GRAPH

eq <- function(x,y) {
  m <- lm(y ~ x)
  as.character(
    as.expression(
      substitute(italic(y) == a + b %.% italic(x)*","~~italic(r)^2~"="~r2,
                list(a = format(coef(m)[1], digits = 4),
                b = format(coef(m)[2], digits = 4),
                r2 = format(summary(m)$r.squared, digits = 3)))
    )
  )
}

alcohol_norm_graph <- alc_2 %>%
  ggplot(mapping = aes(x = log(all_alc+1), y = log(alcohol_norm+1))) + 
    geom_point() +
    theme_classic(base_size = 14) + 
    labs(x = "Log Per-capita monthly alcohol consumption (gal)",
         y = "Log Adjusted Alcohol SVI (normalized vs. anchor)") + 
    geom_smooth(method='lm')

alcohol_graph <- alc_2 %>%
   ggplot(mapping = aes(x = log(all_alc + 1), y = log(alcohol + 1))) + 
    geom_point() +
    theme_classic(base_size = 14) + 
    labs(x = "Log Per-capita monthly alcohol consumption (gal)",
         y = "Log Alcohol SVI (un-normalized)") + 
    geom_smooth(method='lm')
```

```{r}
# Test some regressions

test <- felm(alcohol_norm ~ all_alc|0|0|state, data = alcohol_final_v2)

test2 <- felm(alcohol_norm ~ all_alc|state|0|state, data = alcohol_final_v2)

test3 <- felm(log(alcohol_norm+1) ~ log(all_alc)|0|0|state, data = alcohol_final_v2)

test4 <- felm(log(alcohol_norm+1) ~ log(all_alc)|state|0|state, data = alcohol_final_v2)


# Part 2

test <- felm(alcohol ~ all_alc|0|0|state, data = alcohol_final_v2)

test2 <- felm(alcohol ~ all_alc|state|0|state, data = alcohol_final_v2)

test3 <- felm(log(alcohol+1) ~ log(all_alc)|0|0|state, data = alcohol_final_v2)

test4 <- felm(log(alcohol+1) ~ log(all_alc)|state|0|state, data = alcohol_final_v2)
```

```{r}
# join some populations in so we can do this in a population weighted way

state_dem <- read_csv('data/state_data/R12917460_SL040.csv', col_types = cols()) %>% 
  slice(2:n()) %>% 
  clean_names() %>% 
  select(state_fips_code, total_population_56)

alcohol_final_v2 <- left_join(alcohol_final, state_dem, by = c("fips" = "state_fips_code")) %>%
  mutate(total_population_56 = as.numeric(total_population_56)) 
```


```{r}
# Make a graph!

alcohol_weighted <- alcohol_final_v2 %>%
  group_by(year, month) %>% 
  summarize(all_alc = weighted.mean(all_alc, total_population_56),
            alcohol_norm = weighted.mean(alcohol_norm, total_population_56),
            alcohol = weighted.mean(alcohol, total_population_56)) %>% 
  mutate(date = mdy(paste(month, "01", year, sep = "-"))) %>% 
  ungroup() %>% 
  ggplot(aes(x = date)) + 
    geom_line(aes(y = all_alc, color = "Alcohol Consumption")) + 
    geom_line(aes(y = alcohol / max(alcohol) * 0.3, 
                  color = "Normalized GTrends (rescaled)")) + 
    geom_line(aes(y = alcohol_norm / max(alcohol_norm) * 0.3, 
                  color = "Un-normalized GTrends (rescaled)")) + 
    theme_classic(base_size = 14) + 
    labs(x = "", y = "Gallons per Capita", title = "Alcohol Sales vs. Google Trends") + 
    theme(legend.position = "top") + 
    scale_color_manual(name = c("", "", ""),
                       values = c("Alcohol Consumption" = "blue",
                                  "Normalized GTrends (rescaled)" = "red",
                                  "Un-normalized GTrends (rescaled)" = "green"))
  
```

```{r}
# Tobacco?

tax_rev <- read_excel('data/verification/selected-monthly-sales-tax-collections-data.xlsx') %>%
  clean_names() %>%
  filter(tax_type == "Tobacco") %>% 
  select(-collection_month, -tax_type) %>%
  rename('state' = 'name',
         'month' = 'numeric_month') %>% 
  mutate(state = state2abbr(state)) %>%
  mutate(date = as_date(paste(year, month, '01', sep = "-")))

cig_tax <- read_csv('data/verification/Table_6694.csv') %>% 
  clean_names() %>% 
  filter(year %in% c(2019, 2020, 2021)) %>% 
  select(-data_source, -topic_description, -measure_description,
         -provision_group_description, -provision, -citation, -enacted_date,
         -effective_date, -comments) %>%
  mutate(state_2 = state2abbr(location_description)) %>% 
  select(-location_description) %>%
  mutate(start_date = 
           ifelse(quarter == 1, as_date(paste(year, "01-01", sep = "-")),
                  ifelse(quarter == 2, as_date(paste(year, "04-01", sep = "-")),
                         ifelse(quarter == 3, as_date(paste(year, "07-01")),
                                as_date(paste(year, "10-01", sep = "-"))))),
         end_date = 
           ifelse(quarter == 1, as_date(paste(year, "03-31", sep = "-")),
                           ifelse(quarter == 2, as_date(paste(year, "06-30", sep = "-")),
                                  ifelse(quarter == 3, as_date(paste(year, "09-30", sep = "-")),
                                         as_date(paste(year, "12-31", sep = "-")))))) %>% 
  mutate(start_date = as_date(start_date),
         end_date = as_date(end_date)) %>% 
  select(-quarter, -year)

cigs_combined <- expand_grid(tax_rev, cig_tax) %>% 
  filter(state == state_2,
         date >= start_date,
         date <= end_date) %>%
  select(-state_2, -start_date, -end_date) %>% 
  mutate(amount = as.numeric(amount)) %>% 
  mutate(packs_mil = amount/(value * 1000000))

# combine in the search data

smoking_state <- readRDS('smoking_state.rds')
  
  
```

```{r}
# Graphing HPS data vs. time-series search data

depression_graph <- hps_all_v3 %>% 
  group_by(state) %>% 
  arrange(start_date) %>% 
  ungroup() %>% 
  group_by(start_date) %>% 
  summarize(across(c(hps_depression:depression_norm, anxiety:depression), 
                   ~ weighted.mean(., w = pop))) %>% 
  ggplot(mapping = aes(x = start_date)) + 
    geom_line(aes(y = hps_depression, color = "Household Pulse")) + 
    geom_line(aes(y = depression_norm, color = "Normalized GTrends")) + 
    geom_line(aes(y = depression/2.5, color = "Un-normalized GTrends")) + 
    theme_classic(base_size = 14) + 
    theme(legend.position = "top") +
    labs(x = "", y = "HPS % reporting symptoms & Google SVI",
         title = "Depression GTrends vs. HPS") + 
    scale_color_manual(name = c("", "", ""),
                       values = c("Household Pulse" = "blue",
                                  "Normalized GTrends" = "red",
                                  "Un-normalized GTrends" = "green")) + 
    scale_x_date(date_breaks = "3 months", date_labels = "%m/%y")

anxiety_graph <- hps_all_v3 %>% 
  group_by(state) %>% 
  arrange(start_date) %>% 
  ungroup() %>% 
  group_by(start_date) %>% 
  summarize(across(c(hps_depression:depression_norm, anxiety:depression), 
                   ~ weighted.mean(., w = pop))) %>% 
  ggplot(mapping = aes(x = start_date)) + 
    geom_line(aes(y = hps_anxiety, color = "Household Pulse")) + 
    geom_line(aes(y = anxiety_norm*1.1, color = "Normalized GTrends")) + 
    geom_line(aes(y = anxiety/2.05, color = "Un-normalized GTrends")) + 
    theme_classic(base_size = 14) + 
    theme(legend.position = "top") +
    labs(x = "", y = "HPS % reporting symptoms & Google SVI",
         title = "Anxiety GTrends vs. HPS") + 
    scale_color_manual(name = c("", "", ""),
                       values = c("Household Pulse" = "blue",
                                  "Normalized GTrends" = "red",
                                  "Un-normalized GTrends" = "green")) + 
    scale_x_date(date_breaks = "3 months", date_labels = "%m/%y")
# lagged

hps_all_v1 %>% 
  group_by(state) %>% 
  arrange(start_date) %>% 
  mutate(across(c(anxiety_norm:depression_norm, anxiety:depression), ~ lag(., k = 2))) %>% 
  ungroup() %>% 
  group_by(start_date) %>%
  summarize(across(c(hps_depression:depression_norm, anxiety:depression), 
                   mean)) %>% 
  ggplot(mapping = aes(x = start_date)) + 
    geom_line(aes(y = hps_anxiety, color = "Household Pulse")) + 
    geom_line(aes(y = anxiety_norm, color = "Normalized GTrends")) + 
    geom_line(aes(y = anxiety/2.4, color = "Un-normalized GTrends")) + 
    theme_classic(base_size = 14) + 
    theme(legend.position = "top") + 
    labs(x = "", y = "HPS % reporting symptoms & Google SVI",
         title = "Anxiety") + 
    scale_color_manual(name = c("", "", ""),
                       values = c("Household Pulse" = "blue",
                                  "Normalized GTrends" = "red",
                                  "Un-normalized GTrends" = "green")) + 
    scale_x_date(date_breaks = "3 months", date_labels = "%m/%y")

alcohol_final %>% 
  group_by(year, month) %>% 
  summarize(all_alc = mean(all_alc),
            alcohol = mean(alcohol),
            alcohol_norm = mean(alcohol_norm)) %>% 
  mutate(date = mdy(paste(month, "01", year, sep = "-"))) %>% 
  ggplot(aes(x = date)) + 
    geom_line(aes(y = all_alc, color = "Alcohol Consumption")) + 
    geom_line(aes(y = alcohol / max(alcohol) * 0.35, 
                  color = "Normalized GTrends (rescaled)")) + 
    geom_line(aes(y = alcohol_norm / max(alcohol_norm) * 0.35, 
                  color = "Un-normalized GTrends (rescaled)")) + 
    theme_classic(base_size = 14) + 
    labs(x = "", y = "Gallons per Capita", title = "Alcohol") + 
    theme(legend.position = "top") + 
  scale_color_manual(name = c("", "", ""),
                       values = c("Alcohol Consumption" = "blue",
                                  "Normalized GTrends (rescaled)" = "red",
                                  "Un-normalized GTrends (rescaled)" = "green"))

# New alcohol graphs

hps_all_v19 %>%
  group_by(year, month) 
    

```
```{r}
# Smoking -- read in the smoking final rds

smoking %>% 
  ggplot(mapping = aes(x = log(packs_per_capita+1), y = log(smoking_norm+1))) + 
    geom_point() + 
    geom_smooth(method = "lm") + 
    theme_classic(base_size = 14) + 
    labs(x = "Imputed Packs per Capita", y = "Google Trends SVI (normalized)")

# Use this graph

smoking_graph<- smoking_final %>% 
  group_by(year, month) %>% 
  summarize(packs_per_capita = weighted.mean(packs_per_capita, total_pop),
            smoking = weighted.mean(smoking, total_pop),
            smoking_norm = weighted.mean(smoking_norm, total_pop)) %>% 
  mutate(date = mdy(paste(month, "01", year, sep = "-"))) %>% 
  ggplot(mapping = aes(x = date)) + 
    geom_line(aes(y = packs_per_capita, color = "Packs per Capita")) + 
    geom_line(aes(y = smoking/14, color = "Un-normalized GTrends (rescaled)")) + 
    geom_line(aes(y = smoking_norm/14, color = "Normalized GTrends (rescaled)")) + 
    theme_classic(base_size = 14) + 
    labs(x = "", y = "Packs per Capita", title = "Imputed Cigarette Sales vs. Google Trends") + 
    theme(legend.position = "top") + 
    scale_color_manual(name = c("", "", ""),
                       values = c("Packs per Capita" = "blue",
                                  "Normalized GTrends (rescaled)" = "red",
                                  "Un-normalized GTrends (rescaled)" = "green"))


ggplot(aes(x = date)) + 
    geom_line(aes(y = all_alc, color = "Alcohol Consumption")) + 
    geom_line(aes(y = alcohol / max(alcohol) * 0.35, 
                  color = "Normalized GTrends (rescaled)")) + 
    geom_line(aes(y = alcohol_norm / max(alcohol_norm) * 0.35, 
                  color = "Un-normalized GTrends (rescaled)")) + 
    theme_classic(base_size = 14) + 
    labs(x = "", y = "Gallons per Capita", title = "Alcohol") + 
    theme(legend.position = "top") + 
  scale_color_manual(name = c("", "", ""),
                       values = c("Alcohol Consumption" = "blue",
                                  "Normalized GTrends (rescaled)" = "red",
                                  "Un-normalized GTrends (rescaled)" = "green"))

# smoking_norm <- felm(smoking_norm ~ packs_per_capita|state|0|state, data = smoking)
# 
# smoking_log_norm <- felm(log(smoking_norm+1) ~ log(packs_per_capita+1)|state|0|state, data = smoking)
# 
# smoking_un <- felm(smoking ~ packs_per_capita|state|0|state, data = smoking)
# 
# smoking_log <- felm(log(smoking+1) ~ log(packs_per_capita+1)|state|0|state, data = smoking)
# 
# smoking_norm_nostate <- felm(smoking_norm ~ packs_per_capita|0|0|state, data = smoking)
# 
# smoking_log_norm_nostate <- felm(log(smoking_norm+1) ~ 
#                                    log(packs_per_capita+1)|0|0|state, data = smoking)
# 
# 
# 
# htmlreg(l = list(smoking_norm, smoking_un, smoking_log_norm, smoking_log,
#                  smoking_norm_nostate, smoking_log_norm_nostate), 
#         include.ci = FALSE, digits = 4, file = "smoking_verif.doc")

smoking_1 <- felm(smoking_norm ~ packs_per_capita|0|0|state, data = smoking_final)

smoking_2 <- felm(smoking_norm ~ packs_per_capita|state|0|state, data = smoking_final)

smoking_3 <- felm(log(smoking_norm+1) ~ log(packs_per_capita)|0|0|state, data = smoking_final)

smoking_4 <- felm(log(smoking_norm+1) ~ log(packs_per_capita)|state|0|state, data = smoking_final)

# part 2

smoking_1 <- felm(smoking ~ packs_per_capita|0|0|state, data = smoking_final)

smoking_2 <- felm(smoking ~ packs_per_capita|state|0|state, data = smoking_final)

smoking_3 <- felm(log(smoking+1) ~ log(packs_per_capita)|0|0|state, data = smoking_final)

smoking_4 <- felm(log(smoking+1) ~ log(packs_per_capita)|state|0|state, data = smoking_final)

```
```{r}
# Let's regress depression on anxiety

hps_1 <- felm(hps_depression ~ hps_anxiety|0|0|state, data = hps_all_v1)

hps_2 <- felm(depression_norm ~ anxiety_norm|0|0|state, data = hps_all_v1)

hps_3 <- felm(depression ~ anxiety|0|0|state, data = hps_all_v1)
```

```{r}

# Graphing?

hps_1 <- felm(log(depression_norm+1) ~ log(hps_depression):state|0|0|state,
              data = hps_all_v1)

hps_2 <- felm(log(anxiety_norm+1) ~ log(hps_anxiety):state|0|0|state,
              data = hps_all_v1)

depression_coef <- as_tibble(hps_1$coefficients, rownames = "state") %>% 
  filter(state != "(Intercept)") %>% 
  mutate(state = str_sub(state, start = 26))

anxiety_coef <-  as_tibble(hps_2$coefficients, rownames = "state") %>% 
  filter(state != "(Intercept)") %>% 
  mutate(state = str_sub(state, start = 23))

plot_usmap(data = depression_coef, values = "log(depression_norm + 1)") + 
  scale_fill_continuous(name = "Coefficient", type = "viridis") + 
  labs(title = "HPS vs. Google Trends Depression Searches")

plot_usmap(data = anxiety_coef, values = "log(anxiety_norm + 1)") + 
  scale_fill_continuous(name = "Coefficient", type = "viridis") + 
  labs(title = "HPS vs. Google Trends Anxiety Searches")

 
```

```{r}
# Check out therapy vs. therapy searches - hps all v4

therapy_graph <- hps_all_v19 %>% 
  group_by(start_date) %>% 
  summarize(across(medication:therapy_norm, ~ weighted.mean(., pop, na.rm = TRUE))) %>% 
  ggplot(aes(x = start_date)) + 
    geom_line(aes(y = counseling, color = "HPS % receiving therapy")) + 
    geom_line(aes(y = no_counseling, color = "HPS % needed, but did not get therapy")) + 
    geom_line(aes(y = therapy_norm/9.5, color = "Normalized GTrends (rescaled)")) + 
    theme_classic(base_size = 14) +
    theme(legend.position = "top") + 
    labs(x = "", y = "% of Respondents",
          title = "Therapy GTrends vs. HPS") + 
      scale_color_manual(name = c("", "", ""),
                       values = c("HPS % receiving therapy" = "blue",
                                  "HPS % needed, but did not get therapy" = "red",
                                  "Normalized GTrends (rescaled)" = "green")) +
    scale_x_date(date_breaks = "3 months", date_labels = "%m/%y")
  

therapy_1 <- felm(therapy_norm ~ counseling|0|0|state, data = hps_all_v4)

therapy_2 <- felm(therapy_norm ~ counseling|state|0|state, data = hps_all_v4)

therapy_3 <- felm(log(therapy_norm+1) ~ log(counseling)|0|0|state, data = hps_all_v4)

therapy_4 <- felm(log(therapy_norm+1) ~ log(counseling)|state|0|state, data = hps_all_v4)

htmlreg(l = list(therapy_1, therapy_2, therapy_3, therapy_4), include.ci = FALSE, digits = 4, file = "therapy_verif_part1.doc")

therapy_1 <- felm(therapy ~ counseling|0|0|state, data = hps_all_v4)

therapy_2 <- felm(therapy ~ counseling|state|0|state, data = hps_all_v4)

therapy_3 <- felm(log(therapy+1) ~ log(counseling)|0|0|state, data = hps_all_v4)

therapy_4 <- felm(log(therapy+1) ~ log(counseling)|state|0|state, data = hps_all_v4)


therapy_1 <- felm(log(therapy_norm+1) ~ log(counseling)|0|0|state, data = hps_all_v4)

therapy_2 <- felm(I(log(lag(therapy_norm, n = 1)+1)) ~ log(counseling)|0|0|state, data = hps_all_v4)

therapy_3 <- felm(I(log(lag(therapy_norm, n = 2)+1)) ~ log(counseling)|0|0|state, data = hps_all_v4)

therapy_4 <- felm(I(log(lag(therapy_norm, n = 3)+1)) ~ log(counseling)|0|0|state, data = hps_all_v4)

therapy_5 <- felm(I(log(lag(therapy_norm, n = 4)+1)) ~ log(counseling)|0|0|state, data = hps_all_v4)

htmlreg(l = list(therapy_1, therapy_2, therapy_3, therapy_4, therapy_5), include.ci = FALSE, digits = 4, file = "therapy_verif_part3.doc")

therapy_1 <- felm(log(therapy_norm+1) ~ log(counseling)|state|0|state, data = hps_all_v4)

therapy_2 <- felm(I(log(lag(therapy_norm, n = 1)+1)) ~ log(counseling)|state|0|state, data = hps_all_v4)

therapy_3 <- felm(I(log(lag(therapy_norm, n = 2)+1)) ~ log(counseling)|state|0|state, data = hps_all_v4)

therapy_4 <- felm(I(log(lag(therapy_norm, n = 3)+1)) ~ log(counseling)|state|0|state, data = hps_all_v4)

therapy_5 <- felm(I(log(lag(therapy_norm, n = 4)+1)) ~ log(counseling)|state|0|state, data = hps_all_v4)

htmlreg(l = list(therapy_1, therapy_2, therapy_3, therapy_4, therapy_5), include.ci = FALSE, digits = 4, file = "therapy_verif_part4.doc")


```

```{r}
# Some more exploratory stuff for HPS and depression

# Graph all depression terms against depression HPS? 

# HERE WE ARE

test <- hps_all_v10 %>% 
  group_by(start_date) %>% 
  summarize(across(c(hps_depression, depression_norm, crying_norm:sad_norm),
                   ~ weighted.mean(., pop))) %>% 
  rowwise() %>% 
  mutate(google_average = mean(crying_norm:sad_norm)) %>% 
  ungroup() %>% 
  ggplot(aes(x = start_date)) + 
    geom_line(aes(y = hps_depression, color = "HPS")) + 
    geom_line(aes(y = google_average/1.7, color = "Avg. related SVI")) + 
    geom_line(aes(y = depression_norm/1.3, color = "Depression SVI")) + 
    labs(x = "", y = "% Reporting Depression") + 
    scale_color_manual(name = c("", "", ""),
                       values = c("HPS" = "blue",
                                  "Avg. related SVI" = "red",
                                  "Depression SVI" = "green")) + 
    scale_x_date(date_breaks = "3 months", date_labels = "%m/%y") + 
    theme_classic(base_size = 14) + 
    theme(legend.position = "top")

hps_all_v10 %>% 
  group_by(start_date) %>% 
  summarize(across(c(hps_anxiety, anxiety_norm, stress_norm:worry_norm),
                   ~ weighted.mean(., pop))) %>% 
  group_by(start_date) %>% 
  mutate(google_average = mean(c(stress_norm:worry_norm))) %>% 
  ungroup() %>% 
  ggplot(aes(x = start_date)) + 
    geom_line(aes(y = hps_anxiety, color = "HPS")) + 
    geom_line(aes(y = google_average/1.2, color = "Avg. related SVI")) + 
    geom_line(aes(y = anxiety_norm/1.4, color = "Anxiety SVI")) + 
    labs(x = "", y = "% Reporting Anxiety") + 
    scale_color_manual(name = c("", "", ""),
                       values = c("HPS" = "blue",
                                  "Avg. related SVI" = "red",
                                  "Anxiety SVI" = "green")) + 
    scale_x_date(date_breaks = "3 months", date_labels = "%m/%y") + 
    theme_classic(base_size = 14) +
    theme(legend.position = "top")
  
# ggsave(width = 5.85, height = 4.8, units = "in", file = "depression.jpg")


```

```{r}
# Let me see some state graphs

hps_all_v19 %>%
  filter(state == "NY") %>% 
  group_by(start_date) %>% 
  summarize(across(c(hps_depression, depression_norm, crying_norm:sad_norm),
                   ~ weighted.mean(., pop))) %>% 
  rowwise() %>% 
  mutate(google_average = mean(crying_norm:sad_norm)) %>% 
  ungroup() %>% 
  ggplot(aes(x = start_date)) + 
    geom_line(aes(y = hps_depression, color = "Household Pulse")) + 
    geom_line(aes(y = google_average/1.7, color = "Avg. related SVIs")) + 
    geom_line(aes(y = depression_norm/1.3, color = "Depression SVI")) + 
    labs(x = "", y = "% Reporting Depression") + 
    scale_color_manual(name = c("", "", ""),
                       values = c("Household Pulse" = "blue",
                                  "Avg. related SVIs" = "red",
                                  "Depression SVI" = "green")) + 
    theme_classic(base_size = 14) + 
    theme(legend.position = "top")
```

```{r}
# Let's try this more rigorously

hps_all_v7 <- hps_all_v6 %>% 
  ungroup() %>% 
  mutate(anxiety_norm = anxiety_norm * 100 / max(anxiety_norm),
         depression_norm = depression_norm * 100 / max(depression_norm))

```

```{r}
# Graph again with demographic groups

hps_all_v8 %>%
    group_by(start_date) %>% 
  summarize(across(c(hps_depression, depression_norm, depression, depression_less25k, 
                     depression_75to100k, depression_black,
                     depression_18to29),
                   ~ weighted.mean(., pop))) %>% 
  ggplot(aes(x = start_date)) + 
    geom_line(aes(y = hps_depression, color = "HPS")) + 
    geom_line(aes(y = depression_less25k*100, color = "HPS (less than $25k income)")) + 
    geom_line(aes(y = depression_75to100k*100, color = "HPS ($75-100k income)")) +
    geom_line(aes(y = depression_black*100, color = "HPS (black)")) +
    geom_line(aes(y = depression_18to29*100, color = "HPS (18 to 29)")) +
    theme_classic(base_size = 14) + 
    labs(x = "", y = "% reporting depression symptoms") +
    theme(legend.position = "top")


# geom_line(aes(y = anxiety_norm/1.5, color = "Anxiety SVI")) +
#     geom_line(aes(y = anxiety/2, color = "Anxiety SVI (re-scaled, un-normalized)"))
  
```

```{r}
# New regressions

hps_temp <- hps_all_v8 %>% 
  rowwise() %>% 
  mutate(depression_avg_norm = mean(crying_norm:sad_norm),
         anxiety_avg_norm = mean(stress_norm:worry_norm)) %>% 
  group_by(state) %>% 
  arrange(start_date) %>% 
  mutate(hps_anxiety_1 = lag(hps_anxiety),
         hps_anxiety_2 = lag(hps_anxiety, n = 2),
         hps_anxiety_4 = lag(hps_anxiety, n = 4),
         hps_anxiety_8 = lag(hps_anxiety, n = 8),
         hps_depression_1 = lag(hps_depression),
         hps_depression_2 = lag(hps_depression, n = 2),
         hps_depression_4 = lag(hps_depression, n = 4),
        hps_depression_8 = lag(hps_depression, n = 8))

anxiety_1 <- felm(log(hps_anxiety_1) ~ log(anxiety_avg_norm)|0|0|state, data = hps_temp)

anxiety_2 <- felm(log(hps_anxiety_2) ~ log(anxiety_avg_norm)|0|0|state, data = hps_temp)

anxiety_3 <- felm(log(hps_anxiety_4) ~ log(anxiety_avg_norm)|0|0|state, data = hps_temp)

anxiety_4 <- felm(log(hps_anxiety_8) ~ log(anxiety_avg_norm)|0|0|state, data = hps_temp)

htmlreg(l = list(anxiety_1, anxiety_2, anxiety_3, anxiety_4), digits = 4, include.ci = FALSE,
        file = "anxiety_lag.doc")

depression_1 <- felm(log(hps_depression_1) ~ log(depression_avg_norm)|state|0|state, data = hps_temp)

depression_2 <- felm(log(hps_depression_2) ~ log(depression_avg_norm)|state|0|state, data = hps_temp)

depression_3 <- felm(log(hps_depression_4) ~ log(depression_avg_norm)|state|0|state, data = hps_temp)

depression_4 <- felm(log(hps_depression_8) ~ log(depression_avg_norm)|state|0|state, data = hps_temp)

htmlreg(l = list(depression_1, depression_2, depression_3, depression_4), 
        digits = 4, include.ci = FALSE,
        file = "depression_lag.doc")




```

```{r}
# New verification regressions - hopefully final
#depression_1 <- felm(log(hps_depression) ~ state|0|0|state, data = hps_all_v19)

depression_1 <- felm(log(depression_norm) ~ log(hps_depression)|0|0|state, data = hps_all_v19)

depression_2 <- felm(I(log(0.353 * crying_norm + 
                                                   0.163 * depression_norm + 
                                                   0.529 * lonely_norm + 
                                                   0.586 * sad_norm + 
                                                   0.475 * tired_norm))
                     ~  log(hps_depression)|0|0|state, 
                     data = hps_all_v19)

depression_3 <- felm(log(depression_norm) ~ log(hps_depression)|state|0|state, data = hps_all_v19)

depression_4 <- felm(I(log(0.353 * crying_norm + 
                                                   0.163 * depression_norm + 
                                                   0.529 * lonely_norm + 
                                                   0.586 * sad_norm + 
                                                   0.475 * tired_norm))
                     ~ log(hps_depression)|state|0|state, 
                     data = hps_all_v19)

htmlreg(l = list(depression_1, depression_2, depression_3, depression_4), include.ci = FALSE,
        digits = 4, file = "depression.doc")

```

```{r}
# With population weights

depression_1 <- felm(log(depression_norm) ~ log(hps_depression)|0|0|state, data = hps_all_v19, weights = hps_all_v19$pop)

depression_2 <- felm(log(depression_norm) ~ log(hps_depression)|state|0|state, data = hps_all_v19, weights = hps_all_v19$pop)

depression_3 <- felm(I(log(0.353 * crying_norm + 
                                                   0.163 * depression_norm + 
                                                   0.529 * lonely_norm + 
                                                   0.586 * sad_norm + 
                                                   0.475 * tired_norm))
                     ~  log(hps_depression)|0|0|state, 
                     data = hps_all_v19, weights = hps_all_v19$pop)



depression_4 <- felm(I(log(0.353 * crying_norm + 
                                                   0.163 * depression_norm + 
                                                   0.529 * lonely_norm + 
                                                   0.586 * sad_norm + 
                                                   0.475 * tired_norm))
                     ~ log(hps_depression)|state|0|state, 
                     data = hps_all_v19, weights = hps_all_v19$pop)

htmlreg(l = list(depression_1, depression_2, depression_3, depression_4), include.ci = FALSE,
        digits = 4, file = "depression.doc")
```



```{r}
anxiety_1 <- felm(log(anxiety_norm) ~ log(hps_anxiety)|0|0|state, data = hps_all_v19)

anxiety_2 <- felm(I(log(0.509 * anxiety + 
                                                   0.528 * nervous + 
                                                   0.504 * stress + 
                                                   0.399 * worry + 
                                                   0.223 * panic))
                     ~  log(hps_anxiety)|0|0|state, 
                     data = hps_all_v19)

anxiety_3 <- felm(log(anxiety_norm) ~ log(hps_anxiety)|state|0|state, data = hps_all_v19)

anxiety_4 <- felm(I(log(0.509 * anxiety + 
                                                   0.528 * nervous + 
                                                   0.504 * stress + 
                                                   0.399 * worry + 
                                                   0.223 * panic))
                     ~ log(hps_anxiety)|state|0|state, 
                     data = hps_all_v19)

htmlreg(l = list(anxiety_1, anxiety_2, anxiety_3, anxiety_4), include.ci = FALSE,
        digits = 4, file = "anxiety.doc")
```

```{r}
# weighted

anxiety_1 <- felm(log(anxiety_norm) ~ log(hps_anxiety)|0|0|state, data = hps_all_v19,
                  weights = hps_all_v19$pop)


anxiety_2 <- felm(log(anxiety_norm) ~ log(hps_anxiety)|state|0|state, data = hps_all_v19, weights = hps_all_v19$pop)

anxiety_3 <- felm(I(log(0.509 * anxiety + 
                                                  0.528 * nervous + 
                                                   0.504 * stress + 
                                                   0.399 * worry + 
                                                   0.223 * panic))
                     ~  log(hps_anxiety)|0|0|state, 
                     data = hps_all_v19, weights = hps_all_v19$pop)


anxiety_4 <- felm(I(log(0.509 * anxiety + 
                                                   0.528 * nervous + 
                                                   0.504 * stress + 
                                                   0.399 * worry + 
                                                   0.223 * panic))
                     ~ log(hps_anxiety)|state|0|state, 
                     data = hps_all_v19, weights = hps_all_v19$pop)

htmlreg(l = list(anxiety_1, anxiety_2, anxiety_3, anxiety_4), include.ci = FALSE,
        digits = 4, file = "anxiety.doc")
```



```{r}
# New therapy

therapy_1 <- felm(log(counseling) ~ log(therapy_norm+1)|0|0|state, data = hps_all_v19)

therapy_2 <- felm(log(counseling) ~ log(therapy_norm+1)|state|0|state, data = hps_all_v19)

therapy_3 <- felm(log(no_counseling) ~ log(therapy_norm+1)|0|0|state, data = hps_all_v19)

therapy_4 <- felm(log(no_counseling) ~ log(therapy_norm+1)|state|0|state, data = hps_all_v19)

therapy_5 <- felm(log(medication) ~ log(therapy_norm+1)|0|0|state, data = hps_all_v19)

therapy_6 <- felm(log(medication) ~ log(therapy_norm+1)|state|0|state, data = hps_all_v19)

htmlreg(l = list(therapy_1, therapy_2, therapy_3, therapy_4, therapy_5, therapy_6), include.ci = FALSE, digits = 4, file = "therapy.doc")

therapy_1 <- felm(log(counseling) ~ factor(state)|0|0|state, data = hps_all_v19)

therapy_2 <- felm(log(no_counseling) ~ factor(state)|0|0|state, data = hps_all_v19)
```


```{r}
# Therapy weighted

therapy_1 <- felm(log(counseling) ~ log(therapy_norm+1)|0|0|state, data = hps_all_v19,
                  weights = hps_all_v19$pop)

therapy_2 <- felm(log(counseling) ~ log(therapy_norm+1)|state|0|state, data = hps_all_v19, weights = hps_all_v19$pop)

therapy_3 <- felm(log(no_counseling) ~ log(therapy_norm+1)|0|0|state, data = hps_all_v19, weights = hps_all_v19$pop)

therapy_4 <- felm(log(no_counseling) ~ log(therapy_norm+1)|state|0|state, data = hps_all_v19, weights = hps_all_v19$pop)

therapy_5 <- felm(log(medication) ~ log(therapy_norm+1)|0|0|state, data = hps_all_v19)

therapy_6 <- felm(log(medication) ~ log(therapy_norm+1)|state|0|state, data = hps_all_v19, weights = hps_all_v19$pop)

htmlreg(l = list(therapy_1, therapy_2, therapy_3, therapy_4, therapy_5, therapy_6), include.ci = FALSE, digits = 4, file = "therapy.doc")
```

```{r}
# Another graph

therapy_graph_1 <- hps_all_v19 %>% 
  group_by(start_date) %>% 
  summarize(across(medication:therapy_norm, ~ weighted.mean(., pop, na.rm = TRUE))) %>% 
  ggplot(aes(x = start_date)) + 
    geom_line(aes(y = counseling, color = "HPS % receiving counseling")) + 
    geom_line(aes(y = no_counseling, color = "HPS % needed, but did not get counseling")) + 
    geom_line(aes(y = therapy_norm/9.5, color = "Therapy SVI")) + 
    theme_classic(base_size = 14) +
    theme(legend.position = "top") + 
    labs(x = "", y = "% of Respondents") + 
      scale_color_manual(name = c("", "", ""),
                       values = c("HPS % receiving counseling" = "blue",
                                  "HPS % needed, but did not get counseling" = "red",
                                  "Therapy SVI" = "green")) +
    scale_x_date(date_breaks = "3 months", date_labels = "%m/%y") + 
    guides(color = guide_legend(nrow = 2))

therapy_graph_2 <- hps_all_v19 %>% 
  group_by(start_date) %>% 
  summarize(across(medication:therapy_norm, ~ weighted.mean(., pop, na.rm = TRUE))) %>% 
  ggplot(aes(x = start_date)) + 
    geom_line(aes(y = medication, color = "HPS % using medication")) + 
    geom_line(aes(y = medication_or_counseling, color = "HPS % medication and/or counseling")) + 
    geom_line(aes(y = therapy_norm/4, color = "Therapy SVI")) + 
    theme_classic(base_size = 14) +
    theme(legend.position = "top") + 
    labs(x = "", y = "% of Respondents") + 
    scale_color_manual(name = c("", "", ""),
                       values = c("HPS % using medication" = "blue",
                                  "HPS % medication and/or counseling" = "red",
                                  "Therapy SVI" = "green")) +
    scale_x_date(date_breaks = "3 months", date_labels = "%m/%y") + 
  guides(color = guide_legend(nrow = 2))
```


```{r}
# New smoking and alcohol regressions

alc_1 <- felm(log(alcohol_norm + 1) ~ log(all_alc)|0|0|state, data = alcohol_v2)

alc_2 <- felm(log(alcohol_norm + 1) ~ log(all_alc)|state|0|state, data = alcohol_v2)

alc_3<- felm(log(alcohol_norm + 1) ~ log(all_alc)|state + year|0|state, data = alcohol_v2)

alc_4 <- felm(log(alcoholism_norm + 1) ~ log(all_alc)|0|0|state, data = alcohol_v2)

alc_5 <- felm(log(alcoholism_norm + 1) ~ log(all_alc)|state|0|state, data = alcohol_v2)

alc_7 <- felm(log(alcoholism_norm + 1) ~ log(all_alc)|state + year|0|state, data = alcohol_v2)

htmlreg(l = list(alc_1, alc_2, alc_3, alc_4, alc_5, alc_6), include.ci = FALSE, digits = 4, file = "alc_verif.doc")
```

```{r}
# Weighted
pop <- hps_all_v19 %>% select(pop, state) %>% distinct()

alcohol_v3 <- alcohol_v3 %>% left_join(pop, by = "state")

alc_1 <- felm(log(alcohol_norm + 1) ~ log(all_alc)|0|0|state, data = alcohol_v3,
              weights = alcohol_v3$pop)

alc_2 <- felm(log(alcohol_norm + 1) ~ log(all_alc)|state|0|state, data = alcohol_v3,
              weights = alcohol_v3$pop)

alc_3<- felm(log(alcohol_norm + 1) ~ log(all_alc)|state + year|0|state, data = alcohol_v3, weights = alcohol_v3$pop)

alc_4 <- felm(log(alcoholism_norm + 1) ~ log(all_alc)|0|0|state, data = alcohol_v3,
              weights = alcohol_v3$pop)

alc_5 <- felm(log(alcoholism_norm + 1) ~ log(all_alc)|state|0|state, data = alcohol_v3,
              weights = alcohol_v3$pop)

alc_6 <- felm(log(alcoholism_norm + 1) ~ log(all_alc)|state + year|0|state, data = alcohol_v3, weights = alcohol_v3$pop)

htmlreg(l = list(alc_1, alc_2, alc_3, alc_4, alc_5, alc_6), include.ci = FALSE, digits = 4, file = "alc_verif.doc")
```


```{r}
cig_1 <- felm(log(cigarettes_norm+1) ~ log(packs_per_capita)|0|0|state, data = smoking_v3, weights = smoking_v3$total_pop)

cig_2 <- felm(log(cigarettes_norm+1) ~ log(packs_per_capita)|state|0|state, data = smoking_v3, weights = smoking_v3$total_pop)

cig_3 <- felm(log(cigarettes_norm+1) ~ log(packs_per_capita)|state + year|0|state, data = smoking_v3, weights = smoking_v3$total_pop)

cig_4 <- felm(log(e_cig_norm+1)~ log(packs_per_capita)|0|0|state, data = smoking_v3,
              weights = smoking_v3$total_pop)

cig_5 <- felm(log(e_cig_norm+1) ~log(packs_per_capita)|state|0|state, data = smoking_v3,
              weights = smoking_v3$total_pop)

cig_6 <- felm(log(e_cig_norm+1) ~log(packs_per_capita)|state+year|0|state, data = smoking_v3, weights = smoking_v3$total_pop)

cig_7 <- felm(log(e_cig_norm + cigarettes_norm + smoking_norm+1) 
              ~log(packs_per_capita)|state+year|0|state, data = smoking_v3,
              weights = smoking_v3$total_pop)

htmlreg(l = list(cig_1, cig_2, cig_3, cig_4, cig_5, cig_6, cig_7), include.ci = FALSE, digits = 4,
        file = "smoking_verif.doc")



```

```{r}
cig_1 <- felm(log(cigarettes_norm+1) ~ log(packs_per_capita)|0|0|state, data = smoking_v3)

cig_2 <- felm(log(cigarettes_norm+1) ~ log(packs_per_capita)|state|0|state, data = smoking_v3)

cig_3 <- felm(log(cigarettes_norm+1) ~ log(packs_per_capita)|state + year|0|state, data = smoking_v3)

cig_4 <- felm(log(e_cig_norm+1)~ log(packs_per_capita)|0|0|state, data = smoking_v3)

cig_5 <- felm(log(e_cig_norm+1) ~log(packs_per_capita)|state|0|state, data = smoking_v3)

cig_6 <- felm(log(e_cig_norm+1) ~log(packs_per_capita)|state+year|0|state, data = smoking_v3)

cig_7 <- felm(log(e_cig_norm + cigarettes_norm + smoking_norm+1) 
              ~log(packs_per_capita)|state+year|0|state, data = smoking_v3)

htmlreg(l = list(cig_1, cig_2, cig_3, cig_4, cig_5, cig_6, cig_7), include.ci = FALSE, digits = 4,
        file = "smoking_verif.doc")
```

```{r}
# Remaking the graphs

test<- alcohol_v2 %>% 
  group_by(year, month) %>% 
  summarize(all_alc = mean(all_alc),,
            alcohol_norm = mean(alcohol_norm),
            alcoholism_norm = mean(alcoholism_norm)) %>% 
  mutate(date = mdy(paste(month, "01", year, sep = "-"))) %>% 
  ggplot(aes(x = date)) + 
    geom_line(aes(y = all_alc, color = "Sale Volume")) + 
    geom_line(aes(y = alcohol_norm  * 0.007, 
                  color = "Alcohol SVI")) + 
    geom_line(aes(y = alcoholism_norm * 0.01, 
                  color = "Alcoholism SVI")) + 
    theme_classic(base_size = 14) + 
    labs(x = "", y = "Gallons per Capita") + 
    theme(legend.position = "top",
          plot.margin = unit(c(0.2, 1, 0.2, 0.2), "cm")) + 
  scale_color_manual(name = c("", "", ""),
                       values = c("Sale Volume" = "blue",
                                  "Alcohol SVI" = "red",
                                 "Alcoholism SVI" = "green"))

test2 <- smoking_v3 %>%
  group_by(year, month) %>% 
  summarize(packs_per_capita = mean(packs_per_capita),,
            smoking = mean(smoking_norm),
            cigarettes = mean(cigarettes_norm),
            e_cig = mean(e_cig_norm)) %>% 
  mutate(date = mdy(paste(month, "01", year, sep = "-"))) %>% 
  ggplot(aes(x = date)) + 
    geom_line(aes(y = packs_per_capita, color = "Imputed Sales")) + 
    geom_line(aes(y = smoking * 0.1, 
                  color = "Smoking SVI")) + 
    geom_line(aes(y = cigarettes * 0.09, 
                  color = "Cigarettes SVI")) + 
    geom_line(aes(y = e_cig * 0.08, 
                  color = "E-Cigarettes SVI")) + 
    theme_classic(base_size = 14) + 
    labs(x = "", y = "Packs per Capita") + 
    theme(legend.position = "top",
          plot.margin = unit(c(0.2, 1, 0.2, 0.2), "cm")) + 
  scale_color_manual(name = c("", "", ""),
                       values = c("Imputed Sales" = "blue",
                                  "Smoking SVI" = "red",
                                 "Cigarettes SVI" = "green",
                                 "E-Cigarettes SVI" = "orange")) + 
  scale_x_date(date_breaks = "6 months", date_labels = "%m/%y") + 
  guides(color = guide_legend(nrow = 2))

```
