==> 10_states.Rmd <==
---
title: "10-states"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(tidyverse)
library(janitor)
library(openintro)
library(lubridate)
```

```{r}
# read in the HPS all data

hps <- read_rds('clean_data/search/hps_all_v1.rds')

hps_dates <- hps %>% 
  ungroup() %>% 
  select(start_date, end_date) %>% 
  distinct() %>%
  mutate(days = as.numeric(end_date - start_date))
```

```{r}
# Lockdown data read in from CGRT -- and COVID stuff too

lockdowns <- read_csv('data/policy/OxCGRT_latest.csv', col_types = cols()) %>% 
  clean_names() %>%
  filter(country_code == "USA",
         jurisdiction == "STATE_TOTAL") %>% 
  select(-country_name, -country_code, -jurisdiction, -confirmed_cases, -confirmed_deaths,
         -c8_international_travel_controls,
         -e4_international_support, -e3_fiscal_measures, -stringency_legacy_index,
         -stringency_legacy_index_for_display, -stringency_index_for_display,
         -economic_support_index_for_display, -h8_protection_of_elderly_people,
         -h8_flag, -m1_wildcard, -containment_health_index_for_display,
         -government_response_index_for_display,
         -h4_emergency_investment_in_healthcare, -h5_investment_in_vaccines) %>% 
  mutate(date = ymd(date)) %>%
  mutate(state_code = state2abbr(region_name)) %>% 
  select(-region_code, -region_name)

lockdowns_weekly <- expand_grid(lockdowns, hps_dates) %>% 
  filter(date >= start_date, date <= end_date) %>% 
  group_by(start_date, end_date, state_code) %>% 
  summarize(across(c1_school_closing:economic_support_index, ~ median(., na.rm = TRUE))) %>% 
  mutate(across(c1_school_closing:h7_flag, round))

```

```{r}
# New COVID cases

covid <- read_csv('data/state_data/us-states.csv', col_types = cols()) %>% 
  mutate(state = state2abbr(state)) %>% 
  mutate(date = as_date(date))

hps_dates_2 <- hps_dates %>%
  mutate(new_start_date = start_date - days(5),
         new_end_date = end_date - days(5))
  
covid_weekly <- expand_grid(hps_dates_2, covid) %>% 
  filter(date <= new_end_date,
         date >= new_start_date) %>% 
  drop_na(state) %>% 
  group_by(state, fips, start_date, end_date, days) %>% 
  summarize(cases = sum(cases),
            deaths = sum(deaths))

state_dem <- read_csv('data/state_data/R12917460_SL040.csv', col_types = cols()) %>% 
  slice(2:n()) %>% 
  clean_names() %>% 
  select(state_fips_code, total_population_56, population_density_per_sq_mile,
         total_population_white_alone, population_25_years_and_over,
         population_25_years_and_over_bachelors_degree_or_better,
         civilian_population_in_labor_force_16_years_and_over,
         median_household_income_in_2019_inflation_adjusted_dollars,
         gini_index, total, total_no_health_insurance_coverage) %>% 
  rename('fips' = 'state_fips_code',
         'pop' = 'total_population_56',
         'pop_den' = 'population_density_per_sq_mile',
         'white_pop' = 'total_population_white_alone',
         "bach_degree" = 'population_25_years_and_over_bachelors_degree_or_better',
         'labor_force' = 'civilian_population_in_labor_force_16_years_and_over',
         'median_income' = 'median_household_income_in_2019_inflation_adjusted_dollars',
         'gini' = 'gini_index') %>%
  mutate(uninsured_pct = as.numeric(total_no_health_insurance_coverage)/as.numeric(total),
         white_pct = as.numeric(white_pop)/as.numeric(pop),
         bach_pct = as.numeric(bach_degree)/as.numeric(population_25_years_and_over)) %>% 
  select(-white_pop, bach_degree, -population_25_years_and_over, -total,
         -total_no_health_insurance_coverage) %>%
  mutate(across(pop:median_income, as.numeric))

pop <- state_dem %>% select(fips, pop)

covid_per_capita <- covid_weekly %>%
  left_join(pop, by = "fips") %>% 
  drop_na(pop) %>% 
  mutate(cases_per_100k_5day = cases * 100000 / (pop * days/5),
         deaths_per_100k_5day = deaths * 100000 / (pop * days/5)) %>% 
  select(-cases, -deaths)

hps_2 <- left_join(hps, covid_per_capita, by = c("state", "start_date",
                                                 "end_date")) %>%
  left_join(state_dem %>% select(-pop), by = "fips")
  

```

```{r}
# UI claims

ui <- read_csv('data/state_data/UI Claims - State - Weekly.csv', 
               col_types = cols()) %>% 
  clean_names() %>% 
  mutate(ui_end_date = mdy(paste(month, day_endofweek, year, sep = "-")),
         statefips = as.character(str_pad(statefips, width = 2, side = "left", pad = "0"))) %>% 
  select(ui_end_date, statefips, initclaims_count_regular:contclaims_count_combined) %>%
  mutate(ui_start_date = ui_end_date - days(7))

ui_weekly <- expand_grid(ui, hps_dates_2) %>%
  filter((ui_start_date <= new_end_date & ui_start_date >= new_start_date) | 
           (ui_end_date <= new_end_date & ui_end_date >= new_start_date)) %>% 
  group_by(start_date, end_date, statefips, days) %>% 
  summarize(weeks = n(),
            initclaims_count_regular = sum(initclaims_count_regular),
            contclaims_count_regular = sum(contclaims_count_regular),
            initclaims_count_pua = sum(as.numeric(initclaims_count_pua)),
            contclaims_count_pua = sum(as.numeric(contclaims_count_pua)),
            initclaims_count_combined = sum(initclaims_count_combined),
            contclaims_count_combined = sum(as.numeric(contclaims_count_combined))) %>% 
  mutate(across(contclaims_count_regular:contclaims_count_combined,
                ~ ./weeks))

hps_3 <- left_join(hps_2, ui_weekly, by = c("fips" = "statefips", "start_date", "end_date",
                                            "days")) %>% 
  mutate(across(contclaims_count_regular:contclaims_count_combined,
                ~ ./labor_force)) %>% 
  rename('initclaims_regular' = 'initclaims_count_regular',
         'contclaims_regular' = 'contclaims_count_regular',
         'initclaims_pua' = 'initclaims_count_pua',
         'contclaims_pua' = 'contclaims_count_pua',
         'initclaims_combined' = 'initclaims_count_combined',
         'contclaims_combined' = 'contclaims_count_combined')

# Adding in lockdowns sadly makes this no longer relevant :(

hps_4 <- left_join(hps_3, lockdowns_weekly, by = c("start_date", "end_date", 
                                                   "state" = "state_code")) %>% 
  group_by(fips) %>% 
  arrange(start_date) %>% 
  mutate(period = 1:n())

```
```{r}
# Bring in covid and lockdown searches -- from the search file covid_lockdown_statesearches

searches_weekly <- expand_grid(hps_dates, covid_lockdown_statesearches %>% 
                                 rename('state_2' = 'state')) %>% 
  mutate(state_2 = str_sub(state_2, start = 4)) %>% 
  filter(date + days(6) >= start_date,
         date <= end_date) %>% 
  select(-days, -date) %>% 
  group_by(start_date, end_date, state_2) %>% 
  summarize(across(lockdown:covid, ~ mean(., na.rm = TRUE)))

hps_5 <- left_join(hps_4, searches_weekly, by = c("start_date", "end_date",
                                                  "state" = "state_2")) %>% 
  mutate(lockdown_norm = lockdown / average_3,
         covid_norm = covid / average_3) %>% 
  mutate(across(lockdown_norm:covid_norm, ~ . * 100 / max(.))) %>% 
  mutate(year = year(end_date))
```


```{r}
# Bringing in additional anxiety/depression searches -- these are the _state_terms.rds files

anxiety_depression <- left_join(anxiety_terms_state, depression_terms_state, by = c("state", "date"))

searches_weekly <- expand_grid(hps_dates, anxiety_depression %>% 
                                 rename('state_2' = 'state')) %>% 
  mutate(state_2 = str_sub(state_2, start = 4)) %>% 
  filter(date + days(6) >= start_date,
         date <= end_date) %>% 
  select(-days, -date) %>% 
  group_by(start_date, end_date, state_2) %>% 
  summarize(across(nervous:suicide, ~ mean(., na.rm = TRUE)))

hps_all_v6 <- left_join(hps_all_v5, searches_weekly, by = c("start_date", "end_date",
                                                  "state" = "state_2")) %>% 
  ungroup() %>% 
  mutate(nervous_norm = nervous / average_3,
         stress_norm = stress / average_3,
         panic_norm = panic / average_3,
         worry_norm = worry / average_3,
         crying_norm = crying / average_3,
         lonely_norm = lonely / average_3,
         sad_norm = sad / average_3,
         suicide_norm = suicide / average_3) %>% 
  mutate(across(nervous_norm:suicide_norm, ~ . * 100 / max(.))) %>% 
  mutate(year = year(end_date))
```


```{r}
# Bringing in tiredness stuff 

searches_weekly <- expand_grid(hps_dates, tired_clean %>% 
                                 rename('state_2' = 'state')) %>% 
  mutate(state_2 = str_sub(state_2, start = 4)) %>% 
  filter(date + days(6) >= start_date,
         date <= end_date) %>% 
  select(-days, -date) %>% 
  group_by(start_date, end_date, state_2) %>% 
  summarize(across(tired, ~ mean(., na.rm = TRUE)))

hps_all_v11 <- left_join(hps_all_v10, searches_weekly, by = c("start_date", "end_date",
                                                  "state" = "state_2")) %>% 
  ungroup() %>% 
  mutate(tired_norm = tired / average_3) %>% 
  mutate(across(tired_norm, ~ . * 100 / max(.))) %>% 
  mutate(year = year(end_date))

```

```{r}
#curiosity - actually, just the panel regressions on a broader set of data

depression_1 <- felm(log(hps_depression) ~ factor(c2_workplace_closing) + factor(c1_school_closing) +
               initclaims_combined + 
               contclaims_combined + cases_per_100k_5day + 
               factor(e1_income_support) + factor(e2_debt_contract_relief) + 
                 log(covid_norm+1) + log(lockdown_norm+1)
             |period + factor(state):factor(year)|0|state,
             data = hps_all_v5)

depression_2 <- felm(log(depression_norm + 1) ~ 
                       factor(c2_workplace_closing) + factor(c1_school_closing) + 
               initclaims_combined + 
               contclaims_combined + cases_per_100k_5day + 
               economic_support_index + log(covid_norm+1) + log(lockdown_norm+1)
             |period + factor(state):factor(year)|0|state,
             data = hps_all_v5)

anxiety_1 <- felm(log(hps_anxiety) ~ 
                       factor(c2_workplace_closing) + factor(c1_school_closing) + 
               initclaims_combined + 
               contclaims_combined + cases_per_100k_5day + 
               economic_support_index + log(covid_norm+1) + log(lockdown_norm+1)
             |period + factor(state):factor(year)|0|state,
             data = hps_5)

anxiety_2 <- felm(log(anxiety_norm + 1) ~ 
                       factor(c2_workplace_closing) + factor(c1_school_closing) + 
               initclaims_combined + 
               contclaims_combined + cases_per_100k_5day + 
               economic_support_index + log(covid_norm+1) + log(lockdown_norm+1)
             |period + factor(state):factor(year)|0|state,
             data = hps_5)

htmlreg(l = list(depression_1, depression_2, anxiety_1, anxiety_2),
        include.ci = FALSE, digits = 4, file = "hps_v1.doc")
```

```{r}
# Graphing comparisons for a bunch of terms together?


```


```{r}
# Join in therapy data
# Therapy stuff

healthcare <- read_csv('data/state_data/Mental_Health_Care_in_the_Last_4_Weeks.csv',
                       col_types = cols()) %>% 
  clean_names() %>% 
  filter(state != "United States") %>% 
  select(-group, -phase, -time_period, -time_period_label, -low_ci, -high_ci, -confidence_interval,
         -quartile_range, -suppression_flag) %>% 
  pivot_wider(names_from = indicator, values_from = value) %>% 
  clean_names() %>% 
  rename("medication" = "took_prescription_medication_for_mental_health_last_4_weeks",
         "counseling" = "received_counseling_or_therapy_last_4_weeks",
         "medication_or_counseling" = 
           "took_prescription_medication_for_mental_health_and_or_received_counseling_or_therapy_last_4_weeks",
         "no_counseling" = "needed_counseling_or_therapy_but_did_not_get_it_last_4_weeks") %>% 
  select(-subgroup) %>% 
  mutate(state = state2abbr(state),
         across(time_period_start_date:time_period_end_date, mdy))

hps_6 <- left_join(hps_all_v3, healthcare, by = c("state", "start_date" = "time_period_start_date",
                                                  "end_date" = "time_period_end_date"))

# Join in google search data

therapy_weekly <- expand_grid(hps_dates, therapy_clean %>% 
                                 rename('state_2' = 'state')) %>% 
  mutate(state_2 = str_sub(state_2, start = 4)) %>% 
  filter(date + days(6) >= start_date,
         date <= end_date) %>% 
  select(-days, -date) %>% 
  group_by(start_date, end_date, state_2) %>% 
  summarize(across(therapy, ~ mean(., na.rm = TRUE)))

hps_7 <- left_join(hps_6, therapy_weekly, by =c("state" = "state_2",
                                                "start_date", "end_date")) %>% 
  mutate(therapy_norm = therapy / average_3) %>% 
  mutate(therapy_norm = therapy_norm * 100/ max(therapy_norm))
  
```

```{r}
# Join in the either column -- from the verification file

hps_all_v5 <- hps_all_v4 %>% 
  left_join(hps_state %>% select(start_date, state_2, hps_either), by = c("state" = "state_2",
                                                                          "start_date")) %>%
  mutate(share_no_counseling = no_counseling / (counseling + medication_or_counseling),
         share_counseling = counseling / (counseling + medication_or_counseling),
         share_medication = medication / (counseling + medication_or_counseling),
         share_medication_or_counseling = medication_or_counseling / 
           (counseling + medication_or_counseling))

# redoing quickly for other parts
hps_all_v10 <- hps_all_v9 %>% 
  mutate(share_no_counseling = no_counseling / (counseling + medication_or_counseling),
         share_counseling = counseling / (counseling + medication_or_counseling),
         share_medication = medication / (counseling + medication_or_counseling),
         share_medication_or_counseling = medication_or_counseling / 
           (counseling + medication_or_counseling))
```

```{r}
#Exploratory stuff on care seeking

no_counseling_1 <- felm(log(share_no_counseling) ~ hps_either + cases_per_100k_5day + 
                       initclaims_combined + contclaims_combined + factor(e1_income_support) + 
                       factor(e2_debt_contract_relief) + 
                       factor(c2_workplace_closing) + 
                       log(therapy_norm +1) + 
                       log(covid+1) + log(lockdown +1)|factor(state):factor(year)
                     |0|state, data = hps_all_v5)

counseling_1 <- felm(log(share_counseling) ~ hps_either + cases_per_100k_5day + 
                       initclaims_combined + contclaims_combined + factor(e1_income_support) + 
                       factor(e2_debt_contract_relief) + 
                       factor(c2_workplace_closing) + 
                       log(therapy_norm +1) +
                       log(covid+1) + log(lockdown +1)|factor(state):factor(year)
                     |0|state, data = hps_all_v5)

medication_1 <- felm(log(share_medication) ~ hps_either + cases_per_100k_5day + 
                       initclaims_combined + contclaims_combined + factor(e1_income_support) + 
                       factor(e2_debt_contract_relief) + 
                       factor(c2_workplace_closing) + 
                       log(therapy_norm +1) +
                       log(covid+1) + log(lockdown +1)|factor(state):factor(year)
                     |0|state, data = hps_all_v5)

both_1 <- felm(log(share_medication_or_counseling) ~ hps_either + cases_per_100k_5day + 
                       initclaims_combined + contclaims_combined + factor(e1_income_support) + 
                       factor(e2_debt_contract_relief) + 
                       factor(c2_workplace_closing) + 
                       log(therapy_norm +1) +
                       log(covid+1) + log(lockdown +1)|factor(state):factor(year)
                     |0|state, data = hps_all_v5)

htmlreg(l = list(no_counseling_1, counseling_1, medication_1, both_1),
        include.ci = FALSE, digits = 4, file = "counseling_test1.doc")
```

```{r}
# processing big HPS file

hps_1 <- read_csv('data/state_data/pulse_puf_most_recent.csv', col_types = cols())

hps_2 <- hps_1 %>%
  select(week_x, est_st, tbirth_year, egender, rhispanic, rrace, eeduc, wrkloss, 
         expctloss, hlthstatus, anxious, worry, interest, down) %>% 
  mutate(
    across(c(wrkloss, expctloss, hlthstatus, anxious, worry, interest, down), 
           ~ ifelse(. == -99 | . == -88, NA, .)))


hps_3 <- read_csv('data/state_data/phase2_pulse_puf_most_recent.csv', col_types = cols())

hps_full_part2 <- hps_full_part2 %>%
   select(week_x, est_st, tbirth_year, egender, rhispanic, rrace, eeduc, wrkloss, 
         expctloss, hlthstatus, anxious, worry, interest, down, prescript, mh_svcs, mh_notget) %>% 
  mutate(across(c(wrkloss:mh_notget), 
           ~ ifelse(. == -99 | . == -88, NA, .)))

hps_combined <- bind_rows(hps_full_part1, hps_full_part2)

hps_combined_2 <- hps_combined %>% 
  mutate(across(prescript:mh_notget, ~ ifelse(. == 1, 0, ifelse(. == 2, 1, .))))
```


```{r}
# hps_race <- hps_combined_2 %>%
#   mutate(across(prescript:mh_notget, ~ ifelse(is.na(.), 0, .))) %>% 
#   rowwise() %>% 
#   mutate(depression = ifelse(sum(interest, down, na.rm = TRUE) >= 3 
#                              & is.na(interest) == FALSE & is.na(down) == FALSE, 1, 0),
#          anxiety = ifelse(sum(anxious, worry, na.rm = TRUE) >= 3 &
#                             is.na(anxious) == FALSE & is.na(worry) == FALSE, 1, 0),
#          depression_num = ifelse(is.na(interest) == FALSE & is.na(down) == FALSE,
#                                  1, 0),
#          anxiety_num = ifelse(is.na(anxious) == FALSE & is.na(worry) == FALSE,
#                               1, 0),
#          prescript_num = ifelse(is.na(prescript) == FALSE, 1, 0),
#          mh_svcs_num = ifelse(is.na(mh_svcs) == FALSE, 1, 0),
#          mh_notget_num = ifelse(is.na(mh_notget) == FALSE, 1, 0)) %>%
#   ungroup() %>% 
#   group_by(week_x, est_st, rrace) %>% 
#   summarize(across(c(prescript:mh_notget, depression:mh_notget_num), sum))
# 
# hps_race_2 <- hps_race %>%
#   mutate(depression_pct = depression / depression_num,
#          anxiety_pct = anxiety / anxiety_num) %>% 
#   rowwise() %>% 
#   mutate(mental_health_all = sum(prescript:mh_notget)) %>% 
#   ungroup() %>% 
#   mutate(prescript_pct = prescript / prescript_num,
#          mh_svcs_pct = mh_svcs / mh_svcs_num,
#          mh_notget_pct = mh_notget / mh_notget_num)

# This data just isn't good enough - try the tables!
```

```{r}
# DOWNLOAD ALL THE TABLES

hps_1 <- "https://www2.census.gov/programs-surveys/demo/tables/hhp/"

year <- c("2020", "2021")

therapy <- "/health4_week"

anxiety <- "/health2b_week"
  
depression <- "/health2a_week"

ending <- ".xlsx"

depression_urls <- c()

depression_filenames <- c()

for (i in 1:39) {
  if (i <= 21) {
    depression_urls[i] = paste(hps_1, year[1], "/wk", as.character(i), 
                            depression, as.character(i), ending, sep = "")
    depression_filenames[i] = paste("depression", "_week", as.character(i), ending, sep = "")
  }
  
  else {
    depression_urls[i] =  paste(hps_1, year[2], "/wk", as.character(i), 
                            depression, as.character(i), ending, sep = "")
    depression_filenames[i] = paste("depression", "_week", as.character(i), ending, sep = "")
  }
  
}

#https://www2.census.gov/programs-surveys/demo/tables/hhp/2021/wk25/health2a_week25.xlsx


for (i in 1:length(depression_urls)){
  download.file(depression_urls[i], 
                destfile = paste0("data/state_data/hps/",depression_filenames[i]))
}

# For therapy

therapy_urls <- c()

therapy_filenames <- c()

for (i in 13:39) {
  if (i <= 21) {
    therapy_urls[i-12] = paste(hps_1, year[1], "/wk", as.character(i), 
                            therapy, as.character(i), ending, sep = "")
    therapy_filenames[i-12] = paste("therapy", "_week", as.character(i), ending, sep = "")
  }
  
  else {
    therapy_urls[i-12] =  paste(hps_1, year[2], "/wk", as.character(i), 
                            therapy, as.character(i), ending, sep = "")
    therapy_filenames[i-12] = paste("therapy", "_week", as.character(i), ending, sep = "")
  }
  
}

for (i in 1:length(therapy_urls)){
  download.file(therapy_urls[i], 
                destfile = paste0("data/state_data/hps/",therapy_filenames[i]))
}

```

```{r}
# Let's try processing thedata -- note that we accidentally flipped the names oops

test <- read_excel('data/state_data/hps/depression/anxiety_week1.xlsx',
                   sheet = 52) %>% 
  clean_names()
  
state_name <- str_sub(colnames(test)[1], 97)

test2 <- test %>%
  mutate(state = state_name) %>% 
  filter(.[[1]] == "White alone, not Hispanic") %>% 
  mutate(across(x2:x11, ~ ifelse(. == "-", 0, .))) %>% 
  mutate(interest = as.numeric(x4) + as.numeric(x5),
         interest_total = as.numeric(x2) + as.numeric(x3) + as.numeric(x4) + as.numeric(x5),
         down = as.numeric(x9) + as.numeric(x10),
         down_total = as.numeric(x7) + as.numeric(x8) + as.numeric(x9) + as.numeric(x10))
  # mutate week name equals i
```

```{r}
depression_filenames <- list.files(path = "data/state_data/hps/depression")

depression_income <- tibble()

depression_terms <- c("18 - 29", "30 - 39", "40 - 49", "50 - 59", "60 - 69",
                   "70 - 79", "White alone, not Hispanic", "Black alone, not Hispanic",
                   "Less than $25,000", "$25,000 - $34,999", "$35,000 - $49,999",
                   "$50,000 - $74,999", "$75,000 - $99,999", "$100,000 - $149,999",
                   	"$150,000 - $199,999")

depression_names <- c("18to29", "30to39", "40to49", "50to59", "60to69", "70to79",
                   "white", "black", "less25k", "25to35k", "35to50k", 
                   "50to75k", "75to100k", "100to150k", "150to200k")

depression_total <- tibble()

for (k in 1:length(depression_terms)) {
  for (i in 1:length(depression_filenames)) {
    for (j in 2:52) {
      test <- suppressMessages(
        read_excel(paste("data/state_data/hps/depression/", depression_filenames[i], sep = ""),
                     sheet = j)) %>% 
        clean_names()
      
      
      state_name <- str_sub(colnames(test)[1], 97)
      
      # Pull the week name
      
      week_name <- str_sub(depression_filenames[i], start = 13, 
                           end = str_length(depression_filenames[i]) - 5)
      
      # pull the state name in a more organized way idk
      
      test2 <- test %>%
        mutate(state = state_name) %>% 
        filter(.[[1]] == depression_terms[k]) %>% 
        mutate(across(x2:x11, ~ ifelse(. == "-", 0, .))) %>% 
        mutate(interest = as.numeric(x4) + as.numeric(x5),
               interest_total = as.numeric(x2) + as.numeric(x3) + as.numeric(x4) + as.numeric(x5),
               down = as.numeric(x9) + as.numeric(x10),
               down_total = as.numeric(x7) + as.numeric(x8) + as.numeric(x9) + as.numeric(x10)) %>% 
        mutate(week = as.numeric(week_name)) %>% 
        select(state, week, down, interest, down_total, interest_total)
      
      if(i == 1 & j == 2) {
        depression_income <- test2
      }
      
      else {
        depression_income <- bind_rows(depression_income, test2)
      }
    }
  }
  
  interest_name <- paste("interest", depression_names[k], sep = "_")

  down_name <- paste("down", depression_names[k], sep = "_") 
    
  depression_income_2 <- depression_income %>%
      mutate(state = state2abbr(state)) %>% 
      mutate(interest_pct = interest/interest_total,
             down_pct = down/down_total) %>% 
      select(state, week, interest_pct, down_pct) %>% 
      rename(!!interest_name := "interest_pct",
           !!down_name := "down_pct")
  
  if(k == 1) {
    depression_total <- depression_income_2
  }
  
  else {
    depression_total <- left_join(depression_total, depression_income_2, by = c("week", "state"))
  }
}
```

```{r}
depression_filenames <- list.files(path = "data/state_data/hps/depression")

depression_income <- tibble()

for (i in 1:length(depression_filenames)) {
  for (j in 2:52) {
    test <- suppressMessages(
      read_excel(paste("data/state_data/hps/depression/", depression_filenames[i], sep = ""),
                   sheet = j)) %>% 
      clean_names()
    
    # This is fine even tho it ends up misspelling names
    
    state_name <- str_sub(colnames(test)[1], 97)
    
    # Pull the week name
    
    week_name <- str_sub(depression_filenames[i], start = 13, 
                         end = str_length(depression_filenames[i]) - 5)
    
    test2 <- test %>%
      mutate(state = state_name) %>% 
      filter(.[[1]] == 
      "$150,000 - $199,999" # "Black alone, not Hispanic
             ) %>% 
      mutate(across(x2:x11, ~ ifelse(. == "-", 0, .))) %>% 
      mutate(interest = as.numeric(x4) + as.numeric(x5),
             interest_total = as.numeric(x2) + as.numeric(x3) + as.numeric(x4) + as.numeric(x5),
             down = as.numeric(x9) + as.numeric(x10),
             down_total = as.numeric(x7) + as.numeric(x8) + as.numeric(x9) + as.numeric(x10)) %>% 
      mutate(week = as.numeric(week_name)) %>% 
      select(state, week, interest, interest_total, down, down_total)
    
    if(i == 1 & j == 2) {
      depression_income <- test2
    }
    
    else {
      depression_income <- bind_rows(depression_income, test2)
    }
  }
  
}

depression_income_2 <- depression_income %>%
  mutate(state = state2abbr(state)) %>% 
  mutate(down_pct = down/down_total,
         interest_pct = interest/interest_total) %>% 
  select(state, week, down_pct, interest_pct)
```

```{r}
anxiety_filenames <- list.files(path = "data/state_data/hps/anxiety")

anxiety_income <- tibble()

anxiety_terms <- c("18 - 29", "30 - 39", "40 - 49", "50 - 59", "60 - 69",
                   "70 - 79", "White alone, not Hispanic", "Black alone, not Hispanic",
                   "Less than $25,000", "$25,000 - $34,999", "$35,000 - $49,999",
                   "$50,000 - $74,999", "$75,000 - $99,999", "$100,000 - $149,999",
                   	"$150,000 - $199,999")

anxiety_names <- c("18to29", "30to39", "40to49", "50to59", "60to69", "70to79",
                   "white", "black", "less25k", "25to35k", "35to50k", 
                   "50to75k", "75to100k", "100to150k", "150to200k")

anxiety_total <- tibble()

for (k in 1:length(anxiety_terms)) {
  for (i in 1:length(anxiety_filenames)) {
    for (j in 2:52) {
      test <- suppressMessages(
        read_excel(paste("data/state_data/hps/anxiety/", anxiety_filenames[i], sep = ""),
                     sheet = j)) %>% 
        clean_names()
      
      
      state_name <- str_sub(colnames(test)[1], 94)
      
      # Pull the week name
      
      week_name <- str_sub(anxiety_filenames[i], start = 16, 
                           end = str_length(anxiety_filenames[i]) - 5)
      
      # pull the state name in a more organized way idk
      
      test2 <- test %>%
        mutate(state = state_name) %>% 
        filter(.[[1]] == anxiety_terms[k]) %>% 
        mutate(across(x2:x11, ~ ifelse(. == "-", 0, .))) %>% 
        mutate(anxiety = as.numeric(x4) + as.numeric(x5),
               anxiety_total = as.numeric(x2) + as.numeric(x3) + as.numeric(x4) + as.numeric(x5),
               worry = as.numeric(x9) + as.numeric(x10),
               worry_total = as.numeric(x7) + as.numeric(x8) + as.numeric(x9) + as.numeric(x10)) %>% 
        mutate(week = as.numeric(week_name)) %>% 
        select(state, week, anxiety, anxiety_total, worry, worry_total)
      
      if(i == 1 & j == 2) {
        anxiety_income <- test2
      }
      
      else {
        anxiety_income <- bind_rows(anxiety_income, test2)
      }
    }
  }
  
  anxiety_name <- paste("anxiety", anxiety_names[k], sep = "_")

  worry_name <- paste("worry", anxiety_names[k], sep = "_")
    
  anxiety_income_2 <- anxiety_income %>%
      mutate(state = state2abbr(state)) %>% 
      mutate(anxiety_pct = anxiety/anxiety_total,
             worry_pct = worry/worry_total) %>% 
      select(state, week, anxiety_pct, worry_pct) %>% 
      rename(!!anxiety_name := "anxiety_pct",
           !!worry_name := "worry_pct")
  
  if(k == 1) {
    anxiety_total <- anxiety_income_2
  }
  
  else {
    anxiety_total <- left_join(anxiety_total, anxiety_income_2, by = c("week", "state"))
  }
}

```

```{r}
# Doing this for therapy

therapy_filenames <- list.files(path = "data/state_data/hps/therapy")

therapy_income <- tibble()

therapy_terms <- c("18 - 29", "30 - 39", "40 - 49", "50 - 59", "60 - 69",
                   "70 - 79", "White alone, not Hispanic", "Black alone, not Hispanic",
                   "Less than $25,000", "$25,000 - $34,999", "$35,000 - $49,999",
                   "$50,000 - $74,999", "$75,000 - $99,999", "$100,000 - $149,999",
                   	"$150,000 - $199,999")

therapy_names <- c("18to29", "30to39", "40to49", "50to59", "60to69", "70to79",
                   "white", "black", "less25k", "25to35k", "35to50k", 
                   "50to75k", "75to100k", "100to150k", "150to200k")

therapy_total <- tibble()

for (k in 1:length(therapy_terms)) {
  for (i in 1:length(therapy_filenames)) {
    for (j in 2:52) {
      test <- suppressMessages(
        read_excel(paste("data/state_data/hps/therapy/", therapy_filenames[i], sep = ""),
                     sheet = j)) %>% 
        clean_names()
      
      
      state_name <- str_sub(colnames(test)[1], 87)
      
      # Pull the week name
      
      week_name <- str_sub(therapy_filenames[i], start = 13, 
                           end = str_length(therapy_filenames[i]) - 5)
      
      # pull the state name in a more organized way idk
      
      test2 <- test %>%
        mutate(state = state_name) %>% 
        filter(.[[1]] == therapy_terms[k]) %>% 
        mutate(across(x2:x11, ~ ifelse(. == "-", 0, .))) %>% 
        mutate(counseling = as.numeric(x3),
               counseling_total = as.numeric(x3) + as.numeric(x4),
               no_counseling = as.numeric(x6),
               no_counseling_total = 
                 as.numeric(x6) + as.numeric(x7),
               medication = as.numeric(x9),
               medication_total = as.numeric(x9) + as.numeric(x10)) %>% 
        mutate(week = as.numeric(week_name)) %>% 
        select(state, week, counseling:medication_total)
      
      if(i == 1 & j == 2) {
        therapy_income <- test2
      }
      
      else {
        therapy_income <- bind_rows(therapy_income, test2)
      }
    }
  }
  
  counseling_name <- paste("counseling", therapy_names[k], sep = "_")
  
  no_counseling_name <- paste("no_counseling", therapy_names[k], sep = "_")

  medication_name <- paste("medication", therapy_names[k], sep = "_")
    
  therapy_income_2 <- therapy_income %>%
      mutate(state = state2abbr(state)) %>% 
      mutate(counseling_pct = counseling/counseling_total,
             medication_pct = medication/medication_total,
             no_counseling_pct = no_counseling/no_counseling_total) %>% 
      select(state, week, counseling_pct:no_counseling_pct) %>% 
      rename(!!counseling_name := "counseling_pct",
           !!no_counseling_name := "no_counseling_pct",
           !!medication_name := "medication_pct")
  
  if(k == 1) {
    therapy_total <- therapy_income_2
  }
  
  else {
    therapy_total <- left_join(therapy_total, therapy_income_2, by = c("week", "state"))
  }
}
```


```{r}
# Make an average depression thing

anxiety_total_2 <- anxiety_total %>%
  rowwise() %>% 
  mutate(anxiety_agg_18to29 = mean(anxiety_18to29, worry_18to29, na.rm = TRUE),
         anxiety_agg_30to39 = mean(anxiety_30to39, worry_30to39, na.rm = TRUE),
         anxiety_agg_40to49 = mean(anxiety_40to49, worry_40to49, na.rm = TRUE),
         anxiety_agg_50to59 = mean(anxiety_50to59, worry_50to59, na.rm = TRUE),
         anxiety_agg_60to69 = mean(anxiety_60to69, worry_60to69, na.rm = TRUE),
         anxiety_agg_70to79 = mean(anxiety_70to79, worry_70to79, na.rm = TRUE),
         anxiety_agg_black = mean(anxiety_black, worry_black, na.rm = TRUE),
         anxiety_agg_white = mean(anxiety_white, worry_white, na.rm = TRUE),
         anxiety_agg_less25k = mean(anxiety_less25k, worry_less25k, na.rm = TRUE),
         anxiety_agg_25to35k = mean(anxiety_25to35k, worry_25to35k, na.rm = TRUE),
         anxiety_agg_35to50k = mean(anxiety_35to50k, worry_35to50k, na.rm = TRUE),
         anxiety_agg_50to75k = mean(anxiety_50to75k, worry_50to75k, na.rm = TRUE),
         anxiety_agg_75to100k = mean(anxiety_75to100k, worry_75to100k, na.rm = TRUE),
         anxiety_agg_100to150k = mean(anxiety_100to150k, worry_100to150k, na.rm = TRUE),
         anxiety_agg_150to200k = mean(anxiety_150to200k, worry_150to200k, na.rm = TRUE)) %>% 
  ungroup() %>% 
  select(state, week, anxiety_agg_18to29:anxiety_agg_150to200k)
```

```{r}
# Join things in a coherent way?

hps_all_v8 <- left_join(hps_all_v7, depression_hps_est, by = c("state", "period" = "week")) %>% 
  left_join(., anxiety_hps_est, by = c("state", "period" = "week"))

hps_all_v9 <- left_join(hps_all_v8, therapy_total, by = c("state", "period" = "week"))
```

```{r}
# Let's try to make a subsequent lockdown dummy

first_lockdown <- hps_all_v16 %>%
  select(start_date, state, c1_school_closing, c2_workplace_closing, c6_stay_at_home_requirements) %>% 
  group_by(state) %>% 
  arrange(start_date)

c1 <- first_lockdown %>%
  filter(c1_school_closing > 1) %>% 
  slice(1) %>%
  select(start_date, state) %>% 
  rename("c1_start" = "start_date")

c2 <- first_lockdown %>%
  filter(c2_workplace_closing > 1) %>% 
  slice(1) %>%
  select(start_date, state) %>% 
  rename("c2_start" = "start_date")

c6 <- first_lockdown %>%
  filter(c6_stay_at_home_requirements > 1) %>% 
  slice(1) %>%
  select(start_date, state) %>% 
  rename("c6_start" = "start_date")

hps_all_v12 <- hps_all_v11 %>%
  left_join(c1, by = "state") %>% 
  left_join(c2, by = "state") %>% 
  left_join(c6, by = "state") %>% 
  # This codes dates 1 if they're non-lockdown periods after the first lockdown 
  mutate(c1 = ifelse(c1_flag == 0 & start_date > c1_start, 1, NA),
         c2 = ifelse(c2_flag == 0 & start_date > c2_start, 1, NA),
         c6 = ifelse(c6_flag == 0 & start_date > c6_start, 1, NA)) %>% 
  # This fills everything that's after the first lockdown as a 1
  fill(c1:c6, .direction = "down") %>% 
  # And this fills any initial lockdown/pre-lockdown as 0
  mutate(across(c1:c6, ~ replace_na(., 0))) %>% 
  # Sub c1 is: 1 if you're in the first lockdown, 0 if you're in the first pre-lockdown OR subsequent   # lockdowns, -1 if you're in any post-lockdown period w/o a lockdown
  mutate(sub_c1 = c1_flag - c1,
         sub_c2 = c2_flag - c2,
         sub_c6 = c6_flag - c6) %>% 
  mutate(across(sub_c1:sub_c6, ~ ifelse(. == -1 | . == 0, 0, .)))
  # mutate(across(sub_c1:sub_c6, as.factor))

hps_all_v12 <- hps_all_v11 %>%
  left_join(c1, by = "state") %>% 
  left_join(c2, by = "state") %>% 
  left_join(c6, by = "state") %>% 
  group_by(state) %>% 
  # This makes all subsequent periods after the first lockdown ends a 1
  mutate(c1 = ifelse(c1_flag == 0 & start_date > c1_start, 1, NA),
         c2 = ifelse(c2_flag == 0 & start_date > c2_start, 1, NA),
         c6 = ifelse(c6_flag == 0 & start_date > c6_start, 1, NA)) %>% 
  fill(c1:c6, .direction = "down") %>% 
  mutate(across(c1:c6, ~ replace_na(., 0))) %>% 
  # Pick out all the repeated lockdowns
  mutate(c1 = ifelse(c1 == 1 & c1_flag == 1, 2, c1),
         c2 = ifelse(c2 == 1 & c2_flag == 1, 2, c2),
         c6 = ifelse(c6 == 1 & c6_flag == 1, 2, c6)) %>% 
  mutate(across(c1:c6, ~ ifelse(. == 1, NA, .))) %>% 
  # Make only periods after the second lockdown starts equal to 2
  fill(c1:c6, .direction = "down") %>%
  mutate(across(c1:c6, ~ replace_na(., 0))) %>% 
  mutate(across(c1:c6, ~ ifelse(. == 2, 1, .))) %>%
  # And now we are back to just making only lockdown periods that aren't the first equal to 1
  mutate(c1_sub = ifelse(c1_flag == 0 & c1 == 1, 0, c1),
         c2_sub = ifelse(c2_flag == 0 & c2 == 1, 0, c2),
         c6_sub = ifelse(c6_flag == 0 & c6 == 1, 0, c6)) %>% 
  mutate(c1_first = c1_flag - c1_sub,
         c2_first = c2_flag - c2_sub,
         c6_first = c6_flag-c6_sub) %>% 
  select(-c1, -c2, -c6) %>% 
  mutate(c1_non = ifelse(c1_flag == 0 & start_date > c1_start, 1, 0),
         c2_non = ifelse(c2_flag == 0 & start_date > c2_start, 1, 0),
         c6_non = ifelse(c6_flag == 0 & start_date > c6_start, 1, 0))

# Re code using only transitions to level 2 or above!

hps_all_v18 <- hps_all_v17 %>%
  left_join(c1, by = "state") %>% 
  left_join(c2, by = "state") %>% 
  left_join(c6, by = "state") %>% 
  group_by(state) %>% 
  arrange(period) %>% 
  # This makes all subsequent periods after the first lockdown ends a 1
  mutate(c1_post_2only = ifelse(c1_school_closing <= 1 & start_date > c1_start, 1, NA),
         c2_post_2only = ifelse(c2_workplace_closing <= 1 & start_date > c2_start, 1, NA),
         c6_post_2only = ifelse(c6_stay_at_home_requirements <= 1 & start_date > c6_start, 1, NA)) %>% 
  fill(c1_post_2only:c6_post_2only, .direction = "down") %>% 
  mutate(across(c1_post_2only:c6_post_2only, ~ replace_na(., 0))) %>% 
  # Pick out all the repeated lockdowns
  mutate(c1_future_2 = ifelse(c1_post_2only == 1 & c1_school_closing > 1, 1, 0),
         c2_future_2 = ifelse(c2_post_2only == 1 & c2_workplace_closing > 1, 1, 0),
         c6_future_2 = ifelse(c6_post_2only == 1 & c6_stay_at_home_requirements > 1, 1, 0)) %>% 
  fill(c1_future_2:c6_future_2, .direction = "down") %>%
  mutate(across(c1_future_2:c6_future_2, ~ replace_na(., 0)))
  

  
  


```


```{r}
# Testing out some new stuff

export <- read_excel('export.xlsx') %>% 
  select(-c1_flag, -c2_flag, -c6_flag)

hps_all_v13 <- left_join(hps_all_v11, export, by =c("state", "start_date"))

depression_1 <- felm(log(hps_depression) ~ c1_post + c1_future +
               initclaims_combined + 
               contclaims_combined + cases_per_100k_5day + 
               factor(e1_income_support) + factor(e2_debt_contract_relief) + 
                log(covid_norm+1) + log(lockdown_norm+1)
             |factor(period) + factor(state) + factor(year)|0|state,
             data = hps_all_v12 %>% drop_na(pop),
             weights = na.omit(hps_all_v12$pop))
```


```{r}
# Visualizing some lockdown dates idk

lockdown_dates <- hps_all_v13 %>% 
  select(state, start_date, c1_school_closing, c2_workplace_closing, 
         c6_stay_at_home_requirements) %>% 
  distinct() %>% 
  group_by(start_date) %>% 
  summarize(c1 = mean(c1_school_closing, na.rm = TRUE),
            c2 = mean(c2_workplace_closing, na.rm = TRUE),
            c6 = mean(c6_stay_at_home_requirements, na.rm = TRUE)) %>% 
  pivot_longer(cols = c1:c6, names_to = "type", values_to = "status")
  

lockdown_dates %>% 
  ggplot(aes(x = start_date, y = status, color = type)) + geom_line() +
    scale_color_discrete(name = "NPI Type", labels = c("School Closure", 
                                                     "Workplace Closure",
                                                     "Stay-at-home Order")) + 
    scale_x_date(breaks = "3 months", date_labels = "%b %y") + 
    labs(x = "Date", y = "Mean lockdown status (higher is stricter)") + 
        theme_classic() + 
    theme(legend.position = "top")
  


```

```{r}
test <- felm()
```
