==> 11_pretrends.Rmd <==
---
title: "11_pretrends"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(tidyverse)
library(lfe)
library(texreg)
library(zoo)

searches_norm3_v20_5 <- readRDS('clean_data/search/searches_norm3_v20_5.rds')
```

```{r}

searches_norm3_v22 <- searches_norm3_v21 %>%
  mutate(across(c(busclose_first_counter, busopen_most_pop_counter, 
                  sip_first_counter, no_sip_most_pop_counter), ~ relevel(factor(.), ref = "0")))

searches_norm3_v22_5 <- searches_norm3_v22 %>%
  filter(dma %in% dma_5)

anxiety_bus <- felm(I(log(anxiety + 1 )) ~  
                 busclose_first_counter * year_indicator + 
                busopen_most_pop_counter * year_indicator + 
                        online_only * year_indicator + 
                        hybrid * year_indicator + 
                        other * year_indicator + 
                        cases_per_100k + unemploy_rate + temp + precip + 
                       log(lockdown + 1) + log(covid + 1) + 
                       factor(income_support) + factor(debt_contract_relief)
                       |factor(week) + factor(dma) + factor(year)|0|dma, 
              weights = na.omit(searches_norm3_v22_5$pop),
                  data = searches_norm3_v22_5 %>% filter(is.na(pop) == FALSE))

# Reason traditional DiD works better is b/c data fluctuation?
```

```{r}
# Pre-trends graph

searches_norm3_v23_5 %>% 
  mutate(anxiety_index = 0.924 * anxiety + 0.213 * nervous + 
                            0.133 * stress + 0.198 * worry + 0.211 * panic,
         depression_index = 0.485 * crying + 0.782 * depression + 
                                0.127 * lonely + 0.320 * sad + 0.184 * tired) %>% 
  select(week, year_indicator, year, depression, depression_index, anxiety, anxiety_index, pop) %>% 
  rename("Depression" = "depression",
         "Anxiety" = "anxiety",
         "Depression Index" = "depression_index",
         "Anxiety Index" = "anxiety_index") %>% 
  pivot_longer(cols = `Depression`:`Anxiety Index`, names_to = "term", values_to = "svi") %>% 
  drop_na(pop) %>% 
  group_by(week, year_indicator, term) %>% 
  summarize(svi = weighted.mean(svi, pop), .groups = "drop") %>%
  # group_by(year_indicator, term) %>% 
  # arrange(week) %>% 
  mutate(svi = rollmean(svi, k = 3, fill = "extend")) %>% 
  ggplot(aes(x = week, y = svi, color = factor(year_indicator))) +
    geom_line() + 
    #xlim(NA, 0) + 
    scale_color_discrete(name = "Year", labels = c("2016-2019", "2020")) + 
    facet_wrap(~term, scales = "free") + 
    theme_classic() + 
   theme(legend.position = "top") + 
   labs(y = "Population-weighted SVI (3-week moving avg.)",
        x = "Week of Year")

# SIP start
searches_norm3_v23_5 %>% 
  mutate(sip_first_counter = as.numeric(as.character(sip_first_counter))) %>% 
  mutate(anxiety_index = 0.924 * anxiety + 0.213 * nervous + 
                            0.133 * stress + 0.198 * worry + 0.211 * panic,
         depression_index = 0.485 * crying + 0.782 * depression + 
                                0.127 * lonely + 0.320 * sad + 0.184 * tired) %>% 
  select(sip_first_counter, 
         year_indicator, year, depression, depression_index, anxiety, anxiety_index, pop) %>% 
  rename("Depression" = "depression",
         "Anxiety" = "anxiety",
         "Depression Index" = "depression_index",
         "Anxiety Index" = "anxiety_index") %>% 
  pivot_longer(cols = `Depression`:`Anxiety Index`, names_to = "term", values_to = "svi") %>% 
  drop_na(pop) %>% 
  group_by(sip_first_counter, year_indicator, term) %>% 
  summarize(svi = weighted.mean(svi, pop), .groups = "drop") %>%
  # group_by(year_indicator, term) %>% 
  # arrange(week) %>% 
  #mutate(svi = rollmean(svi, k = 3, fill = 'extend')) %>% 
  ggplot(aes(x = sip_first_counter, y = svi, color = factor(year_indicator))) +
    geom_line() + 
      facet_wrap(~term, scales = "free") + 
    xlim(NA, 1) + 
    scale_color_discrete(name = "Year", labels = c("2016-2019", "2020")) + 
    theme_classic() + 
   theme(legend.position = "top") + 
   labs(y = "Population-weighted SVI",
        x = "Weeks from SIP start")

searches_norm3_v23_5 %>% 
  mutate(busclose_first_counter = as.numeric(as.character(busclose_first_counter))) %>% 
  mutate(anxiety_index = 0.924 * anxiety + 0.213 * nervous + 
                            0.133 * stress + 0.198 * worry + 0.211 * panic,
         depression_index = 0.485 * crying + 0.782 * depression + 
                                0.127 * lonely + 0.320 * sad + 0.184 * tired) %>% 
  select(busclose_first_counter, 
         year_indicator, year, depression, depression_index, anxiety, anxiety_index, pop) %>% 
  rename("Depression" = "depression",
         "Anxiety" = "anxiety",
         "Depression Index" = "depression_index",
         "Anxiety Index" = "anxiety_index") %>% 
  pivot_longer(cols = `Depression`:`Anxiety Index`, names_to = "term", values_to = "svi") %>% 
  drop_na(pop) %>% 
  group_by(busclose_first_counter, year_indicator, term) %>% 
  summarize(svi = weighted.mean(svi, pop), .groups = "drop") %>%
  # group_by(year_indicator, term) %>% 
  # arrange(week) %>% 
  #mutate(svi = rollmean(svi, k = 3, fill = 'extend')) %>% 
  ggplot(aes(x = busclose_first_counter, y = svi, color = factor(year_indicator))) +
    geom_line() + 
    geom_vline(xintercept = 0, colour = "grey60") + 
      facet_wrap(~term, scales = "free") + 
    xlim(NA, 5) + 
    scale_color_discrete(name = "Year", labels = c("2016-2019", "2020")) + 
    theme_classic() + 
   theme(legend.position = "top") + 
   labs(y = "Population-weighted SVI",
        x = "Weeks from business closure start")

# SIP end
searches_norm3_v21_5 %>% 
  mutate(anxiety_index = 0.924 * anxiety + 0.213 * nervous + 
                            0.133 * stress + 0.198 * worry + 0.211 * panic,
         depression_index = 0.485 * crying + 0.782 * depression + 
                                0.127 * lonely + 0.320 * sad + 0.184 * tired) %>% 
  select(no_sip_most_pop_counter, 
         year_indicator, year, depression, depression_index, anxiety, anxiety_index, pop) %>% 
  rename("Depression" = "depression",
         "Anxiety" = "anxiety",
         "Depression Index" = "depression_index",
         "Anxiety Index" = "anxiety_index") %>% 
  pivot_longer(cols = `Depression`:`Anxiety Index`, names_to = "term", values_to = "svi") %>% 
  drop_na(pop) %>% 
  group_by(no_sip_most_pop_counter, year_indicator, term) %>% 
  summarize(svi = weighted.mean(svi, pop), .groups = "drop") %>%
  # group_by(year_indicator, term) %>% 
  # arrange(week) %>% 
  mutate(svi = rollmean(svi, k = 3, fill = 'extend')) %>% 
  ggplot(aes(x = no_sip_most_pop_counter, y = svi, color = year_indicator))) +
    geom_line() + 
    # xlim(0, 20) + 
    scale_color_discrete(name = "Year", labels = c("2016-2019", "2020")) + 
    facet_wrap(~term, scales = "free") + 
    theme_classic() + 
   theme(legend.position = "top") + 
   labs(y = "Population-weighted SVI (3-week moving avg.)",
        x = "Weeks from SIP end")
```

```{r}
# State pretrends

# Modify HPS to get some counters for each one
first_lockdown <- hps_all_v11 %>%
  select(start_date, state, c1_flag, c2_flag, c6_flag) %>% 
  group_by(state) %>% 
  arrange(start_date)

c1 <- first_lockdown %>%
  filter(c1_flag == 1) %>% 
  slice(1) %>%
  select(start_date, state) %>% 
  rename("c1_start" = "start_date")

c2 <- first_lockdown %>%
  filter(c2_flag == 1) %>% 
  slice(1) %>%
  select(start_date, state) %>% 
  rename("c2_start" = "start_date")

c6 <- first_lockdown %>%
  filter(c6_flag == 1) %>% 
  slice(1) %>%
  select(start_date, state) %>% 
  rename("c6_start" = "start_date")

first_nolockdown <- hps_all_v11 %>%
  left_join(c1, by = "state") %>% 
  left_join(c2, by = "state") %>% 
  left_join(c6, by = "state") %>% 
  # This codes dates 1 if they're non-lockdown periods after the first lockdown 
  mutate(c1 = ifelse(c1_flag == 0 & start_date > c1_start, 1, NA),
         c2 = ifelse(c2_flag == 0 & start_date > c2_start, 1, NA),
         c6 = ifelse(c6_flag == 0 & start_date > c6_start, 1, NA)) %>% 
  group_by(state) %>% 
  arrange(start_date)

c1_first <- first_nolockdown %>% 
  filter(c1 == 1) %>% 
  slice(1) %>% 
  select(state, period) %>% 
  rename("c1_period" = "period")

c2_first <- first_nolockdown %>% 
  filter(c2 == 1) %>% 
  slice(1) %>% 
  select(state, period) %>% 
  rename("c2_period" = "period")

c6_first <- first_nolockdown %>%
  filter(c6 == 1) %>% 
  slice(1) %>% 
  select(state, period) %>% 
  rename("c6_period" = "period")

hps_all_v12 <- hps_all_v11 %>% 
  left_join(c1_first, by = "state") %>% 
  left_join(c2_first, by = "state") %>% 
  left_join(c6_first, by = "state") %>% 
  mutate(c1_counter = period - c1_period,
         c2_counter = period - c2_period,
         c6_counter = period - c6_period) %>% 
  select(-c1_period, -c2_period, -c6_period)

hps_all_v12 %>%
  group_by(c1_counter) %>% 
  summarize(hps_depression = mean(hps_depression)) %>% 
  ggplot(aes(x = c1_counter, y = hps_depression)) + geom_line()
```
```{r}
# Trying this again: 1st lockdown, no lockdown after the first, all subsequent lockdowns
# 
first_lockdown <- hps_all_v11 %>%
  group_by(state) %>% 
  arrange(start_date) %>% 
  fill(c(c1_flag, c2_flag, c6_flag), .direction = "down") %>% 
  ungroup() %>% 
  select(start_date, state, c1_school_closing, c2_workplace_closing, 
         c6_stay_at_home_requirements) %>%
  group_by(state) %>%
  arrange(start_date)

c1 <- first_lockdown %>%
  filter(c1_school_closing > 0) %>%
  slice(1) %>%
  select(start_date, state) %>%
  rename("c1_start" = "start_date")

c2 <- first_lockdown %>%
  filter(c2_workplace_closing > 0) %>%
  slice(1) %>%
  select(start_date, state) %>%
  rename("c2_start" = "start_date")

c6 <- first_lockdown %>%
  filter(c6_stay_at_home_requirements > 0) %>%
  slice(1) %>%
  select(start_date, state) %>%
  rename("c6_start" = "start_date")

# This is the first period when the lockdown is not in place
c1_end <- hps_all_v11 %>%
  select(period, state, c1_school_closing) %>% 
  filter(c1_school_closing == 0 & period != 1) %>% 
  group_by(state) %>% 
  arrange(period) %>%
  rename("c1_end" = "period") %>% 
  select(-c1_school_closing) %>% 
  slice(1)
  
c2_end <- hps_all_v11 %>%
  select(period, state, c2_workplace_closing) %>% 
  filter(c2_workplace_closing == 0 & period != 1) %>% 
  group_by(state) %>% 
  arrange(period) %>%
  rename("c2_end" = "period") %>% 
  select(-c2_workplace_closing) %>% 
  slice(1)

c6_end <- hps_all_v11 %>%
  select(period, state, c6_stay_at_home_requirements) %>% 
  filter(c6_stay_at_home_requirements == 0 & period != 1) %>% 
  group_by(state) %>% 
  arrange(period) %>%
  rename("c6_end" = "period") %>% 
  select(-c6_stay_at_home_requirements) %>% 
  slice(1)
  

hps_all_v13 <- hps_all_v11 %>%
  group_by(state) %>% 
  arrange(start_date) %>% 
  fill(c(c1_school_closing, c2_workplace_closing, c6_stay_at_home_requirements), 
       .direction = "down") %>% 
  ungroup() %>% 
  left_join(c1, by = "state") %>% 
  left_join(c2, by = "state") %>% 
  left_join(c6, by = "state") %>% 
  # This codes dates 1 if they're  periods after the first lockdown ends
  mutate(c1_post = ifelse(c1_school_closing == 0 & start_date > c1_start, 1, NA),
         c2_post = ifelse(c2_workplace_closing == 0 & start_date > c2_start, 1, NA),
         c6_post = ifelse(c6_stay_at_home_requirements == 0 & start_date > c6_start, 1, NA)) %>% 
  group_by(state) %>% 
  arrange(start_date) %>% 
  fill(c1_post:c6_post, .direction = "down") %>% 
  mutate(across(c1_post:c6_post, ~ ifelse(is.na(.), 0, .))) %>% 
  ungroup() %>% 
  left_join(c1_end, by = "state") %>% 
  left_join(c2_end, by = "state") %>% 
  left_join(c6_end, by = "state") %>% 
  mutate(c1_future = ifelse(c1_school_closing >= 1 & period >= c1_end, 1, 0),
         c2_future = ifelse(c2_workplace_closing >= 1 & period >= c2_end, 1, 0),
         c6_future = ifelse(c6_stay_at_home_requirements >= 1 & period >= c6_end, 1, 0)) %>% 
  # if there is no period end date then you are never in the subsequent lockdown, only the original
  mutate(across(c1_future:c6_future, ~ ifelse(is.na(.), 0, .))) %>% 
  mutate(c1_first = ifelse(c1_school_closing >= 1, 1, 0) - c1_future,
         c2_first = ifelse(c2_workplace_closing >= 1, 1, 0) - c2_future,
         c6_first = ifelse(c6_stay_at_home_requirements >= 1, 1, 0) - c6_future) %>% 
  # mutate(c1_future = ifelse(is.na(c1_future), ifelse(c1_flag == 1, 1, 0), c1_future),
  #        c2_future = ifelse(is.na(c2_future), ifelse(c2_flag == 1, 1, 0), c2_future),
  #        c6_future = ifelse(is.na(c6_future), ifelse(c6_flag == 1, 1, 0), c1_future)) %>% 
  select(-c1_start, -c2_start, -c6_start, -c1_end, -c2_end, -c6_end)
  
```


```{r}
depression_1 <- felm(log(hps_depression) ~ factor(c2_workplace_closing)  +
               initclaims_combined + 
               contclaims_combined + cases_per_100k_5day + 
               factor(e1_income_support) + factor(e2_debt_contract_relief) + 
                log(covid_norm+1) + log(lockdown_norm+1)
             |factor(period) + factor(state) + factor(year)|0|state,
             data = hps_all_v11 %>% drop_na(pop),
             weights = na.omit(hps_all_v11$pop))
```

```{r}
# pretrends graphing!!!

c1_end <- hps_all_v13 %>% 
  filter(c1_post == 1) %>% 
  group_by(state) %>% 
  arrange(period) %>% 
  slice(1) %>% 
  select(state, period) %>% 
  rename("c1_end_period" = "period")

c2_end <-  hps_all_v13 %>% 
  filter(c2_post == 1) %>% 
  group_by(state) %>% 
  arrange(period) %>% 
  slice(1) %>% 
  select(state, period) %>% 
  rename("c2_end_period" = "period")

c6_end <-  hps_all_v13 %>% 
  filter(c6_post == 1) %>% 
  group_by(state) %>% 
  arrange(period) %>% 
  slice(1) %>% 
  select(state, period) %>% 
  rename("c6_end_period" = "period")

hps_all_v14 <- hps_all_v13 %>% 
  left_join(c1_end, by = "state") %>% 
  left_join(c2_end, by = "state") %>% 
  left_join(c6_end, by = "state") %>% 
  group_by(state) %>% 
  arrange(period) %>% 
  mutate(c1_counter = period - c1_end_period,
         c2_counter = period - c2_end_period,
         c6_counter = period - c6_end_period) %>% 
  select(-c1_end_period, -c2_end_period, -c6_end_period) %>% 
  ungroup()

c1_graph <- hps_all_v14 %>% 
  group_by(c1_counter) %>% 
  summarize(c1_depression = weighted.mean(hps_depression, w = pop),
            c1_anxiety = weighted.mean(hps_anxiety, w = pop))


c2_graph <- hps_all_v14 %>% 
  group_by(c2_counter) %>% 
  summarize(c2_depression = weighted.mean(hps_depression, w = pop),
            c2_anxiety = weighted.mean(hps_anxiety, w = pop))

c6_graph <- hps_all_v14 %>% 
  group_by(c6_counter) %>% 
  summarize(c6_depression = weighted.mean(hps_depression, w = pop),
            c6_anxiety = weighted.mean(hps_anxiety, w = pop))

hps_graph <- c1_graph %>% 
  left_join(c2_graph, by = c("c1_counter" = "c2_counter")) %>% 
  left_join(c6_graph, by = c("c1_counter" = "c6_counter")) %>% 
  pivot_longer(cols = c1_depression:c6_anxiety, values_to = "value",
               names_to = "type") %>% 
  mutate(lockdown = str_sub(type, 1, 2),
         symptom = str_sub(type, 4)) %>% 
  select(-type)

hps_graph %>% 
  filter(symptom == "depression") %>% 
  ggplot(aes(x = c1_counter, y = value, color = lockdown)) + 
     geom_line() +
     labs(x = "Weeks vs. first lockdown end",
          y = "% reporting symptoms") + 
     scale_color_discrete(name = "NPI type",
                          labels = c("School closure", "Workplace closure",
                                     "Stay-at-home order")) + 
     theme_classic()

```
```{r}
# Disaggregating by location

west <- c("WA", "OR", "CA", "NV", "AZ", "ID", "MT", "WY",
      "CO", "NM", "UT", "HI", "AK")
south <- c("TX", "OK", "AR", "LA", "MS", "AL", "TN", "KY",
       "GA", "FL", "SC", "NC", "VA", "WV")
midwest <- c("KS", "NE", "SD", "ND", "MN", "MO", "IA", "IL",
         "IN", "MI", "WI", "OH")
northeast <- c("ME", "NH", "NY", "MA", "RI", "VT", "PA", 
           "NJ", "CT", "DE", "MD", "DC")

hps_all_v15 <- hps_all_v14 %>% 
  mutate(region = ifelse(state %in% west, "west", 
                         ifelse(state %in% south, "south", 
                                ifelse(state %in% midwest, "midwest",
                                       ifelse(state %in% northeast, "northeast", NA)))))

hps_all_v15 %>% 
  group_by(state) %>% 
  arrange(period) %>% 
  mutate(hps_depression = rollmean(hps_depression, k = 3, fill = "extend"),
         hps_anxiety = rollmean(hps_anxiety, k = 3, fill = "extend")) %>% 
  ungroup() %>% 
  group_by(c2_counter, region) %>% 
  summarize(c1_depression = weighted.mean(hps_depression, w = pop),
            c1_anxiety = weighted.mean(hps_anxiety, w = pop)) %>% 
  pivot_longer(cols = c1_depression:c1_anxiety, names_to = "symptom", values_to = "value") %>% 
  mutate(symptom = ifelse(symptom == "c1_anxiety", "Anxiety", "Depression")) %>%
  mutate(symptom = relevel(factor(symptom), ref = "Depression")) %>% 
  ggplot(aes(x = c2_counter, y = value, color = region)) + 
    geom_line() + 
    xlim(-20, 10) +
    theme_classic() +
    labs(x = "Weeks vs. lockdown end",
         y = "% reporting depression symptoms") + 
    scale_color_discrete(name = "Region", labels = c("Midwest",
                                                     "Northeast",
                                                     "South",
                                                     "West")) + 
    theme(legend.position = "top", panel.border = element_rect(color = "black",
                                    fill = NA,
                                    size = 0.5)) + 
    facet_wrap(~symptom)
```
