==> 12_mobility.Rmd <==
---
title: "12_mobility"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(tidyverse)
library(janitor)
library(lubridate)
library(zoo)
library(lfe)
library(texreg)
library(broom)
library(dotwhisker)
```


```{r}
# Mobility data first

mobility <- read_csv('data/2020_US_Region_Mobility_Report.csv') %>% 
  filter(country_region_code == "US", is.na(sub_region_2) == FALSE) %>% 
  select(sub_region_1, census_fips_code, retail_and_recreation_percent_change_from_baseline,
         grocery_and_pharmacy_percent_change_from_baseline,
         workplaces_percent_change_from_baseline, date)

```

```{r}
# Lockdown dates by county

goolsbee_interventions <- readRDS('clean_data/county/goolsbee_interventions.rds') %>%
  mutate(across(c(`sipstart`:`busopen`), ~ ifelse(. == as_date("2021-01-01"),
                                                   as_date("2030-01-01"), .))) %>%
  mutate(across(c(`sipstart`:`busopen`), as_date))
```

```{r}
# COVID data

covid <- read_csv('data/covid_county.csv') %>% 
  select(fips, cases, deaths, date) %>% 
  group_by(fips) %>% 
  arrange(date) %>% 
  # fill in missing cases/deaths downwards
  fill(cases, .direction = "down") %>% 
  fill(deaths, .direction = "down") %>% 
  mutate(new_cases = cases - lag(cases),
         new_deaths = deaths - lag(deaths)) %>% 
  mutate(new_cases_avg = rollmean(new_cases, k= 7, fill = NA),
         new_deaths_avg = rollmean(new_deaths, k = 7, fill = NA)) %>% 
  filter(date >= as_date("2020-02-15"))

# population

pop <- read_csv('data/county_stats/pop_race.csv', skip = 1) %>% 
  clean_names() %>%
  select(id, estimate_sex_and_age_total_population) %>%
  mutate(fips = str_sub(id, str_length(id) - 4, str_length(id))) %>%
  select(-id) %>% 
  rename("pop" = "estimate_sex_and_age_total_population")
```


```{r}
# Join this all so far

mobility_v0 <- left_join(mobility, covid, by = c("census_fips_code" = "fips",
                                                 "date")) %>% 
  left_join(goolsbee_interventions, by = c("census_fips_code" = "fips")) %>% 
  left_join(pop, by = c("census_fips_code" = "fips")) %>% 
  drop_na(census_fips_code) %>% 
  mutate(pop = pop/100000,
         sip = ifelse(date >= sipstart, 1, 0),
         no_sip = ifelse(date >= sipend, 1, 0),
         bus = ifelse(date >= busclose, 1, 0),
         no_bus = ifelse(date >= busopen, 1, 0),
         new_cases_avg = ifelse(is.na(new_cases_avg), 0, new_cases_avg),
         new_deaths_avg = ifelse(is.na(new_deaths_avg), 0, new_deaths_avg)) %>% 
  mutate(new_cases_avg = new_cases_avg/pop,
         new_deaths_avg = new_deaths_avg/pop)
```


```{r}
# Now we want searches lol

dma_codes <- read_csv('dma_codes.csv') %>%
  select(google_code, statefp, cntyfp) %>%
  mutate(statefp = str_pad(statefp, width = 2, side = "left", pad = "0"),
         cntyfp = str_pad(cntyfp, width = 3, side = "left", pad = 0)) %>%
  mutate(fips = str_c(statefp, cntyfp, sep = "")) %>%
  select(google_code, fips)

searches <- searches_norm3_v22 %>% 
  select(covid, lockdown, date, dma) %>% 
  filter(date >= as_date("2020-01-01")) %>% 
  left_join(dma_codes, by = c("dma" = "google_code")) %>% 
  rename("week_date" = "date")
```


```{r}
# Date matching
dates <- mobility_v0 %>% 
  select(date) %>% 
  distinct()

week_dates <- searches %>% 
  select(week_date) %>% 
  distinct()

date_match <- expand_grid(dates, week_dates) %>% 
  mutate(date = as_date(date), week_date = as_date(week_date)) %>% 
  mutate(diff = abs(as.numeric(date - week_date))) %>% 
  group_by(date) %>% 
  arrange(diff) %>% 
  slice(1) %>% 
  select(-diff)
  

mobility_v1 <- left_join(mobility_v0, date_match, by = "date") %>% 
  left_join(searches, by = c("census_fips_code" = "fips",
                             "week_date")) %>% 
  rename("retail_recreation" = "retail_and_recreation_percent_change_from_baseline",
         "grocery_pharmacy" = "grocery_and_pharmacy_percent_change_from_baseline",
         "workplaces" = "workplaces_percent_change_from_baseline",
         "fips" = "census_fips_code") %>% 
  mutate(day = wday(date),
         month = month(date),
         week = week(date)) %>% 
  mutate(across(c(`sipstart`:`busopen`), ~ ifelse(. == as_date("2030-01-01"), NA, .))) %>% 
  mutate(sipcounter = as.numeric(date - sipstart),
         no_sipcounter = as.numeric(date - sipend),
         buscounter = as.numeric(date - busclose),
         no_buscounter = as.numeric(date - busopen))
```

```{r}
# Let's round the sipcounter too

mobility_v2 <- mobility_v1 %>% 
  mutate(sipcounter_wk = round(sipcounter/7),
         no_sipcounter_wk = round(no_sipcounter/7),
         buscounter_wk = round(buscounter/7),
         no_buscounter_wk = round(no_buscounter/7))
```


```{r}
# Try this

retail <- felm(retail_recreation ~ sip + no_sip + new_cases_avg + 
                   log(covid+1) + log(lockdown+1)|day + week + fips|0|fips, data = mobility_v1)

grocery <- felm(grocery_pharmacy ~ sip + no_sip + new_cases_avg + 
                   log(covid+1) + log(lockdown+1)|day + week + fips|0|fips, data = mobility_v1)

workplaces <- felm(workplaces ~ sip + no_sip + new_cases_avg + 
                   log(covid+1) + log(lockdown+1)|day + week + fips|0|fips, data = mobility_v1)
```

```{r}
retail <- felm(retail_recreation ~ bus + no_bus + new_cases_avg + 
                   log(covid+1) + log(lockdown+1)|day + week + fips|0|fips, data = mobility_v5)

grocery <- felm(grocery_pharmacy ~ bus + no_bus + new_cases_avg + 
                   log(covid+1) + log(lockdown+1)|day + week + fips|0|fips, data = mobility_v5)

workplaces <- felm(workplaces ~ bus + no_bus + new_cases_avg + 
                   log(covid+1) + log(lockdown+1)|day + week + fips|0|fips, data = mobility_v5)
```


```{r}
# GRAPH

dwplot(list(retail, grocery, workplaces),
       vline = geom_vline(
           xintercept = 0,
           colour = "grey60",
       ),
       vars_order = c("bus", "no_bus"),
       dodge_size = 0.5) %>% 
    relabel_predictors(c("bus" = "Closure start",
                         "no_bus" = "Closure end")) +
    # theme_classic(base_size = 14) +
      # scale_x_continuous(labels = scales::percent) + 
      # labs(x = "Effect size (approx. % change)") + 
  theme_classic() + 
  theme(legend.position = "top") + 
  # guides(color=guide_legend(nrow=2, byrow=TRUE)) + 
      labs(x = "% change in mobility") + 
      scale_color_discrete(name = "Mobility type", labels = c("Retail/recreation",
                                                              "Grocery/pharmacy",
                                                              "Workplaces"))
```

```{r}
mobility_v1 %>% 
  group_by(fips) %>% 
  arrange(date) %>% 
  mutate(across(retail_recreation:workplaces, ~ rollmean(., k = 7, fill = NA))) %>% 
  ungroup() %>% 
  group_by(no_sipcounter) %>% 
  summarize(across(retail_recreation:workplaces, ~ weighted.mean(., w = pop, na.rm = TRUE))) %>% 
  pivot_longer(cols = retail_recreation:workplaces, names_to = "mobility", values_to = "value") %>% 
  ggplot(aes(x = no_sipcounter, y = value, color = mobility)) + 
    geom_line() + 
    theme_classic() + 
    theme(legend.position = "top") + 
    labs(x = "Days vs. SIP end",
         y = "% mobility change vs. 2019 baseline") + 
    scale_color_discrete(name = "Location", labels = c("Grocery & pharmacy",
                                                       "Retail & recreation",
                                                       "Workplaces"))

mobility_v3 %>% 
  group_by(fips) %>% 
  arrange(date) %>% 
  mutate(across(retail_recreation:workplaces, ~ rollmean(., k = 7, fill = NA))) %>% 
  ungroup() %>% 
  group_by(no_buscounter) %>% 
  summarize(across(retail_recreation:workplaces, ~ weighted.mean(., w = pop, na.rm = TRUE))) %>% 
  pivot_longer(cols = retail_recreation:workplaces, names_to = "mobility", values_to = "value") %>% 
  ggplot(aes(x = no_buscounter, y = value, color = mobility)) + 
    geom_line() + 
    theme_classic() + 
    theme(legend.position = "top") + 
    labs(x = "Days vs. business reopening ",
         y = "% mobility change vs. 2019 baseline") + 
    scale_color_discrete(name = "Location", labels = c("Grocery & pharmacy",
                                                       "Retail & recreation",
                                                       "Workplaces"))

mobility_v3 %>% 
  group_by(fips) %>% 
  arrange(date) %>% 
  mutate(across(retail_recreation:workplaces, ~ rollmean(., k = 7, fill = NA))) %>% 
  ungroup() %>% 
  group_by(buscounter) %>% 
  summarize(across(retail_recreation:workplaces, ~ weighted.mean(., w = pop, na.rm = TRUE))) %>% 
  pivot_longer(cols = retail_recreation:workplaces, names_to = "mobility", values_to = "value") %>% 
  ggplot(aes(x = buscounter, y = value, color = mobility)) + 
    geom_line() + 
    theme_classic() + 
    theme(legend.position = "top") + 
    labs(x = "Days vs. business closure ",
         y = "% mobility change vs. 2019 baseline") + 
    scale_color_discrete(name = "Location", labels = c("Grocery & pharmacy",
                                                       "Retail & recreation",
                                                       "Workplaces"))
  
  
```

```{r}
# Adding in some more stuff

unemployment <- unemployment_county_2016_2020 %>% 
  filter(year == "2020") %>% 
  mutate(unemploy = as.numeric(unemploy) * 100/as.numeric(labor_force)) %>% 
  select(-labor_force)

mobility_v3 <- mobility_v2 %>% 
  mutate(month = month(date)) %>% 
  left_join(unemployment, by = c("month" = "period", "fips")) %>% 
  select(-year)
```

```{r}
# event study - imposition
retail <- felm(retail_recreation ~  
                 i(sipcounter_wk, sip_treat, ref = -1) + 
                 no_sip + 
                 new_cases_avg + 
                   log(covid+1) + log(lockdown+1) + unemploy|day + week + fips|0|fips, 
               data = mobility_v6, weights = mobility_v6$pop)

grocery <- felm(grocery_pharmacy ~ i(sipcounter_wk, sip_treat, ref = -1) + 
                  no_sip + new_cases_avg + 
                   log(covid+1) + log(lockdown+1) + unemploy|day + week + fips|0|fips, 
                data = mobility_v6, weights = mobility_v6$pop)

workplaces <- felm(workplaces ~ i(sipcounter_wk, sip_treat, ref = -1) + no_sip + 
                     new_cases_avg + 
                   log(covid+1) + log(lockdown+1) + unemploy|day + week + fips|0|fips, 
                   data = mobility_v6, weights = mobility_v6$pop)

retail <- felm(retail_recreation ~ i(buscounter_wk, bus_treat, ref = -1) + no_bus + 
                 new_cases_avg + 
                   log(covid+1) + log(lockdown+1) + unemploy|day + week + fips|0|fips, 
               data = mobility_v6, weights = mobility_v6$pop)

grocery <- felm(grocery_pharmacy ~ i(buscounter_wk, bus_treat, ref = -1) + no_bus + 
                  new_cases_avg + 
                   log(covid+1) + log(lockdown+1) + unemploy|day + week + fips|0|fips, 
                data = mobility_v6, weights = mobility_v6$pop)

workplaces <- felm(workplaces ~ i(buscounter_wk, bus_treat, ref = -1) + no_bus + 
                     new_cases_avg + 
                   log(covid+1) + log(lockdown+1) + unemploy|day + week + fips|0|fips, 
                   data = mobility_v6, weights = mobility_v6$pop)

retail_coefs <- tidy(retail) %>% 
  slice(1:48) %>% 
  mutate(term = as.numeric(str_sub(term, start = 38)),
         ci = 1.96 * `std.error`) %>% 
  mutate(type = "retail")

grocery_coefs <- tidy(grocery) %>% 
  slice(1:48) %>% 
  mutate(term = as.numeric(str_sub(term, start = 38)),
         ci = 1.96 * `std.error`) %>% 
  mutate(type = "grocery")

workplaces_coefs <- tidy(workplaces) %>% 
  slice(1:48) %>% 
  mutate(term = as.numeric(str_sub(term, start = 38)),
         ci = 1.96 * `std.error`) %>% 
  mutate(type = "workplaces")

total_coefs <- bind_rows(retail_coefs, grocery_coefs, workplaces_coefs)


total_coefs %>% 
    ggplot(aes(x = term, y = estimate, color = type)) + 
      geom_line() + 
      geom_pointrange(aes(ymin=estimate-ci, ymax=estimate+ci), alpha = 0.6) + 
      geom_hline(yintercept = 0, colour = "grey60") + 
      xlim(NA, 20) + 
      ylim(NA, 10) + 
      theme_classic() + 
      labs(x = "Weeks vs. lockdown start",
           y = "Estimated % change in mobility") + 
      scale_color_discrete(name = "Mobility type",
                           labels = c("Grocery/pharmacy",
                                      "Retail/recreation",
                                      "Workplaces")) + 
     theme(legend.position = "top")

```

```{r}
# event study - end
retail <- felm(retail_recreation ~ i(no_sipcounter_wk, no_sip_treat, ref = -1) + sip + 
                 new_cases_avg + 
                   log(covid+1) + log(lockdown+1) + unemploy|day + week + fips|0|fips, 
               data = mobility_v6, weights = mobility_v6$pop)

grocery <- felm(grocery_pharmacy ~ i(no_sipcounter_wk, no_sip_treat, ref = -1) + sip +  
                  new_cases_avg + 
                   log(covid+1) + log(lockdown+1) + unemploy|day + week + fips|0|fips, 
                data = mobility_v6, weights = mobility_v6$pop)

workplaces <- felm(workplaces ~ i(no_sipcounter_wk, no_sip_treat, ref = -1) + sip + 
                     new_cases_avg + 
                   log(covid+1) + log(lockdown+1) + unemploy|day + week + fips|0|fips, 
                   data = mobility_v6, weights = mobility_v6$pop)

retail <- felm(retail_recreation ~ i(no_buscounter_wk, no_bus_treat, ref = -1) + bus + 
                 new_cases_avg + 
                   log(covid+1) + log(lockdown+1) + unemploy|day + week + fips|0|fips, 
               data = mobility_v6, weights = mobility_v6$pop)

grocery <- felm(grocery_pharmacy ~ i(no_buscounter_wk, no_bus_treat, ref = -1) + bus + 
                  new_cases_avg + 
                   log(covid+1) + log(lockdown+1) + unemploy|day + week + fips|0|fips, 
                data = mobility_v6, weights = mobility_v6$pop)

workplaces <- felm(workplaces ~ i(no_buscounter_wk, no_bus_treat, ref = -1) + bus + 
                     new_cases_avg + 
                   log(covid+1) + log(lockdown+1) + unemploy|day + week + fips|0|fips, 
                   data = mobility_v6, weights = mobility_v6$pop)

retail_coefs <- tidy(retail) %>% 
  slice(1:53) %>% 
  mutate(term = as.numeric(str_sub(term, start = 44)),
         ci = 1.96 * `std.error`) %>% 
  mutate(type = "retail")

grocery_coefs <- tidy(grocery) %>% 
  slice(1:53) %>% 
  mutate(term = as.numeric(str_sub(term, start = 44)),
         ci = 1.96 * `std.error`) %>% 
  mutate(type = "grocery")

workplaces_coefs <- tidy(workplaces) %>% 
  slice(1:53) %>% 
  mutate(term = as.numeric(str_sub(term, start = 44)),
         ci = 1.96 * `std.error`) %>% 
  mutate(type = "workplaces")

total_coefs <- bind_rows(retail_coefs, grocery_coefs, workplaces_coefs)


total_coefs %>% 
    ggplot(aes(x = term, y = estimate, color = type)) + 
      geom_line() + 
      geom_pointrange(aes(ymin=estimate-ci, ymax=estimate+ci), alpha = 0.6) + 
      geom_hline(yintercept = 0, colour = "grey60") + 
      xlim(-10, 20) + 
      ylim(-10, 10) + 
      theme_classic() + 
      labs(x = "Weeks vs. lockdown end",
           y = "Estimated % change in mobility") + 
      scale_color_discrete(name = "Mobility type",
                           labels = c("Grocery/pharmacy",
                                      "Retail/recreation",
                                      "Workplaces")) + 
     theme(legend.position = "top")
```

```{r}

retail_1 <- felm(retail_recreation ~ sip + no_sip + 
                 new_cases_avg + 
                   log(covid+1) + log(lockdown+1) + unemploy|day + week + fips|0|fips, 
               data = mobility_v3, weights = mobility_v3$pop)

grocery_1 <- felm(grocery_pharmacy ~ sip + no_sip + 
                 new_cases_avg + 
                   log(covid+1) + log(lockdown+1) + unemploy|day + week + fips|0|fips, 
               data = mobility_v3, weights = mobility_v3$pop)

workplace_1 <- felm(workplaces ~ sip + no_sip + 
                 new_cases_avg + 
                   log(covid+1) + log(lockdown+1) + unemploy|day + week + fips|0|fips, 
               data = mobility_v3, weights = mobility_v3$pop)

retail_2 <- felm(retail_recreation ~ bus + no_bus + 
                 new_cases_avg + 
                   log(covid+1) + log(lockdown+1) + unemploy|day + week + fips|0|fips, 
               data = mobility_v3, weights = mobility_v3$pop)

grocery_2 <- felm(grocery_pharmacy ~ bus + no_bus + 
                 new_cases_avg + 
                   log(covid+1) + log(lockdown+1) + unemploy|day + week + fips|0|fips, 
               data = mobility_v3, weights = mobility_v3$pop)

workplace_2 <- felm(workplaces ~ bus + no_bus + 
                 new_cases_avg + 
                   log(covid+1) + log(lockdown+1) + unemploy|day + week + fips|0|fips, 
               data = mobility_v3, weights = mobility_v3$pop)

htmlreg(l = list(retail_1, grocery_1, workplace_1, retail_2, grocery_2, workplace_2),
        include.ci = FALSE, digits = 4, file = "mobility.doc")


```

```{r}
# mobility_v4

mobility_v4 <- mobility_v3 %>% 
  mutate(sipcounter_wk_1 = sipcounter_wk-1,
         sipcounter_wk_2 = sipcounter_wk-2,
         sipcounter_wk_3 = sipcounter_wk-3,
         sipcounter_wk_4 = sipcounter_wk-4,
         sipcounter_wk_n2 = sipcounter_wk + 2,
         no_sipcounter_wk_1 = no_sipcounter_wk-1,
         no_sipcounter_wk_2 = no_sipcounter_wk-2,
         no_sipcounter_wk_3 = no_sipcounter_wk-3,
         no_sipcounter_wk_4 = no_sipcounter_wk-4,
         no_sipcounter_wk_n2 = no_sipcounter_wk + 2,
         buscounter_wk_1 = buscounter_wk-1,
         buscounter_wk_2 = buscounter_wk-2,
         buscounter_wk_3 = buscounter_wk-3,
         buscounter_wk_4 = buscounter_wk-4,
         buscounter_wk_n2 = buscounter_wk + 2,
         no_buscounter_wk_1 = no_buscounter_wk-1,
         no_buscounter_wk_2 = no_buscounter_wk-2,
         no_buscounter_wk_3 = no_buscounter_wk-3,
         no_buscounter_wk_4 = no_buscounter_wk-4,
         no_buscounter_wk_n2 = no_buscounter_wk + 2,) %>% 
  mutate(across(sipstart:busopen, as_date)) %>% 
  mutate(sip_n2 = ifelse(sipcounter_wk_n2 > 0, 1, 0),
         sip_1 = ifelse(sipcounter_wk_1 > 0, 1, 0),
         sip_2 = ifelse(sipcounter_wk_2 > 0, 1, 0),
         sip_3 = ifelse(sipcounter_wk_3 > 0, 1, 0),
         sip_4 = ifelse(sipcounter_wk_4 > 0, 1, 0),
         no_sip_n2 = ifelse(no_sipcounter_wk_n2 > 0, 1, 0),
         no_sip_1 = ifelse(no_sipcounter_wk_1 > 0, 1, 0),
         no_sip_2 = ifelse(no_sipcounter_wk_2 > 0, 1, 0),
         no_sip_3 = ifelse(no_sipcounter_wk_3 > 0, 1, 0),
         no_sip_4 = ifelse(no_sipcounter_wk_4 > 0, 1, 0),
         bus_n2 = ifelse(buscounter_wk_n2 > 0, 1, 0),
         bus_1 = ifelse(buscounter_wk_1 > 0, 1, 0),
         bus_2 = ifelse(buscounter_wk_2 > 0, 1, 0),
         bus_3 = ifelse(buscounter_wk_3 > 0, 1, 0),
         bus_4 = ifelse(buscounter_wk_4 > 0, 1, 0),
         nobus_n2 = ifelse(no_buscounter_wk_n2> 0, 1, 0),
         no_bus_1 = ifelse(no_buscounter_wk_1 > 0, 1, 0),
         no_bus_2 = ifelse(no_buscounter_wk_2 > 0, 1, 0),
         no_bus_3 = ifelse(no_buscounter_wk_3 > 0, 1, 0),
         no_bus_4 = ifelse(no_buscounter_wk_4 > 0, 1, 0))

# Edit this

mobility_v6 <- mobility_v5 %>% 
  mutate(sip_treat = ifelse(is.na(sipcounter), 0, 1),
         bus_treat = ifelse(is.na(buscounter), 0, 1),
         no_sip_treat = ifelse(is.na(no_sipcounter), 0, 1),
         no_bus_treat = ifelse(is.na(no_buscounter), 0, 1)) %>% 
  mutate(across(sipcounter_wk:no_buscounter_n5, ~ ifelse(is.na(.), 0, .)))
```

```{r}
# Changing window



```

```{r}
# Changing date thresholds

retail_n2 <- felm(retail_recreation ~ bus_n2 + no_bus + 
                 new_cases_avg + 
                   log(covid+1) + log(lockdown+1) + unemploy|day + week + fips|0|fips, 
               data = mobility_v4, weights = mobility_v4$pop)

retail_0 <- felm(retail_recreation ~ bus + no_bus + 
                 new_cases_avg + 
                   log(covid+1) + log(lockdown+1) + unemploy|day + week + fips|0|fips, 
               data = mobility_v4, weights = mobility_v4$pop)

retail_2 <- felm(retail_recreation ~ bus_2 + no_bus + 
                 new_cases_avg + 
                   log(covid+1) + log(lockdown+1) + unemploy|day + week + fips|0|fips, 
               data = mobility_v4, weights = mobility_v4$pop)

retail_4 <- felm(retail_recreation ~ bus_4 + no_bus + 
                 new_cases_avg + 
                   log(covid+1) + log(lockdown+1) + unemploy|day + week + fips|0|fips, 
               data = mobility_v4, weights = mobility_v4$pop)


htmlreg(l = list(retail_n2, retail_0, retail_2, retail_4),
        include.ci = FALSE, digits = 4, file = "mobility.doc")

```

```{r}
grocery_n2 <- felm(grocery_pharmacy ~ bus_n2 + no_bus + 
                 new_cases_avg + 
                   log(covid+1) + log(lockdown+1) + unemploy|day + week + fips|0|fips, 
               data = mobility_v4, weights = mobility_v4$pop)

grocery_0 <- felm(grocery_pharmacy ~ bus + no_bus + 
                 new_cases_avg + 
                   log(covid+1) + log(lockdown+1) + unemploy|day + week + fips|0|fips, 
               data = mobility_v4, weights = mobility_v4$pop)

grocery_2 <- felm(grocery_pharmacy ~ bus_2 + no_bus + 
                 new_cases_avg + 
                   log(covid+1) + log(lockdown+1) + unemploy|day + week + fips|0|fips, 
               data = mobility_v4, weights = mobility_v4$pop)

grocery_4 <- felm(grocery_pharmacy ~ bus_4 + no_bus + 
                 new_cases_avg + 
                   log(covid+1) + log(lockdown+1) + unemploy|day + week + fips|0|fips, 
               data = mobility_v4, weights = mobility_v4$pop)

workplaces_n2 <- felm(workplaces ~ bus_n2 + no_bus + 
                 new_cases_avg + 
                   log(covid+1) + log(lockdown+1) + unemploy|day + week + fips|0|fips, 
               data = mobility_v4, weights = mobility_v4$pop)

workplaces_0 <- felm(workplaces ~ bus + no_bus + 
                 new_cases_avg + 
                   log(covid+1) + log(lockdown+1) + unemploy|day + week + fips|0|fips, 
               data = mobility_v4, weights = mobility_v4$pop)

workplaces_2 <- felm(workplaces ~ bus_2 + no_bus + 
                 new_cases_avg + 
                   log(covid+1) + log(lockdown+1) + unemploy|day + week + fips|0|fips, 
               data = mobility_v4, weights = mobility_v4$pop)

workplaces_4 <- felm(workplaces ~ bus_4 + no_bus + 
                 new_cases_avg + 
                   log(covid+1) + log(lockdown+1) + unemploy|day + week + fips|0|fips, 
               data = mobility_v4, weights = mobility_v4$pop)

# https://stackoverflow.com/questions/8197559/emulate-ggplot2-default-color-palette 

dwplot(list(retail_n2, retail_0, retail_2, retail_4),
       vars_order = c("bus_n2", "bus", "bus_2", "bus_4"),
       dodge_size = 0.5) %>% 
    relabel_predictors(c("bus_n2" = "2 weeks after",
                         "bus" = "Actual",
                         "bus_2" = "2 weeks before",
                         "bus_4" = "4 weeks before")) +
    theme_classic(base_size = 14) +
      # scale_x_continuous(labels = scales::percent) + 
      # labs(x = "Effect size (approx. % change)") + 
      theme(legend.position = "none") + 
  # guides(color=guide_legend(nrow=2, byrow=TRUE)) + 
      labs(x = "% change in mobility") + 
     scale_color_manual(values = c("#00BA38", "#00BA38", "#00BA38", "#00BA38"))


dwplot(list(grocery_n2, grocery_0, grocery_2, grocery_4),
       vars_order = c("bus_n2", "bus", "bus_2", "bus_4"),
       dodge_size = 0.5) %>% 
    relabel_predictors(c("bus_n2" = "2 weeks after",
                         "bus" = "Actual",
                         "bus_2" = "2 weeks before",
                         "bus_4" = "4 weeks before")) +
    theme_classic(base_size = 14) +
      # scale_x_continuous(labels = scales::percent) + 
      # labs(x = "Effect size (approx. % change)") + 
      theme(legend.position = "none") + 
  # guides(color=guide_legend(nrow=2, byrow=TRUE)) + 
      labs(x = "% change in mobility")
```


```{r}
# Making our own dwplot

grocery_n2_coef <- tidy(grocery_n2) %>% 
  slice(1) %>% 
  mutate(type = "grocery",
         time = "-2")

grocery_0_coef <- tidy(grocery_0) %>% 
  slice(1) %>% 
  mutate(type = "grocery",
         time = "0")

grocery_2_coef <- tidy(grocery_2) %>% 
  slice(1) %>% 
  mutate(type = "grocery",
         time = "2")

grocery_4_coef <- tidy(grocery_4) %>% 
  slice(1) %>% 
  mutate(type = "grocery",
         time = "4")

workplaces_n2_coef <- tidy(workplaces_n2) %>% 
  slice(1) %>% 
  mutate(type = "workplaces",
         time = "-2")

workplaces_0_coef <- tidy(workplaces_0) %>% 
  slice(1) %>% 
  mutate(type = "workplaces",
         time = "0")

workplaces_2_coef <- tidy(workplaces_2) %>% 
  slice(1) %>% 
  mutate(type = "workplaces",
         time = "2")

workplaces_4_coef <- tidy(workplaces_4) %>% 
  slice(1) %>% 
  mutate(type = "workplaces",
         time = "4")

retail_n2_coef <- tidy(retail_n2) %>% 
  slice(1) %>% 
  mutate(type = "retail",
         time = "-2")

retail_0_coef <- tidy(retail_0) %>% 
  slice(1) %>% 
  mutate(type = "retail",
         time = "0")

retail_2_coef <- tidy(retail_2) %>% 
  slice(1) %>% 
  mutate(type = "retail",
         time = "2")

retail_4_coef <- tidy(retail_4) %>% 
  slice(1) %>% 
  mutate(type = "retail",
         time = "4")

total_coefs <- rbind(retail_n2, retail_0, retail_2, retail_4, 
                     grocery_n2, grocery_0, grocery_2, grocery_4,
                     workplaces_n2, workplaces_0, workplaces_2, workplaces_4)
```
